% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deployment.R
\name{as.dataRobotDeploymentAssociationIdSettings}
\alias{as.dataRobotDeploymentAssociationIdSettings}
\title{Association ID settings for a deployment.}
\usage{
as.dataRobotDeploymentAssociationIdSettings(apiResponse)
}
\arguments{
\item{apiResponse}{List of deployment settings retrieved from the DataRobot
API.}
}
\value{
An object classed \code{dataRobotDeploymentAssociationIdSettings}
that contains:
\describe{
\item{columnNames}{character. The columns that can be used as
association IDs.}
\item{requiredInPredictionRequests}{logical. Whether the association ID is
required in a prediction request.}
}
}
\description{
Helper method to process the response object received from the
\verb{/deployments/\{id\}/settings} DataRobot API endpoint. See
\code{\link[=GetDeploymentSettings]{GetDeploymentSettings()}}.
}
\keyword{internal}
