% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deployment.R
\name{as.dataRobotDeploymentDriftTrackingSettings}
\alias{as.dataRobotDeploymentDriftTrackingSettings}
\title{This function specifically narrows the Deployment Settings objects
to just those related to Drift Tracking.}
\usage{
as.dataRobotDeploymentDriftTrackingSettings(inList)
}
\arguments{
\item{inList}{JSON object retrieved from GetDeploymentSettings()}
}
\description{
This function specifically narrows the Deployment Settings objects
to just those related to Drift Tracking.
}
