% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSeries.R
\name{as.dataRobotMultiSeriesProperties}
\alias{as.dataRobotMultiSeriesProperties}
\title{Return value for GetMultiSeriesProperties() and others}
\usage{
as.dataRobotMultiSeriesProperties(inList)
}
\arguments{
\item{inList}{list. See return value below for expected elements.}
}
\value{
A named list which contains:
\itemize{
  \item timeSeriesEligible logical. Whether or not the series is eligible to be used for
    time series.
  \item crossSeriesEligible logical. Whether or not the cross series group by column is
    eligible for cross-series modeling. Will be NULL if no cross series group by column
    is used.
  \item crossSeriesEligibilityReason character. The type of cross series eligibility
    (or ineligibility).
  \item timeUnit character. For time series eligible features, the time unit covered by a
    single time step, e.g. "HOUR", or NULL for features that are not time series eligible.
  \item timeStep integer. Expected difference in time units between rows in the data.
    Will be NULL for features that are not time series eligible.
}
}
\description{
Return value for GetMultiSeriesProperties() and others
}
\seealso{
Other MultiSeriesProject functions: 
\code{\link{GetMultiSeriesProperties}()},
\code{\link{RequestCrossSeriesDetection}()},
\code{\link{RequestMultiSeriesDetection}()}
}
\concept{MultiSeriesProject functions}
