% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoreDiscardedFeatures.R
\name{as.dataRobotRestoredFeaturesInformation}
\alias{as.dataRobotRestoredFeaturesInformation}
\title{Restore discarded during time series feature generation process features back to the project
After restoration features will be included into All Time Series Features and can be
used to create new feature lists.
Format restoredFeatures response}
\usage{
as.dataRobotRestoredFeaturesInformation(inList)
}
\arguments{
\item{inList}{list. See return value below for expected elements.}
}
\value{
A named list that contains:
\describe{
  \item{featuresToRestore}{List of character strings with features which were accepted for
    restoration}
  \item{warnings}{List of character strings with warnings for the not accepted features}
}
}
\description{
Restore discarded during time series feature generation process features back to the project
After restoration features will be included into All Time Series Features and can be
used to create new feature lists.
Format restoredFeatures response
}
