% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{AutopilotMode}
\alias{AutopilotMode}
\title{Autopilot modes}
\format{An object of class \code{list} of length 4.}
\usage{
AutopilotMode
}
\description{
This is a list that contains the valid values for autopilot mode. If you wish, you can
specify autopilot modes using the list values, e.g. AutopilotMode$FullAuto instead of typing
the string 'auto'. This way you can benefit from autocomplete and not have to remember the valid
optons.
}
\keyword{datasets}
