% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{BlendMethods}
\alias{BlendMethods}
\title{Blend methods}
\format{An object of class \code{list} of length 7.}
\usage{
BlendMethods
}
\description{
This is a list that contains the valid values for Blend methods
}
\keyword{datasets}
