% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blueprints.R
\name{BlueprintChartToGraphviz}
\alias{BlueprintChartToGraphviz}
\title{Convert a blueprint chart into graphviz DOT format}
\usage{
BlueprintChartToGraphviz(blueprintChart)
}
\arguments{
\item{blueprintChart}{List, returned by GetBlueprintChart function}
}
\value{
Character string representation of chart in graphviz DOT language
}
\description{
Convert a blueprint chart into graphviz DOT format
}
