% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataRobotRequests.R
\name{CABundle}
\alias{CABundle}
\title{Return the path to the custom CA bundle (if set).}
\usage{
CABundle()
}
\value{
character. The CA bundle path, or \code{""} if not set.
}
\description{
Reads the \env{CURL_CA_BUNDLE} environment variable set by
\code{SaveCABundlePreference()} and returns its value, or an empty string if unset.
}
\keyword{internal}
