% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataRobotRequests.R
\name{CheckUrl}
\alias{CheckUrl}
\title{Make sure the path is a reasonable URL}
\usage{
CheckUrl(url)
}
\arguments{
\item{url}{character. The URL to check.}
}
\description{
Make sure the path is a reasonable URL
}
