% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{ClassificationDeploymentAccuracyMetric}
\alias{ClassificationDeploymentAccuracyMetric}
\title{Accuracy metrics for classification deployments}
\format{
An object of class \code{list} of length 14.
}
\usage{
ClassificationDeploymentAccuracyMetric
}
\description{
Added in DataRobot API 2.18.
}
\keyword{datasets}
