% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServerData.R
\name{CleanServerData}
\alias{CleanServerData}
\title{Reformat paginated data returned from the server.}
\usage{
CleanServerData(serverData)
}
\arguments{
\item{serverData}{list. Raw JSON parsed list returned from the server.}
}
\description{
Reformat paginated data returned from the server.
}
