% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{CreateApplicationsBuildJobs}
\alias{CreateApplicationsBuildJobs}
\title{CreateApplicationsBuildJobs}
\usage{
CreateApplicationsBuildJobs(body, applicationId, ...)
}
\description{
Schedules a build job
}
\details{
Schedules a build job, which is normally automatically scheduled at the end of DataRobot autopilot. This is currently used in tests only, so this route is for internal use only.
}
\seealso{
The method ApplicationsBuildJobsCreate in [datarobot.apicore::ApplicationsApi], which this function wraps.

Other Applications: 
\code{\link{CreateApplicationsDeployments}()},
\code{\link{CreateApplicationsDuplicate}()},
\code{\link{CreateApplications}()},
\code{\link{CreateDemoApplications}()},
\code{\link{DeleteApplicationsDeployments}()},
\code{\link{DeleteApplications}()},
\code{\link{ListApplicationTypesUsecases}()},
\code{\link{ListApplicationTypes}()},
\code{\link{ListApplicationsAccessControl}()},
\code{\link{ListApplicationsSharedRoles}()},
\code{\link{ListApplications}()},
\code{\link{PatchApplications}()},
\code{\link{PatchManyApplicationsAccessControl}()},
\code{\link{PatchManyApplicationsAuthToken}()},
\code{\link{PatchManyApplicationsDeploymentState}()},
\code{\link{PatchManyApplicationsSharedRoles}()},
\code{\link{PatchManyApplicationsVersion}()},
\code{\link{RetrieveApplicationTypes}()},
\code{\link{RetrieveApplicationUserRole}()},
\code{\link{RetrieveApplications}()}
}
\concept{Applications}
