% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{CreateApplicationsDuplicate}
\alias{CreateApplicationsDuplicate}
\title{CreateApplicationsDuplicate}
\usage{
CreateApplicationsDuplicate(
  name,
  description,
  applicationId,
  duplicatePredictions = FALSE,
  ...
)
}
\description{
Create a duplicate of the application
}
\details{
Create a copy of App Builder application. Note that the number of active applications users can have at the same time is limited.
}
\seealso{
The method ApplicationsDuplicateCreate in [datarobot.apicore::ApplicationsApi], which this function wraps.

Other Applications: 
\code{\link{CreateApplicationsBuildJobs}()},
\code{\link{CreateApplicationsDeployments}()},
\code{\link{CreateApplications}()},
\code{\link{CreateDemoApplications}()},
\code{\link{DeleteApplicationsDeployments}()},
\code{\link{DeleteApplications}()},
\code{\link{ListApplicationTypesUsecases}()},
\code{\link{ListApplicationTypes}()},
\code{\link{ListApplicationsAccessControl}()},
\code{\link{ListApplicationsSharedRoles}()},
\code{\link{ListApplications}()},
\code{\link{PatchApplications}()},
\code{\link{PatchManyApplicationsAccessControl}()},
\code{\link{PatchManyApplicationsAuthToken}()},
\code{\link{PatchManyApplicationsDeploymentState}()},
\code{\link{PatchManyApplicationsSharedRoles}()},
\code{\link{PatchManyApplicationsVersion}()},
\code{\link{RetrieveApplicationTypes}()},
\code{\link{RetrieveApplicationUserRole}()},
\code{\link{RetrieveApplications}()}
}
\concept{Applications}
