% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{CreateApprovalPolicies}
\alias{CreateApprovalPolicies}
\title{CreateApprovalPolicies}
\usage{
CreateApprovalPolicies(
  automaticAction,
  review,
  name,
  trigger,
  active = TRUE,
  ...
)
}
\description{
Create a new Approval Policy.
}
\details{
Create a new Approval Policy.
}
\seealso{
The method ApprovalPoliciesCreate in [datarobot.apicore::GovernanceApi], which this function wraps.

Other Governance: 
\code{\link{CreateChangeRequestsReviews}()},
\code{\link{CreateChangeRequests}()},
\code{\link{DeleteApprovalPolicies}()},
\code{\link{ListApprovalPoliciesShareableChangeRequests}()},
\code{\link{ListApprovalPolicies}()},
\code{\link{ListApprovalPolicyMatch}()},
\code{\link{ListApprovalPolicyTriggers}()},
\code{\link{ListChangeRequestsReviews}()},
\code{\link{ListChangeRequestsSuggestedReviewers}()},
\code{\link{ListChangeRequests}()},
\code{\link{PatchChangeRequests}()},
\code{\link{PatchManyChangeRequestsStatus}()},
\code{\link{PutApprovalPolicies}()},
\code{\link{RetrieveApprovalPolicies}()},
\code{\link{RetrieveChangeRequestsReviews}()},
\code{\link{RetrieveChangeRequests}()}
}
\concept{Governance}
