% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{CreateAutomatedDocuments}
\alias{CreateAutomatedDocuments}
\title{CreateAutomatedDocuments}
\usage{
CreateAutomatedDocuments(
  documentTypeSpecificParameters,
  documentType,
  entityId,
  templateId,
  locale,
  outputFormat,
  ...
)
}
\description{
Request generation of automated document
}
\details{
Request generation of an automated document that's available for your account. Below is an example request body to generate Model Compliance documentation:  .. sourcecode:: json      {         \"documentType\": \"MODEL_COMPLIANCE\",         \"entityId\": \"507f191e810c19729de860ea\",         \"outputFormat\": \"docx\"     }  For Autopilot Summary, set a corresponding document type, ``AUTOPILOT_SUMMARY``, and assign a needed project ID to the ``entityId`` value.  After the request is sent, the jobs needed for document generation are queued. You can see the status of the generation by polling the URL in the ``Location`` headers. After the generation is complete, the status URL will automatically redirect you to the resource location to download the document.
}
\seealso{
The method AutomatedDocumentsCreate in [datarobot.apicore::DocumentationApi], which this function wraps.

Other Documentation: 
\code{\link{CreateOneModelComplianceDocsInitializations}()},
\code{\link{DeleteAutomatedDocuments}()},
\code{\link{DeleteComplianceDocTemplate}()},
\code{\link{GetComplianceDocTemplate}()},
\code{\link{ListAutomatedDocumentOptions}()},
\code{\link{ListAutomatedDocuments}()},
\code{\link{ListComplianceDocTemplatesDefault}()},
\code{\link{ListComplianceDocTemplatesSharedRoles}()},
\code{\link{ListComplianceDocTemplates}()},
\code{\link{PatchManyComplianceDocTemplatesSharedRoles}()},
\code{\link{RetrieveAutomatedDocuments}()},
\code{\link{RetrieveModelComplianceDocsInitializations}()},
\code{\link{UpdateComplianceDocTemplate}()},
\code{\link{UploadComplianceDocTemplate}()}
}
\concept{Documentation}
