% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{CreateBatchPredictionJobDefinitions}
\alias{CreateBatchPredictionJobDefinitions}
\title{CreateBatchPredictionJobDefinitions}
\usage{
CreateBatchPredictionJobDefinitions(
  columnNamesRemapping,
  timeseriesSettings,
  modelId,
  passthroughColumns,
  outputSettings,
  explanationNumTopClasses,
  enabled,
  predictionWarningEnabled,
  intakeSettings,
  deploymentId,
  predictionInstance,
  csvSettings,
  chunkSize,
  explanationClassNames,
  modelPackageId,
  schedule,
  numConcurrent,
  passthroughColumnsSet,
  explanationAlgorithm,
  name,
  thresholdHigh,
  thresholdLow,
  pinnedModelId,
  includeProbabilitiesClasses = c(),
  includePredictionStatus = FALSE,
  skipDriftTracking = FALSE,
  disableRowLevelErrorHandling = FALSE,
  abortOnError = TRUE,
  maxExplanations = 0,
  includeProbabilities = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Creates a new Batch Prediction job definition
}
\details{
Create a Batch Prediction Job definition. A configuration for a Batch Prediction job which can either be executed manually upon request or on scheduled intervals, if enabled. The API payload is the same as for /batchPredictions along with optional \code{enabled} and \code{schedule} items.

This method invokes \verb{POST /batchPredictionJobDefinitions/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionJobDefinitionsCreate in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
