% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{CreateBatchPredictionsFromExisting}
\alias{CreateBatchPredictionsFromExisting}
\title{CreateBatchPredictionsFromExisting}
\usage{
CreateBatchPredictionsFromExisting(predictionJobId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new a Batch Prediction job based on an existing Batch Prediction job.
}
\details{
Copies an existing job and submits it to the queue.

This method invokes \verb{POST /batchPredictions/fromExisting/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsFromExistingCreate in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
