% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{CreateBatchPredictionsFromJobDefinition}
\alias{CreateBatchPredictionsFromJobDefinition}
\title{CreateBatchPredictionsFromJobDefinition}
\usage{
CreateBatchPredictionsFromJobDefinition(jobDefinitionId, ...)
}
\description{
Launch a Batch Prediction job for scoring
}
\details{
Launches a one-time batch prediction job based off of the previously supplied definition referring to the job definition ID and puts it on the queue.
}
\seealso{
The method BatchPredictionsFromJobDefinitionCreate in [datarobot.apicore::PredictionsApi], which this function wraps.

Other Predictions: 
\code{\link{CreateBatchPredictionJobDefinitions}()},
\code{\link{CreateBatchPredictionsCsvUploadFinalizeMultipart}()},
\code{\link{CreateBatchPredictionsFromExisting}()},
\code{\link{CreateBatchPredictions}()},
\code{\link{CreateProjectsPredictionDatasetsFileUploads}()},
\code{\link{DeleteBatchPredictionJobDefinitions}()},
\code{\link{DeleteBatchPredictions}()},
\code{\link{DeletePredictJob}()},
\code{\link{DeletePredictionDataset}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteScheduledJobs}()},
\code{\link{DeleteTransferableModel}()},
\code{\link{GetPredictJobs}()},
\code{\link{GetPredictJob}()},
\code{\link{GetPredictionDataset}()},
\code{\link{GetPredictionExplanationsMetadata}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetTransferableModel}()},
\code{\link{ListBatchPredictionJobDefinitionsPortable}()},
\code{\link{ListBatchPredictionJobDefinitions}()},
\code{\link{ListBatchPredictionsDownload}()},
\code{\link{ListBatchPredictions}()},
\code{\link{ListPredictionDatasets}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListPredictions}()},
\code{\link{ListProjectsPredictionsMetadata}()},
\code{\link{ListScheduledJobs}()},
\code{\link{ListTrainingPredictions}()},
\code{\link{ListTransferableModels}()},
\code{\link{PatchBatchPredictionJobDefinitions}()},
\code{\link{PatchBatchPredictions}()},
\code{\link{PatchScheduledJobs}()},
\code{\link{PutBatchPredictionsCsvUploadPart}()},
\code{\link{PutImportedModels}()},
\code{\link{PutManyBatchPredictionsCsvUpload}()},
\code{\link{RetrieveBatchPredictionJobDefinitions}()},
\code{\link{RetrieveBatchPredictions}()},
\code{\link{RetrieveProjectsPredictionsMetadata}()},
\code{\link{RetrieveProjectsPredictions}()},
\code{\link{RetrieveScheduledJobs}()},
\code{\link{UpdateTransferableModel}()},
\code{\link{UploadPredictionDatasetFromCatalog}()},
\code{\link{UploadPredictionDatasetFromDataSource}()},
\code{\link{UploadPredictionDataset}()},
\code{\link{UploadTransferableModel}()}
}
\concept{Predictions}
