% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendars.R, R/projects_apiWrapper.R
\name{CreateCalendar}
\alias{CreateCalendar}
\title{Create a calendar from an uploaded CSV.}
\usage{
CreateCalendar(
  dataSource,
  name = NULL,
  multiSeriesIdColumn = NULL,
  maxWait = 600,
  ...
)

CreateCalendar(
  dataSource,
  name = NULL,
  multiSeriesIdColumn = NULL,
  maxWait = 600,
  ...
)
}
\arguments{
\item{dataSource}{object. Either (a) the name of a CSV file, or (b) a dataframe.
This parameter identifies the source of the calendar data.}

\item{name}{character. Optional. The name of the calendar.}

\item{multiSeriesIdColumn}{character. Optional. Added in 2.19. The column in
the calendar that defines which series an event belongs to. Only one
column is supported.}

\item{maxWait}{integer. The maximum time (in seconds) to wait for the retrieve to complete.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Create a calendar from a file.
}
\details{
Create a calendar from a file in a csv or xlsx format. The calendar file specifies the dates or events in a dataset such that DataRobot automatically derives and creates special features based on the calendar events (e.g., time until the next event, labeling the most recent event).
}
\examples{
\dontrun{
CreateCalendar("inst/extdata/calendar.csv", name = "intlHolidayCalendar")
}
\dontrun{
holidayCalendarDF <- as.data.frame(myCalendar)
CreateCalendar(holidayCalendarDF, name = "intlHolidayCalendar")
}
\dontrun{
CreateCalendar("inst/extdata/calendar.csv",
  name = "intlHolidayCalendar",
  multiSeriesIdColumn = "Country"
)
}
}
\seealso{
The method CalendarsFileUploadCreate in [datarobot.apicore::ProjectsApi], which this function wraps.

Other Projects: 
\code{\link{BatchFeaturesTypeTransform}()},
\code{\link{CloneProject}()},
\code{\link{CreateCalendarsFromCountryCode}()},
\code{\link{CreateCalendarsFromDataset}()},
\code{\link{CreateFeaturelist}()},
\code{\link{CreateHdfsProjects}()},
\code{\link{CreateModelingFeaturelist}()},
\code{\link{CreateProjectCleanupJobs}()},
\code{\link{CreateProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{CreateProjectsFeatureDiscoveryRecipeSqlExports}()},
\code{\link{CreateProjectsSecondaryDatasetsConfigurations}()},
\code{\link{CreateProjectsSegmentationTasks}()},
\code{\link{CreateRelationshipsConfigurations}()},
\code{\link{DeleteCalendar}()},
\code{\link{DeleteProjectCleanupJobs}()},
\code{\link{DeleteProjectsJobs}()},
\code{\link{DeleteProjectsSecondaryDatasetsConfigurations}()},
\code{\link{DeleteProject}()},
\code{\link{DeleteRelationshipsConfigurations}()},
\code{\link{GetCalendar}()},
\code{\link{GetFeatureHistogram}()},
\code{\link{GetFeatureInfo}()},
\code{\link{GetFeaturelist}()},
\code{\link{GetJob}()},
\code{\link{GetModelingFeaturelist}()},
\code{\link{GetProjectStatus}()},
\code{\link{GetProject}()},
\code{\link{GetValidMetrics}()},
\code{\link{ListCalendarCountryCodes}()},
\code{\link{ListCalendarsAccessControl}()},
\code{\link{ListCalendars}()},
\code{\link{ListDeletedProjectsCount}()},
\code{\link{ListDeletedProjects}()},
\code{\link{ListFeatureInfo}()},
\code{\link{ListFeaturelists}()},
\code{\link{ListJobs}()},
\code{\link{ListModelingFeaturelists}()},
\code{\link{ListProjectCleanupJobsDownload}()},
\code{\link{ListProjectCleanupJobsSummary}()},
\code{\link{ListProjectCleanupJobs}()},
\code{\link{ListProjectsAccessControl}()},
\code{\link{ListProjectsCalendarEvents}()},
\code{\link{ListProjectsFeatureDiscoveryDatasetDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogsDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogs}()},
\code{\link{ListProjectsFeatureDiscoveryRecipeSQLsDownload}()},
\code{\link{ListProjectsModelingFeatures}()},
\code{\link{ListProjectsMultiseriesIdsCrossSeriesProperties}()},
\code{\link{ListProjectsMultiseriesNames}()},
\code{\link{ListProjectsRelationshipsConfiguration}()},
\code{\link{ListProjectsSecondaryDatasetsConfigurations}()},
\code{\link{ListProjectsSegmentationTasksMappings}()},
\code{\link{ListProjectsSegmentationTasks}()},
\code{\link{ListProjects}()},
\code{\link{PatchDeletedProjects}()},
\code{\link{PatchManyCalendarsAccessControl}()},
\code{\link{PatchManyProjectsAccessControl}()},
\code{\link{PatchProjectsSegments}()},
\code{\link{PauseQueue}()},
\code{\link{PutRelationshipsConfigurations}()},
\code{\link{RequestCrossSeriesDetection}()},
\code{\link{RequestMultiSeriesDetection}()},
\code{\link{RestoreDiscardedFeatures}()},
\code{\link{RetrieveDiscardedFeaturesInformation}()},
\code{\link{RetrieveProjectCleanupJobs}()},
\code{\link{RetrieveProjectsBatchTypeTransformFeaturesResult}()},
\code{\link{RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{RetrieveProjectsFeatureLineages}()},
\code{\link{RetrieveProjectsModelingFeatures}()},
\code{\link{RetrieveProjectsSecondaryDatasetsConfigurations}()},
\code{\link{RetrieveProjectsSegmentationTaskJobResults}()},
\code{\link{RetrieveProjectsSegmentationTasks}()},
\code{\link{RetrieveRelationshipsConfigurations}()},
\code{\link{SetTarget}()},
\code{\link{SetupProject}()},
\code{\link{StartNewAutoPilot}()},
\code{\link{UpdateCalendar}()},
\code{\link{UpdateProject}()}
}
\concept{Projects}
