% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateCalendarsFromCountryCode}
\alias{CreateCalendarsFromCountryCode}
\title{CreateCalendarsFromCountryCode}
\usage{
CreateCalendarsFromCountryCode(endDate, countryCode, startDate, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Initialize generation of preloaded calendars.
}
\details{
Initialize generation of preloaded calendars. Preloaded calendars are available only for time series projects. Preloaded calendars do not support multiseries calendars.

This method invokes \verb{POST /calendars/fromCountryCode/} in the DataRobot Public API.
}
\seealso{
The method CalendarsFromCountryCodeCreate in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
