% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateCalendarsFromDataset}
\alias{CreateCalendarsFromDataset}
\title{CreateCalendarsFromDataset}
\usage{
CreateCalendarsFromDataset(
  name,
  deleteOnError,
  multiseriesIdColumns,
  datasetId,
  datasetVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a calendar from the dataset
}
\details{
Create a calendar from the dataset.

This method invokes \verb{POST /calendars/fromDataset/} in the DataRobot Public API.
}
\seealso{
The method CalendarsFromDatasetCreate in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
