% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{CreateChangeRequests}
\alias{CreateChangeRequests}
\title{CreateChangeRequests}
\usage{
CreateChangeRequests(changeRequestCreate, ...)
}
\description{
Create Change Request.
}
\details{
Request changes for a supported product entity. For now, you can request changes for deployments only.
}
\seealso{
The method ChangeRequestsCreate in [datarobot.apicore::GovernanceApi], which this function wraps.

Other Governance: 
\code{\link{CreateApprovalPolicies}()},
\code{\link{CreateChangeRequestsReviews}()},
\code{\link{DeleteApprovalPolicies}()},
\code{\link{ListApprovalPoliciesShareableChangeRequests}()},
\code{\link{ListApprovalPolicies}()},
\code{\link{ListApprovalPolicyMatch}()},
\code{\link{ListApprovalPolicyTriggers}()},
\code{\link{ListChangeRequestsReviews}()},
\code{\link{ListChangeRequestsSuggestedReviewers}()},
\code{\link{ListChangeRequests}()},
\code{\link{PatchChangeRequests}()},
\code{\link{PatchManyChangeRequestsStatus}()},
\code{\link{PutApprovalPolicies}()},
\code{\link{RetrieveApprovalPolicies}()},
\code{\link{RetrieveChangeRequestsReviews}()},
\code{\link{RetrieveChangeRequests}()}
}
\concept{Governance}
