% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_apiWrapper.R
\name{CreateClusterLicenseValidation}
\alias{CreateClusterLicenseValidation}
\title{CreateClusterLicenseValidation}
\usage{
CreateClusterLicenseValidation(licenseKey, ...)
}
\description{
Check if a license is valid
}
\details{
Check if a cluster license is valid.
}
\seealso{
The method ClusterLicenseValidationCreate in [datarobot.apicore::InfrastructureApi], which this function wraps.

Other Infrastructure: 
\code{\link{GetServerVersion}()},
\code{\link{ListClusterLicense}()},
\code{\link{PutManyClusterLicense}()}
}
\concept{Infrastructure}
