% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{CreateComments}
\alias{CreateComments}
\title{CreateComments}
\usage{
CreateComments(entityType, mentions, entityId, content, ...)
}
\description{
Post a comment
}
\details{
Post a comment
}
\seealso{
The method CommentsCreate in [datarobot.apicore::CommentsApi], which this function wraps.

Other Comments: 
\code{\link{DeleteComments}()},
\code{\link{ListComments}()},
\code{\link{PatchComments}()},
\code{\link{RetrieveComments}()}
}
\concept{Comments}
