% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{CreateCredentials}
\alias{CreateCredentials}
\title{CreateCredentials}
\usage{
CreateCredentials(credentialsBody, ...)
}
\description{
Store a new set of credentials which can be used for data sources creation.
}
\details{
Store a new set of credentials.
}
\seealso{
The method CredentialsCreate in [datarobot.apicore::CredentialsApi], which this function wraps.

Other Credentials: 
\code{\link{DeleteCredentials}()},
\code{\link{ListCredentialsAssociations}()},
\code{\link{ListCredentials}()},
\code{\link{PatchCredentials}()},
\code{\link{PatchManyCredentialsAssociations}()},
\code{\link{RetrieveCredentials}()}
}
\concept{Credentials}
