% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModels}
\alias{CreateCustomModels}
\title{CreateCustomModels}
\usage{
CreateCustomModels(
  networkEgressPolicy,
  negativeClassLabel,
  targetName,
  replicas,
  supportsRegression,
  maximumMemory,
  description,
  customModelType,
  targetType,
  language,
  requiresHa,
  classLabels,
  supportsBinaryClassification,
  name,
  positiveClassLabel,
  desiredMemory,
  calibratePredictions = TRUE,
  predictionThreshold = 0.5,
  ...
)
}
\description{
Create custom model.
}
\details{
Creates a new custom model and returns the newly created metadata record for it.  All custom models must support at least one target type (binaryClassification, regression). Custom inference models can only support a single target type.  A regression model is expected to produce predictions that are arbitrary floating-point or integer numbers. A classification model is expected to return predictions with probability scores for each class.  For example, a binary classification model might return:  .. code:: Python      {         positiveClassLabel: probability,         negativeClassLabel: 1.0 - probability     }  For Custom Inference Models, the ``file`` parameter must be either a tarball or zip archive containing, at minimum, a script named ``start_server.sh``.  It may contain additional files, including scripts and precompiled binaries as well as data files.  ``start_server.sh`` may execute these scripts and/or binaries.  When this script is executed, it is run as part of an Environment (specified via subsequent API calls), and all included scripts and binaries can take advantage of any programming language interpreters, compilers, libraries, or other tools included in the Environment. ``start_server.sh`` must be marked as executable (``chmod +x``).  When ``start_server.sh`` is launched, it must launch and maintain (in the foreground) a Web server that listens on two URLs:  * ``GET $URL_PREFIX/``     This route must return a 200 response code with an empty body immediately     if the server is ready to respond to prediction requests.  Otherwise it should     either not accept the request, not respond to the request, or return a     503 response code. * ``POST $URL_PREFIX/predict_no_state/``     This route must accept as input a JSON object of the form:      .. code-block:: Python          {             'X': {                 'col1': [...col1_data...],                 'col2': [...col2_data...],                 'col3': [...col3_data...],                 ...             }         }      The data lists will all be the same length.      It must return a JSON object of the form:      .. code-block:: Python          {             'predictions': [...predictions data...]         }      The predictions data must correspond 1:1 to the rows in the input data lists.      ``$URL_PREFIX`` is provided as an environment variable.  The Web server process must     re-read its value every time the process starts, as it may change.     It is an opaque string that is guaranteed to be a valid URL component,     but may contain path separators (``/``).
}
\seealso{
The method CustomModelsCreate in [datarobot.apicore::ModelsApi], which this function wraps.

Other Models: 
\code{\link{AddEureqaSolution}()},
\code{\link{CreateCustomInferenceImagesFeatureImpact}()},
\code{\link{CreateCustomModelTests}()},
\code{\link{CreateCustomModelsFromCustomModel}()},
\code{\link{CreateCustomModelsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersionsConversions}()},
\code{\link{CreateCustomModelsVersionsDependencyBuild}()},
\code{\link{CreateCustomModelsVersionsFeatureImpact}()},
\code{\link{CreateCustomModelsVersionsFromRepository}()},
\code{\link{CreateCustomModelsVersionsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersions}()},
\code{\link{CreateCustomTrainingBlueprints}()},
\code{\link{CreateFromLatestCustomModelsVersion}()},
\code{\link{CreateModelExports}()},
\code{\link{CreateModelPackagesArchive}()},
\code{\link{CreateModelPackagesFromLearningModel}()},
\code{\link{CreatePrimeCode}()},
\code{\link{CreateProjectsBlenderBuilders}()},
\code{\link{CreateProjectsDatetimeModelsFromModel}()},
\code{\link{CreateProjectsDeploymentReadyModels}()},
\code{\link{CreateProjectsModelsFromModel}()},
\code{\link{CreateProjectsModelsPredictionIntervals}()},
\code{\link{CreateRatingTable}()},
\code{\link{DeleteCustomModelTests}()},
\code{\link{DeleteCustomModelsVersionsConversions}()},
\code{\link{DeleteCustomModels}()},
\code{\link{DeleteManyCustomModelsVersionsDependencyBuild}()},
\code{\link{DeleteModelJob}()},
\code{\link{DownloadPrimeCode}()},
\code{\link{DownloadRatingTable}()},
\code{\link{DownloadScoringCode}()},
\code{\link{DownloadTransferableModel}()},
\code{\link{GetBlenderModel}()},
\code{\link{GetDatetimeModel}()},
\code{\link{GetFrozenModel}()},
\code{\link{GetMissingValuesReport}()},
\code{\link{GetModelJob}()},
\code{\link{GetModelParameters}()},
\code{\link{GetModel}()},
\code{\link{GetPrimeEligibility}()},
\code{\link{GetPrimeFile}()},
\code{\link{GetPrimeModel}()},
\code{\link{GetRatingTableModel}()},
\code{\link{GetRatingTable}()},
\code{\link{GetRulesets}()},
\code{\link{IsBlenderEligible}()},
\code{\link{ListCustomInferenceImagesFeatureImpact}()},
\code{\link{ListCustomModelDeployments}()},
\code{\link{ListCustomModelLimits}()},
\code{\link{ListCustomModelTestsLog}()},
\code{\link{ListCustomModelTestsTail}()},
\code{\link{ListCustomModelTests}()},
\code{\link{ListCustomModelsAccessControl}()},
\code{\link{ListCustomModelsDownload}()},
\code{\link{ListCustomModelsVersionsConversions}()},
\code{\link{ListCustomModelsVersionsDependencyBuildLog}()},
\code{\link{ListCustomModelsVersionsDependencyBuild}()},
\code{\link{ListCustomModelsVersionsDownload}()},
\code{\link{ListCustomModelsVersionsFeatureImpact}()},
\code{\link{ListCustomModelsVersions}()},
\code{\link{ListCustomModels}()},
\code{\link{ListCustomTrainingBlueprints}()},
\code{\link{ListModelJobs}()},
\code{\link{ListModelPackagesFeatures}()},
\code{\link{ListModelPackagesSharedRoles}()},
\code{\link{ListModelPackages}()},
\code{\link{ListModelRecommendations}()},
\code{\link{ListModels}()},
\code{\link{ListPrimeFiles}()},
\code{\link{ListPrimeModels}()},
\code{\link{ListProjectsBlenderBuilders}()},
\code{\link{ListProjectsBlenderModels}()},
\code{\link{ListProjectsCombinedModelsSegmentsDownload}()},
\code{\link{ListProjectsCombinedModelsSegments}()},
\code{\link{ListProjectsCombinedModels}()},
\code{\link{ListProjectsDatetimeModels}()},
\code{\link{ListProjectsFrozenModels}()},
\code{\link{ListProjectsModelsClusterNames}()},
\code{\link{ListProjectsModelsNumIterationsTrained}()},
\code{\link{ListProjectsModelsPredictionIntervals}()},
\code{\link{ListProjectsRecommendedModelsRecommendedModel}()},
\code{\link{ListRatingTableModels}()},
\code{\link{ListRatingTables}()},
\code{\link{PatchCustomModelsVersions}()},
\code{\link{PatchCustomModels}()},
\code{\link{PatchManyCustomModelsAccessControl}()},
\code{\link{PatchManyCustomModelsTrainingData}()},
\code{\link{PatchManyCustomModelsVersionsFromRepository}()},
\code{\link{PatchManyProjectsModelsClusterNames}()},
\code{\link{PutManyProjectsSegmentChampion}()},
\code{\link{RenameRatingTable}()},
\code{\link{RequestApproximation}()},
\code{\link{RequestBlender}()},
\code{\link{RequestNewDatetimeModel}()},
\code{\link{RequestNewModel}()},
\code{\link{RequestNewRatingTableModel}()},
\code{\link{RequestPrimeModel}()},
\code{\link{RetrieveCustomModelTests}()},
\code{\link{RetrieveCustomModelsVersionsConversions}()},
\code{\link{RetrieveCustomModelsVersions}()},
\code{\link{RetrieveCustomModels}()},
\code{\link{RetrieveModelPackages}()},
\code{\link{RetrieveProjectsBlenderBuilders}()},
\code{\link{RetrieveProjectsCombinedModels}()},
\code{\link{RetrieveProjectsEureqaDistributionPlot}()},
\code{\link{RetrieveProjectsEureqaModelDetail}()}
}
\concept{Models}
