% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModelsFromCustomModel}
\alias{CreateCustomModelsFromCustomModel}
\title{CreateCustomModelsFromCustomModel}
\usage{
CreateCustomModelsFromCustomModel(customModelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Clone custom model.
}
\details{
Creates a copy of the provided custom model, including metadata, versions of that model, and uploaded files. Associates the new versions with files owned by the custom model.

This method invokes \verb{POST /customModels/fromCustomModel/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsFromCustomModelCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
