% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModelsPredictionExplanationsInitialization}
\alias{CreateCustomModelsPredictionExplanationsInitialization}
\title{CreateCustomModelsPredictionExplanationsInitialization}
\usage{
CreateCustomModelsPredictionExplanationsInitialization(
  environmentId,
  customModelId,
  environmentVersionId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new prediction explanations initialization for custom model.
}
\details{
Create a new prediction explanations initialization for custom model. This is a necessary prerequisite for generating prediction explanations.  v2.23 DEPRECATED please use custom model version route instead: :\url{http:post}:/api/v2/customModels/(customModelId)/versions/(customModelVersionId)/predictionExplanationsInitialization/

This method invokes \verb{POST /customModels/predictionExplanationsInitialization/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsPredictionExplanationsInitializationCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
