% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModelsVersions}
\alias{CreateCustomModelsVersions}
\title{CreateCustomModelsVersions}
\usage{
CreateCustomModelsVersions(
  requiredMetadata,
  networkEgressPolicy,
  requiresHa,
  file,
  baseEnvironmentVersionId,
  baseEnvironmentId,
  customModelId,
  replicas,
  filePath,
  maximumMemory,
  requiredMetadataValues,
  desiredMemory,
  isMajorUpdate = "true",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create custom model version.
}
\details{
Create a new custom model version with attached files if supplied.

This method invokes \verb{POST /customModels/\{customModelId\}/versions/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
