% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModelsVersionsDependencyBuild}
\alias{CreateCustomModelsVersionsDependencyBuild}
\title{CreateCustomModelsVersionsDependencyBuild}
\usage{
CreateCustomModelsVersionsDependencyBuild(
  customModelId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Start a custom model version's dependency build.
}
\details{
Start a custom model version's dependency build. This is required to test, deploy, or train custom models.

This method invokes \verb{POST /customModels/\{customModelId\}/versions/\{customModelVersionId\}/dependencyBuild/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsDependencyBuildCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
