% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModelsVersionsFeatureImpact}
\alias{CreateCustomModelsVersionsFeatureImpact}
\title{CreateCustomModelsVersionsFeatureImpact}
\usage{
CreateCustomModelsVersionsFeatureImpact(
  customModelId,
  rowCount,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create custom model feature impact.
}
\details{
Add a request to calculate feature impact for a custom inference model image to the queue.  This route is a counterpart of a corresponding endpoint for native models: :\url{http:post}:/api/v2/projects/(projectId)/models/(modelId)/featureImpact/

This method invokes \verb{POST /customModels/\{customModelId\}/versions/\{customModelVersionId\}/featureImpact/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsFeatureImpactCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
