% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{CreateCustomTasksVersions}
\alias{CreateCustomTasksVersions}
\title{CreateCustomTasksVersions}
\usage{
CreateCustomTasksVersions(
  requiredMetadata,
  file,
  baseEnvironmentId,
  filePath,
  customTaskId,
  maximumMemory,
  requiredMetadataValues,
  isMajorUpdate = "true",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create custom task version.
}
\details{
Create a new custom task version with attached files if supplied.

This method invokes \verb{POST /customTasks/\{customTaskId\}/versions/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsCreate in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
