% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{CreateCustomTasksVersionsDependencyBuild}
\alias{CreateCustomTasksVersionsDependencyBuild}
\title{CreateCustomTasksVersionsDependencyBuild}
\usage{
CreateCustomTasksVersionsDependencyBuild(
  customTaskVersionId,
  customTaskId,
  ...
)
}
\description{
Start a custom task version's dependency build.
}
\details{
Start a custom task version's dependency build. This is required to test, deploy, or train custom tasks.
}
\seealso{
The method CustomTasksVersionsDependencyBuildCreate in [datarobot.apicore::CustomTasksApi], which this function wraps.

Other CustomTasks: 
\code{\link{CreateCustomTasksFromCustomTask}()},
\code{\link{CreateCustomTasksVersionsFromRepository}()},
\code{\link{CreateCustomTasksVersions}()},
\code{\link{CreateCustomTasks}()},
\code{\link{CreateFromLatestCustomTaskVersion}()},
\code{\link{DeleteCustomTasks}()},
\code{\link{DeleteManyCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksAccessControl}()},
\code{\link{ListCustomTasksDownload}()},
\code{\link{ListCustomTasksVersionsDependencyBuildLog}()},
\code{\link{ListCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksVersionsDownload}()},
\code{\link{ListCustomTasksVersions}()},
\code{\link{ListCustomTasks}()},
\code{\link{PatchCustomTasksVersions}()},
\code{\link{PatchCustomTasks}()},
\code{\link{PatchManyCustomTasksAccessControl}()},
\code{\link{PatchManyCustomTasksVersionsFromRepository}()},
\code{\link{RetrieveCustomTasksVersions}()},
\code{\link{RetrieveCustomTasks}()}
}
\concept{CustomTasks}
