% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{CreateCustomTasksVersionsFromRepository}
\alias{CreateCustomTasksVersionsFromRepository}
\title{CreateCustomTasksVersionsFromRepository}
\usage{
CreateCustomTasksVersionsFromRepository(
  requiredMetadata,
  ref,
  baseEnvironmentId,
  customTaskId,
  repositoryId,
  requiredMetadataValues,
  sourcePath,
  isMajorUpdate = TRUE,
  ...
)
}
\description{
Create custom task version from remote repository.
}
\details{
Create a new custom task version with only files added from the specified remote repository.
}
\seealso{
The method CustomTasksVersionsFromRepositoryCreate in [datarobot.apicore::CustomTasksApi], which this function wraps.

Other CustomTasks: 
\code{\link{CreateCustomTasksFromCustomTask}()},
\code{\link{CreateCustomTasksVersionsDependencyBuild}()},
\code{\link{CreateCustomTasksVersions}()},
\code{\link{CreateCustomTasks}()},
\code{\link{CreateFromLatestCustomTaskVersion}()},
\code{\link{DeleteCustomTasks}()},
\code{\link{DeleteManyCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksAccessControl}()},
\code{\link{ListCustomTasksDownload}()},
\code{\link{ListCustomTasksVersionsDependencyBuildLog}()},
\code{\link{ListCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksVersionsDownload}()},
\code{\link{ListCustomTasksVersions}()},
\code{\link{ListCustomTasks}()},
\code{\link{PatchCustomTasksVersions}()},
\code{\link{PatchCustomTasks}()},
\code{\link{PatchManyCustomTasksAccessControl}()},
\code{\link{PatchManyCustomTasksVersionsFromRepository}()},
\code{\link{RetrieveCustomTasksVersions}()},
\code{\link{RetrieveCustomTasks}()}
}
\concept{CustomTasks}
