% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataEngineQueryGenerator.R,
%   R/ai_catalog_apiWrapper.R
\name{CreateDatasetFromDataEngineQueryGenerator}
\alias{CreateDatasetFromDataEngineQueryGenerator}
\title{Create a Dataset from a DataEngineQueryGenerator for time series data prep.}
\usage{
CreateDatasetFromDataEngineQueryGenerator(
  generatorId,
  datasetId = NULL,
  datasetVersionId = NULL,
  ...
)

CreateDatasetFromDataEngineQueryGenerator(
  generatorId,
  datasetId = NULL,
  datasetVersionId = NULL,
  ...
)
}
\arguments{
\item{generatorId}{character. The ID of the query generator to use.}

\item{datasetId}{character. Optional. The ID of the dataset to apply the query to.}

\item{datasetVersionId}{character. Optional. The ID of the dataset version to apply the query to.}
}
\value{
A dataset object:
\itemize{
\item datasetId. character. The ID of the created dataset.
\item versionId. character. The ID of the created dataset version.
\item name. character. The name of the created dataset.
\item categories. list of character. The categories of the created dataset.
\item creationDate. POSIXct. The timestamp the dataset was created.
\item createdBy. character. The user who created the dataset.
\item dataPersisted. logical. Whether the dataset has extended data profile available.
\item isDataEngineEligible. logical. Whether the dataset is eligible for data engine queries.
\item isLatestVersion. logical. Whether the dataset is the latest version.
\item isSnapshot. logical. Whether the dataset is a static snapshot.
\item datasetSize. numeric. Size of the dataset as a CSV in bytes.
\item rowCount. numeric. The number of rows in the dataset.
\item processingState. character. The current ingestion process state of the dataset.
}
}
\description{
Create Data Engine workspace state from a query generator
}
\details{
Create Data Engine workspace state in database from a query generator.
}
\examples{
\dontrun{
generatorId <- "59a5af20c80891534e3c2bde"
dataset <- CreateDatasetFromDataEngineQueryGenerator(generatorId)
}
}
\seealso{
The method DataEngineWorkspaceStatesFromDataEngineQueryGeneratorCreate in [datarobot.apicore::AiCatalogApi], which this function wraps.

Other Data Engine Query Generator functions: 
\code{\link{CreateDataEngineQueryGenerator}()},
\code{\link{GetDataEngineQueryGenerator}()}

Other AiCatalog: 
\code{\link{CreateDataEngineQueryGenerator}()},
\code{\link{CreateDataEngineWorkspaceStates}()},
\code{\link{CreateDatasetsFeatureTransforms}()},
\code{\link{CreateDatasetsFeaturelists}()},
\code{\link{CreateDatasetsFromDataEngineWorkspaceState}()},
\code{\link{CreateDatasetsFromDataSource}()},
\code{\link{CreateDatasetsFromFile}()},
\code{\link{CreateDatasetsFromHDFS}()},
\code{\link{CreateDatasetsFromURL}()},
\code{\link{CreateDatasetsRefreshJobs}()},
\code{\link{CreateDatasetsRelationships}()},
\code{\link{CreateDatasetsVersionsFromDataEngineWorkspaceState}()},
\code{\link{CreateDatasetsVersionsFromDataSource}()},
\code{\link{CreateDatasetsVersionsFromFile}()},
\code{\link{CreateDatasetsVersionsFromHDFS}()},
\code{\link{CreateDatasetsVersionsFromLatestVersion}()},
\code{\link{CreateDatasetsVersionsFromURL}()},
\code{\link{CreateDatasetsVersionsFromVersion}()},
\code{\link{CreateUserBlueprintsBulkValidations}()},
\code{\link{CreateUserBlueprintsFromBlueprintId}()},
\code{\link{CreateUserBlueprintsFromCustomTaskVersionId}()},
\code{\link{CreateUserBlueprintsFromUserBlueprintId}()},
\code{\link{CreateUserBlueprintsProjectBlueprints}()},
\code{\link{CreateUserBlueprintsTaskParameters}()},
\code{\link{CreateUserBlueprintsValidations}()},
\code{\link{CreateUserBlueprints}()},
\code{\link{DeleteDatasetsFeaturelists}()},
\code{\link{DeleteDatasetsRefreshJobs}()},
\code{\link{DeleteDatasetsRelationships}()},
\code{\link{DeleteDatasetsVersions}()},
\code{\link{DeleteDatasets}()},
\code{\link{DeleteManyUserBlueprints}()},
\code{\link{DeleteUserBlueprints}()},
\code{\link{DownloadDatasetAsCsv}()},
\code{\link{ListCatalogItems}()},
\code{\link{ListDatasetsAccessControl}()},
\code{\link{ListDatasetsAllFeaturesDetails}()},
\code{\link{ListDatasetsFeatureTransforms}()},
\code{\link{ListDatasetsFeaturelists}()},
\code{\link{ListDatasetsFile}()},
\code{\link{ListDatasetsPermissions}()},
\code{\link{ListDatasetsProjects}()},
\code{\link{ListDatasetsRefreshJobsExecutionResults}()},
\code{\link{ListDatasetsRefreshJobs}()},
\code{\link{ListDatasetsRelationships}()},
\code{\link{ListDatasetsSharedRoles}()},
\code{\link{ListDatasetsVersionsAllFeaturesDetails}()},
\code{\link{ListDatasetsVersionsFeaturelists}()},
\code{\link{ListDatasetsVersionsFile}()},
\code{\link{ListDatasetsVersionsProjects}()},
\code{\link{ListDatasetsVersions}()},
\code{\link{ListDatasets}()},
\code{\link{ListUserBlueprintsInputTypes}()},
\code{\link{ListUserBlueprintsSharedRoles}()},
\code{\link{ListUserBlueprintsTasks}()},
\code{\link{ListUserBlueprints}()},
\code{\link{PatchCatalogItems}()},
\code{\link{PatchDatasetsFeaturelists}()},
\code{\link{PatchDatasetsRefreshJobs}()},
\code{\link{PatchDatasetsRelationships}()},
\code{\link{PatchDatasets}()},
\code{\link{PatchManyDatasetsAccessControl}()},
\code{\link{PatchManyDatasetsDeleted}()},
\code{\link{PatchManyDatasetsSharedRoles}()},
\code{\link{PatchManyDatasetsVersionsDeleted}()},
\code{\link{PatchManyDatasets}()},
\code{\link{PatchManyUserBlueprintsSharedRoles}()},
\code{\link{PatchUserBlueprints}()},
\code{\link{RetrieveCatalogItems}()},
\code{\link{RetrieveDataEngineQueryGenerators}()},
\code{\link{RetrieveDataEngineWorkspaceStates}()},
\code{\link{RetrieveDatasetsFeatureHistograms}()},
\code{\link{RetrieveDatasetsFeatureTransforms}()},
\code{\link{RetrieveDatasetsFeaturelists}()},
\code{\link{RetrieveDatasetsRefreshJobs}()},
\code{\link{RetrieveDatasetsVersionsFeatureHistograms}()},
\code{\link{RetrieveDatasetsVersionsFeaturelists}()},
\code{\link{RetrieveDatasetsVersions}()},
\code{\link{RetrieveDatasets}()},
\code{\link{RetrieveUserBlueprints}()}
}
\concept{AiCatalog}
\concept{Data Engine Query Generator functions}
