% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsFromDataEngineWorkspaceState}
\alias{CreateDatasetsFromDataEngineWorkspaceState}
\title{CreateDatasetsFromDataEngineWorkspaceState}
\usage{
CreateDatasetsFromDataEngineWorkspaceState(
  credentials,
  datasetName,
  workspaceStateId,
  doSnapshot,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset from Data Engine workspace
}
\details{
Create a dataset from a Data Engine workspace state.

This method invokes \verb{POST /datasets/fromDataEngineWorkspaceState/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFromDataEngineWorkspaceStateCreate in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
