% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsVersionsFromDataSource}
\alias{CreateDatasetsVersionsFromDataSource}
\title{CreateDatasetsVersionsFromDataSource}
\usage{
CreateDatasetsVersionsFromDataSource(
  password,
  dataSourceId,
  persistDataAfterIngestion,
  credentialId,
  datasetId,
  credentialData,
  categories,
  sampleSize,
  user,
  doSnapshot = TRUE,
  useKerberos = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset version from Data Source
}
\details{
Create a new version for the specified dataset from specified Data Source. The dataset must have been created from a compatible data source originally.

This method invokes \verb{POST /datasets/\{datasetId\}/versions/fromDataSource/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsFromDataSourceCreate in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
