% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsVersionsFromHDFS}
\alias{CreateDatasetsVersionsFromHDFS}
\title{CreateDatasetsVersionsFromHDFS}
\usage{
CreateDatasetsVersionsFromHDFS(
  password,
  persistDataAfterIngestion,
  datasetId,
  namenodeWebhdfsPort,
  categories,
  user,
  url,
  doSnapshot = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset version from HDFS URL
}
\details{
Create a new version for the specified dataset from a HDFS URL. The dataset must have been created from the same HDFS URL originally.

This method invokes \verb{POST /datasets/\{datasetId\}/versions/fromHDFS/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsFromHDFSCreate in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
