% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsVersionsFromURL}
\alias{CreateDatasetsVersionsFromURL}
\title{CreateDatasetsVersionsFromURL}
\usage{
CreateDatasetsVersionsFromURL(
  persistDataAfterIngestion,
  datasetId,
  categories,
  sampleSize,
  url,
  doSnapshot = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset version from URL
}
\details{
Create a new version for the specified dataset from specified URL. The dataset must have been created from the same URL originally.

This method invokes \verb{POST /datasets/\{datasetId\}/versions/fromURL/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsFromURLCreate in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
