% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataSources.R
\name{CreateDataSource}
\alias{CreateDataSource}
\title{CreateDataSource}
\usage{
CreateDataSource(
  type,
  canonicalName,
  dataStoreId,
  query = NULL,
  table = NULL,
  schema = NULL,
  partitionColumn = NULL,
  fetchSize = NULL
)

CreateDataSource(
  type,
  canonicalName,
  dataStoreId,
  query = NULL,
  table = NULL,
  schema = NULL,
  partitionColumn = NULL,
  fetchSize = NULL
)
}
\arguments{
\item{type}{character. The type of data source.}

\item{canonicalName}{character. The user-friendly name of the data source.}

\item{dataStoreId}{character. The ID of the data store to connect to.}

\item{query}{character. A query to execute on the data store to get the data. Optional.}

\item{table}{character. The specified database table. Optional.}

\item{schema}{character. The specified database schema. Optional.}

\item{partitionColumn}{character. The name of the partition column. Optional.}

\item{fetchSize}{integer. a user specified fetch size in the range [1, 20000]. Optional.
By default a fetchSize will be assigned to balance throughput and memory usage}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a data source.
}
\details{
Create a fully configured source of data which could be used for datasets and projects creation.A data source specifies, via SQL query or selected table and schema data, which data to extract from the data connection (the location of data within a given endpoint) to use for modeling or predictions. A data source has one data connection and one connector but can have many datasets. To test the SQL query before creating the data source,use :\url{http:post}:/api/v2/externalDataStores/(dataStoreId)/verifySQL/.

This method invokes \verb{POST /externalDataSources/} in the DataRobot Public API.

Create a data source.
}
\examples{
\dontrun{
dataStoreId <- "5c1303269300d900016b41a7"
CreateDataSource(
  type = "jdbc",
  canonicalName = "Airline stats after 1995",
  dataStoreId = dataStoreId,
  query = 'SELECT * FROM airlines10mb WHERE "Year" >= 1995;'
)
}
}
\seealso{
The method ExternalDataSourcesCreate in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
