% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataStores.R
\name{CreateDataStore}
\alias{CreateDataStore}
\title{CreateDataStore}
\usage{
CreateDataStore(type, canonicalName, driverId, jdbcUrl)

CreateDataStore(type, canonicalName, driverId, jdbcUrl)
}
\arguments{
\item{type}{character. The type of data store.}

\item{canonicalName}{character. The user-friendly name of the data store.}

\item{driverId}{character. The ID of the driver to use.}

\item{jdbcUrl}{character. The full JDBC url.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a data store.
}
\details{
Create a data store which includes a name and a driver ID or a connector ID. The driver would be configured by a JDBC URL or by jdbc fields; The connector would be configured by connection fields.

This method invokes \verb{POST /externalDataStores/} in the DataRobot Public API.

Create a data store.
}
\examples{
\dontrun{
CreateDataStore(
  type = "jdbc",
  canonicalName = "Demo DB",
  driverId = "57a7c978c808916f4a630f89",
  jdbcUrl = "jdbc:postgresql://my.db.address.org:5432/my_db"
)
}
}
\seealso{
The method ExternalDataStoresCreate in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
