% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsMonitoringDataDeletions}
\alias{CreateDeploymentsMonitoringDataDeletions}
\title{CreateDeploymentsMonitoringDataDeletions}
\usage{
CreateDeploymentsMonitoringDataDeletions(
  modelId,
  deploymentId,
  start,
  end,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for deleting deployment monitoring data.
}
\details{
Delete deployment monitoring data.

This method invokes \verb{POST /deployments/\{deploymentId\}/monitoringDataDeletions/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsMonitoringDataDeletionsCreate in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
