% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsOnDemandReports}
\alias{CreateDeploymentsOnDemandReports}
\title{CreateDeploymentsOnDemandReports}
\usage{
CreateDeploymentsOnDemandReports(deploymentId, id, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add report to execution queue.
}
\details{
Add report to execution queue.

This method invokes \verb{POST /deployments/\{deploymentId\}/onDemandReports/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsOnDemandReportsCreate in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
