% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsPredictionInputsFromDataset}
\alias{CreateDeploymentsPredictionInputsFromDataset}
\title{CreateDeploymentsPredictionInputsFromDataset}
\usage{
CreateDeploymentsPredictionInputsFromDataset(
  deploymentId,
  datasetId,
  datasetVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Submit external deployment prediction data.
}
\details{
Assigns prediction dataset to the external deployment to enable the analysis of historical model performance. Multiple datasets containing historical predictions for the external deployment can be uploaded. This requires one request for each dataset. For a regression deployment, predictions can be either an int or float. For a classification (binary/multiclass) deployment, predictions must be lists with each list containing probabilities for each class.

This method invokes \verb{POST /deployments/\{deploymentId\}/predictionInputs/fromDataset/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsPredictionInputsFromDatasetCreate in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
