% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsRetrainingPolicies}
\alias{CreateDeploymentsRetrainingPolicies}
\title{CreateDeploymentsRetrainingPolicies}
\usage{
CreateDeploymentsRetrainingPolicies(
  projectOptionsStrategy,
  projectOptions,
  autopilotOptions,
  modelSelectionStrategy,
  deploymentId,
  name,
  description,
  action,
  trigger,
  timeSeriesOptions,
  featureListStrategy,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for creating a deployment retraining policy.
}
\details{
Create a deployment retraining policy.

This method invokes \verb{POST /deployments/\{deploymentId\}/retrainingPolicies/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingPoliciesCreate in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
