% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsScoringCodeBuilds}
\alias{CreateDeploymentsScoringCodeBuilds}
\title{CreateDeploymentsScoringCodeBuilds}
\usage{
CreateDeploymentsScoringCodeBuilds(
  includeAgent,
  includePredictionExplanations,
  deploymentId,
  includePredictionIntervals,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Build Java package containing Scoring Code with agent integration.
}
\details{
Build Java package containing Scoring Code with agent integration.

This method invokes \verb{POST /deployments/\{deploymentId\}/scoringCodeBuilds/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsScoringCodeBuildsCreate in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
