% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateFromLatestCustomModelsVersion}
\alias{CreateFromLatestCustomModelsVersion}
\title{CreateFromLatestCustomModelsVersion}
\usage{
CreateFromLatestCustomModelsVersion(
  requiredMetadata,
  networkEgressPolicy,
  requiresHa,
  file,
  baseEnvironmentId,
  customModelId,
  replicas,
  filePath,
  maximumMemory,
  filesToDelete,
  requiredMetadataValues,
  desiredMemory,
  isMajorUpdate = "true",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update custom model version files.
}
\details{
Create a new custom model version with files added, replaced or deleted. Files from the previous version of a custom models will be used as a basis.

This method invokes \verb{PATCH /customModels/\{customModelId\}/versions/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionCreateFromLatest in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
