% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{CreateFromLatestCustomTaskVersion}
\alias{CreateFromLatestCustomTaskVersion}
\title{CreateFromLatestCustomTaskVersion}
\usage{
CreateFromLatestCustomTaskVersion(
  requiredMetadata,
  file,
  baseEnvironmentId,
  filePath,
  customTaskId,
  maximumMemory,
  filesToDelete,
  requiredMetadataValues,
  isMajorUpdate = "true",
  ...
)
}
\description{
Update custom task version files.
}
\details{
Create a new custom task version with files added, replaced or deleted. Files from the previous version of a custom task will be used as a basis.
}
\seealso{
The method CustomTaskVersionCreateFromLatest in [datarobot.apicore::CustomTasksApi], which this function wraps.

Other CustomTasks: 
\code{\link{CreateCustomTasksFromCustomTask}()},
\code{\link{CreateCustomTasksVersionsDependencyBuild}()},
\code{\link{CreateCustomTasksVersionsFromRepository}()},
\code{\link{CreateCustomTasksVersions}()},
\code{\link{CreateCustomTasks}()},
\code{\link{DeleteCustomTasks}()},
\code{\link{DeleteManyCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksAccessControl}()},
\code{\link{ListCustomTasksDownload}()},
\code{\link{ListCustomTasksVersionsDependencyBuildLog}()},
\code{\link{ListCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksVersionsDownload}()},
\code{\link{ListCustomTasksVersions}()},
\code{\link{ListCustomTasks}()},
\code{\link{PatchCustomTasksVersions}()},
\code{\link{PatchCustomTasks}()},
\code{\link{PatchManyCustomTasksAccessControl}()},
\code{\link{PatchManyCustomTasksVersionsFromRepository}()},
\code{\link{RetrieveCustomTasksVersions}()},
\code{\link{RetrieveCustomTasks}()}
}
\concept{CustomTasks}
