% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{CreateGroups}
\alias{CreateGroups}
\title{CreateGroups}
\usage{
CreateGroups(accessRoleId, name, description, orgId, email, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create user group
}
\details{
Create a new user group.

This method invokes \verb{POST /groups/} in the DataRobot Public API.
}
\seealso{
The method GroupsCreate in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
