% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{CreateImageAugmentationLists}
\alias{CreateImageAugmentationLists}
\title{CreateImageAugmentationLists}
\usage{
CreateImageAugmentationLists(
  transformationProbability,
  samplesId,
  featureName,
  transformations,
  name,
  numberOfNewImages,
  projectId,
  initialList = FALSE,
  inUse = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Creates a new augmentation list based on the posted payload data.
}
\details{
Creates a new augmentation list based on the posted payload data.

This method invokes \verb{POST /imageAugmentationLists/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationListsCreate in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
