% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{CreateImageAugmentationListsSamples}
\alias{CreateImageAugmentationListsSamples}
\title{CreateImageAugmentationListsSamples}
\usage{
CreateImageAugmentationListsSamples(augmentationId, numberOfRows, ...)
}
\description{
Requests the creation of sample augmentations based on the augmentation list passed in.
}
\details{
This endpoint will schedule a job to augment the specified images from a project's dataset and return a link to monitor the status of the job, as well as a link to retrieve the resulting augmented images.
}
\seealso{
The method ImageAugmentationListsSamplesCreate in [datarobot.apicore::ImagesApi], which this function wraps.

Other Images: 
\code{\link{CreateImageAugmentationLists}()},
\code{\link{CreateImageAugmentationSamples}()},
\code{\link{CreateProjectsModelsImageActivationMaps}()},
\code{\link{CreateProjectsModelsImageEmbeddings}()},
\code{\link{DeleteImageAugmentationLists}()},
\code{\link{ListDatasetsImagesDataQualityLogFile}()},
\code{\link{ListDatasetsImagesDataQualityLog}()},
\code{\link{ListImageAugmentationListsSamples}()},
\code{\link{ListImageAugmentationLists}()},
\code{\link{ListImageAugmentationSamples}()},
\code{\link{ListProjectsDuplicateImages}()},
\code{\link{ListProjectsImageActivationMaps}()},
\code{\link{ListProjectsImageBins}()},
\code{\link{ListProjectsImageEmbeddings}()},
\code{\link{ListProjectsImageSamples}()},
\code{\link{ListProjectsImagesDataQualityLogFile}()},
\code{\link{ListProjectsImagesDataQualityLog}()},
\code{\link{ListProjectsImagesFile}()},
\code{\link{ListProjectsImages}()},
\code{\link{ListProjectsModelsImageActivationMaps}()},
\code{\link{ListProjectsModelsImageEmbeddings}()},
\code{\link{PatchImageAugmentationLists}()},
\code{\link{RetrieveImageAugmentationLists}()},
\code{\link{RetrieveImageAugmentationOptions}()},
\code{\link{RetrieveProjectsImages}()}
}
\concept{Images}
