% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateMultilabelInsightsPairwiseManualSelections}
\alias{CreateMultilabelInsightsPairwiseManualSelections}
\title{CreateMultilabelInsightsPairwiseManualSelections}
\usage{
CreateMultilabelInsightsPairwiseManualSelections(
  columnLabels,
  featureName,
  name,
  projectId,
  multilabelInsightsKey,
  rowLabels,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Save a list of manually selected labels for Feature Statistics matrix.
}
\details{
Save a list of manually selected labels for Feature Statistics matrix.

This method invokes \verb{POST /multilabelInsights/\{multilabelInsightsKey\}/pairwiseManualSelections/} in the DataRobot Public API.
}
\seealso{
The method MultilabelInsightsPairwiseManualSelectionsCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
