% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateNotificationChannels}
\alias{CreateNotificationChannels}
\title{CreateNotificationChannels}
\usage{
CreateNotificationChannels(
  emailAddress,
  secretToken,
  name,
  validateSsl,
  channelType,
  languageCode,
  contentType,
  orgId,
  customHeaders,
  verificationCode,
  payloadUrl,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create notification channel
}
\details{
Create a new notification channel.

This method invokes \verb{POST /notificationChannels/} in the DataRobot Public API.
}
\seealso{
The method NotificationChannelsCreate in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
