% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateNotificationPolicies}
\alias{CreateNotificationPolicies}
\title{CreateNotificationPolicies}
\usage{
CreateNotificationPolicies(
  name,
  active,
  eventType,
  eventGroup,
  orgId,
  channelId,
  ...
)
}
\description{
Create notification policy
}
\details{
Create a new notification policy.
}
\seealso{
The method NotificationPoliciesCreate in [datarobot.apicore::NotificationsApi], which this function wraps.

Other Notifications: 
\code{\link{CreateNotificationChannels}()},
\code{\link{CreateNotificationEmailChannelVerificationStatus}()},
\code{\link{CreateNotificationEmailChannelVerification}()},
\code{\link{CreateNotificationPolicyMutes}()},
\code{\link{CreateNotificationWebhookChannelTests}()},
\code{\link{CreateNotifications}()},
\code{\link{CreateRemoteEvents}()},
\code{\link{DeleteManyUserNotifications}()},
\code{\link{DeleteNotificationChannels}()},
\code{\link{DeleteNotificationPolicies}()},
\code{\link{DeleteNotificationPolicyMutes}()},
\code{\link{DeleteUserNotifications}()},
\code{\link{ListNotificationChannels}()},
\code{\link{ListNotificationEvents}()},
\code{\link{ListNotificationLogs}()},
\code{\link{ListNotificationPolicies}()},
\code{\link{ListNotificationPolicyMutes}()},
\code{\link{ListUserNotifications}()},
\code{\link{PatchManyUserNotifications}()},
\code{\link{PatchUserNotifications}()},
\code{\link{PutNotificationChannels}()},
\code{\link{PutNotificationPolicies}()},
\code{\link{RetrieveNotificationChannels}()},
\code{\link{RetrieveNotificationPolicies}()},
\code{\link{RetrieveNotificationWebhookChannelTests}()}
}
\concept{Notifications}
