% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateNotificationWebhookChannelTests}
\alias{CreateNotificationWebhookChannelTests}
\title{CreateNotificationWebhookChannelTests}
\usage{
CreateNotificationWebhookChannelTests(
  emailAddress,
  secretToken,
  name,
  validateSsl,
  channelType,
  contentType,
  orgId,
  customHeaders,
  payloadUrl,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Test webhook notification channel
}
\details{
Test the webhook notification channel.

This method invokes \verb{POST /notificationWebhookChannelTests/} in the DataRobot Public API.
}
\seealso{
The method NotificationWebhookChannelTestsCreate in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
