% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{CreateOneModelComplianceDocsInitializations}
\alias{CreateOneModelComplianceDocsInitializations}
\title{CreateOneModelComplianceDocsInitializations}
\usage{
CreateOneModelComplianceDocsInitializations(entityId, body, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Initialize compliance documentation pre-processing for the model
}
\details{
Initialize compliance documentation pre-processing for the current model. This route must be called before generating documentation for a custom model.

This method invokes \verb{POST /modelComplianceDocsInitializations/\{entityId\}/} in the DataRobot Public API.
}
\seealso{
The method ModelComplianceDocsInitializationsCreateOne in \link[datarobot.apicore:DocumentationApi]{datarobot.apicore::DocumentationApi}, which invokes the same underlying API endpoint.
}
