% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/CreatePrimeCode.R
\name{CreatePrimeCode}
\alias{CreatePrimeCode}
\title{CreatePrimeCode}
\usage{
CreatePrimeCode(project, primeModelId, language, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{primeModelId}{character. Id returned by GetPrimeModel(s) functions.}

\item{language}{character. Programming language to use for downloadable code (see PrimeLanguage).}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
job Id
}
\description{
Create a Prime File
}
\details{
Request creation and validation of source code from a Prime model

This method invokes \verb{POST /projects/\{projectId\}/primeFiles/} in the DataRobot Public API.

Create and validate the downloadable code for the ruleset associated with this model
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
CreatePrimeCode(projectId, modelId, "Python")
}
}
\seealso{
The method ProjectsPrimeFilesCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
