% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateProjectsBiasMitigatedModels}
\alias{CreateProjectsBiasMitigatedModels}
\title{CreateProjectsBiasMitigatedModels}
\usage{
CreateProjectsBiasMitigatedModels(
  biasMitigationTechnique,
  biasMitigationParentLid,
  biasMitigationFeature,
  projectId,
  includeBiasMitigationFeatureAsPredictorVariable,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add a request to the queue to train a model with bias mitigation applied.
}
\details{
Add a request to the queue to train a model with bias mitigation applied. If the job has been previously submitted, the request will return the jobId of the previously submitted job. Use this jobId to check status of the previously submitted job.

This method invokes \verb{POST /projects/\{projectId\}/biasMitigatedModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBiasMitigatedModelsCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
