% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsDatetimeModelsFeatureEffects}
\alias{CreateProjectsDatetimeModelsFeatureEffects}
\title{CreateProjectsDatetimeModelsFeatureEffects}
\usage{
CreateProjectsDatetimeModelsFeatureEffects(
  backtestIndex,
  modelId,
  rowCount,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add a request to the queue to calculate Feature Effects for a backtest.
}
\details{
Add a request to the queue to calculate Feature Effects for a backtest. If the job has been previously submitted, the request fails, returning the jobId of the previously submitted job. Use this jobId to check status of the previously submitted job.

This method invokes \verb{POST /projects/\{projectId\}/datetimeModels/\{modelId\}/featureEffects/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsFeatureEffectsCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
