% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsDatetimeModelsMulticlassFeatureEffects}
\alias{CreateProjectsDatetimeModelsMulticlassFeatureEffects}
\title{CreateProjectsDatetimeModelsMulticlassFeatureEffects}
\usage{
CreateProjectsDatetimeModelsMulticlassFeatureEffects(
  features,
  backtestIndex,
  modelId,
  rowCount,
  topNFeatures,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Compute feature effects for a multiclass datetime model.
}
\details{
Compute feature effects for a multiclass datetime model. If the job has been previously submitted, the request fails, returning the \code{jobId} of the previously submitted job. Use this \code{jobId} to check status of the previously submitted job. NOTE: feature effects are computed for top 100 classes.

This method invokes \verb{POST /projects/\{projectId\}/datetimeModels/\{modelId\}/multiclassFeatureEffects/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsMulticlassFeatureEffectsCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
