% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateProjectsDeploymentReadyModels}
\alias{CreateProjectsDeploymentReadyModels}
\title{CreateProjectsDeploymentReadyModels}
\usage{
CreateProjectsDeploymentReadyModels(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Prepare a model for deployment
}
\details{
Prepare a specific model for deployment. This model will go through the recommendation stages.

This method invokes \verb{POST /projects/\{projectId\}/deploymentReadyModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDeploymentReadyModelsCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
