% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsExternalScores}
\alias{CreateProjectsExternalScores}
\title{CreateProjectsExternalScores}
\usage{
CreateProjectsExternalScores(
  modelId,
  actualValueColumn,
  datasetId,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Compute model scores for external dataset.
}
\details{
Compute model scores for external dataset, first upload your dataset to the project, and then using the corresponding datasetId, compute scores against that dataset. Computing external scores and insights depends on computed prediction, predictions will be computed if they are not available for this dataset. In order to compute scores and insights, uploaded dataset should contain actual value column. This api is not available in time series projects.

This method invokes \verb{POST /projects/\{projectId\}/externalScores/} in the DataRobot Public API.
}
\seealso{
The method ProjectsExternalScoresCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
