% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateProjectsExternalTimeSeriesBaselineDataValidationJobs}
\alias{CreateProjectsExternalTimeSeriesBaselineDataValidationJobs}
\title{CreateProjectsExternalTimeSeriesBaselineDataValidationJobs}
\usage{
CreateProjectsExternalTimeSeriesBaselineDataValidationJobs(
  forecastWindowStart,
  forecastWindowEnd,
  catalogVersionId,
  holdoutEndDate,
  multiseriesIdColumns,
  holdoutStartDate,
  backtests,
  projectId,
  target,
  datetimePartitionColumn,
  ...
)
}
\description{
Validate baseline data
}
\details{
This route validates if a provided catalog version id can be used as baseline for calculating metrics. This functionality is available only for time series projects.For a baseline dataset to be valid, the number of unique date amd multiseries_id columnrows must match the unique number of date and multiseries_id column rows in the uploadedtraining dataset. This functionality is limited to one forecast distance. Additionally, the catalog must be a snapshot.
}
\seealso{
The method ExternalTimeSeriesBaselineDataValidationJobsCreate in [datarobot.apicore::ProjectsApi], which this function wraps.

Other Projects: 
\code{\link{BatchFeaturesTypeTransform}()},
\code{\link{CloneProject}()},
\code{\link{CreateCalendarsFromCountryCode}()},
\code{\link{CreateCalendarsFromDataset}()},
\code{\link{CreateCalendar}()},
\code{\link{CreateFeaturelist}()},
\code{\link{CreateHdfsProjects}()},
\code{\link{CreateModelingFeaturelist}()},
\code{\link{CreateProjectCleanupJobs}()},
\code{\link{CreateProjectsFeatureDiscoveryRecipeSqlExports}()},
\code{\link{CreateProjectsSecondaryDatasetsConfigurations}()},
\code{\link{CreateProjectsSegmentationTasks}()},
\code{\link{CreateRelationshipsConfigurations}()},
\code{\link{DeleteCalendar}()},
\code{\link{DeleteProjectCleanupJobs}()},
\code{\link{DeleteProjectsJobs}()},
\code{\link{DeleteProjectsSecondaryDatasetsConfigurations}()},
\code{\link{DeleteProject}()},
\code{\link{DeleteRelationshipsConfigurations}()},
\code{\link{GetCalendar}()},
\code{\link{GetFeatureHistogram}()},
\code{\link{GetFeatureInfo}()},
\code{\link{GetFeaturelist}()},
\code{\link{GetJob}()},
\code{\link{GetModelingFeaturelist}()},
\code{\link{GetProjectStatus}()},
\code{\link{GetProject}()},
\code{\link{GetValidMetrics}()},
\code{\link{ListCalendarCountryCodes}()},
\code{\link{ListCalendarsAccessControl}()},
\code{\link{ListCalendars}()},
\code{\link{ListDeletedProjectsCount}()},
\code{\link{ListDeletedProjects}()},
\code{\link{ListFeatureInfo}()},
\code{\link{ListFeaturelists}()},
\code{\link{ListJobs}()},
\code{\link{ListModelingFeaturelists}()},
\code{\link{ListProjectCleanupJobsDownload}()},
\code{\link{ListProjectCleanupJobsSummary}()},
\code{\link{ListProjectCleanupJobs}()},
\code{\link{ListProjectsAccessControl}()},
\code{\link{ListProjectsCalendarEvents}()},
\code{\link{ListProjectsFeatureDiscoveryDatasetDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogsDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogs}()},
\code{\link{ListProjectsFeatureDiscoveryRecipeSQLsDownload}()},
\code{\link{ListProjectsModelingFeatures}()},
\code{\link{ListProjectsMultiseriesIdsCrossSeriesProperties}()},
\code{\link{ListProjectsMultiseriesNames}()},
\code{\link{ListProjectsRelationshipsConfiguration}()},
\code{\link{ListProjectsSecondaryDatasetsConfigurations}()},
\code{\link{ListProjectsSegmentationTasksMappings}()},
\code{\link{ListProjectsSegmentationTasks}()},
\code{\link{ListProjects}()},
\code{\link{PatchDeletedProjects}()},
\code{\link{PatchManyCalendarsAccessControl}()},
\code{\link{PatchManyProjectsAccessControl}()},
\code{\link{PatchProjectsSegments}()},
\code{\link{PauseQueue}()},
\code{\link{PutRelationshipsConfigurations}()},
\code{\link{RequestCrossSeriesDetection}()},
\code{\link{RequestMultiSeriesDetection}()},
\code{\link{RestoreDiscardedFeatures}()},
\code{\link{RetrieveDiscardedFeaturesInformation}()},
\code{\link{RetrieveProjectCleanupJobs}()},
\code{\link{RetrieveProjectsBatchTypeTransformFeaturesResult}()},
\code{\link{RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{RetrieveProjectsFeatureLineages}()},
\code{\link{RetrieveProjectsModelingFeatures}()},
\code{\link{RetrieveProjectsSecondaryDatasetsConfigurations}()},
\code{\link{RetrieveProjectsSegmentationTaskJobResults}()},
\code{\link{RetrieveProjectsSegmentationTasks}()},
\code{\link{RetrieveRelationshipsConfigurations}()},
\code{\link{SetTarget}()},
\code{\link{SetupProject}()},
\code{\link{StartNewAutoPilot}()},
\code{\link{UpdateCalendar}()},
\code{\link{UpdateProject}()}
}
\concept{Projects}
