% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsFeatureAssociationMatrix}
\alias{CreateProjectsFeatureAssociationMatrix}
\title{CreateProjectsFeatureAssociationMatrix}
\usage{
CreateProjectsFeatureAssociationMatrix(featurelistId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Compute feature association matrix.
}
\details{
Compute feature association matrix for given featurelist.

This method invokes \verb{POST /projects/\{projectId\}/featureAssociationMatrix/} in the DataRobot Public API.
}
\seealso{
The method ProjectsFeatureAssociationMatrixCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
