% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateProjectsFeatureDiscoveryRecipeSqlExports}
\alias{CreateProjectsFeatureDiscoveryRecipeSqlExports}
\title{CreateProjectsFeatureDiscoveryRecipeSqlExports}
\usage{
CreateProjectsFeatureDiscoveryRecipeSqlExports(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Generate feature discovery SQL recipe
}
\details{
Generate feature discovery SQL recipe for a project

This method invokes \verb{POST /projects/\{projectId\}/featureDiscoveryRecipeSqlExports/} in the DataRobot Public API.
}
\seealso{
The method FeatureDiscoveryRecipeSqlExportsCreate in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
