% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsGeometryFeaturePlots}
\alias{CreateProjectsGeometryFeaturePlots}
\title{CreateProjectsGeometryFeaturePlots}
\usage{
CreateProjectsGeometryFeaturePlots(feature, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a map of one location feature
}
\details{
Create a map of one location feature

This method invokes \verb{POST /projects/\{projectId\}/geometryFeaturePlots/} in the DataRobot Public API.
}
\seealso{
The method ProjectsGeometryFeaturePlotsCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
