% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsModelsCrossClassAccuracyScores}
\alias{CreateProjectsModelsCrossClassAccuracyScores}
\title{CreateProjectsModelsCrossClassAccuracyScores}
\usage{
CreateProjectsModelsCrossClassAccuracyScores(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Start Cross Class Accuracy calculations.
}
\details{
Submits a job to start Cross Class Accuracy scores calculations for the model.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/crossClassAccuracyScores/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsCrossClassAccuracyScoresCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
