% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsModelsDataDisparityInsights}
\alias{CreateProjectsModelsDataDisparityInsights}
\title{CreateProjectsModelsDataDisparityInsights}
\usage{
CreateProjectsModelsDataDisparityInsights(
  feature,
  modelId,
  comparedClassNames,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Start insight calculations.
}
\details{
Submits a job to start Cross Class Data Disparity insight calculations.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/dataDisparityInsights/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDataDisparityInsightsCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
