% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateProjectsModelsFromModel}
\alias{CreateProjectsModelsFromModel}
\title{CreateProjectsModelsFromModel}
\usage{
CreateProjectsModelsFromModel(
  monotonicIncreasingFeaturelistId,
  monotonicDecreasingFeaturelistId,
  samplePct,
  scoringType,
  modelId,
  trainingRowCount,
  featurelistId,
  projectId,
  nClusters,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrain a model
}
\details{
Retrain an existing model using a new sample size and/or feature list.When configuring retraining sample sizes for models in projects with large row counts, DataRobot recommends requesting sample sizes using integer row counts instead of percentages. This is because percentages map to many actual possible row counts and only one of which is the actual sample size for up to validation. For example, if a project has 199,408 rows and you request a 64\% sample size, any number of rows between 126,625 rows and 128,618 rows maps to 64\% of the data. Using actual integer row counts (or project.max_training_rows) avoids ambiguity around how many rows of data you want the model to use. Note that only one of samplePct or trainingRowCount should be specified.

This method invokes \verb{POST /projects/\{projectId\}/models/fromModel/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsFromModelCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
