% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsModelsMulticlassFeatureEffects}
\alias{CreateProjectsModelsMulticlassFeatureEffects}
\title{CreateProjectsModelsMulticlassFeatureEffects}
\usage{
CreateProjectsModelsMulticlassFeatureEffects(
  features,
  modelId,
  rowCount,
  topNFeatures,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Compute feature effects for a multiclass model.
}
\details{
Compute feature effects for a multiclass model. If the job has been previously submitted, the request fails, returning the jobId of the previously submitted job. Use this jobId to check status of the previously submitted job. NOTE: feature effects are computed for top 100 classes.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/multiclassFeatureEffects/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsMulticlassFeatureEffectsCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
