% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsModelsShapImpact}
\alias{CreateProjectsModelsShapImpact}
\title{CreateProjectsModelsShapImpact}
\usage{
CreateProjectsModelsShapImpact(modelId, rowCount, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a Shap based Feature Impact.
}
\details{
Create a Shap based Feature Impact with a model.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/shapImpact/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsShapImpactCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
