% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsPayoffMatrices}
\alias{CreateProjectsPayoffMatrices}
\title{CreateProjectsPayoffMatrices}
\usage{
CreateProjectsPayoffMatrices(
  truePositiveValue,
  trueNegativeValue,
  name,
  falsePositiveValue,
  falseNegativeValue,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a payoff matrix.
}
\details{
Create a payoff matrix associated with a project.

This method invokes \verb{POST /projects/\{projectId\}/payoffMatrices/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPayoffMatricesCreate in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
