% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateProjectsSegmentationTasks}
\alias{CreateProjectsSegmentationTasks}
\title{CreateProjectsSegmentationTasks}
\usage{
CreateProjectsSegmentationTasks(
  userDefinedSegmentIdColumns,
  multiseriesIdColumns,
  modelPackageId,
  projectId,
  target,
  datetimePartitionColumn,
  useAutomatedSegmentation = FALSE,
  useTimeSeries = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create segmentation tasks.
}
\details{
Create segmentation tasks for the dataset used in the project.

This method invokes \verb{POST /projects/\{projectId\}/segmentationTasks/} in the DataRobot Public API.
}
\seealso{
The method SegmentationTasksCreate in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
