% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_configuration_apiWrapper.R
\name{CreateSsoConfigurations}
\alias{CreateSsoConfigurations}
\title{CreateSsoConfigurations}
\usage{
CreateSsoConfigurations(
  autoGenerateUsers,
  certificate,
  securityParameters,
  entityId,
  issuer,
  signOnUrl,
  organizationId,
  idpMetadataUrl,
  roleMapping,
  idpMetadataHttpsVerify,
  idpMetadata,
  attributeMapping,
  signOutUrl,
  enableSso,
  groupMapping,
  name,
  enforceSso,
  configurationType,
  idpResponseMethod = "POST",
  sessionLengthSeconds = 604800,
  spRequestMethod = "REDIRECT",
  ...
)
}
\description{
Create an SSO configuration for a specific organization
}
\details{
Create an SSO configuration for a specific organization
}
\seealso{
The method SsoConfigurationsCreate in [datarobot.apicore::SsoConfigurationApi], which this function wraps.

Other SsoConfiguration: 
\code{\link{ListSsoConfigurations}()},
\code{\link{PatchSsoConfigurations}()},
\code{\link{RetrieveSsoConfigurations}()}
}
\concept{SsoConfiguration}
