% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_apiWrapper.R
\name{CreateStringEncryptions}
\alias{CreateStringEncryptions}
\title{CreateStringEncryptions}
\usage{
CreateStringEncryptions(plainText, ...)
}
\description{
Encrypt a string which DataRobot can decrypt.
}
\details{
DataRobot does not store your credentials when accessing external data stores. Instead you first encrypt your sensitive information with DataRobot and then, when accessing those data stores, supply the encrypted credentials. DataRobot will decrypt your credentials and use them to establish a connection to your data store.
}
\seealso{
The method StringEncryptionsCreate in [datarobot.apicore::UtilitiesApi], which this function wraps.
}
\concept{Utilities}
