% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{CreateUsageDataExports}
\alias{CreateUsageDataExports}
\title{CreateUsageDataExports}
\usage{
CreateUsageDataExports(
  include,
  noCache,
  start,
  end,
  userId,
  projectId,
  includeReport,
  includeIdentifyingFields = TRUE,
  ...
)
}
\description{
Create a customer usage data artifact request. Requires \"CAN_ACCESS_USER_ACTIVITY\" permission.
}
\details{
Create a customer usage data artifact request. Requires \"CAN_ACCESS_USER_ACTIVITY\" permission. Returns async task status URL as a \"Location\" header. Async task status should be polled in order to retrieve artifact id.
}
\seealso{
The method UsageDataExportsCreate in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{DeleteManyUsersRateLimitUsage}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsEvents}()},
\code{\link{ListEventLogsPredictionUsage}()},
\code{\link{ListEventLogs}()},
\code{\link{ListUsageDataExportsSupportedEvents}()},
\code{\link{ListUsersRateLimitUsage}()},
\code{\link{RetrieveEventLogs}()},
\code{\link{RetrieveUsageDataExports}()}
}
\concept{Analytics}
