% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{CreateUseCases}
\alias{CreateUseCases}
\title{CreateUseCases}
\usage{
CreateUseCases(
  owner,
  notes,
  description,
  businessImpact,
  potentialValueTemplate,
  targetDates,
  content,
  feasibility,
  predictionTargets,
  stage,
  realizedValue,
  permissions,
  mentions,
  name,
  commentId,
  id,
  potentialValue,
  ...
)
}
\description{
Create a new use case.
}
\details{
Create a new use case.
}
\seealso{
The method UseCasesCreate in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplatesCalculation}()},
\code{\link{ListUseCaseValueTemplates}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesAttachments}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCasesSharedRoles}()},
\code{\link{ListUseCases}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{PatchUseCases}()},
\code{\link{RetrieveUseCaseValueTemplates}()},
\code{\link{RetrieveUseCasesAttachments}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
