% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{CreateUseCasesAttachments}
\alias{CreateUseCasesAttachments}
\title{CreateUseCasesAttachments}
\usage{
CreateUseCasesAttachments(useCaseId, attachedObjects, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Attach a list of resources to this use case.
}
\details{
Attach a list of resources to this use case.

This method invokes \verb{POST /useCases/\{useCaseId\}/attachments/} in the DataRobot Public API.
}
\seealso{
The method UseCasesAttachmentsCreate in \link[datarobot.apicore:UseCaseApi]{datarobot.apicore::UseCaseApi}, which invokes the same underlying API endpoint.
}
