% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R,
%   R/ai_catalog_apiWrapper.R
\name{CreateUserBlueprints}
\alias{CreateUserBlueprints}
\title{CreateUserBlueprints}
\usage{
CreateUserBlueprints(
  blueprint,
  description,
  modelType,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)

CreateUserBlueprints(
  blueprint,
  description,
  modelType,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a user blueprint.

Create a user blueprint.
}
\details{
Create a user blueprint

This method invokes \verb{POST /userBlueprints/} in the DataRobot Public API.

Create a user blueprint

This method invokes \verb{POST /userBlueprints/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsCreate in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsCreate in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
