% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R,
%   R/ai_catalog_apiWrapper.R
\name{CreateUserBlueprintsFromBlueprintId}
\alias{CreateUserBlueprintsFromBlueprintId}
\title{CreateUserBlueprintsFromBlueprintId}
\usage{
CreateUserBlueprintsFromBlueprintId(
  description,
  modelType,
  projectId,
  blueprintId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)

CreateUserBlueprintsFromBlueprintId(
  description,
  modelType,
  projectId,
  blueprintId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Clone a blueprint from a project.

Clone a blueprint from a project.
}
\details{
Clone a blueprint from a project.

This method invokes \verb{POST /userBlueprints/fromBlueprintId/} in the DataRobot Public API.

Clone a blueprint from a project.

This method invokes \verb{POST /userBlueprints/fromBlueprintId/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsFromBlueprintIdCreate in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsFromBlueprintIdCreate in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
