% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R,
%   R/ai_catalog_apiWrapper.R
\name{CreateUserBlueprintsFromUserBlueprintId}
\alias{CreateUserBlueprintsFromUserBlueprintId}
\title{CreateUserBlueprintsFromUserBlueprintId}
\usage{
CreateUserBlueprintsFromUserBlueprintId(
  userBlueprintId,
  description,
  modelType,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)

CreateUserBlueprintsFromUserBlueprintId(
  userBlueprintId,
  description,
  modelType,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Clone a user blueprint.

Clone a user blueprint.
}
\details{
Clone a user blueprint.

This method invokes \verb{POST /userBlueprints/fromUserBlueprintId/} in the DataRobot Public API.

Clone a user blueprint.

This method invokes \verb{POST /userBlueprints/fromUserBlueprintId/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsFromUserBlueprintIdCreate in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsFromUserBlueprintIdCreate in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
