% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{CreateUserCleanupPreviewJobs}
\alias{CreateUserCleanupPreviewJobs}
\title{CreateUserCleanupPreviewJobs}
\usage{
CreateUserCleanupPreviewJobs(orphansOwner, userIds, orgId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Users permanent delete preview.
}
\details{
Schedules a job for building preview for users selected for permanent deletion. Number of users to be deleted in one go is restricted by DELETED_USERS_BATCH_LIMIT system setting.

This method invokes \verb{POST /userCleanupPreviewJobs/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupPreviewJobsCreate in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
