% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{CreateUsers}
\alias{CreateUsers}
\title{CreateUsers}
\usage{
CreateUsers(
  lastName,
  country,
  maxUploadSize,
  maxIdleWorkers,
  language,
  maxGpuWorkers,
  organizationId,
  maxWorkers,
  firstName,
  password,
  maxUploadSizeCatalog,
  requireClickthroughAgreement,
  userType,
  defaultCatalogDatasetSampleSize,
  username,
  expirationDate,
  ...
)
}
\description{
Create a User and add them to an organisation.
}
\details{
Create a User and add them to an organisation.
}
\seealso{
The method UsersCreate in [datarobot.apicore::UserManagementApi], which this function wraps.

Other UserManagement: 
\code{\link{CreateGroupsUsers}()},
\code{\link{CreateGroups}()},
\code{\link{CreateOrganizationsUsers}()},
\code{\link{CreateUserCleanupJobs}()},
\code{\link{CreateUserCleanupPreviewJobs}()},
\code{\link{DeleteGroups}()},
\code{\link{DeleteManyGroupsUsers}()},
\code{\link{DeleteManyGroups}()},
\code{\link{DeleteOrganizationsUsers}()},
\code{\link{DeleteUserCleanupJobs}()},
\code{\link{DeleteUserCleanupPreviewJobs}()},
\code{\link{DeleteUserCleanupPreviews}()},
\code{\link{DeleteUserCleanupSummaries}()},
\code{\link{ListGroupsUsers}()},
\code{\link{ListOrganizationsJobs}()},
\code{\link{ListOrganizationsUsers}()},
\code{\link{ListOrganizations}()},
\code{\link{ListUserCleanupPreviewsContent}()},
\code{\link{ListUserCleanupPreviewsDeleteParams}()},
\code{\link{ListUserCleanupPreviewsStatistics}()},
\code{\link{ListUserCleanupSummariesContent}()},
\code{\link{ListUserCleanupSummariesDeleteParams}()},
\code{\link{ListUserCleanupSummariesStatistics}()},
\code{\link{ListUsersLimits}()},
\code{\link{PatchGroups}()},
\code{\link{PatchManyUsersLimits}()},
\code{\link{PatchOrganizationsUsers}()},
\code{\link{RetrieveGroups}()},
\code{\link{RetrieveOrganizationsUsers}()},
\code{\link{RetrieveOrganizations}()},
\code{\link{RetrieveUserCleanupJobs}()},
\code{\link{RetrieveUserCleanupPreviewJobs}()}
}
\concept{UserManagement}
