% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{CrossValidateModel}
\alias{CrossValidateModel}
\title{Run cross validation on a model.}
\usage{
CrossValidateModel(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModel, or each element of the list returned by
the function ListModels.}
}
\value{
Job ID of the cross validation job.
}
\description{
Note that this runs cross validation on a model as-is. If you would like to run cross-validation
on a model with new parameters, use \code{RequestNewModel} instead.
}
\details{
Note that this is not implemented for prime models or datetime models.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  modelId <- "5996f820af07fc605e81ead4"
  model <- GetModel(projectId, modelId)
  CrossValidateModel(model)
}
}
