% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{cvMethods}
\alias{cvMethods}
\title{CV methods}
\format{An object of class \code{list} of length 5.}
\usage{
cvMethods
}
\description{
This is a list that contains the valid values for CV methods
}
\keyword{datasets}
