% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{DataPartition}
\alias{DataPartition}
\title{Data Partition methods}
\format{
An object of class \code{list} of length 3.
}
\usage{
DataPartition
}
\description{
This is a list that contains the valid values for data partitions
}
\keyword{datasets}
