% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadData.R
\name{DataPathFromDataArg}
\alias{DataPathFromDataArg}
\title{Get the data path.}
\usage{
DataPathFromDataArg(dataSource, saveFile = NULL)
}
\arguments{
\item{dataSource}{object. The dataframe or path to CSV to get data for.}

\item{saveFile}{character. Optional. A file name to write an autosaved dataframe to.}
}
\description{
Verifies that newdata is either an existing datafile or a dataframe
If a dataframe, save as a CSV file
If neither an existing datafile nor a dataframe, halt with error
}
