% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datarobot-package.R
\docType{package}
\name{datarobot}
\alias{datarobot}
\alias{datarobot-package}
\title{datarobot: 'DataRobot' Predictive Modeling API}
\description{
For working with the 'DataRobot' predictive modeling platform's API \url{https://www.datarobot.com/}.
}
\section{Package options}{


The \code{datarobot} package uses the following \code{\link[=options]{options()}} to configure behavior:

\itemize{
\item \code{dataRobotCurlDebugMode}: logical. Set to \code{TRUE} to print the internals of an API call for debugging purposes. Valid only when invoking the functions \code{DataRobotGET/DELETE/PATCH/POST}, not those in \code{datarobot.apicore}.

}
}

\keyword{internal}
