% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{DataSubset}
\alias{DataSubset}
\title{Data subset for training predictions}
\format{An object of class \code{list} of length 3.}
\usage{
DataSubset
}
\description{
This is a list that contains the valid values for the \code{dataSubset} parameter
found in \code{RequestTrainingPredictions}. If you wish, you can specify
\code{dataSubset} using the list values here.
}
\keyword{datasets}
