% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{DatetimeTrendPlotsResolutions}
\alias{DatetimeTrendPlotsResolutions}
\title{Datetime trend plots resolutions}
\format{
An object of class \code{list} of length 9.
}
\usage{
DatetimeTrendPlotsResolutions
}
\description{
Datetime trend plots resolutions
}
\keyword{datasets}
