% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{DatetimeTrendPlotsStatuses}
\alias{DatetimeTrendPlotsStatuses}
\title{Datetime trend plots statuses}
\format{
An object of class \code{list} of length 6.
}
\usage{
DatetimeTrendPlotsStatuses
}
\description{
Datetime trend plots statuses
}
\keyword{datasets}
