% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R, R/AnomalyAssessment.R
\name{DeleteAnomalyAssessmentRecord}
\alias{DeleteAnomalyAssessmentRecord}
\title{DeleteAnomalyAssessmentRecord}
\usage{
DeleteAnomalyAssessmentRecord(projectId, recordId, ...)
}
\arguments{
\item{projectId}{character. The ID of the project.}

\item{recordId}{character. The ID of the anomaly assessment record.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete the anomaly assessment record.
}
\details{
Delete the anomaly assessment record.

This method invokes \verb{DELETE /projects/\{projectId\}/anomalyAssessmentRecords/\{recordId\}/} in the DataRobot Public API.

Delete anomaly assessment record.

Record is deleted with preview and explanations.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
recordId <- "59a5af20c80891534e3c2bdb"
explanations <- DeleteAnomalyAssessmentRecord(projectId, recordId)
}
}
\seealso{
The method ProjectsAnomalyAssessmentRecordsDelete in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.

Other Anomaly Assessment functions: 
\code{\link{GetAnomalyAssessmentExplanations}()},
\code{\link{GetAnomalyAssessmentPredictionsPreview}()},
\code{\link{InitializeAnomalyAssessment}()},
\code{\link{ListAnomalyAssessmentRecords}()}
}
\concept{Anomaly Assessment functions}
