% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{DeleteAutomatedDocuments}
\alias{DeleteAutomatedDocuments}
\title{DeleteAutomatedDocuments}
\usage{
DeleteAutomatedDocuments(documentId, ...)
}
\description{
Delete automated document.
}
\details{
Delete a document using its ID. Example request:      .. code-block:: text              DELETE https://app.datarobot.com/api/v2/automatedDocuments/5ec4ea7e41054c158c5b002f/ HTTP/1.1             Authorization: Bearer DnwzBUSTOtKBO5Sp1hoUByG4YgZwCCw4
}
\seealso{
The method AutomatedDocumentsDelete in [datarobot.apicore::DocumentationApi], which this function wraps.

Other Documentation: 
\code{\link{CreateAutomatedDocuments}()},
\code{\link{CreateOneModelComplianceDocsInitializations}()},
\code{\link{DeleteComplianceDocTemplate}()},
\code{\link{GetComplianceDocTemplate}()},
\code{\link{ListAutomatedDocumentOptions}()},
\code{\link{ListAutomatedDocuments}()},
\code{\link{ListComplianceDocTemplatesDefault}()},
\code{\link{ListComplianceDocTemplatesSharedRoles}()},
\code{\link{ListComplianceDocTemplates}()},
\code{\link{PatchManyComplianceDocTemplatesSharedRoles}()},
\code{\link{RetrieveAutomatedDocuments}()},
\code{\link{RetrieveModelComplianceDocsInitializations}()},
\code{\link{UpdateComplianceDocTemplate}()},
\code{\link{UploadComplianceDocTemplate}()}
}
\concept{Documentation}
