% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{DeleteBatchPredictionJobDefinitions}
\alias{DeleteBatchPredictionJobDefinitions}
\title{DeleteBatchPredictionJobDefinitions}
\usage{
DeleteBatchPredictionJobDefinitions(jobDefinitionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete Batch Prediction job definition
}
\details{
Delete a Batch Prediction job definition

This method invokes \verb{DELETE /batchPredictionJobDefinitions/\{jobDefinitionId\}/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionJobDefinitionsDelete in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
