% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Calendars.R
\name{DeleteCalendar}
\alias{DeleteCalendar}
\title{DeleteCalendar}
\usage{
DeleteCalendar(calendarId, ...)
}
\arguments{
\item{calendarId}{character. The ID of the calendar to retrieve.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a calendar.
}
\details{
Delete a calendar. This can only be done if all projects and deployments using the calendar have been deleted.

This method invokes \verb{DELETE /calendars/\{calendarId\}/} in the DataRobot Public API.

Delete a calendar
}
\examples{
\dontrun{
calendarId <- "5da75da31fb4a45b8a815a53"
DeleteCalendar(calendarId)
}
}
\seealso{
The method CalendarsDelete in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
