% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{DeleteComments}
\alias{DeleteComments}
\title{DeleteComments}
\usage{
DeleteComments(commentId, ...)
}
\description{
Delete a comment
}
\details{
Delete a comment
}
\seealso{
The method CommentsDelete in [datarobot.apicore::CommentsApi], which this function wraps.

Other Comments: 
\code{\link{CreateComments}()},
\code{\link{ListComments}()},
\code{\link{PatchComments}()},
\code{\link{RetrieveComments}()}
}
\concept{Comments}
