% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R,
%   R/ComplianceDocumentation.R
\name{DeleteComplianceDocTemplate}
\alias{DeleteComplianceDocTemplate}
\title{DeleteComplianceDocTemplate}
\usage{
DeleteComplianceDocTemplate(templateId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
Nothing returned, but deletes the compliance doc template.
}
\description{
Delete a compliance documentation template
}
\details{
Delete a compliance documentation template. Documentation previously generated using this template will remain unchanged.

This method invokes \verb{DELETE /complianceDocTemplates/\{templateId\}/} in the DataRobot Public API.

Deletes a compliance doc template.

Note that default templates cannot be deleted.
}
\examples{
\dontrun{
templateId <- "5cf85080d9436e5c310c796d"
DeleteComplianceDocTemplate(templateId)
}
}
\seealso{
The method ComplianceDocTemplatesDelete in \link[datarobot.apicore:DocumentationApi]{datarobot.apicore::DocumentationApi}, which invokes the same underlying API endpoint.
}
