% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{DeleteCredentials}
\alias{DeleteCredentials}
\title{DeleteCredentials}
\usage{
DeleteCredentials(credentialId, ...)
}
\description{
Delete the credentials set.
}
\details{
Delete the credential set matching the specified ID if it is not used by any data connection or batch prediction job.
}
\seealso{
The method CredentialsDelete in [datarobot.apicore::CredentialsApi], which this function wraps.

Other Credentials: 
\code{\link{CreateCredentials}()},
\code{\link{ListCredentialsAssociations}()},
\code{\link{ListCredentials}()},
\code{\link{PatchCredentials}()},
\code{\link{PatchManyCredentialsAssociations}()},
\code{\link{RetrieveCredentials}()}
}
\concept{Credentials}
