% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{DeleteCustomModels}
\alias{DeleteCustomModels}
\title{DeleteCustomModels}
\usage{
DeleteCustomModels(customModelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete custom model.
}
\details{
Delete a custom model. Only users who have permission to edit custom model can delete it. Only custom models which are not currently deployed or undergoing custom model testing can be deleted. Relevant CustomModelImage will be deleted also.

This method invokes \verb{DELETE /customModels/\{customModelId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsDelete in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
