% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSources.R
\name{DeleteDataSource}
\alias{DeleteDataSource}
\title{Delete a data store.}
\usage{
DeleteDataSource(dataSourceId)
}
\arguments{
\item{dataSourceId}{character. The ID of the data store to update.}
}
\description{
Delete a data store.
}
\examples{
\dontrun{
 dataSourceId <- "5c1303269300d900016b41a7"
 DeleteDataSource(dataSourceId)
}
}
