% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R, R/Deployment.R
\name{DeleteDeployment}
\alias{DeleteDeployment}
\title{DeleteDeployment}
\usage{
DeleteDeployment(deploymentId, ignoreManagementAgent = "false", ...)
}
\arguments{
\item{deploymentId}{character. The ID of the deployment.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete deployment
}
\details{
Delete a deployment.

This method invokes \verb{DELETE /deployments/\{deploymentId\}/} in the DataRobot Public API.

Delete a deployment.
}
\examples{
\dontrun{
deploymentId <- "5e319d2e422fbd6b58a5edad"
DeleteDeployment(deploymentId)
}
}
\seealso{
The method DeploymentsDelete in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
