% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{DeleteDeploymentsChallengers}
\alias{DeleteDeploymentsChallengers}
\title{DeleteDeploymentsChallengers}
\usage{
DeleteDeploymentsChallengers(deploymentId, challengerId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete challenger model
}
\details{
Delete challenger model.

This method invokes \verb{DELETE /deployments/\{deploymentId\}/challengers/\{challengerId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsChallengersDelete in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
