% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{DeleteDeploymentsRetrainingPolicies}
\alias{DeleteDeploymentsRetrainingPolicies}
\title{DeleteDeploymentsRetrainingPolicies}
\usage{
DeleteDeploymentsRetrainingPolicies(retrainingPolicyId, deploymentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for deleting a deployment retraining policy.
}
\details{
Delete a deployment retraining policy.

This method invokes \verb{DELETE /deployments/\{deploymentId\}/retrainingPolicies/\{retrainingPolicyId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingPoliciesDelete in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
