% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Featurelists.R
\name{DeleteFeaturelist}
\alias{DeleteFeaturelist}
\title{Delete a featurelist}
\usage{
DeleteFeaturelist(featurelist)
}
\arguments{
\item{featurelist}{list. The featurelist to delete.}
}
\description{
Delete a featurelist
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  featureList <- CreateFeaturelist(projectId, "myFeaturelist", c("feature1", "feature2"))
  DeleteFeaturelist(featurelist)
}
}
