% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{DeleteImageAugmentationLists}
\alias{DeleteImageAugmentationLists}
\title{DeleteImageAugmentationLists}
\usage{
DeleteImageAugmentationLists(augmentationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete an existing augmentation lists by id
}
\details{
Delete an existing augmentation lists by id

This method invokes \verb{DELETE /imageAugmentationLists/\{augmentationId\}/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationListsDelete in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
