% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jobs.R
\name{DeleteJob}
\alias{DeleteJob}
\title{Cancel a running job}
\usage{
DeleteJob(job)
}
\arguments{
\item{job}{The job you want to cancel (one of the items in the list returned from ListJobs)}
}
\description{
Cancel a running job
}
