% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{DeleteManyCustomModelsVersionsDependencyBuild}
\alias{DeleteManyCustomModelsVersionsDependencyBuild}
\title{DeleteManyCustomModelsVersionsDependencyBuild}
\usage{
DeleteManyCustomModelsVersionsDependencyBuild(
  customModelId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel dependency build.
}
\details{
Cancel the custom model version's dependency build.

This method invokes \verb{DELETE /customModels/\{customModelId\}/versions/\{customModelVersionId\}/dependencyBuild/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsDependencyBuildDeleteMany in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
