% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{DeleteManyGroupsUsers}
\alias{DeleteManyGroupsUsers}
\title{DeleteManyGroupsUsers}
\usage{
DeleteManyGroupsUsers(groupId, users, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Remove users from group
}
\details{
Remove users from the group.

This method invokes \verb{DELETE /groups/\{groupId\}/users/} in the DataRobot Public API.
}
\seealso{
The method GroupsUsersDeleteMany in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
