% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R,
%   R/ai_catalog_apiWrapper.R
\name{DeleteManyUserBlueprints}
\alias{DeleteManyUserBlueprints}
\title{DeleteManyUserBlueprints}
\usage{
DeleteManyUserBlueprints(userBlueprintIds, ...)

DeleteManyUserBlueprints(userBlueprintIds, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete user blueprints.

Delete user blueprints.
}
\details{
Delete user blueprints, specified by userBlueprintIds.

This method invokes \verb{DELETE /userBlueprints/} in the DataRobot Public API.

Delete user blueprints, specified by userBlueprintIds.

This method invokes \verb{DELETE /userBlueprints/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsDeleteMany in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsDeleteMany in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
