% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{DeleteManyUserNotifications}
\alias{DeleteManyUserNotifications}
\title{DeleteManyUserNotifications}
\usage{
DeleteManyUserNotifications(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete all user notifications
}
\details{
Delete all notifications associated with the user.

This method invokes \verb{DELETE /userNotifications/} in the DataRobot Public API.
}
\seealso{
The method UserNotificationsDeleteMany in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
