% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{DeleteManyUsersRateLimitUsage}
\alias{DeleteManyUsersRateLimitUsage}
\title{DeleteManyUsersRateLimitUsage}
\usage{
DeleteManyUsersRateLimitUsage(userId, ...)
}
\description{
Reset resource usage for all resources
}
\details{
Reset specified user's rate limit resource usage to zero for all resources. When windows roll over, all limits are automatically reset. Use this route to reset usage sooner.
}
\seealso{
The method UsersRateLimitUsageDeleteMany in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{CreateUsageDataExports}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsEvents}()},
\code{\link{ListEventLogsPredictionUsage}()},
\code{\link{ListEventLogs}()},
\code{\link{ListUsageDataExportsSupportedEvents}()},
\code{\link{ListUsersRateLimitUsage}()},
\code{\link{RetrieveEventLogs}()},
\code{\link{RetrieveUsageDataExports}()}
}
\concept{Analytics}
