% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{DeleteModel}
\alias{DeleteModel}
\title{Delete a specified DataRobot model}
\usage{
DeleteModel(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModel, or each element of the list returned by
the function ListModels.}
}
\description{
This function removes the model specified by the parameter model from its
associated project.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  modelId <- "5996f820af07fc605e81ead4"
  model <- GetModel(projectId, modelId)
  DeleteModel(model)
}
}
