% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Featurelists.R
\name{DeleteModelingFeaturelist}
\alias{DeleteModelingFeaturelist}
\title{Delete a modeling featurelist}
\usage{
DeleteModelingFeaturelist(featurelist)
}
\arguments{
\item{featurelist}{list. The modeling featurelist to delete.}
}
\description{
Delete a modeling featurelist
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  featureList <- CreateModelingFeaturelist(projectId, "myFeaturelist", c("feature1", "feature2"))
  featurelistId <- featureList$featurelistId
  GetModelingFeaturelist(projectId, featurelistId)
  DeleteModelingFeaturelist(projectId, featurelistId)
}
}
