% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/DeleteModelJob.R
\name{DeleteModelJob}
\alias{DeleteModelJob}
\title{DeleteModelJob}
\usage{
DeleteModelJob(project, modelJobId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId
with this identifier.}

\item{modelJobId}{integer. Identifier for the modeling job to be
deleted; can be obtained from the results returned by the function
ListModelJobs.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel a modeling job.
}
\details{
Cancel a modeling job

This method invokes \verb{DELETE /projects/\{projectId\}/modelJobs/\{jobId\}/} in the DataRobot Public API.

Delete a model job from the modeling queue

This function deletes the modeling job specified by modelJobId from
the DataRobot modeling queue.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
initialJobs <- ListModelJobs(project)
job <- initialJobs[[1]]
modelJobId <- job$modelJobId
DeleteModelJob(projectId, modelJobId)
}
}
\seealso{
The method ProjectsModelJobsDelete in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
