% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{DeleteMultilabelInsightsPairwiseManualSelections}
\alias{DeleteMultilabelInsightsPairwiseManualSelections}
\title{DeleteMultilabelInsightsPairwiseManualSelections}
\usage{
DeleteMultilabelInsightsPairwiseManualSelections(
  manualSelectionListId,
  multilabelInsightsKey,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete label list.
}
\details{
Delete label list.

This method invokes \verb{DELETE /multilabelInsights/\{multilabelInsightsKey\}/pairwiseManualSelections/\{manualSelectionListId\}/} in the DataRobot Public API.
}
\seealso{
The method MultilabelInsightsPairwiseManualSelectionsDelete in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
