% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteModelJob.R
\name{DeletePendingJob}
\alias{DeletePendingJob}
\title{Delete a model job from the modeling queue}
\usage{
DeletePendingJob(project, modelJobId)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId
with this identifier.}

\item{modelJobId}{Integer, identifier for the modeling job to be
deleted; can be obtained from the results returned by the function
GetModelJobs.}
}
\description{
(Deprecated in 2.1, will be removed in 2.3. Use GetModelJobs instead.)
}

