% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R,
%   R/PredictionDatasets.R
\name{DeletePredictionDataset}
\alias{DeletePredictionDataset}
\title{DeletePredictionDataset}
\usage{
DeletePredictionDataset(project, datasetId)

DeletePredictionDataset(project, datasetId)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{datasetId}{The id of the dataset to delete}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a dataset that was uploaded for prediction.

This function removes a prediction dataset
}
\details{
Delete a dataset that was uploaded for prediction.

This method invokes \verb{DELETE /projects/\{projectId\}/predictionDatasets/\{datasetId\}/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
datasets <- ListPredictionDatasets(projectId)
dataset <- datasets[[1]]
datasetId <- dataset$id
DeletePredictionDataset(projectId, datasetId)
}
}
\seealso{
The method ProjectsPredictionDatasetsDelete in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
