% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R,
%   R/predictions_apiWrapper.R, R/PredictionExplanations.R
\name{DeletePredictionExplanations}
\alias{DeletePredictionExplanations}
\title{DeletePredictionExplanations}
\usage{
DeletePredictionExplanations(project, predictionExplanationId, ...)

DeletePredictionExplanations(project, predictionExplanationId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{predictionExplanationId}{character. Id of the prediction explanations.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
Logical TRUE and displays a message to the user if the delete
request was successful; otherwise an error message is displayed.
}
\description{
Delete saved Prediction Explanations.

Delete saved Prediction Explanations.
}
\details{
Delete saved Prediction Explanations. Deletes both the actual prediction explanations and the corresponding PredictionExplanationsRecord.

This method invokes \verb{DELETE /projects/\{projectId\}/predictionExplanationsRecords/\{predictionExplanationsId\}/} in the DataRobot Public API.

Delete saved Prediction Explanations. Deletes both the actual prediction explanations and the corresponding PredictionExplanationsRecord.

This method invokes \verb{DELETE /projects/\{projectId\}/predictionExplanationsRecords/\{predictionExplanationsId\}/} in the DataRobot Public API.

Function to delete prediction explanations

This function deletes prediction explanations specified by project and
predictionExplanationId.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
datasets <- ListPredictionDatasets(projectId)
dataset <- datasets[[1]]
datasetId <- dataset$id
model <- GetModel(projectId, modelId)
jobId <- RequestPredictionExplanations(model, datasetId)
predictionExplanationId <- GetPredictionExplanationsMetadataFromJobId(projectId, jobId)$id
DeletePredictionExplanations(projectId, predictionExplanationId)
}
}
\seealso{
The method ProjectsPredictionExplanationsRecordsDelete in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.

The method ProjectsPredictionExplanationsRecordsDelete in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
