% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionExplanations.R
\name{DeletePredictionExplanationsInitialization}
\alias{DeletePredictionExplanationsInitialization}
\title{Delete the prediction explanations initialization for a model.}
\usage{
DeletePredictionExplanationsInitialization(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModel, or each element of the list returned by
the function ListModels.}
}
\value{
Logical TRUE and displays a message to the user if the delete
request was successful; otherwise an error message is displayed.
}
\description{
Delete the prediction explanations initialization for a model.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  modelId <- "5996f820af07fc605e81ead4"
  model <- GetModel(projectId, modelId)
  DeletePredictionExplanationsInitialization(model)
}
}
