% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Projects.R
\name{DeleteProject}
\alias{DeleteProject}
\title{DeleteProject}
\usage{
DeleteProject(project, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a project
}
\details{
Delete a project

This method invokes \verb{DELETE /projects/\{projectId\}/} in the DataRobot Public API.

Delete a specified element from the DataRobot project list

This function deletes the project defined by project,
described under Arguments. This parameter may be obtained
in several ways, including: (1), as one of the projectId
elements of the list returned by ListProjects; (2), as
the S3 object returned by the GetProject function; or (3),
as the list returned by the SetupProject function.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
DeleteProject(projectId)
}
}
\seealso{
The method Delete in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
