% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{DeleteProjectsPayoffMatrices}
\alias{DeleteProjectsPayoffMatrices}
\title{DeleteProjectsPayoffMatrices}
\usage{
DeleteProjectsPayoffMatrices(payoffMatrixId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a payoff matrix in a project.
}
\details{
Delete a payoff matrix in a project.

This method invokes \verb{DELETE /projects/\{projectId\}/payoffMatrices/\{payoffMatrixId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPayoffMatricesDelete in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
