% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReasonCodes.R
\name{DeleteReasonCodes}
\alias{DeleteReasonCodes}
\title{Function to delete reason codes}
\usage{
DeleteReasonCodes(project, reasonCodeId)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{reasonCodeId}{Character string. id of the reason codes}
}
\value{
Logical TRUE and displays a message to the user if the delete
request was successful; otherwise an error message is displayed.
}
\description{
This function deletes reason codes specified by project and reasonCodeId
}
