% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReasonCodes.R
\name{DeleteReasonCodesInitialization}
\alias{DeleteReasonCodesInitialization}
\title{Delete the reason codes initialization for a model.}
\usage{
DeleteReasonCodesInitialization(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModelObject, or each element of the list returned by
the function GetAllModels.}
}
\description{
Delete the reason codes initialization for a model.
}
