% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{DeleteRelationshipsConfigurations}
\alias{DeleteRelationshipsConfigurations}
\title{DeleteRelationshipsConfigurations}
\usage{
DeleteRelationshipsConfigurations(relationshipsConfigurationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a relationships configuration
}
\details{
Delete a relationships configuration

This method invokes \verb{DELETE /relationshipsConfigurations/\{relationshipsConfigurationId\}/} in the DataRobot Public API.
}
\seealso{
The method RelationshipsConfigurationsDelete in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
