% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{DeleteScheduledJobs}
\alias{DeleteScheduledJobs}
\title{DeleteScheduledJobs}
\usage{
DeleteScheduledJobs(jobId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete scheduled job
}
\details{
Delete scheduled job

This method invokes \verb{DELETE /scheduledJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method ScheduledJobsDelete in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
