% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_apiWrapper.R
\name{DeleteStatus}
\alias{DeleteStatus}
\title{DeleteStatus}
\usage{
DeleteStatus(statusId, ...)
}
\description{
Delete a task
}
\details{
Destroy an async status object
}
\seealso{
The method StatusDelete in [datarobot.apicore::JobsApi], which this function wraps.

Other Jobs: 
\code{\link{ListStatus}()},
\code{\link{RetrieveStatus}()}
}
\concept{Jobs}
