% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferrableModels.R
\name{DeleteTransferrableModel}
\alias{DeleteTransferrableModel}
\title{Delete this imported model.}
\usage{
DeleteTransferrableModel(importId)
}
\arguments{
\item{importId}{Character string unique id of the import}
}
\description{
Delete this imported model.
}

