% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{DeleteUseCasesAttachments}
\alias{DeleteUseCasesAttachments}
\title{DeleteUseCasesAttachments}
\usage{
DeleteUseCasesAttachments(useCaseId, attachmentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Removes a resource from a use case.
}
\details{
Removes a resource from a use case.

This method invokes \verb{DELETE /useCases/\{useCaseId\}/attachments/\{attachmentId\}/} in the DataRobot Public API.
}
\seealso{
The method UseCasesAttachmentsDelete in \link[datarobot.apicore:UseCaseApi]{datarobot.apicore::UseCaseApi}, which invokes the same underlying API endpoint.
}
