% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{DeleteUserNotifications}
\alias{DeleteUserNotifications}
\title{DeleteUserNotifications}
\usage{
DeleteUserNotifications(userNotificationId, ...)
}
\description{
Delete user notification
}
\details{
Delete one notification associated with the user.
}
\seealso{
The method UserNotificationsDelete in [datarobot.apicore::NotificationsApi], which this function wraps.

Other Notifications: 
\code{\link{CreateNotificationChannels}()},
\code{\link{CreateNotificationEmailChannelVerificationStatus}()},
\code{\link{CreateNotificationEmailChannelVerification}()},
\code{\link{CreateNotificationPolicies}()},
\code{\link{CreateNotificationPolicyMutes}()},
\code{\link{CreateNotificationWebhookChannelTests}()},
\code{\link{CreateNotifications}()},
\code{\link{CreateRemoteEvents}()},
\code{\link{DeleteManyUserNotifications}()},
\code{\link{DeleteNotificationChannels}()},
\code{\link{DeleteNotificationPolicies}()},
\code{\link{DeleteNotificationPolicyMutes}()},
\code{\link{ListNotificationChannels}()},
\code{\link{ListNotificationEvents}()},
\code{\link{ListNotificationLogs}()},
\code{\link{ListNotificationPolicies}()},
\code{\link{ListNotificationPolicyMutes}()},
\code{\link{ListUserNotifications}()},
\code{\link{PatchManyUserNotifications}()},
\code{\link{PatchUserNotifications}()},
\code{\link{PutNotificationChannels}()},
\code{\link{PutNotificationPolicies}()},
\code{\link{RetrieveNotificationChannels}()},
\code{\link{RetrieveNotificationPolicies}()},
\code{\link{RetrieveNotificationWebhookChannelTests}()}
}
\concept{Notifications}
