% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{DeploymentAccuracyMetric}
\alias{DeploymentAccuracyMetric}
\title{Deployment accuracy metrics}
\format{
An object of class \code{list} of length 27.
}
\usage{
DeploymentAccuracyMetric
}
\description{
All possible deployment accuracy metrics. Added in DataRobot API 2.18.
}
\details{
For usage, see `\code{DeploymentAccuracy} and `code{DeploymentAccuracyOverTime}.
}
\keyword{datasets}
