% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{DeploymentServiceHealthMetric}
\alias{DeploymentServiceHealthMetric}
\title{Deployment service health metrics}
\format{
An object of class \code{list} of length 11.
}
\usage{
DeploymentServiceHealthMetric
}
\description{
Added in DataRobot API 2.18.
}
\details{
For usage, see \code{GetDeploymentServiceStats}.
}
\keyword{datasets}
