% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{DeprecatedHeaderMessage}
\alias{DeprecatedHeaderMessage}
\title{Handler for Deprecated header in API responses}
\usage{
DeprecatedHeaderMessage(rawResponse)
}
\arguments{
\item{rawResponse}{An httr response object.}
}
\description{
Handler for Deprecated header in API responses
}
\keyword{internal}
