% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{DifferencingMethod}
\alias{DifferencingMethod}
\title{Differencing method}
\format{An object of class \code{list} of length 4.}
\usage{
DifferencingMethod
}
\description{
Differencing method
}
\keyword{datasets}
