% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.createApiConvenienceWrapper}
\alias{.createApiConvenienceWrapper}
\title{Find all generators related to Api classes, e.g. `ProjectsApi`, and
instantiate them in a list in the `datarobot` namespace for convenience.}
\usage{
.createApiConvenienceWrapper()
}
\description{
Find all generators related to Api classes, e.g. `ProjectsApi`, and
instantiate them in a list in the `datarobot` namespace for convenience.
}
\keyword{internal}
