% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AiCatalog.R
\name{DownloadDatasetAsCsv}
\alias{DownloadDatasetAsCsv}
\title{DownloadDatasetAsCsv}
\usage{
DownloadDatasetAsCsv(catalogId)
}
\arguments{
\item{catalogId}{character. The id of the catalog item.}
}
\description{
Downloads the dataset CSV file corresponding to catalogId
}
\details{
To Note: While using this method a call to httr::content() requires the installation of the readr package
}
\seealso{
\link[readr]{https://cran.r-project.org/web/packages/readr/index.html}
}
\concept{AiCatalog}
