% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/DownloadPrimeCode.R
\name{DownloadPrimeCode}
\alias{DownloadPrimeCode}
\title{DownloadPrimeCode}
\usage{
DownloadPrimeCode(project, primeFileId, filepath, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{primeFileId}{numeric. Prime file Id (can be acquired using ListPrimeFiles function)}

\item{filepath}{character. The location to save the file to.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download Code
}
\details{
Download code from an existing Prime file

This method invokes \verb{GET /projects/\{projectId\}/primeFiles/\{primeFileId\}/download/} in the DataRobot Public API.

Download the code of DataRobot Prime model and save it to a file.

Training a model using a ruleset is a necessary prerequisite for being able to download the
  code for a ruleset.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
primeFiles <- ListPrimeFiles(projectId)
primeFile <- primeFiles[[1]]
primeFileId <- primeFile$id
file <- file.path(tempdir(), "primeCode.py")
DownloadPrimeCode(projectId, primeFileId, file)
}
}
\seealso{
The method ProjectsPrimeFilesDownloadList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
