% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RatingTables.R
\name{DownloadRatingTable}
\alias{DownloadRatingTable}
\title{DownloadRatingTable}
\usage{
DownloadRatingTable(project, ratingTableId, filename)

DownloadRatingTable(project, ratingTableId, filename)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{ratingTableId}{character. The ID of the rating table.}

\item{filename}{character. Filename of file to save the rating table to.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
Nothing returned, but downloads the file to the stated filename.
}
\description{
Retrieve Rating Table File
}
\details{
Retrieve the CSV file for the rating table

This method invokes \verb{GET /projects/\{projectId\}/ratingTables/\{ratingTableId\}/file/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
ratingTableId <- "5984b4d7100d2b31c1166529"
file <- file.path(tempdir(), "ratingTable.csv")
DownloadRatingTable(projectId, ratingTableId, file)
}
}
\seealso{
The method ProjectsRatingTablesFileList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
