% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R, R/models_apiWrapper.R
\name{DownloadScoringCode}
\alias{DownloadScoringCode}
\title{Download scoring code JAR}
\usage{
DownloadScoringCode(project, modelId, fileName, sourceCode = FALSE, ...)

DownloadScoringCode(project, modelId, fileName, sourceCode = FALSE, ...)
}
\arguments{
\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{fileName}{character. File path where scoring code will be saved.}

\item{sourceCode}{logical. Optional. Set to TRUE to download source code archive.
It will not be executable.}
}
\description{
Retrieve Scoring Code
}
\details{
Retrieve Scoring Code for making new predictions from an existing model offline. You need the \"Scoring Code\" feature enabled to use this route.  By default, returns a compiled executable JAR that can be executed locally to calculate model predictions, or it can be used as a library for a Java application. Execute it with the '--help` parameters to learn how to use it as a command-line utility. See model API documentation (https://javadoc.io/doc/com.datarobot/datarobot-prediction/latest/index.html) to be able to use it inside an existing Java application.  With the sourceCode query parameter set to 'true', returns a source code archive that can be used to review internal calculations of the model. This JAR is NOT executable.  See \"https://app.datarobot.com/docs/predictions/scoring-code/scorecode.html\" in DataRobot application for more information.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
file <- file.path(tempdir(), "scoringCode.jar")
DownloadScoringCode(projectId, modelId, file)
}
}
\seealso{
The method ProjectsModelsScoringCodeList in [datarobot.apicore::ModelsApi], which this function wraps.

Other Models: 
\code{\link{AddEureqaSolution}()},
\code{\link{CreateCustomInferenceImagesFeatureImpact}()},
\code{\link{CreateCustomModelTests}()},
\code{\link{CreateCustomModelsFromCustomModel}()},
\code{\link{CreateCustomModelsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersionsConversions}()},
\code{\link{CreateCustomModelsVersionsDependencyBuild}()},
\code{\link{CreateCustomModelsVersionsFeatureImpact}()},
\code{\link{CreateCustomModelsVersionsFromRepository}()},
\code{\link{CreateCustomModelsVersionsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersions}()},
\code{\link{CreateCustomModels}()},
\code{\link{CreateCustomTrainingBlueprints}()},
\code{\link{CreateFromLatestCustomModelsVersion}()},
\code{\link{CreateModelExports}()},
\code{\link{CreateModelPackagesArchive}()},
\code{\link{CreateModelPackagesFromLearningModel}()},
\code{\link{CreatePrimeCode}()},
\code{\link{CreateProjectsBlenderBuilders}()},
\code{\link{CreateProjectsDatetimeModelsFromModel}()},
\code{\link{CreateProjectsDeploymentReadyModels}()},
\code{\link{CreateProjectsModelsFromModel}()},
\code{\link{CreateProjectsModelsPredictionIntervals}()},
\code{\link{CreateRatingTable}()},
\code{\link{DeleteCustomModelTests}()},
\code{\link{DeleteCustomModelsVersionsConversions}()},
\code{\link{DeleteCustomModels}()},
\code{\link{DeleteManyCustomModelsVersionsDependencyBuild}()},
\code{\link{DeleteModelJob}()},
\code{\link{DownloadPrimeCode}()},
\code{\link{DownloadRatingTable}()},
\code{\link{DownloadTransferableModel}()},
\code{\link{GetBlenderModel}()},
\code{\link{GetDatetimeModel}()},
\code{\link{GetFrozenModel}()},
\code{\link{GetMissingValuesReport}()},
\code{\link{GetModelJob}()},
\code{\link{GetModelParameters}()},
\code{\link{GetModel}()},
\code{\link{GetPrimeEligibility}()},
\code{\link{GetPrimeFile}()},
\code{\link{GetPrimeModel}()},
\code{\link{GetRatingTableModel}()},
\code{\link{GetRatingTable}()},
\code{\link{GetRulesets}()},
\code{\link{IsBlenderEligible}()},
\code{\link{ListCustomInferenceImagesFeatureImpact}()},
\code{\link{ListCustomModelDeployments}()},
\code{\link{ListCustomModelLimits}()},
\code{\link{ListCustomModelTestsLog}()},
\code{\link{ListCustomModelTestsTail}()},
\code{\link{ListCustomModelTests}()},
\code{\link{ListCustomModelsAccessControl}()},
\code{\link{ListCustomModelsDownload}()},
\code{\link{ListCustomModelsVersionsConversions}()},
\code{\link{ListCustomModelsVersionsDependencyBuildLog}()},
\code{\link{ListCustomModelsVersionsDependencyBuild}()},
\code{\link{ListCustomModelsVersionsDownload}()},
\code{\link{ListCustomModelsVersionsFeatureImpact}()},
\code{\link{ListCustomModelsVersions}()},
\code{\link{ListCustomModels}()},
\code{\link{ListCustomTrainingBlueprints}()},
\code{\link{ListModelJobs}()},
\code{\link{ListModelPackagesFeatures}()},
\code{\link{ListModelPackagesSharedRoles}()},
\code{\link{ListModelPackages}()},
\code{\link{ListModelRecommendations}()},
\code{\link{ListModels}()},
\code{\link{ListPrimeFiles}()},
\code{\link{ListPrimeModels}()},
\code{\link{ListProjectsBlenderBuilders}()},
\code{\link{ListProjectsBlenderModels}()},
\code{\link{ListProjectsCombinedModelsSegmentsDownload}()},
\code{\link{ListProjectsCombinedModelsSegments}()},
\code{\link{ListProjectsCombinedModels}()},
\code{\link{ListProjectsDatetimeModels}()},
\code{\link{ListProjectsFrozenModels}()},
\code{\link{ListProjectsModelsClusterNames}()},
\code{\link{ListProjectsModelsNumIterationsTrained}()},
\code{\link{ListProjectsModelsPredictionIntervals}()},
\code{\link{ListProjectsRecommendedModelsRecommendedModel}()},
\code{\link{ListRatingTableModels}()},
\code{\link{ListRatingTables}()},
\code{\link{PatchCustomModelsVersions}()},
\code{\link{PatchCustomModels}()},
\code{\link{PatchManyCustomModelsAccessControl}()},
\code{\link{PatchManyCustomModelsTrainingData}()},
\code{\link{PatchManyCustomModelsVersionsFromRepository}()},
\code{\link{PatchManyProjectsModelsClusterNames}()},
\code{\link{PutManyProjectsSegmentChampion}()},
\code{\link{RenameRatingTable}()},
\code{\link{RequestApproximation}()},
\code{\link{RequestBlender}()},
\code{\link{RequestNewDatetimeModel}()},
\code{\link{RequestNewModel}()},
\code{\link{RequestNewRatingTableModel}()},
\code{\link{RequestPrimeModel}()},
\code{\link{RetrieveCustomModelTests}()},
\code{\link{RetrieveCustomModelsVersionsConversions}()},
\code{\link{RetrieveCustomModelsVersions}()},
\code{\link{RetrieveCustomModels}()},
\code{\link{RetrieveModelPackages}()},
\code{\link{RetrieveProjectsBlenderBuilders}()},
\code{\link{RetrieveProjectsCombinedModels}()},
\code{\link{RetrieveProjectsEureqaDistributionPlot}()},
\code{\link{RetrieveProjectsEureqaModelDetail}()}
}
\concept{Models}
