% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureLogs.R,
%   R/datetime_partitioning_apiWrapper.R
\name{DownloadTimeSeriesFeatureDerivationLog}
\alias{DownloadTimeSeriesFeatureDerivationLog}
\title{Download the time series feature derivation log as a text file.}
\usage{
DownloadTimeSeriesFeatureDerivationLog(project, file, ...)

DownloadTimeSeriesFeatureDerivationLog(project, file, ...)
}
\arguments{
\item{file}{character. The name or path of the file to download to.}
}
\value{
Nothing, but writes the output to the desired file.
}
\description{
Retrieve a text file containing the time series project feature log
}
\details{
Retrieve a text file containing the time series project feature log.  The Time Series Feature Log provides details about the feature generation process for a time series project. It includes information about which features are generated and their priority,as well as the detected properties of the time series data such as whether the series is stationary, and periodicities detected.  This route is only supported for time series projects that have finished partitioning.  The feature derivation log will include information about:  * Maximum number of feature to be generated, e.g., ``Limit on the maximum number of feature in this project is 500`` * Number of derived features tested during the feature generation process, e.g., ``Total number of derived features during the feature generation process is 571`` * Number of generated features removed during the feature reduction process e.g. ``Total number of features removed during the feature reduction process is 472`` * Number of remaining features after the combined feature generation and reduction process, e.g., ``The finalized number of features is 99`` * Detected stationarity of the series, e.g., ``Series detected as non-stationary`` * Detected presence of multiplicative trend in the series, e.g., ``Multiplicative trend detected`` * Detected periodicities in the series, e.g., ``Detected periodicities: 7 day`` * Window sizes used in rolling statistics / lag extractors, e.g., ``The window sizes chosen to be: 2 months (because the time step is 1 month and Feature Derivation Window is 2 months)`` * Features that are specified as known-in-advance, e.g., ``Variables treated as apriori: holiday`` * Details about why certain variables are transformed in the input data, e.g., ``Generating variable \"y (log)\" from \"y\" because multiplicative trend is detected`` * Details about features generated as time series features, and their priority, e.g., ``Generating feature \"date (actual)\" from \"date\" (priority: 1)``
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
DownloadTimeSeriesFeatureDerivationLog(projectId, "featureLog.txt")
}
}
\seealso{
\code{\link{GetTimeSeriesFeatureDerivationLog}}

The method ProjectsTimeSeriesFeatureLogFileList in [datarobot.apicore::DatetimePartitioningApi], which this function wraps.

Other DatetimePartitioning: 
\code{\link{CreateProjectsDatetimePartitioning}()},
\code{\link{CreateProjectsOptimizedDatetimePartitionings}()},
\code{\link{GetDatetimePartition}()},
\code{\link{GetTimeSeriesFeatureDerivationLog}()},
\code{\link{ListProjectsOptimizedDatetimePartitionings}()},
\code{\link{RetrieveProjectsOptimizedDatetimePartitionings}()}
}
\concept{DatetimePartitioning}
