% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferrableModels.R
\name{DownloadTransferrableModel}
\alias{DownloadTransferrableModel}
\title{Download an transferrable model file for use in an on-premise DataRobot standalone
prediction environment.
This function can only be used if model export is enabled, and will only be useful
if you have an on-premise environment in which to import it.}
\usage{
DownloadTransferrableModel(project, modelId, modelFile)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{modelId}{Unique alphanumeric identifier for the model of interest.}

\item{modelFile}{file name to be use for tranferrable model}
}
\description{
Download an transferrable model file for use in an on-premise DataRobot standalone
prediction environment.
This function can only be used if model export is enabled, and will only be useful
if you have an on-premise environment in which to import it.
}
