% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToDataRobot.R
\name{EditConfig}
\alias{EditConfig}
\title{Edit drconfig}
\usage{
EditConfig(configPath = GetDefaultConfigPath(), reauth = FALSE)
}
\arguments{
\item{configPath}{character. Defaults to the standard location of drconfig.yaml.}

\item{reauth}{logical. Updates the current R session to use the updated auth scheme.}
}
\description{
Opens for editing the DataRobot configuration file used to manage API client authentication.
Once editing is complete, this function will optionally update the R client to use the new
auth scheme.
}
\details{
The specific editor used is system-dependent -- see \code{\link[utils:file.edit]{utils::file.edit()}} for more details.
Keep in mind that any changes to this file may impact other DataRobot API clients you have
installed.
}
\seealso{
\code{\link[=ConnectToDataRobot]{ConnectToDataRobot()}}, \code{\link[utils:file.edit]{utils::file.edit()}}
}
