% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExpectHasKeys}
\alias{ExpectHasKeys}
\title{Make sure that the object has all of the keys specified. Also tests that there
are not additional keys if \code{allowAdditional} is FALSE (default).}
\usage{
ExpectHasKeys(obj, keys, allowAdditional = FALSE)
}
\arguments{
\item{obj}{object. A list, vector, or data.frame to check names.}

\item{keys}{character. A vector of names of keys to check.}

\item{allowAdditional}{logical. Should we allow there to be more keys than specified?}
}
\description{
Make sure that the object has all of the keys specified. Also tests that there
are not additional keys if \code{allowAdditional} is FALSE (default).
}
