% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{FeatureFromAsyncUrl}
\alias{FeatureFromAsyncUrl}
\title{Retrieve a feature from the creation URL}
\usage{
FeatureFromAsyncUrl(asyncUrl, maxWait = 600)
}
\arguments{
\item{asyncUrl}{The temporary status URL}

\item{maxWait}{The maximum time to wait (in seconds) for project creation before aborting.}
}
\description{
If feature creation times out, the error message includes a URL corresponding to the
creation task. That URL can be passed to this function (which will return the feature
details when finished) to resume waiting for feature creation.
}
