% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatMixedList.R
\name{FormatMixedList}
\alias{FormatMixedList}
\title{Format mixed JSON required for group-based partitioning}
\usage{
FormatMixedList(mixedList, specialCase)
}
\arguments{
\item{mixedList}{list. The list to format JSON for.}

\item{specialCase}{character. A vector of names of columns where JSON
unboxing should apply.}
}
\description{
Format mixed JSON required for group-based partitioning
}
