% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeUtils.R
\name{formatRFC3339Timestamp}
\alias{formatRFC3339Timestamp}
\title{formatRFC3339Timestamp}
\usage{
formatRFC3339Timestamp(date)
}
\arguments{
\item{date}{POSIXt or date. The date(s) to be formatted.}
}
\description{
The DataRobot APIs expect dates formatted as RFC 3339 strings. This is the
same as ISO 8601. To be safe, use UTC as the timezone (and format it with a
'Z' suffix), and use 'T' as the date/time separator.
}
\seealso{
Other API datetime functions: 
\code{\link{RFC3339DateTimeFormat}},
\code{\link{parseRFC3339Timestamp}()},
\code{\link{transformRFC3339Period}()},
\code{\link{validateReportingPeriodTime}()}
}
\concept{API datetime functions}
