% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiftChart.R
\name{GetAllLiftCharts}
\alias{GetAllLiftCharts}
\title{Retrieve lift chart data for a model for all available data partitions (deprecated)}
\usage{
GetAllLiftCharts(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by the function
GetModel, or each element of the list returned by the function ListModels.}
}
\description{
Retrieve lift chart data for a model for all available data partitions (deprecated)
}
