% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{GetAllModels}
\alias{GetAllModels}
\title{Retrieve all available model information for a DataRobot project}
\usage{
GetAllModels(project)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}
}
\description{
(Deprecated in 2.8, will be removed in 2.10. Use ListModels instead.)
}
