% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RocCurve.R
\name{GetAllRocCurves}
\alias{GetAllRocCurves}
\title{Retrieve ROC curve data for a model for all available data partitions (deprecated version)}
\usage{
GetAllRocCurves(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by the function
GetModel, or each element of the list returned by the function ListModels.}
}
\description{
Retrieve ROC curve data for a model for all available data partitions (deprecated version)
}
\seealso{
ListRocCurves
}
