% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatetimeTrendPlots.R
\name{GetAnomalyOverTimePlotPreview}
\alias{GetAnomalyOverTimePlotPreview}
\title{Retrieve Anomaly over Time preview plot for a model.}
\usage{
GetAnomalyOverTimePlotPreview(
  model,
  predictionThreshold = 0.5,
  backtest = 0,
  source = SourceType$Validation,
  seriesId = NULL,
  maxWait = 600
)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModel, or each element of the list returned by
the function ListModels.}

\item{predictionThreshold}{numeric. Optional. Only bins with predictions exceeding
this threshold will be returned in the response.}

\item{backtest}{integer or character. Optional. Retrieve plots for a specific backtest.
Use the backtest index starting from zero.
To retrieve plots for holdout, use \code{DataSubset$Holdout}.}

\item{source}{character. Optional. The source of the data for the backtest/holdout.
Must be one of \code{SourceType}.}

\item{seriesId}{character. Optional. The name of the series to retrieve for multiseries projects.
If not provided an average plot for the first 1000 series will be retrieved.}

\item{maxWait}{integer. Optional. The maximum time to wait for a compute job to complete
before retrieving the plots. Default is 600. If 0, the plots would be retrieved
without attempting the computation.}
}
\value{
list with the following components:
\itemize{
  \item startDate. POSIXct: The datetime of the start of the chartdata (inclusive).
  \item endDate. POSIXct: The datetime of the end of the chartdata (exclusive).
  \item predictionThreshold. numeric: Only bins with predictions exceeding
        this threshold are returned in the response.
  \item bins. data.frame: Each row represents a bin in the plot. Dataframe has following columns:
  \itemize{
    \item startDate. POSIXct: The datetime of the start of the bin (inclusive).
    \item endDate. POSIXct: The datetime of the end of the bin (exclusive).
  }
}
}
\description{
Retrieve Anomaly over Time preview plot for a model.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
model <- GetModel(projectId, modelId)
plot <- GetAnomalyOverTimePlotPreview(model, predictionThreshold = 0.1)

# The code below is an example demonstration on how to plot the data
x <- seq(plot$startDate, plot$endDate, plot$bins$endDate[1] - plot$bins$startDate[1])
png("anomaly_over_time_preview.png", width = 1200, height = 600, units = "px")
par(mar = c(10, 5, 5, 5))
plot(x, rep(0, length(x)), type = "l", yaxt = "n", ylab = "", xaxt = "n", xlab = "")
points(plot$bins$startDate, rep(0, length(plot$bins$startDate)), pch = 19, col = "red")
axis(1, x, format(x, "\%Y-\%m-\%d"), las = 3)
title(xlab = "Date", mgp = c(7, 1, 0))
legendString <- sprintf("Anomaly for prediction threshold \%.2f", plot$predictionThreshold)
legend("topright", legend = c(legendString), col = c("red"), pch = c(19))
dev.off()
}
}
