% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blenders.R
\name{GetBlenderModelFromJobId}
\alias{GetBlenderModelFromJobId}
\title{Retrieve a new or updated blender model defined by modelJobId}
\usage{
GetBlenderModelFromJobId(project, modelJobId, maxWait = 600)
}
\arguments{
\item{modelJobId}{integer. The integer returned by RequestBlender.}

\item{maxWait}{integer. The maximum time (in seconds) to wait for the model job to
complete.}
}
\description{
The function RequestBlender initiates the creation of new blender models in a
DataRobot project.
}
\details{
It submits requests to the DataRobot modeling
engine and returns an integer-valued modelJobId. The
GetBlenderModelFromJobId function polls the modeling engine until
the model has been built or a specified time limit is exceeded,
returning an S3 object of class 'dataRobotBlenderModel' when the model
is available.

Motivation for this function is the fact that some models -
e.g., very complex machine learning models fit to large datasets -
may take a long time to complete. Splitting the model creation
request from model retrieval in these cases allows the user to
perform other interactive R session tasks between the time the
model creation/update request is made and the time the final
model is available.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelsToBlend <- c("5996f820af07fc605e81ead4", "59a5ce3301e9f0296721c64c")
blendJobId <- RequestBlender(projectId, modelsToBlend, "GLM")
GetBlenderModelFromJobId(projectId, blendJobId)
}
}
