% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{GetBlueprintChart}
\alias{GetBlueprintChart}
\title{GetBlueprintChart}
\usage{
GetBlueprintChart(project, blueprintId, ...)
}
\description{
Retrieve a blueprint chart by blueprint id.
}
\details{
Retrieve a blueprint chart by blueprint id.
}
\seealso{
The method ProjectsBlueprintsBlueprintChartList in [datarobot.apicore::BlueprintsApi], which this function wraps.

Other Blueprints: 
\code{\link{CreateUserBlueprintsBulkValidations}()},
\code{\link{CreateUserBlueprintsFromBlueprintId}()},
\code{\link{CreateUserBlueprintsFromCustomTaskVersionId}()},
\code{\link{CreateUserBlueprintsFromUserBlueprintId}()},
\code{\link{CreateUserBlueprintsProjectBlueprints}()},
\code{\link{CreateUserBlueprintsTaskParameters}()},
\code{\link{CreateUserBlueprintsValidations}()},
\code{\link{CreateUserBlueprints}()},
\code{\link{DeleteManyUserBlueprints}()},
\code{\link{DeleteUserBlueprints}()},
\code{\link{GetBlueprintDocumentation}()},
\code{\link{GetBlueprint}()},
\code{\link{GetModelBlueprintChart}()},
\code{\link{GetModelBlueprintDocumentation}()},
\code{\link{ListBlueprints}()},
\code{\link{ListProjectsModelsLogs}()},
\code{\link{ListProjectsModelsTrainingArtifact}()},
\code{\link{ListUserBlueprintsInputTypes}()},
\code{\link{ListUserBlueprintsSharedRoles}()},
\code{\link{ListUserBlueprintsTasks}()},
\code{\link{ListUserBlueprints}()},
\code{\link{PatchManyUserBlueprintsSharedRoles}()},
\code{\link{PatchUserBlueprints}()},
\code{\link{RetrieveUserBlueprints}()}
}
\concept{Blueprints}
