% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendars.R, R/projects_apiWrapper.R
\name{GetCalendar}
\alias{GetCalendar}
\title{Retrieve a calendar}
\usage{
GetCalendar(calendarId, ...)

GetCalendar(calendarId, ...)
}
\arguments{
\item{calendarId}{character. The ID of the calendar to retrieve.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Retrieve information about a calendar.
}
\details{
List all the information about a calendar such as the total number of event dates, the earliest calendar event date, the IDs of projects currently using this calendar and the others.
}
\examples{
\dontrun{
calendarId <- "5da75da31fb4a45b8a815a53"
GetCalendar(calendarId)
}
}
\seealso{
The method CalendarsRetrieve in [datarobot.apicore::ProjectsApi], which this function wraps.

Other Projects: 
\code{\link{BatchFeaturesTypeTransform}()},
\code{\link{CloneProject}()},
\code{\link{CreateCalendarsFromCountryCode}()},
\code{\link{CreateCalendarsFromDataset}()},
\code{\link{CreateCalendar}()},
\code{\link{CreateFeaturelist}()},
\code{\link{CreateHdfsProjects}()},
\code{\link{CreateModelingFeaturelist}()},
\code{\link{CreateProjectCleanupJobs}()},
\code{\link{CreateProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{CreateProjectsFeatureDiscoveryRecipeSqlExports}()},
\code{\link{CreateProjectsSecondaryDatasetsConfigurations}()},
\code{\link{CreateProjectsSegmentationTasks}()},
\code{\link{CreateRelationshipsConfigurations}()},
\code{\link{DeleteCalendar}()},
\code{\link{DeleteProjectCleanupJobs}()},
\code{\link{DeleteProjectsJobs}()},
\code{\link{DeleteProjectsSecondaryDatasetsConfigurations}()},
\code{\link{DeleteProject}()},
\code{\link{DeleteRelationshipsConfigurations}()},
\code{\link{GetFeatureHistogram}()},
\code{\link{GetFeatureInfo}()},
\code{\link{GetFeaturelist}()},
\code{\link{GetJob}()},
\code{\link{GetModelingFeaturelist}()},
\code{\link{GetProjectStatus}()},
\code{\link{GetProject}()},
\code{\link{GetValidMetrics}()},
\code{\link{ListCalendarCountryCodes}()},
\code{\link{ListCalendarsAccessControl}()},
\code{\link{ListCalendars}()},
\code{\link{ListDeletedProjectsCount}()},
\code{\link{ListDeletedProjects}()},
\code{\link{ListFeatureInfo}()},
\code{\link{ListFeaturelists}()},
\code{\link{ListJobs}()},
\code{\link{ListModelingFeaturelists}()},
\code{\link{ListProjectCleanupJobsDownload}()},
\code{\link{ListProjectCleanupJobsSummary}()},
\code{\link{ListProjectCleanupJobs}()},
\code{\link{ListProjectsAccessControl}()},
\code{\link{ListProjectsCalendarEvents}()},
\code{\link{ListProjectsFeatureDiscoveryDatasetDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogsDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogs}()},
\code{\link{ListProjectsFeatureDiscoveryRecipeSQLsDownload}()},
\code{\link{ListProjectsModelingFeatures}()},
\code{\link{ListProjectsMultiseriesIdsCrossSeriesProperties}()},
\code{\link{ListProjectsMultiseriesNames}()},
\code{\link{ListProjectsRelationshipsConfiguration}()},
\code{\link{ListProjectsSecondaryDatasetsConfigurations}()},
\code{\link{ListProjectsSegmentationTasksMappings}()},
\code{\link{ListProjectsSegmentationTasks}()},
\code{\link{ListProjects}()},
\code{\link{PatchDeletedProjects}()},
\code{\link{PatchManyCalendarsAccessControl}()},
\code{\link{PatchManyProjectsAccessControl}()},
\code{\link{PatchProjectsSegments}()},
\code{\link{PauseQueue}()},
\code{\link{PutRelationshipsConfigurations}()},
\code{\link{RequestCrossSeriesDetection}()},
\code{\link{RequestMultiSeriesDetection}()},
\code{\link{RestoreDiscardedFeatures}()},
\code{\link{RetrieveDiscardedFeaturesInformation}()},
\code{\link{RetrieveProjectCleanupJobs}()},
\code{\link{RetrieveProjectsBatchTypeTransformFeaturesResult}()},
\code{\link{RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{RetrieveProjectsFeatureLineages}()},
\code{\link{RetrieveProjectsModelingFeatures}()},
\code{\link{RetrieveProjectsSecondaryDatasetsConfigurations}()},
\code{\link{RetrieveProjectsSegmentationTaskJobResults}()},
\code{\link{RetrieveProjectsSegmentationTasks}()},
\code{\link{RetrieveRelationshipsConfigurations}()},
\code{\link{SetTarget}()},
\code{\link{SetupProject}()},
\code{\link{StartNewAutoPilot}()},
\code{\link{UpdateCalendar}()},
\code{\link{UpdateProject}()}
}
\concept{Projects}
