% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{GetDataStoreTables}
\alias{GetDataStoreTables}
\title{GetDataStoreTables}
\usage{
GetDataStoreTables(dataStoreId, username, password, schema = NULL, ...)
}
\description{
Retrieves a data store's database tables (including views).
}
\details{
Retrieves a data store's database tables (including views).
}
\seealso{
The method ExternalDataStoresTablesCreate in [datarobot.apicore::DataConnectivityApi], which this function wraps.

Other DataConnectivity: 
\code{\link{CreateDataSource}()},
\code{\link{CreateDataStore}()},
\code{\link{CreateExternalDataDriverFile}()},
\code{\link{CreateExternalDataDrivers}()},
\code{\link{CreateExternalDataStoresColumns}()},
\code{\link{CreateExternalDataStoresVerifySQL}()},
\code{\link{DeleteDataSource}()},
\code{\link{DeleteDataStore}()},
\code{\link{DeleteExternalDataDrivers}()},
\code{\link{GetDataSource}()},
\code{\link{GetDataStoreSchemas}()},
\code{\link{GetDataStore}()},
\code{\link{GetDriver}()},
\code{\link{ListDataSources}()},
\code{\link{ListDataStores}()},
\code{\link{ListDrivers}()},
\code{\link{ListExternalDataDriversConfiguration}()},
\code{\link{ListExternalDataSourcesAccessControl}()},
\code{\link{ListExternalDataSourcesPermissions}()},
\code{\link{ListExternalDataSourcesSharedRoles}()},
\code{\link{ListExternalDataStoresCredentials}()},
\code{\link{ListExternalDataStoresPermissions}()},
\code{\link{ListExternalDataStoresSharedRoles}()},
\code{\link{ListExternalDriverConfigurations}()},
\code{\link{PatchExternalDataDrivers}()},
\code{\link{PatchManyExternalDataSourcesAccessControl}()},
\code{\link{PatchManyExternalDataSourcesSharedRoles}()},
\code{\link{PatchManyExternalDataStoresAccessControl}()},
\code{\link{PatchManyExternalDataStoresSharedRoles}()},
\code{\link{RetrieveExternalDriverConfigurations}()},
\code{\link{TestDataStore}()},
\code{\link{UpdateDataSource}()},
\code{\link{UpdateDataStore}()}
}
\concept{DataConnectivity}
