% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{GetDatetimeModelObject}
\alias{GetDatetimeModelObject}
\title{Get a datetime model object (deprecated).}
\usage{
GetDatetimeModelObject(projectId, modelId)
}
\arguments{
\item{projectId}{character. Unique alphanumeric identifier for the project of interest.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}
}
\description{
Deprecated. Use \link{GetDatetimeModel} instead.
}
