% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_partitioning_apiWrapper.R
\name{GetDatetimePartition}
\alias{GetDatetimePartition}
\title{GetDatetimePartition}
\usage{
GetDatetimePartition(project, ...)
}
\description{
Retrieve datetime partitioning configuration.
}
\details{
Retrieve datetime partitioning configuration  The datetime partition object in the response describes the full partitioning parameters. Since it becomes available after the target has been fully specified and the project is ready for modeling, there are some additional fields available compared to the response from :http:post:`/api/v2/projects/(projectId)/datetimePartitioning/`.  The available training data corresponds to all the data available for training, while the primary training data corresponds to the data that can be used to train while ensuring that all backtests are available. If a model is trained with more data than is available in the primary training data, then all backtests may not have scores available.  All durations and datetimes will be specified in accordance with the :ref:`timestamp and duration formatting rules<time_format>`.
}
\seealso{
The method ProjectsDatetimePartitioningList in [datarobot.apicore::DatetimePartitioningApi], which this function wraps.

Other DatetimePartitioning: 
\code{\link{CreateProjectsDatetimePartitioning}()},
\code{\link{CreateProjectsOptimizedDatetimePartitionings}()},
\code{\link{DownloadTimeSeriesFeatureDerivationLog}()},
\code{\link{GetTimeSeriesFeatureDerivationLog}()},
\code{\link{ListProjectsOptimizedDatetimePartitionings}()},
\code{\link{RetrieveProjectsOptimizedDatetimePartitionings}()}
}
\concept{DatetimePartitioning}
