% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R, R/Deployment.R
\name{GetDeploymentSettings}
\alias{GetDeploymentSettings}
\title{GetDeploymentSettings}
\usage{
GetDeploymentSettings(deployment, ...)
}
\arguments{
\item{deployment}{An S3 object representing a model deployment, or the unique ID of such a
deployment.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
List representing the various settings to be configured on a
deployment, including:
\describe{
\item{associationId}{object. Information on association ID for tracking deployment accuracy.
See \code{\link[=GetDeploymentAssociationId]{GetDeploymentAssociationId()}}}
\item{challengerModels}{logical. Whether challenger models are enabled.}
\item{featureDrift}{logical. Whether feature drift tracking is enabled. See
\code{\link[=GetDeploymentDriftTrackingSettings]{GetDeploymentDriftTrackingSettings()}}}
\item{humility}{logical. Whether humility rules are enabled.}
\item{predictionIntervals}{object. Information on prediction intervals.}
\item{predictionWarning}{object. Information on prediction warning settings.}
\item{predictionsByForecastDate}{object. Information on predictions by forecast date.}
\item{predictionsDataCollection}{logical. Whether predictions data is stored.}
\item{targetDrift}{logical. Whether target drift tracking is enabled.}
\item{segmentAnalysis}{object. Information on segment analysis settings.}
}
}
\description{
Get deployment settings
}
\details{
Retrieve deployment settings.

This method invokes \verb{GET /deployments/\{deploymentId\}/settings/} in the DataRobot Public API.

Retrieves all settings for a deployed model.
}
\seealso{
The method DeploymentsSettingsList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.

Other deployment configuration functions: 
\code{\link{UpdateDeploymentSettings}()}
}
\concept{deployment configuration functions}
\keyword{internal}
