% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/Drivers.R
\name{GetDriver}
\alias{GetDriver}
\title{GetDriver}
\usage{
GetDriver(driverId, ...)
}
\arguments{
\item{driverId}{character. The id of the driver.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list containing information on the particular driver:
\itemize{
  \item className character. The Java class name of the driver.
  \item baseNames character. A vector of the file name(s) of the jar files.
  \item canonicalName character. The user-friendly name of the driver.
  \item id character. The driverId of the driver.
  \item creator character. The userId of the user who created the driver.
}
}
\description{
Retrieve driver details.
}
\details{
Retrieve driver details by ID.

This method invokes \verb{GET /externalDataDrivers/\{driverId\}/} in the DataRobot Public API.

Returns information about a particular driver.
}
\examples{
\dontrun{
driverId <- "57a7c978c808916f4a630f89"
GetDriver(driverId)
}
}
\seealso{
The method ExternalDataDriversRetrieve in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
