% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{GetFeatureDiscoveryRelationships}
\alias{GetFeatureDiscoveryRelationships}
\title{GetFeatureDiscoveryRelationships}
\usage{
GetFeatureDiscoveryRelationships(project)
}
\arguments{
\item{project}{object. Either list with projectId element or projectId value}
}
\description{
Gets the feature discovery relationships for the specified Project.
}
\concept{Features}
