% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Features.R
\name{GetFeatureHistogram}
\alias{GetFeatureHistogram}
\title{GetFeatureHistogram}
\usage{
GetFeatureHistogram(project, featureName, binLimit = NULL, key = NULL, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{featureName}{Name of the feature to retrieve. Note: DataRobot renames some features, so
the feature name may not be the one from your original data. You can use ListFeatureInfo to list
the features and check the name.}

\item{binLimit}{integer. Optional. Desired max number of histogram bins. The default is 60.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
list containing:
  \itemize{
    \item count numeric. The number of values in this bin's range. If a project is using weights,
      the value is equal to the sum of weights of all feature values in the bin's range.
    \item target numeric. Average of the target feature for values in this bin. It may be NULL
      if the feature is deemed uninformative, if the target has not yet been set
      (see \code{SetTarget}), or if the project is multiclass.
    \item label character. The value of the feature if categorical, otherwise the low end of the
      bin range such that the difference between two consecutive bin labels is the length of the
      bin.
  }
}
\description{
Get feature histogram
}
\details{
Get histogram chart data for a specific feature. Information that can be used to build histogram charts. Plot data returned is based on raw data that is calculated during initial project creation and updated after the project's target variable has been selected. The number of bins in the histogram is no greater than the requested limit.

This method invokes \verb{GET /projects/\{projectId\}/featureHistograms/\{featureName\}/} in the DataRobot Public API.

Retrieve histogram plot data for a specific feature

A histogram is a popular way of visual representation of a feature values
distribution in a series of bins. For categorical features every bin represents
exactly one of feature values plus the number of occurrences of that value.
For numeric features every bin represents a range of values (low end inclusive,
high end exclusive) plus the total number of occurrences of all values in this range.
In addition to that, with every bin for categorical and numeric features there is also
included a target feature average for values in that bin (though it can be missing
if the feature is deemed uninformative, if the project target has not been selected
yet using \code{SetTarget}, or if the project is a multiclass project).
}
\seealso{
The method FeatureHistogramsRetrieve in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
