% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{GetFeatures}
\alias{GetFeatures}
\title{Deprecated (use ListModelFeatures instead)}
\usage{
GetFeatures(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModelObject, or each element of the list returned by
the function GetAllModels.}
}
\description{
Deprecated (use ListModelFeatures instead)
}

