% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatetimeTrendPlots.R
\name{GetForecastVsActualPlotsMetadata}
\alias{GetForecastVsActualPlotsMetadata}
\title{Retrieve Forecast vs Actual plots metadata for a model.}
\usage{
GetForecastVsActualPlotsMetadata(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModel, or each element of the list returned by
the function ListModels.}
}
\value{
list with the following components:
\itemize{
  \item resolutions. list: A list of \code{DatetimeTrendPlotsResolutions},
        which represents available time resolutions for which plots can be retrieved.
  \item backtestStatuses. data.frame: Each row represents a list of forecast distances
        for status for the backtest \code{SourceType}.
        The row index corresponds to the backtest index via the relation
        \code{rowIndex <- backtestIndex + 1}. Status should be one of
        \code{DatetimeTrendPlotsStatuses}. If there are no forecast distances for some status
        it would be missing from the dataframe.
  \item backtestMetadata. data.frame: Each row represents a metadata for the backtest
        \code{SourceType} start and end date. The row index corresponds to the
        backtest index via the relation \code{rowIndex <- backtestIndex + 1}.
        Each cell contains a POSIXct timestamp for start date (inclusive)
        and end date (exclusive) if the corresponding source type
        for the backtest is computed, and NA otherwise.
  \item holdoutStatuses. list: Contains statuses for holdout.
  \itemize{
    \item training. list: list with the statuses, one of \code{DatetimeTrendPlotsStatuses}, and
          corresponding forecast distances for each status.
    \item validation. list: list with the statuses, one of \code{DatetimeTrendPlotsStatuses}, and
          corresponding forecast distances for each status.
  }
  \item holdoutMetadata. list. Contains metadata for holdout.
  \itemize{
    \item training. list. Contains start and end date for holdout training.
    \item validation. list. Contains start and end date for holdout validation.
      \itemize{
        \item startDate. POSIXct or NA:
              The datetime of the start of the holdout training/validation (inclusive).
              NA if the data is not computed.
        \item endDate. POSIXct or NA:
              The datetime of the end of the holdout training/validation (exclusive).
              NA if the data is not computed.
        }
  }
}
}
\description{
Retrieve Forecast vs Actual plots metadata for a model.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
model <- GetModel(projectId, modelId)
GetForecastVsActualPlotsMetadata(model)
}
}
