% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insights.R
\name{GetGeneralizedInsight}
\alias{GetGeneralizedInsight}
\title{An internal function to help fetch insights.}
\usage{
GetGeneralizedInsight(method, model, source = DataPartition$VALIDATION,
  fallbackToParentInsights = FALSE)
}
\arguments{
\item{method}{character. The API URL to use to get insight information.}

\item{model}{dataRobotModel. A DataRobot model object like that returned by \code{GetModel}.}

\item{source}{character. The data partition for which data would be returned. Default is
\code{DataPartition$VALIDATION}. See \code{DataPartition} for details.}

\item{fallbackToParentInsights}{logical. If TRUE, this will return the lift chart data for the
model's parent if the lift chart is not available for the model and the model has a parent
model.}
}
\description{
See \code{GetLiftChart}, \code{GetRocCurve} for details.
}
