% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetMissingValuesReport}
\alias{GetMissingValuesReport}
\title{GetMissingValuesReport}
\usage{
GetMissingValuesReport(project, modelId, ...)
}
\description{
Retrieve a summary of how the model's subtasks handle missing values.
}
\details{
Retrieve a summary of how the model's subtasks handle missing values         Only models built after the missing value report feature was added will have reports,         and only models with at least one imputation or encoding task, e.g. ordinal encoding,         missing value imputation. Blenders and scaleout models do not support Missing Value reports.          The report will describe how each feature's missing values were treated, and report how many         missing values were present in the training data. Features which were not processed by a         given blueprint task will not mention it: for instance, a categorical feature with many         unique values may not be considered eligible for processing by a One-Hot Encoding          Report is collected for those features which are considered eligible by given         blueprint task. For instance, categorical feature with a lot of unique values may not be         considered as eligible in One-Hot Encoding task.
}
\seealso{
The method ProjectsModelsMissingReportList in [datarobot.apicore::ModelsApi], which this function wraps.

Other Models: 
\code{\link{AddEureqaSolution}()},
\code{\link{CreateCustomInferenceImagesFeatureImpact}()},
\code{\link{CreateCustomModelTests}()},
\code{\link{CreateCustomModelsFromCustomModel}()},
\code{\link{CreateCustomModelsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersionsConversions}()},
\code{\link{CreateCustomModelsVersionsDependencyBuild}()},
\code{\link{CreateCustomModelsVersionsFeatureImpact}()},
\code{\link{CreateCustomModelsVersionsFromRepository}()},
\code{\link{CreateCustomModelsVersionsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersions}()},
\code{\link{CreateCustomModels}()},
\code{\link{CreateCustomTrainingBlueprints}()},
\code{\link{CreateFromLatestCustomModelsVersion}()},
\code{\link{CreateModelExports}()},
\code{\link{CreateModelPackagesArchive}()},
\code{\link{CreateModelPackagesFromLearningModel}()},
\code{\link{CreatePrimeCode}()},
\code{\link{CreateProjectsBlenderBuilders}()},
\code{\link{CreateProjectsDatetimeModelsFromModel}()},
\code{\link{CreateProjectsDeploymentReadyModels}()},
\code{\link{CreateProjectsModelsFromModel}()},
\code{\link{CreateProjectsModelsPredictionIntervals}()},
\code{\link{CreateRatingTable}()},
\code{\link{DeleteCustomModelTests}()},
\code{\link{DeleteCustomModelsVersionsConversions}()},
\code{\link{DeleteCustomModels}()},
\code{\link{DeleteManyCustomModelsVersionsDependencyBuild}()},
\code{\link{DeleteModelJob}()},
\code{\link{DownloadPrimeCode}()},
\code{\link{DownloadRatingTable}()},
\code{\link{DownloadScoringCode}()},
\code{\link{DownloadTransferableModel}()},
\code{\link{GetBlenderModel}()},
\code{\link{GetDatetimeModel}()},
\code{\link{GetFrozenModel}()},
\code{\link{GetModelJob}()},
\code{\link{GetModelParameters}()},
\code{\link{GetModel}()},
\code{\link{GetPrimeEligibility}()},
\code{\link{GetPrimeFile}()},
\code{\link{GetPrimeModel}()},
\code{\link{GetRatingTableModel}()},
\code{\link{GetRatingTable}()},
\code{\link{GetRulesets}()},
\code{\link{IsBlenderEligible}()},
\code{\link{ListCustomInferenceImagesFeatureImpact}()},
\code{\link{ListCustomModelDeployments}()},
\code{\link{ListCustomModelLimits}()},
\code{\link{ListCustomModelTestsLog}()},
\code{\link{ListCustomModelTestsTail}()},
\code{\link{ListCustomModelTests}()},
\code{\link{ListCustomModelsAccessControl}()},
\code{\link{ListCustomModelsDownload}()},
\code{\link{ListCustomModelsVersionsConversions}()},
\code{\link{ListCustomModelsVersionsDependencyBuildLog}()},
\code{\link{ListCustomModelsVersionsDependencyBuild}()},
\code{\link{ListCustomModelsVersionsDownload}()},
\code{\link{ListCustomModelsVersionsFeatureImpact}()},
\code{\link{ListCustomModelsVersions}()},
\code{\link{ListCustomModels}()},
\code{\link{ListCustomTrainingBlueprints}()},
\code{\link{ListModelJobs}()},
\code{\link{ListModelPackagesFeatures}()},
\code{\link{ListModelPackagesSharedRoles}()},
\code{\link{ListModelPackages}()},
\code{\link{ListModelRecommendations}()},
\code{\link{ListModels}()},
\code{\link{ListPrimeFiles}()},
\code{\link{ListPrimeModels}()},
\code{\link{ListProjectsBlenderBuilders}()},
\code{\link{ListProjectsBlenderModels}()},
\code{\link{ListProjectsCombinedModelsSegmentsDownload}()},
\code{\link{ListProjectsCombinedModelsSegments}()},
\code{\link{ListProjectsCombinedModels}()},
\code{\link{ListProjectsDatetimeModels}()},
\code{\link{ListProjectsFrozenModels}()},
\code{\link{ListProjectsModelsClusterNames}()},
\code{\link{ListProjectsModelsNumIterationsTrained}()},
\code{\link{ListProjectsModelsPredictionIntervals}()},
\code{\link{ListProjectsRecommendedModelsRecommendedModel}()},
\code{\link{ListRatingTableModels}()},
\code{\link{ListRatingTables}()},
\code{\link{PatchCustomModelsVersions}()},
\code{\link{PatchCustomModels}()},
\code{\link{PatchManyCustomModelsAccessControl}()},
\code{\link{PatchManyCustomModelsTrainingData}()},
\code{\link{PatchManyCustomModelsVersionsFromRepository}()},
\code{\link{PatchManyProjectsModelsClusterNames}()},
\code{\link{PutManyProjectsSegmentChampion}()},
\code{\link{RenameRatingTable}()},
\code{\link{RequestApproximation}()},
\code{\link{RequestBlender}()},
\code{\link{RequestNewDatetimeModel}()},
\code{\link{RequestNewModel}()},
\code{\link{RequestNewRatingTableModel}()},
\code{\link{RequestPrimeModel}()},
\code{\link{RetrieveCustomModelTests}()},
\code{\link{RetrieveCustomModelsVersionsConversions}()},
\code{\link{RetrieveCustomModelsVersions}()},
\code{\link{RetrieveCustomModels}()},
\code{\link{RetrieveModelPackages}()},
\code{\link{RetrieveProjectsBlenderBuilders}()},
\code{\link{RetrieveProjectsCombinedModels}()},
\code{\link{RetrieveProjectsEureqaDistributionPlot}()},
\code{\link{RetrieveProjectsEureqaModelDetail}()}
}
\concept{Models}
