% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{GetModelBlueprintChart}
\alias{GetModelBlueprintChart}
\title{GetModelBlueprintChart}
\usage{
GetModelBlueprintChart(project, modelId, ...)
}
\description{
Retrieve a reduced model blueprint chart by model id.
}
\details{
Retrieve a reduced model blueprint chart by model id. The model blueprint charts are reduced from the full blueprint charts to show only those sections of the blueprint that were actually used in the model, given the selected featurelist.
}
\seealso{
The method ProjectsModelsBlueprintChartList in [datarobot.apicore::BlueprintsApi], which this function wraps.

Other Blueprints: 
\code{\link{CreateUserBlueprintsBulkValidations}()},
\code{\link{CreateUserBlueprintsFromBlueprintId}()},
\code{\link{CreateUserBlueprintsFromCustomTaskVersionId}()},
\code{\link{CreateUserBlueprintsFromUserBlueprintId}()},
\code{\link{CreateUserBlueprintsProjectBlueprints}()},
\code{\link{CreateUserBlueprintsTaskParameters}()},
\code{\link{CreateUserBlueprintsValidations}()},
\code{\link{CreateUserBlueprints}()},
\code{\link{DeleteManyUserBlueprints}()},
\code{\link{DeleteUserBlueprints}()},
\code{\link{GetBlueprintChart}()},
\code{\link{GetBlueprintDocumentation}()},
\code{\link{GetBlueprint}()},
\code{\link{GetModelBlueprintDocumentation}()},
\code{\link{ListBlueprints}()},
\code{\link{ListProjectsModelsLogs}()},
\code{\link{ListProjectsModelsTrainingArtifact}()},
\code{\link{ListUserBlueprintsInputTypes}()},
\code{\link{ListUserBlueprintsSharedRoles}()},
\code{\link{ListUserBlueprintsTasks}()},
\code{\link{ListUserBlueprints}()},
\code{\link{PatchManyUserBlueprintsSharedRoles}()},
\code{\link{PatchUserBlueprints}()},
\code{\link{RetrieveUserBlueprints}()}
}
\concept{Blueprints}
