% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R, R/Blueprints.R
\name{GetModelBlueprintDocumentation}
\alias{GetModelBlueprintDocumentation}
\title{GetModelBlueprintDocumentation}
\usage{
GetModelBlueprintDocumentation(project, modelId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
list with following components
\describe{
  \item{task}{Character string name of the task described in document}
  \item{description}{Character string task description}
  \item{title}{Character string title of document}
  \item{parameters}{List of parameters that task can received in human-readable
  format with following components: name, type, description}
  \item{links}{List of external links used in document with following components: name, url}
  \item{references}{List of references used in document with following components: name, url}
}
}
\description{
Retrieve task documentation for a reduced model blueprint.
}
\details{
Retrieve task documentation for a reduced model blueprint. The model blueprint is reduced from the full blueprint to show only those sections of the blueprint that were actually used in the model, given the selected featurelist.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/blueprintDocs/} in the DataRobot Public API.

Get documentation for tasks used in the model blueprint

A model blueprint is a "pruned down" blueprint representing what was actually run for the model.
This is solely the branches of the blueprint that were executed based on the featurelist.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
GetModelBlueprintDocumentation(projectId, modelId)
}
}
\seealso{
The method ProjectsModelsBlueprintDocsList in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.
}
