% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Featurelists.R
\name{GetModelingFeaturelist}
\alias{GetModelingFeaturelist}
\title{GetModelingFeaturelist}
\usage{
GetModelingFeaturelist(project, featurelistId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{featurelistId}{Unique alphanumeric identifier for the featurelist
to be retrieved.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list with the following elements describing the requested featurelist:
\itemize{
  \item featurelistId character. The unique alphanumeric identifier for the featurelist.
  \item projectId character. The project to which the featurelist belongs.
  \item features character. The names of the variables included in the featurelist.
  \item name character. The name of the featurelist.
  \item created character. A timestamp of when the featurelist was created.
  \item isUserCreated logical. Whether or not the featurelist was created by a user
    (as opposed to DataRobot automation).
  \item numModels numeric. The number of models that currently use this featurelist.
}
}
\description{
Retrieve a single modeling featurelist by ID
}
\details{
Retrieve a single modeling featurelist by ID. When reporting the number of models that \"use\" a featurelist, a model is considered to use a featurelist if it is used as to train the model or as a monotonic constraint featurelist, or if the model is a blender with component models that use the featurelist. This route will only become available after the target and partitioning options have been set for a project. Modeling featurelists are featurelists of modeling features, and are the appropriate featurelists to use when creating models or restarting the autopilot. In a time series project, these will be distinct from those returned from :\url{http:get}:/api/v2/projects/(projectId)/featurelists/ while in other projects these will be identical.

This method invokes \verb{GET /projects/\{projectId\}/modelingFeaturelists/\{featurelistId\}/} in the DataRobot Public API.

Retrieve a specific modeling featurelist from a DataRobot project

In time series projects, a new set of modeling features is created after setting the
partitioning options. These features are automatically derived from those in the project's
dataset and are the features used for modeling. Modeling features are only accessible once
the target and partitioning options have been set. In projects that don't use time series
modeling, once the target has been set, ModelingFeaturelists and Featurelists will behave
the same.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
featureList <- CreateModelingFeaturelist(projectId, "myFeaturelist", c("feature1", "feature2"))
featurelistId <- featureList$featurelistId
GetModelingFeaturelist(projectId, featurelistId)
}
}
\seealso{
The method ModelingFeaturelistsRetrieve in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
