% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListModelJobs.R
\name{GetModelJobs}
\alias{GetModelJobs}
\title{Retrieve status of Autopilot modeling jobs that are not complete (deprecated)}
\usage{
GetModelJobs(project, status = NULL)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{status}{character. The status of the desired jobs: one of JobStatus$Queue,
JobStatus$InProgress, orJobStatus$Error. If NULL (default), queued and inprogress jobs
are returned.}
}
\description{
Retrieve status of Autopilot modeling jobs that are not complete (deprecated)
}
\seealso{
ListModelJobs
}
