% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{GetModelObject}
\alias{GetModelObject}
\title{Retrieve the details of a specified model}
\usage{
GetModelObject(project, modelId)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}
}
\description{
(Deprecated in 2.8, will be removed in 2.10. Use GetModel instead.)
}
