% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetModelParameters}
\alias{GetModelParameters}
\title{GetModelParameters}
\usage{
GetModelParameters(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve model parameters.
}
\details{
Retrieve model parameters. These are the parameters that appear in the webapp on the Coefficients tab. Note that they are only available for some models.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/parameters/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsParametersList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
