% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetModelJobs.R
\name{GetPendingJobs}
\alias{GetPendingJobs}
\title{Retrieve status of Autopilot modeling jobs that are not complete}
\usage{
GetPendingJobs(project, jobStatus = NULL)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric identifier for the
project, or (2) a list containing the element projectId with this identifier.}

\item{jobStatus}{(deprecated - use status instead) The status of the desired jobs: one of
JobStatus$Queue, JobStatus$InProgress, or JobStatus$Error. If NULL (default), queued and
inprogress jobs are returned.}
}
\description{
(Deprecated in 2.1, will be removed in 2.3. Use GetModelJobs instead.)
}

