% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R, R/GetPredictJobs.R
\name{GetPredictJobs}
\alias{GetPredictJobs}
\title{GetPredictJobs}
\usage{
GetPredictJobs(project, status = NULL, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{status}{character. The status of the desired jobs: one of JobStatus$Queue,
JobStatus$InProgress, orJobStatus$Error. If NULL (default), queued and inprogress jobs
are returned.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
Dataframe with one row for each prediction job in the queue,
  with the following columns:
  \describe{
    \item{status}{Prediction job status; one of JobStatus$Queue, JobStatus$InProgress, or
  JobStatus$Error}
    \item{predictJobId}{Character string specifying the job id}
    \item{modelId}{Character string specifying the model from which
    predictions have been requested}
    \item{projectId}{Character string specifying the project that contains the model}
}
}
\description{
List all prediction jobs for a project
}
\details{
List all prediction jobs for a project

This method invokes \verb{GET /projects/\{projectId\}/predictJobs/} in the DataRobot Public API.

Function to list all prediction jobs in a project
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
GetPredictJobs(projectId)
}
}
\seealso{
The method ProjectsPredictJobsList in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
