% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/GetPrimeEligibility.R
\name{GetPrimeEligibility}
\alias{GetPrimeEligibility}
\title{GetPrimeEligibility}
\usage{
GetPrimeEligibility(project, modelId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
list with two members:
\itemize{
  \item canMakePrime logical. TRUE if model can be approximated using DataRobot Prime,
    FALSE if model can not be approximated.
  \item message character. Provides information why model may not be approximated with
    DataRobot Prime.
}
}
\description{
Check a Model for Prime Eligibility
}
\details{
Check if a model can be approximated by DataRobot Prime

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/primeInfo/} in the DataRobot Public API.

Check if model can be approximated with DataRobot Prime
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
GetPrimeEligibility(projectId, modelId)
}
}
\seealso{
The method ProjectsModelsPrimeInfoList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
