% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/PrimeModels.R
\name{GetPrimeModel}
\alias{GetPrimeModel}
\title{GetPrimeModel}
\usage{
GetPrimeModel(project, modelId)

GetPrimeModel(project, modelId)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
list (classed as \code{dataRobotPrimeModel}) containing information
  about specified DataRobot Prime model.
}
\description{
Retrieve a Prime model details

This function requests the DataRobot Prime model information for the DataRobot
project specified by the project argument, and modelId.
}
\details{
Retrieve a Prime model details

This method invokes \verb{GET /projects/\{projectId\}/primeModels/\{modelId\}/} in the DataRobot Public API.

The function returns list containing information about specified DataRobot Prime model.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
GetPrimeModel(projectId, modelId)
}
}
\seealso{
The method ProjectsPrimeModelsRetrieve in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
