% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projects.R
\name{GetProjectList}
\alias{GetProjectList}
\title{Retrieve a list of all DataRobot projects (deprecated version)}
\usage{
GetProjectList()
}
\description{
Retrieve a list of all DataRobot projects (deprecated version)
}
\seealso{
ListProjects
}
