% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RatingTables.R
\name{GetRatingTable}
\alias{GetRatingTable}
\title{GetRatingTable}
\usage{
GetRatingTable(project, ratingTableId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{ratingTableId}{character. The ID of the rating table.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An S3 object of class 'dataRobotRatingTable' summarizing all
  available information about the rating table.
}
\description{
Retrieve Rating Table Information
}
\details{
Retrieves a rating table

This method invokes \verb{GET /projects/\{projectId\}/ratingTables/\{ratingTableId\}/} in the DataRobot Public API.

Retrieve a single rating table.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
ratingTableId <- "5984b4d7100d2b31c1166529"
GetRatingTable(projectId, ratingTableId)
}
}
\seealso{
The method ProjectsRatingTablesRetrieve in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
