% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RatingTables.R
\name{GetRatingTableModel}
\alias{GetRatingTableModel}
\title{GetRatingTableModel}
\usage{
GetRatingTableModel(project, modelId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
list containing information about specified model with a rating table.
}
\description{
Retrieve Rating Table Model
}
\details{
Look up a particular rating table model

This method invokes \verb{GET /projects/\{projectId\}/ratingTableModels/\{modelId\}/} in the DataRobot Public API.

Retrieve information about specified model with a rating table.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
modelId <- "5984b4d7100d2b31c1166529"
GetRatingTableModel(projectId, modelId)
}
}
\seealso{
The method ProjectsRatingTableModelsRetrieve in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
