% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jobs.R
\name{GetRedirectFromResponse}
\alias{GetRedirectFromResponse}
\title{Helper method to retrieve the redirect from the response headers to any
request that puts or queries for a DR job. Use this in your job-related methods
to simplify testing.}
\usage{
GetRedirectFromResponse(rawResponse)
}
\description{
Helper method to retrieve the redirect from the response headers to any
request that puts or queries for a DR job. Use this in your job-related methods
to simplify testing.
}
\keyword{internal}
