% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/GetRulesets.R
\name{GetRulesets}
\alias{GetRulesets}
\title{GetRulesets}
\usage{
GetRulesets(project, modelId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{Unique alphanumeric identifier for the model of interest.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list of lists with one element for each ruleset. If there are no rulesets created for
  a model then an empty list is returned
If the group is not empty,a list is returned with the following elements:
\itemize{
  \item projectId. Character string giving the unique identifier for the project.
  \item rulesetId. Integer number giving the identifier for the ruleset.
  \item score. Score of ruleset (using project leaderboard metric).
  \item parentModelId. Character string giving the unique identifier for the parent model.
  \item ruleCount. integer: number of rules in ruleset.
  \item modelId. Character string giving the unique identifier for a model using the ruleset.
    May be NULL if no model using the ruleset has been created yet.
}
}
\description{
List Rulesets
}
\details{
List all the rulesets approximating a model  When rulesets are created for the parent model, all of the rulesets are created at once, but not all rulesets have corresponding Prime models (until they are directly requested).

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/primeRulesets/} in the DataRobot Public API.

List the rulesets approximating a model generated by DataRobot Prime

This function will return list of rulesets that could be used to approximate the specified model.
Rulesets are created using the RequestApproximation function. If model hasn't been approximated
yet, will return empty list
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
GetRulesets(projectId, modelId)
}
}
\seealso{
The method ProjectsModelsPrimeRulesetsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
