% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServerData.R
\name{GetServerDataInRows}
\alias{GetServerDataInRows}
\title{Handle server side pagination.}
\usage{
GetServerDataInRows(serverData, batchSize = 50)
}
\arguments{
\item{serverData}{list. Raw JSON parsed list returned from the server.}

\item{batchSize}{integer. The number of requests per page to expect.}
}
\description{
Handle server side pagination.
}
