% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_apiWrapper.R
\name{GetServerVersion}
\alias{GetServerVersion}
\title{GetServerVersion}
\usage{
GetServerVersion(...)
}
\description{
Retrieve version information.
}
\details{
Provides the version information for the current version of the API.
}
\seealso{
The method VersionList in [datarobot.apicore::InfrastructureApi], which this function wraps.

Other Infrastructure: 
\code{\link{CreateClusterLicenseValidation}()},
\code{\link{ListClusterLicense}()},
\code{\link{PutManyClusterLicense}()}
}
\concept{Infrastructure}
