% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainingPredictions.R
\name{GetTrainingPredictionDataFrame}
\alias{GetTrainingPredictionDataFrame}
\title{Simplify the training prediction rows into a tidy format dataframe.}
\usage{
GetTrainingPredictionDataFrame(rows)
}
\arguments{
\item{rows}{data.frame. The dataframe to tidy.}
}
\description{
Simplify the training prediction rows into a tidy format dataframe.
}
