% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/GetValidMetrics.R
\name{GetValidMetrics}
\alias{GetValidMetrics}
\title{GetValidMetrics}
\usage{
GetValidMetrics(project, target)

GetValidMetrics(project, target)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{target}{character. String giving the name of the response variable
to be predicted by all project models.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
Character vector containing the names of the metric values
that are valid for a subsequent call to the SetTarget function.
}
\description{
List feature metrics

For the response variable defined by the character string target
and the project defined by the parameter project, return the vector
of metric names that can be specified for fitting models in this project.
This function is intended for use after SetupProject has been run but
before SetTarget, allowing the user to specify valid non-default
values for the metric parameter.
}
\details{
List the appropriate metrics if a feature were chosen as the target. The metrics listed will include both weighted and unweighted metrics - which are appropriate will depend on whether a weights column is used.

This method invokes \verb{GET /projects/\{projectId\}/features/metrics/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
GetValidMetrics(projectId, "targetFeature")
}
}
\seealso{
The method FeaturesMetricsList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
