% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{IsId}
\alias{IsId}
\title{Checks if an id is a valid DataRobot ID (24 character string)}
\usage{
IsId(id)
}
\arguments{
\item{id}{character. An ID to test whether it is a valid DataRobot ID.}
}
\description{
Checks if an id is a valid DataRobot ID (24 character string)
}
