% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{JobStatus}
\alias{JobStatus}
\title{Job statuses}
\format{An object of class \code{list} of length 5.}
\usage{
JobStatus
}
\description{
This is a list that contains the valid values for job status when querying the list of jobs mode.
If you wish, you can specify job status modes using the list values, e.g. JobStatus$InProgress
instead of typing the string 'inprogress'. This way you can benefit from autocomplete and not have to
remember the valid optons.
}
\keyword{datasets}

