% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{JobType}
\alias{JobType}
\title{Job type}
\format{
An object of class \code{list} of length 10.
}
\usage{
JobType
}
\description{
This is a list that contains the valid values for job type when querying the list of jobs.
}
\keyword{datasets}
