% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{ListApplicationTypes}
\alias{ListApplicationTypes}
\title{ListApplicationTypes}
\usage{
ListApplicationTypes(offset = 0, limit = 0, ...)
}
\description{
List of available application types to deploy.
}
\details{
List of available application types to deploy.
}
\seealso{
The method ApplicationTypesList in [datarobot.apicore::ApplicationsApi], which this function wraps.

Other Applications: 
\code{\link{CreateApplicationsBuildJobs}()},
\code{\link{CreateApplicationsDeployments}()},
\code{\link{CreateApplicationsDuplicate}()},
\code{\link{CreateApplications}()},
\code{\link{CreateDemoApplications}()},
\code{\link{DeleteApplicationsDeployments}()},
\code{\link{DeleteApplications}()},
\code{\link{ListApplicationTypesUsecases}()},
\code{\link{ListApplicationsAccessControl}()},
\code{\link{ListApplicationsSharedRoles}()},
\code{\link{ListApplications}()},
\code{\link{PatchApplications}()},
\code{\link{PatchManyApplicationsAccessControl}()},
\code{\link{PatchManyApplicationsAuthToken}()},
\code{\link{PatchManyApplicationsDeploymentState}()},
\code{\link{PatchManyApplicationsSharedRoles}()},
\code{\link{PatchManyApplicationsVersion}()},
\code{\link{RetrieveApplicationTypes}()},
\code{\link{RetrieveApplicationUserRole}()},
\code{\link{RetrieveApplications}()}
}
\concept{Applications}
