% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{ListApplicationTypesUsecases}
\alias{ListApplicationTypesUsecases}
\title{ListApplicationTypesUsecases}
\usage{
ListApplicationTypesUsecases(applicationTypeId, offset = 0, limit = 100, ...)
}
\description{
List application type usecases
}
\details{
List of available usecases for application type.
}
\seealso{
The method ApplicationTypesUsecasesList in [datarobot.apicore::ApplicationsApi], which this function wraps.

Other Applications: 
\code{\link{CreateApplicationsBuildJobs}()},
\code{\link{CreateApplicationsDeployments}()},
\code{\link{CreateApplicationsDuplicate}()},
\code{\link{CreateApplications}()},
\code{\link{CreateDemoApplications}()},
\code{\link{DeleteApplicationsDeployments}()},
\code{\link{DeleteApplications}()},
\code{\link{ListApplicationTypes}()},
\code{\link{ListApplicationsAccessControl}()},
\code{\link{ListApplicationsSharedRoles}()},
\code{\link{ListApplications}()},
\code{\link{PatchApplications}()},
\code{\link{PatchManyApplicationsAccessControl}()},
\code{\link{PatchManyApplicationsAuthToken}()},
\code{\link{PatchManyApplicationsDeploymentState}()},
\code{\link{PatchManyApplicationsSharedRoles}()},
\code{\link{PatchManyApplicationsVersion}()},
\code{\link{RetrieveApplicationTypes}()},
\code{\link{RetrieveApplicationUserRole}()},
\code{\link{RetrieveApplications}()}
}
\concept{Applications}
