% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListApprovalPolicies}
\alias{ListApprovalPolicies}
\title{ListApprovalPolicies}
\usage{
ListApprovalPolicies(entityType, namePart, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List Approval Policies.
}
\details{
List Approval Policies.

This method invokes \verb{GET /approvalPolicies/} in the DataRobot Public API.
}
\seealso{
The method ApprovalPoliciesList in \link[datarobot.apicore:GovernanceApi]{datarobot.apicore::GovernanceApi}, which invokes the same underlying API endpoint.
}
