% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListApprovalPoliciesShareableChangeRequests}
\alias{ListApprovalPoliciesShareableChangeRequests}
\title{ListApprovalPoliciesShareableChangeRequests}
\usage{
ListApprovalPoliciesShareableChangeRequests(
  approvalPolicyId,
  offset = 0,
  limit = 100,
  orderBy = "",
  ...
)
}
\description{
Retrieve associated Change Requests Info.
}
\details{
Get information about Change Requests submitted for a certain Approval Policy.
}
\seealso{
The method ApprovalPoliciesShareableChangeRequestsList in [datarobot.apicore::GovernanceApi], which this function wraps.

Other Governance: 
\code{\link{CreateApprovalPolicies}()},
\code{\link{CreateChangeRequestsReviews}()},
\code{\link{CreateChangeRequests}()},
\code{\link{DeleteApprovalPolicies}()},
\code{\link{ListApprovalPolicies}()},
\code{\link{ListApprovalPolicyMatch}()},
\code{\link{ListApprovalPolicyTriggers}()},
\code{\link{ListChangeRequestsReviews}()},
\code{\link{ListChangeRequestsSuggestedReviewers}()},
\code{\link{ListChangeRequests}()},
\code{\link{PatchChangeRequests}()},
\code{\link{PatchManyChangeRequestsStatus}()},
\code{\link{PutApprovalPolicies}()},
\code{\link{RetrieveApprovalPolicies}()},
\code{\link{RetrieveChangeRequestsReviews}()},
\code{\link{RetrieveChangeRequests}()}
}
\concept{Governance}
