% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListApprovalPolicyTriggers}
\alias{ListApprovalPolicyTriggers}
\title{ListApprovalPolicyTriggers}
\usage{
ListApprovalPolicyTriggers(...)
}
\description{
Get a list of available policy triggers.
}
\details{
Get a list of available policy triggers.
}
\seealso{
The method ApprovalPolicyTriggersList in [datarobot.apicore::GovernanceApi], which this function wraps.

Other Governance: 
\code{\link{CreateApprovalPolicies}()},
\code{\link{CreateChangeRequestsReviews}()},
\code{\link{CreateChangeRequests}()},
\code{\link{DeleteApprovalPolicies}()},
\code{\link{ListApprovalPoliciesShareableChangeRequests}()},
\code{\link{ListApprovalPolicies}()},
\code{\link{ListApprovalPolicyMatch}()},
\code{\link{ListChangeRequestsReviews}()},
\code{\link{ListChangeRequestsSuggestedReviewers}()},
\code{\link{ListChangeRequests}()},
\code{\link{PatchChangeRequests}()},
\code{\link{PatchManyChangeRequestsStatus}()},
\code{\link{PutApprovalPolicies}()},
\code{\link{RetrieveApprovalPolicies}()},
\code{\link{RetrieveChangeRequestsReviews}()},
\code{\link{RetrieveChangeRequests}()}
}
\concept{Governance}
