% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{ListAutomatedDocumentOptions}
\alias{ListAutomatedDocumentOptions}
\title{ListAutomatedDocumentOptions}
\usage{
ListAutomatedDocumentOptions(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all available document types and locales.
}
\details{
Check which document types and locales are available for generation with your account.

This method invokes \verb{GET /automatedDocumentOptions/} in the DataRobot Public API.
}
\seealso{
The method AutomatedDocumentOptionsList in \link[datarobot.apicore:DocumentationApi]{datarobot.apicore::DocumentationApi}, which invokes the same underlying API endpoint.
}
