% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListBatchPredictionsDownload}
\alias{ListBatchPredictionsDownload}
\title{ListBatchPredictionsDownload}
\usage{
ListBatchPredictionsDownload(predictionJobId, partNumber = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download the scored data set of a batch prediction job
}
\details{
This is only valid for jobs scored using the \"localFile\" output option

This method invokes \verb{GET /batchPredictions/\{predictionJobId\}/download/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsDownloadList in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
