% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Calendars.R
\name{ListCalendars}
\alias{ListCalendars}
\title{ListCalendars}
\usage{
ListCalendars(offset = 0, limit = 0, projectId = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list of S3 objects of class "dataRobotCalendar"
}
\description{
List all available calendars for a user.
}
\details{
List all the calendars which the user has access to.

This method invokes \verb{GET /calendars/} in the DataRobot Public API.

List all available calendars.
}
\examples{
\dontrun{
ListCalendars()
}
}
\seealso{
The method CalendarsList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
