% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListCatalogItems}
\alias{ListCatalogItems}
\title{ListCatalogItems}
\usage{
ListCatalogItems(
  orderBy,
  type,
  datasourceType,
  ownerUsername,
  ownerUserId,
  tag,
  category,
  searchFor,
  offset = 0,
  accessType = "any",
  useCache = "false",
  limit = 0,
  initialCacheSize = 500,
  filterFailed = "false",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all catalog items accessible by the user.
}
\details{
List all catalog items accessible by the user.

This method invokes \verb{GET /catalogItems/} in the DataRobot Public API.
}
\seealso{
The method CatalogItemsList in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
