% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListChangeRequests}
\alias{ListChangeRequests}
\title{ListChangeRequests}
\usage{
ListChangeRequests(
  entityType,
  myRequests,
  orderBy,
  showApproved,
  showCancelled,
  entityId,
  status,
  offset = 0,
  limit = 100,
  ...
)
}
\description{
List Change Requests.
}
\details{
List all Change Requests accessible by the user for the given product entity type.
}
\seealso{
The method ChangeRequestsList in [datarobot.apicore::GovernanceApi], which this function wraps.

Other Governance: 
\code{\link{CreateApprovalPolicies}()},
\code{\link{CreateChangeRequestsReviews}()},
\code{\link{CreateChangeRequests}()},
\code{\link{DeleteApprovalPolicies}()},
\code{\link{ListApprovalPoliciesShareableChangeRequests}()},
\code{\link{ListApprovalPolicies}()},
\code{\link{ListApprovalPolicyMatch}()},
\code{\link{ListApprovalPolicyTriggers}()},
\code{\link{ListChangeRequestsReviews}()},
\code{\link{ListChangeRequestsSuggestedReviewers}()},
\code{\link{PatchChangeRequests}()},
\code{\link{PatchManyChangeRequestsStatus}()},
\code{\link{PutApprovalPolicies}()},
\code{\link{RetrieveApprovalPolicies}()},
\code{\link{RetrieveChangeRequestsReviews}()},
\code{\link{RetrieveChangeRequests}()}
}
\concept{Governance}
