% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListChangeRequestsReviews}
\alias{ListChangeRequestsReviews}
\title{ListChangeRequestsReviews}
\usage{
ListChangeRequestsReviews(
  changeRequestId,
  changeVersionId,
  status,
  offset = 0,
  limit = 100,
  ...
)
}
\description{
List Change Request reviews.
}
\details{
List Change Request reviews.
}
\seealso{
The method ChangeRequestsReviewsList in [datarobot.apicore::GovernanceApi], which this function wraps.

Other Governance: 
\code{\link{CreateApprovalPolicies}()},
\code{\link{CreateChangeRequestsReviews}()},
\code{\link{CreateChangeRequests}()},
\code{\link{DeleteApprovalPolicies}()},
\code{\link{ListApprovalPoliciesShareableChangeRequests}()},
\code{\link{ListApprovalPolicies}()},
\code{\link{ListApprovalPolicyMatch}()},
\code{\link{ListApprovalPolicyTriggers}()},
\code{\link{ListChangeRequestsSuggestedReviewers}()},
\code{\link{ListChangeRequests}()},
\code{\link{PatchChangeRequests}()},
\code{\link{PatchManyChangeRequestsStatus}()},
\code{\link{PutApprovalPolicies}()},
\code{\link{RetrieveApprovalPolicies}()},
\code{\link{RetrieveChangeRequestsReviews}()},
\code{\link{RetrieveChangeRequests}()}
}
\concept{Governance}
