% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_apiWrapper.R
\name{ListClusterLicense}
\alias{ListClusterLicense}
\title{ListClusterLicense}
\usage{
ListClusterLicense(...)
}
\description{
Retrieve the information about cluster license
}
\details{
Retrieve the information about the currently deployed cluster license.
}
\seealso{
The method ClusterLicenseList in [datarobot.apicore::InfrastructureApi], which this function wraps.

Other Infrastructure: 
\code{\link{CreateClusterLicenseValidation}()},
\code{\link{GetServerVersion}()},
\code{\link{PutManyClusterLicense}()}
}
\concept{Infrastructure}
