% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{ListComments}
\alias{ListComments}
\title{ListComments}
\usage{
ListComments(entityType, orderBy, entityId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List comments
}
\details{
List comments

This method invokes \verb{GET /comments/\{entityType\}/\{entityId\}/} in the DataRobot Public API.
}
\seealso{
The method CommentsList in \link[datarobot.apicore:CommentsApi]{datarobot.apicore::CommentsApi}, which invokes the same underlying API endpoint.
}
