% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{ListComplianceDocTemplatesSharedRoles}
\alias{ListComplianceDocTemplatesSharedRoles}
\title{ListComplianceDocTemplatesSharedRoles}
\usage{
ListComplianceDocTemplatesSharedRoles(
  shareRecipientType,
  name,
  id,
  templateId,
  offset = 0,
  limit = 10,
  ...
)
}
\description{
Get template's access control list
}
\details{
Get a list of users, groups and organizations who have access to this template and their roles on the template.
}
\seealso{
The method ComplianceDocTemplatesSharedRolesList in [datarobot.apicore::DocumentationApi], which this function wraps.

Other Documentation: 
\code{\link{CreateAutomatedDocuments}()},
\code{\link{CreateOneModelComplianceDocsInitializations}()},
\code{\link{DeleteAutomatedDocuments}()},
\code{\link{DeleteComplianceDocTemplate}()},
\code{\link{GetComplianceDocTemplate}()},
\code{\link{ListAutomatedDocumentOptions}()},
\code{\link{ListAutomatedDocuments}()},
\code{\link{ListComplianceDocTemplatesDefault}()},
\code{\link{ListComplianceDocTemplates}()},
\code{\link{PatchManyComplianceDocTemplatesSharedRoles}()},
\code{\link{RetrieveAutomatedDocuments}()},
\code{\link{RetrieveModelComplianceDocsInitializations}()},
\code{\link{UpdateComplianceDocTemplate}()},
\code{\link{UploadComplianceDocTemplate}()}
}
\concept{Documentation}
