% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomInferenceImagesFeatureImpact}
\alias{ListCustomInferenceImagesFeatureImpact}
\title{ListCustomInferenceImagesFeatureImpact}
\usage{
ListCustomInferenceImagesFeatureImpact(imageId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get custom model feature impact.
}
\details{
Retrieve feature impact scores for features in a custom inference model image.  v2.23 DEPRECATED: please use version route instead: :\url{http:get}:/api/v2/customModels/(customModelId)/versions/(customModelVersionId)/featureImpact/  This route is a counterpart of a corresponding endpoint for native models: :\url{http:get}:/api/v2/projects/(projectId)/models/(modelId)/featureImpact/

This method invokes \verb{GET /customInferenceImages/\{imageId\}/featureImpact/} in the DataRobot Public API.
}
\seealso{
The method CustomInferenceImagesFeatureImpactList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
