% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelDeployments}
\alias{ListCustomModelDeployments}
\title{ListCustomModelDeployments}
\usage{
ListCustomModelDeployments(
  customModelIds,
  environmentIds,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List custom model deployments.
}
\details{
List of model deployments for user sorted by creation time descending.

This method invokes \verb{GET /customModelDeployments/} in the DataRobot Public API.
}
\seealso{
The method CustomModelDeploymentsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
