% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelLimits}
\alias{ListCustomModelLimits}
\title{ListCustomModelLimits}
\usage{
ListCustomModelLimits(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get custom model resource limits.
}
\details{
Retrieve custom model resource limits the user has access to.

This method invokes \verb{GET /customModelLimits/} in the DataRobot Public API.
}
\seealso{
The method CustomModelLimitsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
