% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelsDownload}
\alias{ListCustomModelsDownload}
\title{ListCustomModelsDownload}
\usage{
ListCustomModelsDownload(customModelId, pps = "False", ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download the latest custom model version content.
}
\details{
Download the latest item bundle from a custom model as a zip compressed archive.

This method invokes \verb{GET /customModels/\{customModelId\}/download/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsDownloadList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
