% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelsVersionsDependencyBuild}
\alias{ListCustomModelsVersionsDependencyBuild}
\title{ListCustomModelsVersionsDependencyBuild}
\usage{
ListCustomModelsVersionsDependencyBuild(
  customModelId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the custom model version's dependency build status.
}
\details{
Retrieve the custom model version's dependency build status.

This method invokes \verb{GET /customModels/\{customModelId\}/versions/\{customModelVersionId\}/dependencyBuild/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsDependencyBuildList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
