% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelsVersionsDownload}
\alias{ListCustomModelsVersionsDownload}
\title{ListCustomModelsVersionsDownload}
\usage{
ListCustomModelsVersionsDownload(
  customModelId,
  customModelVersionId,
  pps = "False",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download custom model version content.
}
\details{
Download a specific item bundle from a custom model as a zip compressed archive.

This method invokes \verb{GET /customModels/\{customModelId\}/versions/\{customModelVersionId\}/download/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsDownloadList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
