% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelTests}
\alias{ListCustomModelTests}
\title{ListCustomModelTests}
\usage{
ListCustomModelTests(
  networkEgressPolicy,
  requiresHa,
  customModelId,
  replicas,
  maximumMemory,
  desiredMemory,
  offset = 0,
  limit = 1000,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List custom model tests.
}
\details{
Retrieve the testing history for a model.

This method invokes \verb{GET /customModelTests/} in the DataRobot Public API.
}
\seealso{
The method CustomModelTestsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
