% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{ListCustomTasksAccessControl}
\alias{ListCustomTasksAccessControl}
\title{ListCustomTasksAccessControl}
\usage{
ListCustomTasksAccessControl(customTaskId, offset = 0, limit = 1000, ...)
}
\description{
Get a list of users who have access to this custom task and their roles on it.
}
\details{
Get a list of users who have access to this custom task and their roles on it.
}
\seealso{
The method CustomTasksAccessControlList in [datarobot.apicore::CustomTasksApi], which this function wraps.

Other CustomTasks: 
\code{\link{CreateCustomTasksFromCustomTask}()},
\code{\link{CreateCustomTasksVersionsDependencyBuild}()},
\code{\link{CreateCustomTasksVersionsFromRepository}()},
\code{\link{CreateCustomTasksVersions}()},
\code{\link{CreateCustomTasks}()},
\code{\link{CreateFromLatestCustomTaskVersion}()},
\code{\link{DeleteCustomTasks}()},
\code{\link{DeleteManyCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksDownload}()},
\code{\link{ListCustomTasksVersionsDependencyBuildLog}()},
\code{\link{ListCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksVersionsDownload}()},
\code{\link{ListCustomTasksVersions}()},
\code{\link{ListCustomTasks}()},
\code{\link{PatchCustomTasksVersions}()},
\code{\link{PatchCustomTasks}()},
\code{\link{PatchManyCustomTasksAccessControl}()},
\code{\link{PatchManyCustomTasksVersionsFromRepository}()},
\code{\link{RetrieveCustomTasksVersions}()},
\code{\link{RetrieveCustomTasks}()}
}
\concept{CustomTasks}
