% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{ListCustomTasksDownload}
\alias{ListCustomTasksDownload}
\title{ListCustomTasksDownload}
\usage{
ListCustomTasksDownload(customTaskId, ...)
}
\description{
Download the latest custom task version content.
}
\details{
Download the latest item bundle from a custom task as a zip compressed archive.
}
\seealso{
The method CustomTasksDownloadList in [datarobot.apicore::CustomTasksApi], which this function wraps.

Other CustomTasks: 
\code{\link{CreateCustomTasksFromCustomTask}()},
\code{\link{CreateCustomTasksVersionsDependencyBuild}()},
\code{\link{CreateCustomTasksVersionsFromRepository}()},
\code{\link{CreateCustomTasksVersions}()},
\code{\link{CreateCustomTasks}()},
\code{\link{CreateFromLatestCustomTaskVersion}()},
\code{\link{DeleteCustomTasks}()},
\code{\link{DeleteManyCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksAccessControl}()},
\code{\link{ListCustomTasksVersionsDependencyBuildLog}()},
\code{\link{ListCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksVersionsDownload}()},
\code{\link{ListCustomTasksVersions}()},
\code{\link{ListCustomTasks}()},
\code{\link{PatchCustomTasksVersions}()},
\code{\link{PatchCustomTasks}()},
\code{\link{PatchManyCustomTasksAccessControl}()},
\code{\link{PatchManyCustomTasksVersionsFromRepository}()},
\code{\link{RetrieveCustomTasksVersions}()},
\code{\link{RetrieveCustomTasks}()}
}
\concept{CustomTasks}
