% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{ListCustomTasksVersionsDependencyBuild}
\alias{ListCustomTasksVersionsDependencyBuild}
\title{ListCustomTasksVersionsDependencyBuild}
\usage{
ListCustomTasksVersionsDependencyBuild(customTaskVersionId, customTaskId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the custom task version's dependency build status.
}
\details{
Retrieve the custom task version's dependency build status.

This method invokes \verb{GET /customTasks/\{customTaskId\}/versions/\{customTaskVersionId\}/dependencyBuild/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsDependencyBuildList in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
