% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsAccessControl}
\alias{ListDatasetsAccessControl}
\title{ListDatasetsAccessControl}
\usage{
ListDatasetsAccessControl(
  datasetId,
  userId,
  username,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List dataset access
}
\details{
List the users and their associated roles for the specified dataset.

This method invokes \verb{GET /datasets/\{datasetId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method DatasetsAccessControlList in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
