% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsFeatureTransforms}
\alias{ListDatasetsFeatureTransforms}
\title{ListDatasetsFeatureTransforms}
\usage{
ListDatasetsFeatureTransforms(datasetId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List dataset feature transforms
}
\details{
Retrieves the transforms of the dataset with given ID.

This method invokes \verb{GET /datasets/\{datasetId\}/featureTransforms/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFeatureTransformsList in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
