% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsPermissions}
\alias{ListDatasetsPermissions}
\title{ListDatasetsPermissions}
\usage{
ListDatasetsPermissions(datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Describe dataset permissions
}
\details{
Describe what permissions current user has for given dataset.

This method invokes \verb{GET /datasets/\{datasetId\}/permissions/} in the DataRobot Public API.
}
\seealso{
The method DatasetsPermissionsList in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
