% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsRefreshJobsExecutionResults}
\alias{ListDatasetsRefreshJobsExecutionResults}
\title{ListDatasetsRefreshJobsExecutionResults}
\usage{
ListDatasetsRefreshJobsExecutionResults(
  jobId,
  datasetId,
  offset = 0,
  limit = 20,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Results of dataset refresh job.
}
\details{
Paginated list of execution results for refresh job with the given ID and dataset with the given ID, sorted from newest to oldest.

This method invokes \verb{GET /datasets/\{datasetId\}/refreshJobs/\{jobId\}/executionResults/} in the DataRobot Public API.
}
\seealso{
The method DatasetsRefreshJobsExecutionResultsList in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
