% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsRelationships}
\alias{ListDatasetsRelationships}
\title{ListDatasetsRelationships}
\usage{
ListDatasetsRelationships(
  linkedDatasetId,
  datasetId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List related datasets
}
\details{
Retrieve a list of the dataset relationships for a specific dataset.

This method invokes \verb{GET /datasets/\{datasetId\}/relationships/} in the DataRobot Public API.
}
\seealso{
The method DatasetsRelationshipsList in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
