% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsVersions}
\alias{ListDatasetsVersions}
\title{ListDatasetsVersions}
\usage{
ListDatasetsVersions(
  orderBy,
  datasetId,
  category,
  offset = 0,
  limit = 100,
  filterFailed = "false",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List dataset versions
}
\details{
List all versions associated with given datasetId and which match the specified query parameters.

This method invokes \verb{GET /datasets/\{datasetId\}/versions/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsList in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
