% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsVersionsAllFeaturesDetails}
\alias{ListDatasetsVersionsAllFeaturesDetails}
\title{ListDatasetsVersionsAllFeaturesDetails}
\usage{
ListDatasetsVersionsAllFeaturesDetails(
  datasetId,
  datasetVersionId,
  offset = 0,
  limit = 100,
  orderBy = "name",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get dataset features
}
\details{
Return detailed information on all the features and transforms for this dataset.If the Dataset Item has attribute snapshot = True, all optional fields also appear

This method invokes \verb{GET /datasets/\{datasetId\}/versions/\{datasetVersionId\}/allFeaturesDetails/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsAllFeaturesDetailsList in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
