% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataSources.R
\name{ListDataSources}
\alias{ListDataSources}
\title{ListDataSources}
\usage{
ListDataSources(type = "jdbc", ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
data.frame containing information on possible data sources.
}
\description{
List data sources.
}
\details{
Return detailed list of available data sources.

This method invokes \verb{GET /externalDataSources/} in the DataRobot Public API.

Returns a dataframe with information on available data sources.
}
\examples{
\dontrun{
ListDataSources()
}
}
\seealso{
The method ExternalDataSourcesList in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
