% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeletedDeployments}
\alias{ListDeletedDeployments}
\title{ListDeletedDeployments}
\usage{
ListDeletedDeployments(offset = 0, limit = 20, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List deleted deployments
}
\details{
List deleted deployments.Only available as part of an enterprise (on-prem) installation. Requires a CAN_DELETE_APP_PROJECTS permission to execute.

This method invokes \verb{GET /deletedDeployments/} in the DataRobot Public API.
}
\seealso{
The method DeletedDeploymentsList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
