% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListDeletedProjectsCount}
\alias{ListDeletedProjectsCount}
\title{ListDeletedProjectsCount}
\usage{
ListDeletedProjectsCount(
  deletedAfter,
  creator,
  organization,
  limit,
  deletedBefore,
  projectId,
  searchFor,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Count soft-deleted projects.
}
\details{
Get current number of deleted projects matching search criteria. Value is limited by DELETED_PROJECTS_BATCH_LIMIT system setting. That means that the actual amount of deleted projects can be greater than the limit, but counting will stop when reaching it.

This method invokes \verb{GET /deletedProjectsCount/} in the DataRobot Public API.
}
\seealso{
The method DeletedProjectsCountList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
