% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsAccuracyOverTime}
\alias{ListDeploymentsAccuracyOverTime}
\title{ListDeploymentsAccuracyOverTime}
\usage{
ListDeploymentsAccuracyOverTime(
  segmentAttribute,
  targetClasses,
  metric,
  modelId,
  deploymentId,
  start,
  end,
  bucketSize,
  segmentValue = "",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve accuracy over time data for one single metric.
}
\details{
Retrieve accuracy over time data for one single metric. > > The following metrics can be retrieved. > > > For classification deployments: > > > > -   Accuracy > > -   AUC > > -   Balanced Accuracy > > -   FVE Binomial > > -   Gini Norm > > -   Kolmogorov-Smirnov > > -   LogLoss (default) > > -   Rate@Top5\% > > -   Rate@Top10\% > > > > For regression deployments: > > > > -   Gamma Deviance (default) > > -   FVE Gamma > > -   FVE Poisson > > -   FVE Tweedie > > -   MAD > > -   MAE > > -   MAPE > > -   Poisson Deviance > > -   R Squared > > -   RMSE > > -   RMSLE > > -   Tweedie Deviance

This method invokes \verb{GET /deployments/\{deploymentId\}/accuracyOverTime/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsAccuracyOverTimeList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
