% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsCapabilities}
\alias{ListDeploymentsCapabilities}
\title{ListDeploymentsCapabilities}
\usage{
ListDeploymentsCapabilities(deploymentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve capabilities.
}
\details{
Retrieve the capabilities for the deployment.

This method invokes \verb{GET /deployments/\{deploymentId\}/capabilities/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsCapabilitiesList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
