% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsFeatureDrift}
\alias{ListDeploymentsFeatureDrift}
\title{ListDeploymentsFeatureDrift}
\usage{
ListDeploymentsFeatureDrift(
  segmentAttribute,
  modelId,
  deploymentId,
  start,
  end,
  offset = 0,
  metric = "psi",
  limit = 200,
  segmentValue = "",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve feature drift scores
}
\details{
Retrieve drift scores for features of the deployment.

This method invokes \verb{GET /deployments/\{deploymentId\}/featureDrift/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsFeatureDriftList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
