% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsFeatureDriftOverTime}
\alias{ListDeploymentsFeatureDriftOverTime}
\title{ListDeploymentsFeatureDriftOverTime}
\usage{
ListDeploymentsFeatureDriftOverTime(
  segmentAttribute,
  modelId,
  deploymentId,
  start,
  featureNames,
  segmentValue,
  end,
  bucketSize,
  metric = "psi",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve drift over time info for a feature of the deployment.
}
\details{
Retrieve drift over time info for a feature of the deployment.

This method invokes \verb{GET /deployments/\{deploymentId\}/featureDriftOverTime/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsFeatureDriftOverTimeList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
