% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsFeatures}
\alias{ListDeploymentsFeatures}
\title{ListDeploymentsFeatures}
\usage{
ListDeploymentsFeatures(
  search,
  deploymentId,
  offset = 0,
  includeNonPredictionFeatures = "false",
  limit = 50,
  forSegmentedAnalysis = "false",
  ...
)
}
\description{
Get deployment features
}
\details{
Retrieve features contained within the universe dataset associated with a specific deployment. By default, this returns all raw features required for predictions.
}
\seealso{
The method DeploymentsFeaturesList in [datarobot.apicore::DeploymentsApi], which this function wraps.

Other Deployments: 
\code{\link{CreateDeploymentsActualsFromDataset}()},
\code{\link{CreateDeploymentsChallengerPredictions}()},
\code{\link{CreateDeploymentsChallengers}()},
\code{\link{CreateDeploymentsFromProjectRecommendedModel}()},
\code{\link{CreateDeploymentsOnDemandReports}()},
\code{\link{CreateDeploymentsPredictionInputsFromDataset}()},
\code{\link{CreateDeploymentsPredictions}()},
\code{\link{CreateDeploymentsScoringCodeBuilds}()},
\code{\link{CreateDeployment}()},
\code{\link{DeleteDeploymentsChallengers}()},
\code{\link{DeleteDeployment}()},
\code{\link{GetDeploymentAccuracy}()},
\code{\link{GetDeploymentServiceStats}()},
\code{\link{GetDeploymentSettings}()},
\code{\link{GetDeployment}()},
\code{\link{ListDeletedDeployments}()},
\code{\link{ListDeploymentsAccuracyOverTime}()},
\code{\link{ListDeploymentsChallengers}()},
\code{\link{ListDeploymentsFeatureDrift}()},
\code{\link{ListDeploymentsHumilityStatsOverTime}()},
\code{\link{ListDeploymentsHumilityStats}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfigurationHistory}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{ListDeploymentsPredictionResults}()},
\code{\link{ListDeploymentsScoringCode}()},
\code{\link{ListDeploymentsServiceStatsOverTime}()},
\code{\link{ListDeploymentsSharedRoles}()},
\code{\link{ListDeploymentsTargetDrift}()},
\code{\link{ListDeployments}()},
\code{\link{ListPredictionServers}()},
\code{\link{PatchDeploymentsChallengers}()},
\code{\link{PatchDeployments}()},
\code{\link{PatchManyDeletedDeployments}()},
\code{\link{PatchManyDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{PatchManyDeploymentsSharedRoles}()},
\code{\link{ReplaceDeployedModel}()},
\code{\link{RetrieveDeploymentsChallengers}()},
\code{\link{SubmitActuals}()},
\code{\link{ValidateReplaceDeployedModel}()}
}
\concept{Deployments}
