% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsHumilityStats}
\alias{ListDeploymentsHumilityStats}
\title{ListDeploymentsHumilityStats}
\usage{
ListDeploymentsHumilityStats(
  segmentAttribute,
  modelId,
  deploymentId,
  start,
  end,
  bucketSize,
  segmentValue = "",
  ...
)
}
\description{
Retrieve humility stats
}
\details{
Retrieve humility rule service triggers statistics overview.
}
\seealso{
The method DeploymentsHumilityStatsList in [datarobot.apicore::DeploymentsApi], which this function wraps.

Other Deployments: 
\code{\link{CreateDeploymentsActualsFromDataset}()},
\code{\link{CreateDeploymentsChallengerPredictions}()},
\code{\link{CreateDeploymentsChallengers}()},
\code{\link{CreateDeploymentsFromProjectRecommendedModel}()},
\code{\link{CreateDeploymentsOnDemandReports}()},
\code{\link{CreateDeploymentsPredictionInputsFromDataset}()},
\code{\link{CreateDeploymentsPredictions}()},
\code{\link{CreateDeploymentsScoringCodeBuilds}()},
\code{\link{CreateDeployment}()},
\code{\link{DeleteDeploymentsChallengers}()},
\code{\link{DeleteDeployment}()},
\code{\link{GetDeploymentAccuracy}()},
\code{\link{GetDeploymentServiceStats}()},
\code{\link{GetDeploymentSettings}()},
\code{\link{GetDeployment}()},
\code{\link{ListDeletedDeployments}()},
\code{\link{ListDeploymentsAccuracyOverTime}()},
\code{\link{ListDeploymentsChallengers}()},
\code{\link{ListDeploymentsFeatureDrift}()},
\code{\link{ListDeploymentsFeatures}()},
\code{\link{ListDeploymentsHumilityStatsOverTime}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfigurationHistory}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{ListDeploymentsPredictionResults}()},
\code{\link{ListDeploymentsScoringCode}()},
\code{\link{ListDeploymentsServiceStatsOverTime}()},
\code{\link{ListDeploymentsSharedRoles}()},
\code{\link{ListDeploymentsTargetDrift}()},
\code{\link{ListDeployments}()},
\code{\link{ListPredictionServers}()},
\code{\link{PatchDeploymentsChallengers}()},
\code{\link{PatchDeployments}()},
\code{\link{PatchManyDeletedDeployments}()},
\code{\link{PatchManyDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{PatchManyDeploymentsSharedRoles}()},
\code{\link{ReplaceDeployedModel}()},
\code{\link{RetrieveDeploymentsChallengers}()},
\code{\link{SubmitActuals}()},
\code{\link{ValidateReplaceDeployedModel}()}
}
\concept{Deployments}
