% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsRetrainingSettings}
\alias{ListDeploymentsRetrainingSettings}
\title{ListDeploymentsRetrainingSettings}
\usage{
ListDeploymentsRetrainingSettings(deploymentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for fetching deployment retraining settings.
}
\details{
Retrieve deployment retraining settings.

This method invokes \verb{GET /deployments/\{deploymentId\}/retrainingSettings/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingSettingsList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
