% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsTargetDrift}
\alias{ListDeploymentsTargetDrift}
\title{ListDeploymentsTargetDrift}
\usage{
ListDeploymentsTargetDrift(
  segmentAttribute,
  modelId,
  deploymentId,
  start,
  end,
  metric = "psi",
  segmentValue = "",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve target drift
}
\details{
Retrieve target drift data.

This method invokes \verb{GET /deployments/\{deploymentId\}/targetDrift/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsTargetDriftList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
