% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListEventLogs}
\alias{ListEventLogs}
\title{ListEventLogs}
\usage{
ListEventLogs(
  maxTimestamp,
  auditReportType,
  event,
  userId,
  projectId,
  minTimestamp,
  orgId,
  offset = 0,
  includeIdentifyingFields = "true",
  order = "desc",
  ...
)
}
\description{
Retrieve one page of audit log records.
}
\details{
Retrieve one page of audit log records.
}
\seealso{
The method EventLogsList in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{CreateUsageDataExports}()},
\code{\link{DeleteManyUsersRateLimitUsage}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsEvents}()},
\code{\link{ListEventLogsPredictionUsage}()},
\code{\link{ListUsageDataExportsSupportedEvents}()},
\code{\link{ListUsersRateLimitUsage}()},
\code{\link{RetrieveEventLogs}()},
\code{\link{RetrieveUsageDataExports}()}
}
\concept{Analytics}
