% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListEventLogsEvents}
\alias{ListEventLogsEvents}
\title{ListEventLogsEvents}
\usage{
ListEventLogsEvents(...)
}
\description{
Retrieve all the available events. DEPRECATED API.
}
\details{
Retrieve all the available events. DEPRECATED API.
}
\seealso{
The method EventLogsEventsList in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{CreateUsageDataExports}()},
\code{\link{DeleteManyUsersRateLimitUsage}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsPredictionUsage}()},
\code{\link{ListEventLogs}()},
\code{\link{ListUsageDataExportsSupportedEvents}()},
\code{\link{ListUsersRateLimitUsage}()},
\code{\link{RetrieveEventLogs}()},
\code{\link{RetrieveUsageDataExports}()}
}
\concept{Analytics}
