% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListEventLogsPredictionUsage}
\alias{ListEventLogsPredictionUsage}
\title{ListEventLogsPredictionUsage}
\usage{
ListEventLogsPredictionUsage(
  maxTimestamp,
  userId,
  projectId,
  minTimestamp,
  offset = 0,
  includeIdentifyingFields = "true",
  order = "desc",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve prediction usage data.
}
\details{
Retrieve prediction usage data. \code{CAN_ACCESS_USER_ACTIVITY} permission is required.

This method invokes \verb{GET /eventLogs/predictionUsage/} in the DataRobot Public API.
}
\seealso{
The method EventLogsPredictionUsageList in \link[datarobot.apicore:AnalyticsApi]{datarobot.apicore::AnalyticsApi}, which invokes the same underlying API endpoint.
}
