% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{ListExternalDataSourcesSharedRoles}
\alias{ListExternalDataSourcesSharedRoles}
\title{ListExternalDataSourcesSharedRoles}
\usage{
ListExternalDataSourcesSharedRoles(
  shareRecipientType,
  dataSourceId,
  name,
  id,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get data source's access control list
}
\details{
Get a list of users, groups and organizations who have access to this data source and their roles.

This method invokes \verb{GET /externalDataSources/\{dataSourceId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataSourcesSharedRolesList in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
