% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{ListExternalDataStoresCredentials}
\alias{ListExternalDataStoresCredentials}
\title{ListExternalDataStoresCredentials}
\usage{
ListExternalDataStoresCredentials(
  types,
  dataStoreId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List credentials associated with the specified data store.
}
\details{
Returns a list of credentials associated with the specified data store.

This method invokes \verb{GET /externalDataStores/\{dataStoreId\}/credentials/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresCredentialsList in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
