% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{ListImageAugmentationLists}
\alias{ListImageAugmentationLists}
\title{ListImageAugmentationLists}
\usage{
ListImageAugmentationLists(
  featureName,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List of augmentation lists that match the specified query.
}
\details{
List of augmentation lists that match the specified query.

This method invokes \verb{GET /imageAugmentationLists/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationListsList in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
