% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{ListImageAugmentationListsSamples}
\alias{ListImageAugmentationListsSamples}
\title{ListImageAugmentationListsSamples}
\usage{
ListImageAugmentationListsSamples(augmentationId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve latest Augmentation Samples generated for list
}
\details{
Retrieve latest Augmentation Samples generated for list

This method invokes \verb{GET /imageAugmentationLists/\{augmentationId\}/samples/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationListsSamplesList in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
