% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{ListImageAugmentationSamples}
\alias{ListImageAugmentationSamples}
\title{ListImageAugmentationSamples}
\usage{
ListImageAugmentationSamples(samplesId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
(Deprecated in v2.28) Retrieve previously generated Augmentation Samples. This route is being deprecated please use \verb{GET /imageAugmentationLists/<augmentationId>/samples} route instead.
}
\details{
(Deprecated in v2.28) Retrieve previously generated Augmentation Samples. This route is being deprecated please use GET /imageAugmentationLists/\\<augmentationId\\>/samples route instead.

This method invokes \verb{GET /imageAugmentationSamples/\{samplesId\}/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationSamplesList in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
