% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Jobs.R
\name{ListJobs}
\alias{ListJobs}
\title{ListJobs}
\usage{
ListJobs(project, status = NULL, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{status}{character. The status of the desired jobs: one of JobStatus$Queue,
JobStatus$InProgress, orJobStatus$Error. If NULL (default), queued and inprogress jobs
are returned.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list of lists with one element for each job. The named list for
each job contains:
\itemize{
  \item status character. Model job status; an element of \code{JobStatus}, e.g.
    \code{JobStatus$Queue.}
  \item url character. URL to request more detail about the job.
  \item id character. The job id.
  \item jobType character. See \code{JobType} for valid values.
  \item projectId character. The project that contains the model.
  \item isBlocked logical. If TRUE, the job is blocked (cannot be executed) until its
    dependencies are resolved.
}
}
\description{
List project jobs
}
\details{
List the project's jobs.

This method invokes \verb{GET /projects/\{projectId\}/jobs/} in the DataRobot Public API.

Retrieve information about jobs

This function requests information about the jobs that go through the DataRobot queue.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ListJobs(projectId)
}
}
\seealso{
The method JobsList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
