% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlops_apiWrapper.R
\name{ListMlopsPortablePredictionServerImageMetadata}
\alias{ListMlopsPortablePredictionServerImageMetadata}
\title{ListMlopsPortablePredictionServerImageMetadata}
\usage{
ListMlopsPortablePredictionServerImageMetadata(...)
}
\description{
Fetches currently active PPS Docker image metadata
}
\details{
Fetches currently active PPS Docker image metadata
}
\seealso{
The method MlopsPortablePredictionServerImageMetadataList in [datarobot.apicore::MlopsApi], which this function wraps.

Other Mlops: 
\code{\link{ListMlopsPortablePredictionServerImage}()}
}
\concept{Mlops}
