% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{ListModelFeatures}
\alias{ListModelFeatures}
\title{Returns the list of features (i.e., variables) on which a specified model is based}
\usage{
ListModelFeatures(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModelObject, or each element of the list returned by
the function GetAllModels.}
}
\value{
A character vector of feature names, with one component for
each model feature.
}
\description{
This function returns the list of features (typically, response variable
and raw covariates) used in building the model specified by model, an S3
object of class 'dataRobotModel'.
}

