% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Featurelists.R
\name{ListModelingFeaturelists}
\alias{ListModelingFeaturelists}
\title{ListModelingFeaturelists}
\usage{
ListModelingFeaturelists(project, offset = 0, limit = 0, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An S3 object of class 'listOfFeaturelists', which is a
list of dataframes: each element of the list corresponds to one
featurelist associated with the project, and each dataframe has
one row and the following four columns:
\itemize{
  \item featurelistId. Unique alphanumeric identifier for the featurelist.
  \item projectId. Unique alphanumeric project identifier.
  \item features. Comma-separated character string listing the variables included in the
    featurelist.
  \item name. Character string giving the name of the featurelist.
}
}
\description{
List all modeling featurelists from a project
}
\details{
List all modeling featurelists from the project requested by ID. This route will only become available after the target and partitioning options have been set for a project. Modeling featurelists are featurelists of modeling features, and are the correct featurelists to use when creating models or restarting the autopilot. In a time series project, these will differ from those returned from :\url{http:get}:/api/v2/projects/(projectId)/featurelists/ while in other projects these will be identical. See the \verb{documentation <input_vs_modeling>} for more information on the distinction between input and modeling data in time series projects.

This method invokes \verb{GET /projects/\{projectId\}/modelingFeaturelists/} in the DataRobot Public API.

Retrieve all modeling featurelists associated with a project

In time series projects, a new set of modeling features is created after setting the
partitioning options. These features are automatically derived from those in the project's
dataset and are the features used for modeling. Modeling features are only accessible once
the target and partitioning options have been set. In projects that don't use time series
modeling, once the target has been set, ModelingFeaturelists and Featurelists will behave
the same.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ListModelingFeaturelists(projectId)
}
}
\seealso{
The method ModelingFeaturelistsList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
