% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListModelPackages}
\alias{ListModelPackages}
\title{ListModelPackages}
\usage{
ListModelPackages(
  predictionThreshold,
  forChallenger,
  search,
  modelId,
  imported,
  predictionEnvironmentId,
  similarTo,
  modelKind,
  buildStatus,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List model packages
}
\details{
Retrieve the list of model packages a user has access to.

This method invokes \verb{GET /modelPackages/} in the DataRobot Public API.
}
\seealso{
The method ModelPackagesList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
