% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListModelPackagesFeatures}
\alias{ListModelPackagesFeatures}
\title{ListModelPackagesFeatures}
\usage{
ListModelPackagesFeatures(
  search,
  orderBy,
  modelPackageId,
  offset = 0,
  includeNonPredictionFeatures = "false",
  limit = 50,
  forSegmentedAnalysis = "false",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve feature list.
}
\details{
Retrieve the feature list for given model package.

This method invokes \verb{GET /modelPackages/\{modelPackageId\}/features/} in the DataRobot Public API.
}
\seealso{
The method ModelPackagesFeaturesList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
