% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/ModelRecommendations.R
\name{ListModelRecommendations}
\alias{ListModelRecommendations}
\title{ListModelRecommendations}
\usage{
ListModelRecommendations(project, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list containing information about each recommendation made by DataRobot, containing:
\itemize{
  \item modelId character. The model ID of the recommended model.
  \item projectId character. The project ID of the project the recommendations were made for.
  \item recommendationType character. The type of recommendation being made.
}
}
\description{
List recommended models for the project
}
\details{
Retrieves all of the current recommended models for the project

This method invokes \verb{GET /projects/\{projectId\}/recommendedModels/} in the DataRobot Public API.

Retrieve information about model recommendation made by DataRobot for your project.

DataRobot will help pick out a few models from your project that meet certain criteria,
such as being the most accurate model or being a model that captures a good blend of both
prediction speed and model accuracy.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
ListModelRecommendations(projectId)
}
}
\seealso{
The method ProjectsRecommendedModelsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
