% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListMultilabelInsightsPairwiseManualSelections}
\alias{ListMultilabelInsightsPairwiseManualSelections}
\title{ListMultilabelInsightsPairwiseManualSelections}
\usage{
ListMultilabelInsightsPairwiseManualSelections(multilabelInsightsKey, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get all label lists.
}
\details{
Get all label lists.

This method invokes \verb{GET /multilabelInsights/\{multilabelInsightsKey\}/pairwiseManualSelections/} in the DataRobot Public API.
}
\seealso{
The method MultilabelInsightsPairwiseManualSelectionsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
