% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{ListNotificationPolicies}
\alias{ListNotificationPolicies}
\title{ListNotificationPolicies}
\usage{
ListNotificationPolicies(
  eventGroup,
  namePart,
  channelId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List notification policies
}
\details{
List the notification policies that satisfy the query condition.

This method invokes \verb{GET /notificationPolicies/} in the DataRobot Public API.
}
\seealso{
The method NotificationPoliciesList in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
