% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{ListNotificationPolicyMutes}
\alias{ListNotificationPolicyMutes}
\title{ListNotificationPolicyMutes}
\usage{
ListNotificationPolicyMutes(entityId, orgId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List the ignored notifications filtered by orgId if provided
}
\details{
List the ignored notifications filtered by orgId if provided

This method invokes \verb{GET /notificationPolicyMutes/} in the DataRobot Public API.
}
\seealso{
The method NotificationPolicyMutesList in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
