% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R,
%   R/predictions_apiWrapper.R, R/PredictionExplanations.R
\name{ListPredictionExplanationsMetadata}
\alias{ListPredictionExplanationsMetadata}
\title{ListPredictionExplanationsMetadata}
\usage{
ListPredictionExplanationsMetadata(
  project,
  modelId = NULL,
  limit = NULL,
  offset = NULL,
  ...
)

ListPredictionExplanationsMetadata(
  project,
  modelId = NULL,
  limit = NULL,
  offset = NULL,
  ...
)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Optional. If specified, only prediction explanations computed
for this model will be returned.}

\item{limit}{integer. Optional. At most this many results are returned, default: no limit}

\item{offset}{integer. This many results will be skipped, default: 0}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
List of metadata for all prediction explanations in the project.
  Each element of list is metadata for one prediction explanations
  (for format see \code{GetPredictionExplanationsMetadata}).
}
\description{
List PredictionExplanationsRecord objects for a project.

List PredictionExplanationsRecord objects for a project.
}
\details{
List PredictionExplanationsRecord objects for a project. These contain metadata about the computed prediction explanations and the location at which the PredictionExplanations can be retrieved.

This method invokes \verb{GET /projects/\{projectId\}/predictionExplanationsRecords/} in the DataRobot Public API.

List PredictionExplanationsRecord objects for a project. These contain metadata about the computed prediction explanations and the location at which the PredictionExplanations can be retrieved.

This method invokes \verb{GET /projects/\{projectId\}/predictionExplanationsRecords/} in the DataRobot Public API.

Retrieve metadata for prediction explanations in specified project
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ListPredictionExplanationsMetadata(projectId)
}
}
\seealso{
The method ProjectsPredictionExplanationsRecordsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.

The method ProjectsPredictionExplanationsRecordsList in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
