% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R, R/GetPredictions.R
\name{ListPredictions}
\alias{ListPredictions}
\title{ListPredictions}
\usage{
ListPredictions(project, modelId = NULL, datasetId = NULL)

ListPredictions(project, modelId = NULL, datasetId = NULL)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{numeric. Optional. Filter returned predictions to only be predictions made
against the model specified by this model ID.}

\item{datasetId}{numeric. Optional. Filter returned predictions to only be predictions made
against the prediction dataset specified by this dataset ID.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A data.frame specifying:
  \itemize{
    \item projectId character. The ID of the project the predictions were made in.
    \item datasetId character. The dataset ID of the dataset used to make predictions
    \item modelId character. The model ID of the model used to make predictions.
    \item predictionId character. The unique ID corresponding to those predictions. Use
      \code{GetPredictions(projectId, predictionId)} to fetch the individual predictions.
    \item includesPredictionIntervals logical. Whether or not the predictions include
      prediction intervals. See \code{Predict} for details.
    \item predictionIntervalsSize integer. Optional. The size, in percent, of prediction
     intervals or NULL if there are no intervals. See \code{Predict} for details.
  }
}
\description{
Get a list of prediction records.

This function itemizes the predictions available for a given project, model, and/or dataset.
Note that this function does not actually return the predictions. Use
\code{GetPredictions(projectId, predictionId)} to get the predictions for a particular
set of predictions.
}
\details{
Get a list of prediction records.  v2.21 Use :\url{http:get}:/api/v2/projects/(projectId)/predictionsMetadata/ instead. The only difference is that parameter datasetId is renamed to predictionDatasetId both in request and response.

This method invokes \verb{GET /projects/\{projectId\}/predictions/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
predictions <- ListPredictions(projectId)
}
}
\seealso{
The method ProjectsPredictionsList in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
