% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R, R/Deployment.R
\name{ListPredictionServers}
\alias{ListPredictionServers}
\title{ListPredictionServers}
\usage{
ListPredictionServers(offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list of DataRobotPredictionServer objects containing:
\itemize{
 \item id character. The ID of the prediction server.
 \item url character. The URL of the prediction server.
 \item dataRobotKey character. The key used to access the prediction server.
}
}
\description{
List prediction servers.
}
\details{
List prediction servers available for a user.

This method invokes \verb{GET /predictionServers/} in the DataRobot Public API.

List all available prediction servers.
}
\examples{
\dontrun{
ListPredictionServers()
}
}
\seealso{
The method PredictionServersList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
