% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/PrimeFiles.R
\name{ListPrimeFiles}
\alias{ListPrimeFiles}
\title{ListPrimeFiles}
\usage{
ListPrimeFiles(
  project,
  parentModelId = NULL,
  modelId = NULL,
  offset = 0,
  limit = 0,
  ...
)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{parentModelId}{numeric. Optional. Filter for only those prime files approximating this
parent model.}

\item{modelId}{numeric. Optional. Filter for only those prime files with code for this
prime model.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
List of lists. Each element of the list corresponds to one Prime file available to
  download. The elements of this list have the same format as the return value of GetPrimeFile.
}
\description{
Get Prime Files
}
\details{
List all DataRobot Prime files available for download

This method invokes \verb{GET /projects/\{projectId\}/primeFiles/} in the DataRobot Public API.

List all downloadable code files from DataRobot Prime for the project

Training a model using a ruleset is a necessary prerequisite for being able to download the code
for a ruleset.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ListPrimeFiles(projectId)
}
}
\seealso{
The method ProjectsPrimeFilesList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
