% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/PrimeModels.R
\name{ListPrimeModels}
\alias{ListPrimeModels}
\title{ListPrimeModels}
\usage{
ListPrimeModels(project, offset = 0, limit = 100, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
data.frame (classed as \code{dataRobotPrimeModels}) containing
  information about each DataRobot Prime model in a project (one row per
  Prime model).
}
\description{
List all Prime models in a project
}
\details{
List all Prime models in a project

This method invokes \verb{GET /projects/\{projectId\}/primeModels/} in the DataRobot Public API.

Retrieve information about all DataRobot Prime models for a DataRobot project

This function requests the DataRobot Prime models information for the DataRobot
project specified by the project argument, described under Arguments.

The function returns data.frame containing information about each DataRobot Prime model in a
project (one row per Prime model)
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ListPrimeModels(projectId)
}
}
\seealso{
The method ProjectsPrimeModelsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
