% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectCleanupJobsSummary}
\alias{ListProjectCleanupJobsSummary}
\title{ListProjectCleanupJobsSummary}
\usage{
ListProjectCleanupJobsSummary(statusId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a projects cleanup jobs summary.
}
\details{
Get number of projects whose deletion finished in particular state

This method invokes \verb{GET /projectCleanupJobs/\{statusId\}/summary/} in the DataRobot Public API.
}
\seealso{
The method CleanupJobsSummaryList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
