% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsBiasMitigatedModels}
\alias{ListProjectsBiasMitigatedModels}
\title{ListProjectsBiasMitigatedModels}
\usage{
ListProjectsBiasMitigatedModels(
  parentModelId,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List of bias mitigated models for the selected project.
}
\details{
List of bias mitigated models for the selected project.

This method invokes \verb{GET /projects/\{projectId\}/biasMitigatedModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBiasMitigatedModelsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
