% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsCombinedModelsSegments}
\alias{ListProjectsCombinedModelsSegments}
\title{ListProjectsCombinedModelsSegments}
\usage{
ListProjectsCombinedModelsSegments(
  combinedModelId,
  searchSegmentName,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Combined Model segments info.
}
\details{
Retrieve Combined Model segments info (name, related project & model details).

This method invokes \verb{GET /projects/\{projectId\}/combinedModels/\{combinedModelId\}/segments/} in the DataRobot Public API.
}
\seealso{
The method ProjectsCombinedModelsSegmentsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
