% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsDatetimeModelsBacktestStabilityPlot}
\alias{ListProjectsDatetimeModelsBacktestStabilityPlot}
\title{ListProjectsDatetimeModelsBacktestStabilityPlot}
\usage{
ListProjectsDatetimeModelsBacktestStabilityPlot(
  metricName,
  modelId,
  forecastDistance,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a plot displaying the stability of the datetime model across different backtests.
}
\details{
Retrieve a plot displaying the stability of the datetime model across different backtests.  All durations and datetimes should be specified in accordance with the \verb{timestamp and duration formatting rules<time_format>}.

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/backtestStabilityPlot/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsBacktestStabilityPlotList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
