% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadata}
\alias{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadata}
\title{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadata}
\usage{
ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadata(
  modelId,
  datasetId,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the metadata of the Accuracy Over Time (AOT) chart for an external dataset.
}
\details{
Retrieve the metadata of the Accuracy Over Time (AOT) chart for an external dataset. Datetimes are specified in accordance with \verb{timestamp and duration formatting rules <time_format>}.

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/datasetAccuracyOverTimePlots/\{datasetId\}/metadata/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadataList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
