% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsDatetimeModelsFeatureEffectsMetadata}
\alias{ListProjectsDatetimeModelsFeatureEffectsMetadata}
\title{ListProjectsDatetimeModelsFeatureEffectsMetadata}
\usage{
ListProjectsDatetimeModelsFeatureEffectsMetadata(modelId, projectId, ...)
}
\description{
Retrieve Feature Effects metadata for each backtest. Response contains status and available sources for each backtest of the model.
}
\details{
Retrieve Feature Effects metadata for each backtest. Response contains status and available sources for each backtest of the model. One of the provided `backtestIndex` indexes used for submitting the compute request and retrieving Feature Effects. * Start/stop models contain a single `backtestIndex` response value of `startstop`. * Other models contain `backtestIndex` of `0`, `1`, ..., `holdout`. One of the provided `source` parameters used for retrieving Feature Effects. * Each backtest source can be, at a minimum, `training` or `validation`. If holdout is configured for the project, `backtestIndex` also includes `holdout` with sources `training` and `holdout`. * Source value of `training` is always available. (versions prior to v2.17 support `validation` only) * When a start/stop model is trained into `validation` or `holdout` without stacked predictions (i.e., no out-of-sample predictions in `validation` or `holdout`), `validation` and `holdout` sources are not available. * Source `holdout` is not available when there is no holdout configured for the project.
}
\seealso{
The method ProjectsDatetimeModelsFeatureEffectsMetadataList in [datarobot.apicore::InsightsApi], which this function wraps.

Other Insights: 
\code{\link{CreateMultilabelInsightsPairwiseManualSelections}()},
\code{\link{CreateProjectsDatetimeModelsFeatureEffects}()},
\code{\link{CreateProjectsDatetimeModelsFeatureFit}()},
\code{\link{CreateProjectsDatetimeModelsMulticlassFeatureEffects}()},
\code{\link{CreateProjectsExternalScores}()},
\code{\link{CreateProjectsFeatureAssociationMatrix}()},
\code{\link{CreateProjectsGeometryFeaturePlots}()},
\code{\link{CreateProjectsModelsClusterInsights}()},
\code{\link{CreateProjectsModelsCrossClassAccuracyScores}()},
\code{\link{CreateProjectsModelsDataDisparityInsights}()},
\code{\link{CreateProjectsModelsFairnessInsights}()},
\code{\link{CreateProjectsModelsFeatureEffects}()},
\code{\link{CreateProjectsModelsFeatureFit}()},
\code{\link{CreateProjectsModelsImageActivationMaps}()},
\code{\link{CreateProjectsModelsImageEmbeddings}()},
\code{\link{CreateProjectsModelsMulticlassFeatureEffects}()},
\code{\link{CreateProjectsModelsShapImpact}()},
\code{\link{CreateProjectsPayoffMatrices}()},
\code{\link{CreateProjectsShapMatrices}()},
\code{\link{DeleteAnomalyAssessmentRecord}()},
\code{\link{DeleteMultilabelInsightsPairwiseManualSelections}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteProjectsPayoffMatrices}()},
\code{\link{GetAnomalyAssessmentExplanations}()},
\code{\link{GetAnomalyAssessmentPredictionsPreview}()},
\code{\link{GetFeatureAssociationMatrixDetails}()},
\code{\link{GetFeatureAssociationMatrix}()},
\code{\link{GetPredictionExplanationsMetadata}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetWordCloud}()},
\code{\link{InitializeAnomalyAssessment}()},
\code{\link{ListAnomalyAssessmentRecords}()},
\code{\link{ListMultilabelInsightsHistogram}()},
\code{\link{ListMultilabelInsightsPairwiseManualSelections}()},
\code{\link{ListMultilabelInsightsPairwiseStatistics}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListProjectsBiasVsAccuracyInsights}()},
\code{\link{ListProjectsDatetimeModelsBacktestStabilityPlot}()},
\code{\link{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadata}()},
\code{\link{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsPreview}()},
\code{\link{ListProjectsDatetimeModelsFeatureEffects}()},
\code{\link{ListProjectsDatetimeModelsFeatureFitMetadata}()},
\code{\link{ListProjectsDatetimeModelsFeatureFit}()},
\code{\link{ListProjectsDatetimeModelsForecastDistanceStabilityPlot}()},
\code{\link{ListProjectsDatetimeModelsMulticlassFeatureEffects}()},
\code{\link{ListProjectsDatetimeModelsMultiseriesHistograms}()},
\code{\link{ListProjectsDatetimeModelsMultiseriesScoresFile}()},
\code{\link{ListProjectsExternalScores}()},
\code{\link{ListProjectsFeatureAssociationFeaturelists}()},
\code{\link{ListProjectsFeaturesFrequentValues}()},
\code{\link{ListProjectsImageActivationMaps}()},
\code{\link{ListProjectsImageEmbeddings}()},
\code{\link{ListProjectsModelsAnomalyInsightsFile}()},
\code{\link{ListProjectsModelsAnomalyInsightsTable}()},
\code{\link{ListProjectsModelsClusterInsightsDownload}()},
\code{\link{ListProjectsModelsClusterInsights}()},
\code{\link{ListProjectsModelsConfusionChartsClassDetails}()},
\code{\link{ListProjectsModelsConfusionChartsMetadata}()},
\code{\link{ListProjectsModelsConfusionCharts}()},
\code{\link{ListProjectsModelsCrossClassAccuracyScores}()},
\code{\link{ListProjectsModelsDataDisparityInsights}()},
\code{\link{ListProjectsModelsDatasetConfusionChartsClassDetails}()},
\code{\link{ListProjectsModelsDatasetConfusionChartsMetadata}()},
\code{\link{ListProjectsModelsDatasetConfusionCharts}()},
\code{\link{ListProjectsModelsDatasetLiftCharts}()},
\code{\link{ListProjectsModelsDatasetMulticlassLiftCharts}()},
\code{\link{ListProjectsModelsDatasetResidualsCharts}()},
\code{\link{ListProjectsModelsDatasetRocCurves}()},
\code{\link{ListProjectsModelsFairnessInsights}()},
\code{\link{ListProjectsModelsFeatureEffectsMetadata}()},
\code{\link{ListProjectsModelsFeatureEffects}()},
\code{\link{ListProjectsModelsFeatureFitMetadata}()},
\code{\link{ListProjectsModelsFeatureFit}()},
\code{\link{ListProjectsModelsFeatureListsClusterInsights}()},
\code{\link{ListProjectsModelsImageActivationMaps}()},
\code{\link{ListProjectsModelsImageEmbeddings}()},
\code{\link{ListProjectsModelsLabelwiseRocCurves}()},
\code{\link{ListProjectsModelsLiftChart}()},
\code{\link{ListProjectsModelsMulticlassFeatureEffects}()},
\code{\link{ListProjectsModelsMulticlassFeatureImpact}()},
\code{\link{ListProjectsModelsMulticlassLiftChart}()},
\code{\link{ListProjectsModelsResiduals}()},
\code{\link{ListProjectsModelsRocCurve}()},
\code{\link{ListProjectsModelsShapImpact}()},
\code{\link{ListProjectsMulticategoricalInvalidFormatFile}()},
\code{\link{ListProjectsMulticategoricalInvalidFormat}()},
\code{\link{ListProjectsPayoffMatrices}()},
\code{\link{ListProjectsShapMatrices}()},
\code{\link{PatchMultilabelInsightsPairwiseManualSelections}()},
\code{\link{PutProjectsPayoffMatrices}()},
\code{\link{RetrieveProjectsDatetimeModelsDatasetAccuracyOverTimePlots}()},
\code{\link{RetrieveProjectsGeometryFeaturePlots}()},
\code{\link{RetrieveProjectsModelsDatasetConfusionCharts}()},
\code{\link{RetrieveProjectsModelsLiftChart}()},
\code{\link{RetrieveProjectsModelsMulticlassLiftChart}()},
\code{\link{RetrieveProjectsModelsMultilabelLiftCharts}()},
\code{\link{RetrieveProjectsModelsResiduals}()},
\code{\link{RetrieveProjectsModelsRocCurve}()},
\code{\link{RetrieveProjectsShapMatrices}()}
}
\concept{Insights}
