% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsDatetimeModelsMultiseriesHistograms}
\alias{ListProjectsDatetimeModelsMultiseriesHistograms}
\title{ListProjectsDatetimeModelsMultiseriesHistograms}
\usage{
ListProjectsDatetimeModelsMultiseriesHistograms(
  metric,
  modelId,
  projectId,
  bins = "10",
  attribute = "rowCount",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the histograms for series insights.
}
\details{
Retrieve the histograms for series insights.  Histogram is computed only for first 1000 series (ordered by name).

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/multiseriesHistograms/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsMultiseriesHistogramsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
