% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsDatetimeModelsMultiseriesScoresFile}
\alias{ListProjectsDatetimeModelsMultiseriesScoresFile}
\title{ListProjectsDatetimeModelsMultiseriesScoresFile}
\usage{
ListProjectsDatetimeModelsMultiseriesScoresFile(
  metric,
  modelId,
  multiseriesValue,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the CSV file for the series accuracy.
}
\details{
Retrieve the CSV file for the series accuracy.

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/multiseriesScores/file/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsMultiseriesScoresFileList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
