% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsExternalScores}
\alias{ListProjectsExternalScores}
\title{ListProjectsExternalScores}
\usage{
ListProjectsExternalScores(
  modelId,
  datasetId,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List of scores on prediction datasets for a project.
}
\details{
List of scores on prediction datasets for a project with filtering option by dataset or model or both of them. Prediction dataset may have scores if it contained a column with actual values and predictions were computed on this dataset.

This method invokes \verb{GET /projects/\{projectId\}/externalScores/} in the DataRobot Public API.
}
\seealso{
The method ProjectsExternalScoresList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
