% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsFeatureDiscoveryLogsDownload}
\alias{ListProjectsFeatureDiscoveryLogsDownload}
\title{ListProjectsFeatureDiscoveryLogsDownload}
\usage{
ListProjectsFeatureDiscoveryLogsDownload(projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a text file containing the feature discovery log
}
\details{
Retrieve a text file containing the feature discovery log. This route is only supported for feature discovery projects that have finished partitioning.

This method invokes \verb{GET /projects/\{projectId\}/featureDiscoveryLogs/download/} in the DataRobot Public API.
}
\seealso{
The method FeatureDiscoveryLogsDownloadList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
