% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsFeaturesFrequentValues}
\alias{ListProjectsFeaturesFrequentValues}
\title{ListProjectsFeaturesFrequentValues}
\usage{
ListProjectsFeaturesFrequentValues(featureName, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the frequent values information for a particular feature.
}
\details{
Retrieve the frequent values information for a particular feature.   Only valid for numeric features. This route returns information about the frequent values seen for a particular feature, based on the EDA sample of the dataset. Up to 60 values will be returned, and when more values are present, they will be bucketed into a level called \"==All Other==\" at the end of the response.

This method invokes \verb{GET /projects/\{projectId\}/features/\{featureName\}/frequentValues/} in the DataRobot Public API.
}
\seealso{
The method ProjectsFeaturesFrequentValuesList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
