% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{ListProjectsImageBins}
\alias{ListProjectsImageBins}
\title{ListProjectsImageBins}
\usage{
ListProjectsImageBins(featureName, projectId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List image bins and covers for every target value or range.
}
\details{
List image bins and covers for every target value or range.

This method invokes \verb{GET /projects/\{projectId\}/imageBins/} in the DataRobot Public API.
}
\seealso{
The method ProjectsImageBinsList in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
