% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{ListProjectsImagesDataQualityLogFile}
\alias{ListProjectsImagesDataQualityLogFile}
\title{ListProjectsImagesDataQualityLogFile}
\usage{
ListProjectsImagesDataQualityLogFile(projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a text file containing the images data quality log
}
\details{
Retrieve a text file containing the images data quality log

This method invokes \verb{GET /projects/\{projectId\}/imagesDataQualityLog/file/} in the DataRobot Public API.
}
\seealso{
The method ProjectsImagesDataQualityLogFileList in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
