% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsModelingFeatures}
\alias{ListProjectsModelingFeatures}
\title{ListProjectsModelingFeatures}
\usage{
ListProjectsModelingFeatures(projectId, offset = 0, limit = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List project modeling features.
}
\details{
List the features from a project that are used for modeling with descriptive information.

This method invokes \verb{GET /projects/\{projectId\}/modelingFeatures/} in the DataRobot Public API.
}
\seealso{
The method ModelingFeaturesList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
