% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsAnomalyInsightsFile}
\alias{ListProjectsModelsAnomalyInsightsFile}
\title{ListProjectsModelsAnomalyInsightsFile}
\usage{
ListProjectsModelsAnomalyInsightsFile(filename, modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a CSV file of the raw data displayed with the anomaly score from the model.
}
\details{
Retrieve a CSV file of the raw data displayed with the anomaly score from the specific model. The number of rows included will be set by the expected outlier fraction but up to a maximum of 1000 rows. Only models built from anomaly detection blueprints have those insights.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/anomalyInsightsFile/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsAnomalyInsightsFileList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
