% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsAnomalyInsightsTable}
\alias{ListProjectsModelsAnomalyInsightsTable}
\title{ListProjectsModelsAnomalyInsightsTable}
\usage{
ListProjectsModelsAnomalyInsightsTable(
  modelId,
  anomalyScoreRounding,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a table of the raw data displayed with the anomaly score from the specific model.
}
\details{
Retrieve a table of the raw data displayed with the anomaly score from the specific model. The number of rows displayed is limited to 100 rows by the ANOMALY_INSIGHT_SAMPLE_ROW_COUNT configuration setting. Additionally, feature column count and the size of data in text fields is also limited. Only models built from anomaly detection blueprints have those insights.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/anomalyInsightsTable/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsAnomalyInsightsTableList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
