% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsClusterInsightsDownload}
\alias{ListProjectsModelsClusterInsightsDownload}
\title{ListProjectsModelsClusterInsightsDownload}
\usage{
ListProjectsModelsClusterInsightsDownload(
  modelId,
  featurelistId,
  projectId,
  format = "CSV",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download Cluster Insights result.
}
\details{
Download all computed Cluster Insights for a clustering project model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/clusterInsights/download/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsClusterInsightsDownloadList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
