% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsConfusionCharts}
\alias{ListProjectsModelsConfusionCharts}
\title{ListProjectsModelsConfusionCharts}
\usage{
ListProjectsModelsConfusionCharts(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve all available confusion charts for model.
}
\details{
Retrieve all available confusion charts for model. The response will > include a json array of all available confusion charts, in the same > format as the response from > :http:get:/api/v2/projects/(projectId)/models/(modelId)/confusionCharts/(source)/. > .. note:: Available for multiclass projects only.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/confusionCharts/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsConfusionChartsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
