% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsConfusionChartsMetadata}
\alias{ListProjectsModelsConfusionChartsMetadata}
\title{ListProjectsModelsConfusionChartsMetadata}
\usage{
ListProjectsModelsConfusionChartsMetadata(
  modelId,
  source,
  projectId,
  thumbnailCellSize = 10,
  orientation = "actual",
  orderBy = "className",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve metadata for the confusion chart of a model.
}
\details{
Retrieve metadata for the confusion chart of a model. .. note:: > Available for multiclass projects only.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/confusionCharts/\{source\}/metadata/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsConfusionChartsMetadataList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
