% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsCrossClassAccuracyScores}
\alias{ListProjectsModelsCrossClassAccuracyScores}
\title{ListProjectsModelsCrossClassAccuracyScores}
\usage{
ListProjectsModelsCrossClassAccuracyScores(
  modelId,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List Cross Class Accuracy scores.
}
\details{
Retrieves a list of Cross Class Accuracy scores for the model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/crossClassAccuracyScores/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsCrossClassAccuracyScoresList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
