% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsDatasetConfusionCharts}
\alias{ListProjectsModelsDatasetConfusionCharts}
\title{ListProjectsModelsDatasetConfusionCharts}
\usage{
ListProjectsModelsDatasetConfusionCharts(
  colEnd,
  modelId,
  datasetId,
  colStart,
  rowEnd,
  rowStart,
  projectId,
  orientation = "actual",
  offset = 0,
  limit = 100,
  orderBy = "className",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List of Confusion Charts objects on external datasets for a project with filtering option by dataset.
}
\details{
List of Confusion Charts objects on external datasets for a project with filtering option by dataset. Prediction dataset may have Confusion Chart for multiclass projects computed if it contained a target with actual values and insights were computed on this dataset. A confusion chart consists of the confusion matrix for all classes, classes frequencies and oneVsAll metrics for all classes. The confusion matrix can be requested in a particular sort order and orientated by rows or columns. Available for multiclass projects only.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/datasetConfusionCharts/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDatasetConfusionChartsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
