% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsDatasetConfusionChartsMetadata}
\alias{ListProjectsModelsDatasetConfusionChartsMetadata}
\title{ListProjectsModelsDatasetConfusionChartsMetadata}
\usage{
ListProjectsModelsDatasetConfusionChartsMetadata(
  modelId,
  datasetId,
  projectId,
  thumbnailCellSize = 10,
  orientation = "actual",
  orderBy = "className",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve metadata for the confusion chart of a model on external dataset for a project.
}
\details{
Retrieve metadata for the confusion chart of a model on external dataset for a project. Available for multiclass projects only.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/datasetConfusionCharts/\{datasetId\}/metadata/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDatasetConfusionChartsMetadataList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
