% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsDatasetLiftCharts}
\alias{ListProjectsModelsDatasetLiftCharts}
\title{ListProjectsModelsDatasetLiftCharts}
\usage{
ListProjectsModelsDatasetLiftCharts(
  modelId,
  datasetId,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve List of Lift chart data on prediction datasets for a project.
}
\details{
List of Lift chart objects on prediction datasets for a project with filtering option by dataset. Prediction dataset may have Lift chart computed if it contained a column with actual values and predictions were computed on this dataset. This controller is not supported for multiclass classification projects. For multiclass, instead use /projects/\\<projectId\\>/models/\\<modelId\\>/datasetMulticlassLiftCharts/.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/datasetLiftCharts/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDatasetLiftChartsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
