% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsDatasetMulticlassLiftCharts}
\alias{ListProjectsModelsDatasetMulticlassLiftCharts}
\title{ListProjectsModelsDatasetMulticlassLiftCharts}
\usage{
ListProjectsModelsDatasetMulticlassLiftCharts(
  modelId,
  datasetId,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve List of Multiclass Lift chart data on prediction datasets for a project.
}
\details{
List of Multiclass Lift chart objects on prediction datasets for a project with filtering option by dataset. Prediction dataset may have Multiclass Lift chart computed if it contained a column with actual values and predictions were computed on this dataset. Multiclass Lift charts are supported for multiclass classification projects only.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/datasetMulticlassLiftCharts/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDatasetMulticlassLiftChartsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
