% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R, R/images_apiWrapper.R
\name{ListProjectsModelsImageActivationMaps}
\alias{ListProjectsModelsImageActivationMaps}
\title{ListProjectsModelsImageActivationMaps}
\usage{
ListProjectsModelsImageActivationMaps(modelId, featureName, projectId, ...)

ListProjectsModelsImageActivationMaps(modelId, featureName, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Image Activation Maps for a feature of a model.

Retrieve Image Activation Maps for a feature of a model.
}
\details{
Retrieve Image Activation Maps for a feature of a model. Image Activation maps are a technique to get the discriminative image regions used by a CNN to identify a specific class in the image. In other words, an image activation map lets us see which regions in the image were relevant to this class. The higher the value in the activation map the greater the effect the region had on the prediction.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/imageActivationMaps/} in the DataRobot Public API.

Retrieve Image Activation Maps for a feature of a model. Image Activation maps are a technique to get the discriminative image regions used by a CNN to identify a specific class in the image. In other words, an image activation map lets us see which regions in the image were relevant to this class. The higher the value in the activation map the greater the effect the region had on the prediction.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/imageActivationMaps/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsImageActivationMapsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.

The method ProjectsModelsImageActivationMapsList in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
