% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsLabelwiseRocCurves}
\alias{ListProjectsModelsLabelwiseRocCurves}
\title{ListProjectsModelsLabelwiseRocCurves}
\usage{
ListProjectsModelsLabelwiseRocCurves(
  modelId,
  searchQuery,
  threshold,
  source,
  projectId,
  labels,
  offset = 0,
  sortOrder = "ascending",
  includeModelAverage = FALSE,
  limit = 100,
  sortBy = "label",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve labelwise ROC curves for model and given source.
}
\details{
Retrieve labelwise ROC curves for model and given source.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/labelwiseRocCurves/\{source\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsLabelwiseRocCurvesList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
