% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsLiftChart}
\alias{ListProjectsModelsLiftChart}
\title{ListProjectsModelsLiftChart}
\usage{
ListProjectsModelsLiftChart(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve all available lift charts for model.
}
\details{
Retrieve all available lift charts for model. The response will include a json list of all available lift charts, in the same format as the response from :\url{http:get}:/api/v2/projects/(projectId)/models/(modelId)/liftChart/(source)/.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/liftChart/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsLiftChartList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
