% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{ListProjectsModelsLogs}
\alias{ListProjectsModelsLogs}
\title{ListProjectsModelsLogs}
\usage{
ListProjectsModelsLogs(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve an archive (tar.gz) of the logs produced and persisted by a model.
}
\details{
Retrieve an archive (tar.gz) of the logs produced and persisted by a model. Note that only blueprints with custom tasks create persistent logs - this will not work with any other type of model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/logs/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsLogsList in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.
}
