% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsMulticlassFeatureImpact}
\alias{ListProjectsModelsMulticlassFeatureImpact}
\title{ListProjectsModelsMulticlassFeatureImpact}
\usage{
ListProjectsModelsMulticlassFeatureImpact(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve feature impact scores for each class in a multiclass model.
}
\details{
Retrieve feature impact scores for each class in a multiclass model. Feature Impact is computed for each column by creating new data with that column randomly permuted (but the others left unchanged), and seeing how the error metric score for the predictions is affected. Elsewhere this technique is sometimes called 'Permutation Importance'. The impactUnnormalized is how much worse the error metric score is when making predictions on this modified data. The impactNormalized is normalized so that the largest value is 1. In both cases, larger values indicate more important features.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/multiclassFeatureImpact/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsMulticlassFeatureImpactList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
