% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsModelsPredictionIntervals}
\alias{ListProjectsModelsPredictionIntervals}
\title{ListProjectsModelsPredictionIntervals}
\usage{
ListProjectsModelsPredictionIntervals(
  modelId,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve prediction intervals that are already calculated for this model.
}
\details{
Retrieve prediction intervals (in descending order) that are already calculated for this model. Note that the project this model belongs to must be a time series project.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/predictionIntervals/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsPredictionIntervalsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
