% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{ListProjectsModelsTrainingArtifact}
\alias{ListProjectsModelsTrainingArtifact}
\title{ListProjectsModelsTrainingArtifact}
\usage{
ListProjectsModelsTrainingArtifact(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve an archive (tar.gz) of the artifacts produced and persisted by a model.
}
\details{
Retrieve an archive (tar.gz) of the artifacts produced and persisted by a model. Note that only blueprints with custom tasks create these artifacts - this will not work with any other type of model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/trainingArtifact/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsTrainingArtifactList in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.
}
