% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsMulticategoricalInvalidFormat}
\alias{ListProjectsMulticategoricalInvalidFormat}
\title{ListProjectsMulticategoricalInvalidFormat}
\usage{
ListProjectsMulticategoricalInvalidFormat(projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve multicategorical data quality log.
}
\details{
Retrieve multicategorical data quality log.

This method invokes \verb{GET /projects/\{projectId\}/multicategoricalInvalidFormat/} in the DataRobot Public API.
}
\seealso{
The method ProjectsMulticategoricalInvalidFormatList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
