% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsMultiseriesNames}
\alias{ListProjectsMultiseriesNames}
\title{ListProjectsMultiseriesNames}
\usage{
ListProjectsMultiseriesNames(projectId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List the names of a multiseries project.
}
\details{
List the individual series names of a multiseries project

This method invokes \verb{GET /projects/\{projectId\}/multiseriesNames/} in the DataRobot Public API.
}
\seealso{
The method MultiseriesNamesList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
