% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_partitioning_apiWrapper.R
\name{ListProjectsOptimizedDatetimePartitionings}
\alias{ListProjectsOptimizedDatetimePartitionings}
\title{ListProjectsOptimizedDatetimePartitionings}
\usage{
ListProjectsOptimizedDatetimePartitionings(
  projectId,
  offset = 0,
  limit = 10,
  ...
)
}
\description{
List all created optimized datetime partitioning configurations
}
\details{
List all created optimized datetime partitioning configurations
}
\seealso{
The method ProjectsOptimizedDatetimePartitioningsList in [datarobot.apicore::DatetimePartitioningApi], which this function wraps.

Other DatetimePartitioning: 
\code{\link{CreateProjectsDatetimePartitioning}()},
\code{\link{CreateProjectsOptimizedDatetimePartitionings}()},
\code{\link{DownloadTimeSeriesFeatureDerivationLog}()},
\code{\link{GetDatetimePartition}()},
\code{\link{GetTimeSeriesFeatureDerivationLog}()},
\code{\link{RetrieveProjectsOptimizedDatetimePartitionings}()}
}
\concept{DatetimePartitioning}
