% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListProjectsPredictionsMetadata}
\alias{ListProjectsPredictionsMetadata}
\title{ListProjectsPredictionsMetadata}
\usage{
ListProjectsPredictionsMetadata(
  predictionDatasetId,
  modelId,
  projectId,
  offset = 0,
  limit = 1000,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of prediction metadata records.
}
\details{
Use the ID of a metadata object to get the complete set of predictions.

This method invokes \verb{GET /projects/\{projectId\}/predictionsMetadata/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionsMetadataList in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
