% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsSecondaryDatasetsConfigurations}
\alias{ListProjectsSecondaryDatasetsConfigurations}
\title{ListProjectsSecondaryDatasetsConfigurations}
\usage{
ListProjectsSecondaryDatasetsConfigurations(
  modelId,
  featurelistId,
  projectId,
  offset = 0,
  limit = 100,
  includeDeleted = "false",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all secondary dataset configurations for a project
}
\details{
List all secondary dataset configurations for a project, optionally filtered by feature list id.

This method invokes \verb{GET /projects/\{projectId\}/secondaryDatasetsConfigurations/} in the DataRobot Public API.
}
\seealso{
The method SecondaryDatasetsConfigurationsList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
