% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsSegmentationTasks}
\alias{ListProjectsSegmentationTasks}
\title{ListProjectsSegmentationTasks}
\usage{
ListProjectsSegmentationTasks(projectId, offset = 0, limit = 10, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List segmentation tasks.
}
\details{
List all segmentation tasks created for the project.

This method invokes \verb{GET /projects/\{projectId\}/segmentationTasks/} in the DataRobot Public API.
}
\seealso{
The method SegmentationTasksList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
