% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RatingTables.R
\name{ListRatingTableModels}
\alias{ListRatingTableModels}
\title{ListRatingTableModels}
\usage{
ListRatingTableModels(
  project,
  bulkOperationId = NULL,
  orderBy = NULL,
  name = NULL,
  samplePct = NULL,
  withMetric = NULL,
  isStarred = NULL,
  ...
)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
data.frame containing information about each model with a rating table in a
  project (one row per model with a rating table).
}
\description{
List Rating Table Models
}
\details{
Lists all the models from a project that have rating tables

This method invokes \verb{GET /projects/\{projectId\}/ratingTableModels/} in the DataRobot Public API.

Retrieve information about all DataRobot models with a rating table.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
ListRatingTableModels(projectId)
}
}
\seealso{
The method ProjectsRatingTableModelsList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
