% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RatingTables.R
\name{ListRatingTables}
\alias{ListRatingTables}
\title{ListRatingTables}
\usage{
ListRatingTables(
  project,
  parentModelId = NULL,
  modelId = NULL,
  offset = 0,
  limit = 0,
  ...
)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
data.frame containing information about each rating table in a
  project (one row per model with a rating table).
}
\description{
List Rating Tables For The Project
}
\details{
List RatingTable objects for a project.   These contain metadata about the rating table and the location at which the corresponding rating   table file can be retrieved.

This method invokes \verb{GET /projects/\{projectId\}/ratingTables/} in the DataRobot Public API.

Retrieve information about all rating tables.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
ListRatingTables(projectId)
}
}
\seealso{
The method ProjectsRatingTablesList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
