% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_configuration_apiWrapper.R
\name{ListSsoConfigurations}
\alias{ListSsoConfigurations}
\title{ListSsoConfigurations}
\usage{
ListSsoConfigurations(orgId, offset = 0, limit = 100, ...)
}
\description{
List sso configurations.
}
\details{
List the sso configurations that correspond to provided conditions.
}
\seealso{
The method SsoConfigurationsList in [datarobot.apicore::SsoConfigurationApi], which this function wraps.

Other SsoConfiguration: 
\code{\link{CreateSsoConfigurations}()},
\code{\link{PatchSsoConfigurations}()},
\code{\link{RetrieveSsoConfigurations}()}
}
\concept{SsoConfiguration}
