% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_apiWrapper.R
\name{ListStatus}
\alias{ListStatus}
\title{ListStatus}
\usage{
ListStatus(offset = 0, limit = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List tasks
}
\details{
List the async tasks that are currently running for your account

This method invokes \verb{GET /status/} in the DataRobot Public API.
}
\seealso{
The method StatusList in \link[datarobot.apicore:JobsApi]{datarobot.apicore::JobsApi}, which invokes the same underlying API endpoint.
}
