% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R,
%   R/TrainingPredictions.R
\name{ListTrainingPredictions}
\alias{ListTrainingPredictions}
\title{ListTrainingPredictions}
\usage{
ListTrainingPredictions(project, offset = 0, limit = 0, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
data.frame containing information about each training prediction.
}
\description{
List training prediction jobs
}
\details{
Get a list of training prediction records

This method invokes \verb{GET /projects/\{projectId\}/trainingPredictions/} in the DataRobot Public API.

Retrieve information about all training prediction datasets in a project.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
ListTrainingPredictions(projectId)
}
}
\seealso{
The method TrainingPredictionsList in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
