% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListUsageDataExportsSupportedEvents}
\alias{ListUsageDataExportsSupportedEvents}
\title{ListUsageDataExportsSupportedEvents}
\usage{
ListUsageDataExportsSupportedEvents(...)
}
\description{
Describe supported available audit events with which to filter result data.
}
\details{
Describe supported available audit events with which to filter result data.
}
\seealso{
The method UsageDataExportsSupportedEventsList in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{CreateUsageDataExports}()},
\code{\link{DeleteManyUsersRateLimitUsage}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsEvents}()},
\code{\link{ListEventLogsPredictionUsage}()},
\code{\link{ListEventLogs}()},
\code{\link{ListUsersRateLimitUsage}()},
\code{\link{RetrieveEventLogs}()},
\code{\link{RetrieveUsageDataExports}()}
}
\concept{Analytics}
