% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCases}
\alias{ListUseCases}
\title{ListUseCases}
\usage{
ListUseCases(stage, orderBy, namePart, offset = 0, limit = 100, ...)
}
\description{
List use cases the requesting user has access to.
}
\details{
List use cases the requesting user has access to.
}
\seealso{
The method UseCasesList in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{CreateUseCases}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplatesCalculation}()},
\code{\link{ListUseCaseValueTemplates}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesAttachments}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCasesSharedRoles}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{PatchUseCases}()},
\code{\link{RetrieveUseCaseValueTemplates}()},
\code{\link{RetrieveUseCasesAttachments}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
