% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCasesAttachments}
\alias{ListUseCasesAttachments}
\title{ListUseCasesAttachments}
\usage{
ListUseCasesAttachments(useCaseId, type, offset = 0, limit = 100, ...)
}
\description{
Get a list of resources attached to this use case.
}
\details{
Get a list of resources attached to this use case.
}
\seealso{
The method UseCasesAttachmentsList in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{CreateUseCases}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplatesCalculation}()},
\code{\link{ListUseCaseValueTemplates}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCasesSharedRoles}()},
\code{\link{ListUseCases}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{PatchUseCases}()},
\code{\link{RetrieveUseCaseValueTemplates}()},
\code{\link{RetrieveUseCasesAttachments}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
