% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCasesSharedRoles}
\alias{ListUseCasesSharedRoles}
\title{ListUseCasesSharedRoles}
\usage{
ListUseCasesSharedRoles(
  shareRecipientType,
  useCaseId,
  name,
  id,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of users, groups and organizations that have an access to this use case
}
\details{
Get a list of users, groups and organizations that have an access to this use case

This method invokes \verb{GET /useCases/\{useCaseId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method UseCasesSharedRolesList in \link[datarobot.apicore:UseCaseApi]{datarobot.apicore::UseCaseApi}, which invokes the same underlying API endpoint.
}
