% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCaseValueTemplates}
\alias{ListUseCaseValueTemplates}
\title{ListUseCaseValueTemplates}
\usage{
ListUseCaseValueTemplates(...)
}
\description{
List available use cases templates.
}
\details{
List available use cases templates.
}
\seealso{
The method UseCaseValueTemplatesList in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{CreateUseCases}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplatesCalculation}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesAttachments}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCasesSharedRoles}()},
\code{\link{ListUseCases}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{PatchUseCases}()},
\code{\link{RetrieveUseCaseValueTemplates}()},
\code{\link{RetrieveUseCasesAttachments}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
