% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCaseValueTemplatesCalculation}
\alias{ListUseCaseValueTemplatesCalculation}
\title{ListUseCaseValueTemplatesCalculation}
\usage{
ListUseCaseValueTemplatesCalculation(
  incorrectDecisionsCount,
  templateType,
  incorrectDecisionCost,
  accuracyImprovement,
  targetValue,
  decisionsCount,
  ...
)
}
\description{
Calculate value of template with given template parameters.
}
\details{
Calculate value of template with given template parameters.
}
\seealso{
The method UseCaseValueTemplatesCalculationList in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{CreateUseCases}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplates}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesAttachments}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCasesSharedRoles}()},
\code{\link{ListUseCases}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{PatchUseCases}()},
\code{\link{RetrieveUseCaseValueTemplates}()},
\code{\link{RetrieveUseCasesAttachments}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
