% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListUserCleanupPreviewsStatistics}
\alias{ListUserCleanupPreviewsStatistics}
\title{ListUserCleanupPreviewsStatistics}
\usage{
ListUserCleanupPreviewsStatistics(reportId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Users permanent delete preview statistics.
}
\details{
Retrieve users permanent delete preview statistics.

This method invokes \verb{GET /userCleanupPreviews/\{reportId\}/statistics/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupPreviewsStatisticsList in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
