% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListUserCleanupSummariesStatistics}
\alias{ListUserCleanupSummariesStatistics}
\title{ListUserCleanupSummariesStatistics}
\usage{
ListUserCleanupSummariesStatistics(reportId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Users permanent delete summary statistics.
}
\details{
Retrieve users permanent delete summary statistics.

This method invokes \verb{GET /userCleanupSummaries/\{reportId\}/statistics/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupSummariesStatisticsList in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
