% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{ListUserNotifications}
\alias{ListUserNotifications}
\title{ListUserNotifications}
\usage{
ListUserNotifications(isRead, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List user notifications
}
\details{
Returns list of user's notifications from latest to oldest.

This method invokes \verb{GET /userNotifications/} in the DataRobot Public API.
}
\seealso{
The method UserNotificationsList in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
