% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListUsersLimits}
\alias{ListUsersLimits}
\title{ListUsersLimits}
\usage{
ListUsersLimits(userId, ...)
}
\description{
Get the rate limits and account limits for a user
}
\details{
User limits are those limits on an account that can be updated per-user. Rate limits that can't be modified per-user are not returned here.
}
\seealso{
The method UsersLimitsList in [datarobot.apicore::UserManagementApi], which this function wraps.

Other UserManagement: 
\code{\link{CreateGroupsUsers}()},
\code{\link{CreateGroups}()},
\code{\link{CreateOrganizationsUsers}()},
\code{\link{CreateUserCleanupJobs}()},
\code{\link{CreateUserCleanupPreviewJobs}()},
\code{\link{CreateUsers}()},
\code{\link{DeleteGroups}()},
\code{\link{DeleteManyGroupsUsers}()},
\code{\link{DeleteManyGroups}()},
\code{\link{DeleteOrganizationsUsers}()},
\code{\link{DeleteUserCleanupJobs}()},
\code{\link{DeleteUserCleanupPreviewJobs}()},
\code{\link{DeleteUserCleanupPreviews}()},
\code{\link{DeleteUserCleanupSummaries}()},
\code{\link{ListGroupsUsers}()},
\code{\link{ListOrganizationsJobs}()},
\code{\link{ListOrganizationsUsers}()},
\code{\link{ListOrganizations}()},
\code{\link{ListUserCleanupPreviewsContent}()},
\code{\link{ListUserCleanupPreviewsDeleteParams}()},
\code{\link{ListUserCleanupPreviewsStatistics}()},
\code{\link{ListUserCleanupSummariesContent}()},
\code{\link{ListUserCleanupSummariesDeleteParams}()},
\code{\link{ListUserCleanupSummariesStatistics}()},
\code{\link{PatchGroups}()},
\code{\link{PatchManyUsersLimits}()},
\code{\link{PatchOrganizationsUsers}()},
\code{\link{RetrieveGroups}()},
\code{\link{RetrieveOrganizationsUsers}()},
\code{\link{RetrieveOrganizations}()},
\code{\link{RetrieveUserCleanupJobs}()},
\code{\link{RetrieveUserCleanupPreviewJobs}()}
}
\concept{UserManagement}
