% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListUsersRateLimitUsage}
\alias{ListUsersRateLimitUsage}
\title{ListUsersRateLimitUsage}
\usage{
ListUsersRateLimitUsage(userId, ...)
}
\description{
List resource usage for a user
}
\details{
List the rate limit resource usage for a user. The usage array returned will have one object corresponding to each rate limit applied to the user.
}
\seealso{
The method UsersRateLimitUsageList in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{CreateUsageDataExports}()},
\code{\link{DeleteManyUsersRateLimitUsage}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsEvents}()},
\code{\link{ListEventLogsPredictionUsage}()},
\code{\link{ListEventLogs}()},
\code{\link{ListUsageDataExportsSupportedEvents}()},
\code{\link{RetrieveEventLogs}()},
\code{\link{RetrieveUsageDataExports}()}
}
\concept{Analytics}
