% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewWebModel.R
\name{MakeUrl}
\alias{MakeUrl}
\title{MakeUrl}
\usage{
MakeUrl(parsedUrl, projectId = NULL, model = NULL)
}
\arguments{
\item{parsedUrl}{list. List returned by httr::parse_url}

\item{projectId}{character. A project ID}

\item{model}{dataRobotModel. A DataRobot model object}
}
\value{
URL string
}
\description{
MakeUrl
}
\seealso{
\link[httr::parse_url]{https://www.rdocumentation.org/packages/httr/versions/1.4.2/topics/parse_url}
}
