% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{ModelCapability}
\alias{ModelCapability}
\title{Model capabilities}
\format{
An object of class \code{list} of length 12.
}
\usage{
ModelCapability
}
\description{
For usage, see `\code{GetModelCapabilities}`.
}
\keyword{datasets}
