% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{ModelDeploymentStatus}
\alias{ModelDeploymentStatus}
\title{Deployment statuses for model deployments.}
\format{An object of class \code{list} of length 3.}
\usage{
ModelDeploymentStatus
}
\description{
Deployment statuses for model deployments.
}
\keyword{datasets}
