% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{ModelReplacementReason}
\alias{ModelReplacementReason}
\title{Model replacement reason}
\format{An object of class \code{list} of length 6.}
\usage{
ModelReplacementReason
}
\description{
Model replacement reason
}
\keyword{datasets}
