% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{MulticlassDeploymentAccuracyMetric}
\alias{MulticlassDeploymentAccuracyMetric}
\title{Accuracy metrics for multiclass deployments}
\format{
An object of class \code{list} of length 3.
}
\usage{
MulticlassDeploymentAccuracyMetric
}
\description{
Added in DataRobot API 2.23.
}
\keyword{datasets}
