% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeUtils.R
\name{parseRFC3339Timestamp}
\alias{parseRFC3339Timestamp}
\title{parseRFC3339Timestamp}
\usage{
parseRFC3339Timestamp(timestampstring)
}
\arguments{
\item{timestampstring}{character. Timestamp in RFC 3339 format.}
}
\value{
The input timestamp as a POSIXt
}
\description{
The DataRobot APIs returns dates in RFC 3339 format.
}
\seealso{
Other API datetime functions: 
\code{\link{RFC3339DateTimeFormat}},
\code{\link{formatRFC3339Timestamp}()},
\code{\link{transformRFC3339Period}()},
\code{\link{validateReportingPeriodTime}()}
}
\concept{API datetime functions}
