% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{PatchBatchPredictions}
\alias{PatchBatchPredictions}
\title{PatchBatchPredictions}
\usage{
PatchBatchPredictions(
  predictionJobId,
  scoredRows,
  jobIntakeSize,
  hidden,
  skippedRows,
  aborted,
  jobOutputSize,
  failedRows,
  started,
  completed,
  logs,
  status,
  partNumber = 0,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update a Batch Prediction job
}
\details{
If a job has finished execution regardless of the result, it can have parameters changed to ensure better filtering in the job list upon retrieval. Another case: updating job scoring status externally.

This method invokes \verb{PATCH /batchPredictions/\{predictionJobId\}/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsPatch in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
