% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchCatalogItems}
\alias{PatchCatalogItems}
\title{PatchCatalogItems}
\usage{
PatchCatalogItems(catalogId, name, description, tags, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update the name, description, or tags for the requested catalog item.
}
\details{
Update the name, description, or tags for the requested catalog item.

This method invokes \verb{PATCH /catalogItems/\{catalogId\}/} in the DataRobot Public API.
}
\seealso{
The method CatalogItemsPatch in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
