% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{PatchComments}
\alias{PatchComments}
\title{PatchComments}
\usage{
PatchComments(mentions, commentId, content, ...)
}
\description{
Update a comment
}
\details{
Update a comment
}
\seealso{
The method CommentsPatch in [datarobot.apicore::CommentsApi], which this function wraps.

Other Comments: 
\code{\link{CreateComments}()},
\code{\link{DeleteComments}()},
\code{\link{ListComments}()},
\code{\link{RetrieveComments}()}
}
\concept{Comments}
