% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{PatchCredentials}
\alias{PatchCredentials}
\title{PatchCredentials}
\usage{
PatchCredentials(
  gcpKey,
  clientId,
  oauthIssuerUrl,
  awsAccessKeyId,
  oauthIssuerType,
  description,
  oauthScopes,
  token,
  password,
  name,
  credentialId,
  clientSecret,
  snowflakeAccountName,
  user,
  azureConnectionString,
  awsSessionToken,
  refreshToken,
  awsSecretAccessKey,
  ...
)
}
\description{
Update specified credentials
}
\details{
Update specified credentials
}
\seealso{
The method CredentialsPatch in [datarobot.apicore::CredentialsApi], which this function wraps.

Other Credentials: 
\code{\link{CreateCredentials}()},
\code{\link{DeleteCredentials}()},
\code{\link{ListCredentialsAssociations}()},
\code{\link{ListCredentials}()},
\code{\link{PatchManyCredentialsAssociations}()},
\code{\link{RetrieveCredentials}()}
}
\concept{Credentials}
