% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{PatchCustomModels}
\alias{PatchCustomModels}
\title{PatchCustomModels}
\usage{
PatchCustomModels(
  networkEgressPolicy,
  negativeClassLabel,
  targetName,
  replicas,
  maximumMemory,
  description,
  language,
  requiresHa,
  classLabels,
  customModelId,
  name,
  positiveClassLabel,
  desiredMemory,
  predictionThreshold = 0.5,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update custom model.
}
\details{
Updates metadata for an existing custom model.  All custom models must support at least one target type (binaryClassification, regression). Custom inference models can only support a single target type.  Setting positiveClassLabel and negativeClassLabel to null will set the labels to their default values (1 and 0 for positiveClassLabel and negativeClassLabel, respectively).  Setting positiveClassLabel, negativeClassLabel, 'targetName\\` is disabled if model has active deployments or assigned training data.

This method invokes \verb{PATCH /customModels/\{customModelId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsPatch in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
