% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{PatchCustomTasks}
\alias{PatchCustomTasks}
\title{PatchCustomTasks}
\usage{
PatchCustomTasks(customTaskId, name, maximumMemory, description, language, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update custom task.
}
\details{
Updates metadata for an existing custom task.  All custom tasks must support one target type (e.g. binaryClassification, regression, transform).  Setting positiveClassLabel and negativeClassLabel to null will set the labels to their default values (1 and 0 for positiveClassLabel and negativeClassLabel, respectively).  Setting positiveClassLabel, negativeClassLabel, targetName is disabled if task has active deployments.

This method invokes \verb{PATCH /customTasks/\{customTaskId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksPatch in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
