% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{PatchDeploymentsRetrainingPoliciesRuns}
\alias{PatchDeploymentsRetrainingPoliciesRuns}
\title{PatchDeploymentsRetrainingPoliciesRuns}
\usage{
PatchDeploymentsRetrainingPoliciesRuns(
  retrainingPolicyId,
  deploymentId,
  runId,
  status,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for updating a single deployment retraining policy run.
}
\details{
Update a single deployment retraining policy run.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/retrainingPolicies/\{retrainingPolicyId\}/runs/\{runId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingPoliciesRunsPatch in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
