% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{PatchGroups}
\alias{PatchGroups}
\title{PatchGroups}
\usage{
PatchGroups(
  accessRoleId,
  maxAppCount,
  maxUploadSize,
  groupId,
  description,
  orgId,
  maxWorkers,
  maxEdaWorkers,
  maxUploadSizeCatalog,
  maxRam,
  name,
  maxCustomDeployments,
  email,
  accountPermissions,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update user group
}
\details{
Update the user group.

This method invokes \verb{PATCH /groups/\{groupId\}/} in the DataRobot Public API.
}
\seealso{
The method GroupsPatch in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
