% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{PatchManyApplicationsAccessControl}
\alias{PatchManyApplicationsAccessControl}
\title{PatchManyApplicationsAccessControl}
\usage{
PatchManyApplicationsAccessControl(data, permissions, applicationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update access control for this application.
}
\details{
Update access control for this application. Request is processed only if updates can be performed on all entries.

This method invokes \verb{PATCH /applications/\{applicationId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsAccessControlPatchMany in \link[datarobot.apicore:ApplicationsApi]{datarobot.apicore::ApplicationsApi}, which invokes the same underlying API endpoint.
}
