% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{PatchManyApplicationsAuthToken}
\alias{PatchManyApplicationsAuthToken}
\title{PatchManyApplicationsAuthToken}
\usage{
PatchManyApplicationsAuthToken(applicationId, ...)
}
\description{
Redeploys application with refreshed auth token.
}
\details{
Redeploys application with refreshed auth token.
}
\seealso{
The method ApplicationsAuthTokenPatchMany in [datarobot.apicore::ApplicationsApi], which this function wraps.

Other Applications: 
\code{\link{CreateApplicationsBuildJobs}()},
\code{\link{CreateApplicationsDeployments}()},
\code{\link{CreateApplicationsDuplicate}()},
\code{\link{CreateApplications}()},
\code{\link{CreateDemoApplications}()},
\code{\link{DeleteApplicationsDeployments}()},
\code{\link{DeleteApplications}()},
\code{\link{ListApplicationTypesUsecases}()},
\code{\link{ListApplicationTypes}()},
\code{\link{ListApplicationsAccessControl}()},
\code{\link{ListApplicationsSharedRoles}()},
\code{\link{ListApplications}()},
\code{\link{PatchApplications}()},
\code{\link{PatchManyApplicationsAccessControl}()},
\code{\link{PatchManyApplicationsDeploymentState}()},
\code{\link{PatchManyApplicationsSharedRoles}()},
\code{\link{PatchManyApplicationsVersion}()},
\code{\link{RetrieveApplicationTypes}()},
\code{\link{RetrieveApplicationUserRole}()},
\code{\link{RetrieveApplications}()}
}
\concept{Applications}
