% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{PatchManyApplicationsDeploymentState}
\alias{PatchManyApplicationsDeploymentState}
\title{PatchManyApplicationsDeploymentState}
\usage{
PatchManyApplicationsDeploymentState(newDeploymentState, applicationId, ...)
}
\description{
Update the application's deployment state.
}
\details{
Update the application's deployment state. Transitioning the app to `deactivated` state will keep the app record in the database, but the app will not be accessible at its URL until changed back to `deployed` state again. When the limit of active apps has been reached, this API can be used to deactivate some applications to make space for new ones.
}
\seealso{
The method ApplicationsDeploymentStatePatchMany in [datarobot.apicore::ApplicationsApi], which this function wraps.

Other Applications: 
\code{\link{CreateApplicationsBuildJobs}()},
\code{\link{CreateApplicationsDeployments}()},
\code{\link{CreateApplicationsDuplicate}()},
\code{\link{CreateApplications}()},
\code{\link{CreateDemoApplications}()},
\code{\link{DeleteApplicationsDeployments}()},
\code{\link{DeleteApplications}()},
\code{\link{ListApplicationTypesUsecases}()},
\code{\link{ListApplicationTypes}()},
\code{\link{ListApplicationsAccessControl}()},
\code{\link{ListApplicationsSharedRoles}()},
\code{\link{ListApplications}()},
\code{\link{PatchApplications}()},
\code{\link{PatchManyApplicationsAccessControl}()},
\code{\link{PatchManyApplicationsAuthToken}()},
\code{\link{PatchManyApplicationsSharedRoles}()},
\code{\link{PatchManyApplicationsVersion}()},
\code{\link{RetrieveApplicationTypes}()},
\code{\link{RetrieveApplicationUserRole}()},
\code{\link{RetrieveApplications}()}
}
\concept{Applications}
