% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{PatchManyCalendarsAccessControl}
\alias{PatchManyCalendarsAccessControl}
\title{PatchManyCalendarsAccessControl}
\usage{
PatchManyCalendarsAccessControl(calendarId, users, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update the access control for this calendar.
}
\details{
Update the access control for this calendar. See the entity sharing documentation \\<sharing\\> for more information.

This method invokes \verb{PATCH /calendars/\{calendarId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method CalendarsAccessControlPatchMany in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
