% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{PatchManyChangeRequestsStatus}
\alias{PatchManyChangeRequestsStatus}
\title{PatchManyChangeRequestsStatus}
\usage{
PatchManyChangeRequestsStatus(changeRequestId, status, ...)
}
\description{
Resolve or Cancel the Change Request.
}
\details{
Resolve or Cancel the Change Request.
}
\seealso{
The method ChangeRequestsStatusPatchMany in [datarobot.apicore::GovernanceApi], which this function wraps.

Other Governance: 
\code{\link{CreateApprovalPolicies}()},
\code{\link{CreateChangeRequestsReviews}()},
\code{\link{CreateChangeRequests}()},
\code{\link{DeleteApprovalPolicies}()},
\code{\link{ListApprovalPoliciesShareableChangeRequests}()},
\code{\link{ListApprovalPolicies}()},
\code{\link{ListApprovalPolicyMatch}()},
\code{\link{ListApprovalPolicyTriggers}()},
\code{\link{ListChangeRequestsReviews}()},
\code{\link{ListChangeRequestsSuggestedReviewers}()},
\code{\link{ListChangeRequests}()},
\code{\link{PatchChangeRequests}()},
\code{\link{PutApprovalPolicies}()},
\code{\link{RetrieveApprovalPolicies}()},
\code{\link{RetrieveChangeRequestsReviews}()},
\code{\link{RetrieveChangeRequests}()}
}
\concept{Governance}
