% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{PatchManyComplianceDocTemplatesSharedRoles}
\alias{PatchManyComplianceDocTemplatesSharedRoles}
\title{PatchManyComplianceDocTemplatesSharedRoles}
\usage{
PatchManyComplianceDocTemplatesSharedRoles(roles, templateId, operation, ...)
}
\description{
Update template's access controls
}
\details{
Set roles for users on this template.
}
\seealso{
The method ComplianceDocTemplatesSharedRolesPatchMany in [datarobot.apicore::DocumentationApi], which this function wraps.

Other Documentation: 
\code{\link{CreateAutomatedDocuments}()},
\code{\link{CreateOneModelComplianceDocsInitializations}()},
\code{\link{DeleteAutomatedDocuments}()},
\code{\link{DeleteComplianceDocTemplate}()},
\code{\link{GetComplianceDocTemplate}()},
\code{\link{ListAutomatedDocumentOptions}()},
\code{\link{ListAutomatedDocuments}()},
\code{\link{ListComplianceDocTemplatesDefault}()},
\code{\link{ListComplianceDocTemplatesSharedRoles}()},
\code{\link{ListComplianceDocTemplates}()},
\code{\link{RetrieveAutomatedDocuments}()},
\code{\link{RetrieveModelComplianceDocsInitializations}()},
\code{\link{UpdateComplianceDocTemplate}()},
\code{\link{UploadComplianceDocTemplate}()}
}
\concept{Documentation}
