% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{PatchManyCredentialsAssociations}
\alias{PatchManyCredentialsAssociations}
\title{PatchManyCredentialsAssociations}
\usage{
PatchManyCredentialsAssociations(
  credentialsToRemove,
  credentialId,
  credentialsToAdd,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add objects associated with credentials
}
\details{
Add objects associated with credentials

This method invokes \verb{PATCH /credentials/\{credentialId\}/associations/} in the DataRobot Public API.
}
\seealso{
The method CredentialsAssociationsPatchMany in \link[datarobot.apicore:CredentialsApi]{datarobot.apicore::CredentialsApi}, which invokes the same underlying API endpoint.
}
