% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{PatchManyCustomModelsTrainingData}
\alias{PatchManyCustomModelsTrainingData}
\title{PatchManyCustomModelsTrainingData}
\usage{
PatchManyCustomModelsTrainingData(
  customModelId,
  datasetName,
  partitionColumn,
  datasetId,
  datasetVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Assign training data to custom model.
}
\details{
Assigns training data from provided dataset to provided Custom Inference Model. For each of custom model's deployments the training data from the specified project is used as baseline, this will enable turning drift tracking on for those deployments.

This method invokes \verb{PATCH /customModels/\{customModelId\}/trainingData/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsTrainingDataPatchMany in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
