% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{PatchManyCustomModelsVersionsFromRepository}
\alias{PatchManyCustomModelsVersionsFromRepository}
\title{PatchManyCustomModelsVersionsFromRepository}
\usage{
PatchManyCustomModelsVersionsFromRepository(
  requiredMetadata,
  ref,
  baseEnvironmentId,
  customModelId,
  repositoryId,
  requiredMetadataValues,
  sourcePath,
  isMajorUpdate = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create custom model version from remote repository with files from previous version.
}
\details{
Create a new custom model version with files added from a remote repository. Files from the previous version of a custom models will be used as a basis.

This method invokes \verb{PATCH /customModels/\{customModelId\}/versions/fromRepository/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsFromRepositoryPatchMany in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
