% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{PatchManyCustomTasksVersionsFromRepository}
\alias{PatchManyCustomTasksVersionsFromRepository}
\title{PatchManyCustomTasksVersionsFromRepository}
\usage{
PatchManyCustomTasksVersionsFromRepository(
  requiredMetadata,
  ref,
  baseEnvironmentId,
  customTaskId,
  repositoryId,
  requiredMetadataValues,
  sourcePath,
  isMajorUpdate = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create custom task version from remote repository with files from previous version.
}
\details{
Create a new custom task version with files added from a remote repository. Files from the previous version of a custom task will be used as a basis.

This method invokes \verb{PATCH /customTasks/\{customTaskId\}/versions/fromRepository/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsFromRepositoryPatchMany in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
