% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{PatchManyDeploymentsSharedRoles}
\alias{PatchManyDeploymentsSharedRoles}
\title{PatchManyDeploymentsSharedRoles}
\usage{
PatchManyDeploymentsSharedRoles(roles, deploymentId, operation, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update model deployment's controls
}
\details{
Set roles for users on this model deployment.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsSharedRolesPatchMany in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
