% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{PatchManyDeploymentsStatus}
\alias{PatchManyDeploymentsStatus}
\title{PatchManyDeploymentsStatus}
\usage{
PatchManyDeploymentsStatus(deploymentId, status, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Change deployment status
}
\details{
Change deployment status.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/status/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsStatusPatchMany in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
