% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{PatchManyProjectsModelsClusterNames}
\alias{PatchManyProjectsModelsClusterNames}
\title{PatchManyProjectsModelsClusterNames}
\usage{
PatchManyProjectsModelsClusterNames(
  modelId,
  clusterNameMappings,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update cluster names assigned to an unsupervised cluster model
}
\details{
Update and then retrieve all cluster names assigned to an unsupervised cluster model

This method invokes \verb{PATCH /projects/\{projectId\}/models/\{modelId\}/clusterNames/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsClusterNamesPatchMany in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
