% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R,
%   R/ai_catalog_apiWrapper.R
\name{PatchManyUserBlueprintsSharedRoles}
\alias{PatchManyUserBlueprintsSharedRoles}
\title{PatchManyUserBlueprintsSharedRoles}
\usage{
PatchManyUserBlueprintsSharedRoles(userBlueprintId, roles, operation, ...)

PatchManyUserBlueprintsSharedRoles(userBlueprintId, roles, operation, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Share a user blueprint with a user, group, or organization

Share a user blueprint with a user, group, or organization
}
\details{
Share a user blueprint with a user, group, or organization

This method invokes \verb{PATCH /userBlueprints/\{userBlueprintId\}/sharedRoles/} in the DataRobot Public API.

Share a user blueprint with a user, group, or organization

This method invokes \verb{PATCH /userBlueprints/\{userBlueprintId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsSharedRolesPatchMany in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsSharedRolesPatchMany in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
