% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{PatchOrganizationsUsers}
\alias{PatchOrganizationsUsers}
\title{PatchOrganizationsUsers}
\usage{
PatchOrganizationsUsers(
  organizationId,
  maxWorkers,
  accessRoleId,
  orgAdmin,
  accessRoleIds,
  userId,
  expirationDate,
  activated,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Patch organization's user
}
\details{
Patch organization's user. Only system or the organization admin can perform this operation.

This method invokes \verb{PATCH /organizations/\{organizationId\}/users/\{userId\}/} in the DataRobot Public API.
}
\seealso{
The method OrganizationsUsersPatch in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
