% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{PatchProjectsSegments}
\alias{PatchProjectsSegments}
\title{PatchProjectsSegments}
\usage{
PatchProjectsSegments(segmentId, projectId, operation = "restart", ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update child segment project.
}
\details{
The only supported operation right now is segment restart, which removes existing child segment project and starts another child project instead for the given segment. Should be only used for child segments which are stuck during project startup or upload.

This method invokes \verb{PATCH /projects/\{projectId\}/segments/\{segmentId\}/} in the DataRobot Public API.
}
\seealso{
The method SegmentsPatch in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
