% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_configuration_apiWrapper.R
\name{PatchSsoConfigurations}
\alias{PatchSsoConfigurations}
\title{PatchSsoConfigurations}
\usage{
PatchSsoConfigurations(
  autoGenerateUsers,
  certificate,
  securityParameters,
  entityId,
  issuer,
  signOnUrl,
  organizationId,
  idpMetadataUrl,
  roleMapping,
  idpMetadataHttpsVerify,
  idpMetadata,
  attributeMapping,
  idpResponseMethod,
  signOutUrl,
  sessionLengthSeconds,
  enableSso,
  groupMapping,
  name,
  enforceSso,
  configurationId,
  spRequestMethod,
  configurationType,
  advancedConfiguration,
  ...
)
}
\description{
Update an SSO configuration for a specific organization.
}
\details{
Update an SSO configuration for a specific organization.
}
\seealso{
The method SsoConfigurationsPatch in [datarobot.apicore::SsoConfigurationApi], which this function wraps.

Other SsoConfiguration: 
\code{\link{CreateSsoConfigurations}()},
\code{\link{ListSsoConfigurations}()},
\code{\link{RetrieveSsoConfigurations}()}
}
\concept{SsoConfiguration}
