% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{PatchUseCases}
\alias{PatchUseCases}
\title{PatchUseCases}
\usage{
PatchUseCases(
  owner,
  feasibility,
  predictionTargets,
  notes,
  stage,
  realizedValue,
  useCaseId,
  name,
  description,
  businessImpact,
  potentialValueTemplate,
  potentialValue,
  targetDates,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update a use case.
}
\details{
Update a use case.

This method invokes \verb{PATCH /useCases/\{useCaseId\}/} in the DataRobot Public API.
}
\seealso{
The method UseCasesPatch in \link[datarobot.apicore:UseCaseApi]{datarobot.apicore::UseCaseApi}, which invokes the same underlying API endpoint.
}
