% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/PauseQueue.R
\name{PauseQueue}
\alias{PauseQueue}
\title{PauseQueue}
\usage{
PauseQueue(project, command = NULL, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Pause or unpause Autopilot
}
\details{
Pause or unpause the autopilot for a project.

This method invokes \verb{POST /projects/\{projectId\}/autopilot/} in the DataRobot Public API.

Pause the DataRobot modeling queue

This function pauses the DataRobot modeling queue for a specified project
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
PauseQueue(projectId)
}
}
\seealso{
The method AutopilotCreate in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
