% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{PeriodicityMaxTimeStep}
\alias{PeriodicityMaxTimeStep}
\title{Periodicity max time step}
\format{An object of class \code{numeric} of length 1.}
\usage{
PeriodicityMaxTimeStep
}
\description{
Periodicity max time step
}
\keyword{datasets}
