% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{PeriodicityTimeUnits}
\alias{PeriodicityTimeUnits}
\title{Periodicity time units}
\format{
An object of class \code{list} of length 8.
}
\usage{
PeriodicityTimeUnits
}
\description{
Same as time units, but kept for backwards compatibility.
}
\keyword{datasets}
