% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{PostgreSQLdrivers}
\alias{PostgreSQLdrivers}
\title{PostgreSQL drivers}
\format{
An object of class \code{list} of length 2.
}
\usage{
PostgreSQLdrivers
}
\description{
This is a list that contains the valid values for PostgreSQL drivers.
}
\keyword{datasets}
