% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionDatasets.R
\name{PredictionDatasetFromAsyncUrl}
\alias{PredictionDatasetFromAsyncUrl}
\title{Retrieve prediction dataset info from the dataset creation URL}
\usage{
PredictionDatasetFromAsyncUrl(asyncUrl, maxWait = 600)
}
\arguments{
\item{asyncUrl}{The temporary status URL}

\item{maxWait}{The maximum time to wait (in seconds) for creation before aborting.}
}
\description{
If dataset creation times out, the error message includes a URL corresponding to the creation
task. That URL can be passed to this function (which will return the completed dataset info
details when finished) to resume waiting for creation.
}
