% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{PrimeLanguage}
\alias{PrimeLanguage}
\title{Prime Language}
\format{An object of class \code{list} of length 2.}
\usage{
PrimeLanguage
}
\description{
This is a list that contains the valid values for downloadable code programming languages.
}
\keyword{datasets}
