% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupProject.R
\name{ProjectFromJobResponse}
\alias{ProjectFromJobResponse}
\title{Retrieve a project from the job response, which has a project-creation URL}
\usage{
ProjectFromJobResponse(jobResponse, maxWait = 600)
}
\arguments{
\item{jobResponse}{An HTTP POST response that includes a redirect to the
temporary status URL.}

\item{maxWait}{The maximum time to wait (in seconds) for project creation before aborting.}
}
\description{
If project creation times out, the error message includes a URL corresponding to the project
creation task. That URL can be passed to this function (which will return the completed project
details when finished) to resume waiting for project creation.
}
