% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{PutApprovalPolicies}
\alias{PutApprovalPolicies}
\title{PutApprovalPolicies}
\usage{
PutApprovalPolicies(
  automaticAction,
  review,
  name,
  trigger,
  approvalPolicyId,
  active = TRUE,
  ...
)
}
\description{
Update an Approval Policy.
}
\details{
Update the policy with the given ID.
}
\seealso{
The method ApprovalPoliciesPut in [datarobot.apicore::GovernanceApi], which this function wraps.

Other Governance: 
\code{\link{CreateApprovalPolicies}()},
\code{\link{CreateChangeRequestsReviews}()},
\code{\link{CreateChangeRequests}()},
\code{\link{DeleteApprovalPolicies}()},
\code{\link{ListApprovalPoliciesShareableChangeRequests}()},
\code{\link{ListApprovalPolicies}()},
\code{\link{ListApprovalPolicyMatch}()},
\code{\link{ListApprovalPolicyTriggers}()},
\code{\link{ListChangeRequestsReviews}()},
\code{\link{ListChangeRequestsSuggestedReviewers}()},
\code{\link{ListChangeRequests}()},
\code{\link{PatchChangeRequests}()},
\code{\link{PatchManyChangeRequestsStatus}()},
\code{\link{RetrieveApprovalPolicies}()},
\code{\link{RetrieveChangeRequestsReviews}()},
\code{\link{RetrieveChangeRequests}()}
}
\concept{Governance}
