% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{PutBatchPredictionsCsvUploadPart}
\alias{PutBatchPredictionsCsvUploadPart}
\title{PutBatchPredictionsCsvUploadPart}
\usage{
PutBatchPredictionsCsvUploadPart(predictionJobId, partNumber = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Upload CSV data in multiple parts
}
\details{
Stream CSV data to the prediction job in many parts.Only available for jobs that uses the localFile intake option.

This method invokes \verb{PUT /batchPredictions/\{predictionJobId\}/csvUpload/part/\{partNumber\}/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsCsvUploadPartPut in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
