% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{PutManyBatchPredictionsCsvUpload}
\alias{PutManyBatchPredictionsCsvUpload}
\title{PutManyBatchPredictionsCsvUpload}
\usage{
PutManyBatchPredictionsCsvUpload(predictionJobId, body, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Creates a new_model_id Batch Prediction job
}
\details{
Stream CSV data to the prediction job. Only available for jobs thatuses the localFile intake option.

This method invokes \verb{PUT /batchPredictions/\{predictionJobId\}/csvUpload/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsCsvUploadPutMany in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
