% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_apiWrapper.R
\name{PutManyClusterLicense}
\alias{PutManyClusterLicense}
\title{PutManyClusterLicense}
\usage{
PutManyClusterLicense(licenseKey, ...)
}
\description{
Create or replace license
}
\details{
Create or replace the cluster license.
}
\seealso{
The method ClusterLicensePutMany in [datarobot.apicore::InfrastructureApi], which this function wraps.

Other Infrastructure: 
\code{\link{CreateClusterLicenseValidation}()},
\code{\link{GetServerVersion}()},
\code{\link{ListClusterLicense}()}
}
\concept{Infrastructure}
