% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{PutNotificationChannels}
\alias{PutNotificationChannels}
\title{PutNotificationChannels}
\usage{
PutNotificationChannels(
  emailAddress,
  secretToken,
  name,
  validateSsl,
  languageCode,
  contentType,
  customHeaders,
  channelId,
  verificationCode,
  payloadUrl,
  ...
)
}
\description{
Update notification channel
}
\details{
Update the notification channel.
}
\seealso{
The method NotificationChannelsPut in [datarobot.apicore::NotificationsApi], which this function wraps.

Other Notifications: 
\code{\link{CreateNotificationChannels}()},
\code{\link{CreateNotificationEmailChannelVerificationStatus}()},
\code{\link{CreateNotificationEmailChannelVerification}()},
\code{\link{CreateNotificationPolicies}()},
\code{\link{CreateNotificationPolicyMutes}()},
\code{\link{CreateNotificationWebhookChannelTests}()},
\code{\link{CreateNotifications}()},
\code{\link{CreateRemoteEvents}()},
\code{\link{DeleteManyUserNotifications}()},
\code{\link{DeleteNotificationChannels}()},
\code{\link{DeleteNotificationPolicies}()},
\code{\link{DeleteNotificationPolicyMutes}()},
\code{\link{DeleteUserNotifications}()},
\code{\link{ListNotificationChannels}()},
\code{\link{ListNotificationEvents}()},
\code{\link{ListNotificationLogs}()},
\code{\link{ListNotificationPolicies}()},
\code{\link{ListNotificationPolicyMutes}()},
\code{\link{ListUserNotifications}()},
\code{\link{PatchManyUserNotifications}()},
\code{\link{PatchUserNotifications}()},
\code{\link{PutNotificationPolicies}()},
\code{\link{RetrieveNotificationChannels}()},
\code{\link{RetrieveNotificationPolicies}()},
\code{\link{RetrieveNotificationWebhookChannelTests}()}
}
\concept{Notifications}
