% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{PutNotificationPolicies}
\alias{PutNotificationPolicies}
\title{PutNotificationPolicies}
\usage{
PutNotificationPolicies(
  policyId,
  name,
  active,
  eventType,
  eventGroup,
  channelId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update notification policy
}
\details{
Update the notification policy.

This method invokes \verb{PUT /notificationPolicies/\{policyId\}/} in the DataRobot Public API.
}
\seealso{
The method NotificationPoliciesPut in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
