% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{PutProjectsPayoffMatrices}
\alias{PutProjectsPayoffMatrices}
\title{PutProjectsPayoffMatrices}
\usage{
PutProjectsPayoffMatrices(
  truePositiveValue,
  trueNegativeValue,
  payoffMatrixId,
  name,
  falsePositiveValue,
  falseNegativeValue,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update a payoff matrix.
}
\details{
Update all fields in a payoff matrix, including values and label.

This method invokes \verb{PUT /projects/\{projectId\}/payoffMatrices/\{payoffMatrixId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPayoffMatricesPut in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
