% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{RecommendedModelType}
\alias{RecommendedModelType}
\title{Recommended model type values}
\format{An object of class \code{list} of length 3.}
\usage{
RecommendedModelType
}
\description{
Recommended model type values
}
\keyword{datasets}
