% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{ReformatListOfModels}
\alias{ReformatListOfModels}
\title{Convert the list-of-lists output from the server to a list-of-modelObjects format.}
\usage{
ReformatListOfModels(listOfLists, projectDetails)
}
\arguments{
\item{listOfLists}{list. The list to reformat.}

\item{projectDetails}{list. Details to reformat with.}
}
\description{
Convert the list-of-lists output from the server to a list-of-modelObjects format.
}
