% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatMetrics.R
\name{ReformatMetrics}
\alias{ReformatMetrics}
\title{replace NULL in $metrics list elements with NA}
\usage{
ReformatMetrics(metricsList)
}
\arguments{
\item{metricsList}{list. List of metrics to reformat.}
}
\description{
replace NULL in $metrics list elements with NA
}
