% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{RegressionDeploymentAccuracyMetric}
\alias{RegressionDeploymentAccuracyMetric}
\title{Accuracy metrics for regression deployments}
\format{
An object of class \code{list} of length 12.
}
\usage{
RegressionDeploymentAccuracyMetric
}
\description{
Added in DataRobot API 2.18.
}
\keyword{datasets}
