% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RatingTables.R
\name{RenameRatingTable}
\alias{RenameRatingTable}
\title{RenameRatingTable}
\usage{
RenameRatingTable(project, ratingTableId, ratingTableName, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{ratingTableId}{character. The ID of the rating table.}

\item{ratingTableName}{character. The new name for the rating table.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An S3 object of class 'dataRobotRatingTable' summarizing all
  available information about the renamed rating table.
}
\description{
Update an uploaded rating table
}
\details{
Rating tables may only be updated if they have not yet been used to create a model.

This method invokes \verb{PATCH /projects/\{projectId\}/ratingTables/\{ratingTableId\}/} in the DataRobot Public API.

Renames a rating table to a different name.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
ratingTableId <- "5984b4d7100d2b31c1166529"
RenameRatingTable(projectId, ratingTableId, "Renamed Table")
}
}
\seealso{
The method ProjectsRatingTablesPatch in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
