% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R, R/Deployment.R
\name{ReplaceDeployedModel}
\alias{ReplaceDeployedModel}
\title{ReplaceDeployedModel}
\usage{
ReplaceDeployedModel(
  deploymentId,
  newModelId,
  replacementReason,
  maxWait = 600,
  ...
)
}
\arguments{
\item{deploymentId}{character. The ID of the deployment.}

\item{newModelId}{character. The ID of the model to use in the deployment. This model
will replace the old model. You can also pass a dataRobotModel object.}

\item{replacementReason}{character. Optional. The reason for replacing the deployment.
See \code{ModelReplacementReason} for a list of reasons.}

\item{maxWait}{integer. How long to wait (in seconds) for the computation to complete
before returning a timeout error? (Default 600 seconds)}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A DataRobotDeployment object containing:
\itemize{
 \item id character. The ID of the deployment.
 \item label character. The label of the deployment.
 \item description character. The description of the deployment.
 \item defaultPredictionServer list. Information on the default prediction
   server connected with the deployment. See \code{ListPredictionServers}
   for details.
 \item model dataRobotModel. The model associated with the deployment.
   See \code{GetModel} for details.
 \item capabilities list. Information on the capabilities of the deployment.
 \item predictionUsage list. Information on the prediction usage of the deployment.
 \item permissions list. User's permissions on the deployment.
 \item serviceHealth list. Information on the service health of the deployment.
 \item modelHealth list. Information on the model health of the deployment.
 \item accuracyHealth list. Information on the accuracy health of the deployment.
}
}
\description{
Model Replacement.
}
\details{
Replace the model used to make predictions for the deployment. A validation process will be performed to make sure the new model is eligible as a replacement. If the validation fails, the model replacement will not occur.The Model Replacement Validation endpoint can be used to confirm the new model is eligible as a replacement.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/model/} in the DataRobot Public API.

Replace a model in a deployment with another model.
}
\examples{
\dontrun{
deploymentId <- "5e319d2e422fbd6b58a5edad"
newModelId <- "5996f820af07fc605e81ead4"
ReplaceDeployedModel(deploymentId, newModelId, ModelReplacementReason$Other)
}
}
\seealso{
The method DeploymentsModelPatchMany in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
