% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RequestApproximation.R
\name{RequestApproximation}
\alias{RequestApproximation}
\title{RequestApproximation}
\usage{
RequestApproximation(project, modelId, body = NULL, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
job Id
}
\description{
Create Rulesets
}
\details{
Approximate an existing model on the leaderboard with DataRobot Prime. A request body should be an empty JSON {}

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/primeRulesets/} in the DataRobot Public API.

Request an approximation of a model using DataRobot Prime

This function will create several rulesets that approximate the specified model.
The code used in the approximation can be downloaded to be run locally.
Currently only Python and Java downloadable code is available

General workflow of creating and downloading Prime code may look like following:
RequestApproximation - create several rulesets that approximate the specified model
GetRulesets - list all rulesets created for the parent model
RequestPrimeModel - create Prime model for specified ruleset (use one of rulesets return by
  GetRulesets)
GetPrimeModelFromJobId - get PrimeModelId using JobId returned by RequestPrimeModel
CreatePrimeCode - create code for one of available Prime models
GetPrimeFileFromJobId - get PrimeFileId using JobId returned by CreatePrimeCode
DownloadPrimeCode - download specified Prime code file
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
RequestApproximation(projectId, modelId)
}
}
\seealso{
The method ProjectsModelsPrimeRulesetsCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
