% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blenders.R, R/models_apiWrapper.R
\name{RequestBlender}
\alias{RequestBlender}
\title{Submit a job for creating blender model. Upon success, the new job will
be added to the end of the queue.}
\usage{
RequestBlender(project, modelsToBlend, blendMethod, ...)

RequestBlender(project, modelsToBlend, blendMethod, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelsToBlend}{character. Vector listing the model Ids to be blended.}

\item{blendMethod}{character. Parameter specifying blending method.
See acceptable values within BlendMethods.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An integer value that can be used as the modelJobId parameter
in subsequent calls to the GetBlenderModelFromJobId function.
}
\description{
This function requests the creation of a blend of several models in specified
DataRobot project. The function also allows the user to specify method used
for blending. This function returns an integer modelJobId value,
which can be used by the GetBlenderModelFromJobId function to return the full
blender model object.

Create a blender from other models using a specified blender method.
}
\details{
Create a blender from other models using a specified blender method. Note: Time Series projects only allow the following blender methods: \"AVG\", \"MED\", \"FORECAST_DISTANCE_ENET\", and \"FORECAST_DISTANCE_AVG\".

This method invokes \verb{POST /projects/\{projectId\}/blenderModels/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelsToBlend <- c("5996f820af07fc605e81ead4", "59a5ce3301e9f0296721c64c")
RequestBlender(projectId, modelsToBlend, "GLM")
}
}
\seealso{
The method ProjectsBlenderModelsCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
