% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{RequestNewDatetimeModel}
\alias{RequestNewDatetimeModel}
\title{Adds a new datetime model of the type specified by the blueprint to a DataRobot project}
\usage{
RequestNewDatetimeModel(
  project,
  blueprint,
  featurelist = NULL,
  trainingRowCount = NULL,
  trainingDuration = NULL,
  timeWindowSamplePct = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  monotonicDecreasingFeaturelistId = NULL
)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{blueprint}{list. A list with at least the following two elements:
blueprintId and projectId.  Note that the individual elements of the
list returned by ListBlueprints are admissible values for this parameter.}

\item{featurelist}{list. A list that contains the element featurelistId that
specifies the featurelist to be used in building the model; if not
specified (i.e., for the default value NULL), the project default
(Informative Features) is used.}

\item{trainingRowCount}{integer. Optional, the number of rows of data
that should be used to train the model. If specified, trainingDuration may not be specified.}

\item{trainingDuration}{character. String (optional) a duration string specifying what
time range the data used to train the model should span.
If specified, trainingRowCount may not be specified.}

\item{timeWindowSamplePct}{integer. Optional. May only be specified when the requested model
is a time window (e.g. duration or start and end dates).
An integer between 1 and 99 indicating the percentage to sample by within the window.
The points kept are determined by a random uniform sample.}

\item{monotonicIncreasingFeaturelistId}{character. Optional. The id of the featurelist
that defines the set of features with a monotonically increasing relationship to the
target. If \code{NULL} (default), the default for the project will be used (if any).
Note that currently there is no way to create a model without monotonic constraints
if there was a project-level default set. If desired, the featurelist itself can
also be passed as this parameter.}

\item{monotonicDecreasingFeaturelistId}{character. Optional. The id of the featurelist
that defines the set of features with a monotonically decreasing relationship to the
target. If \code{NULL}, the default for the project will be used (if any). If empty
(i.e., \code{""}), no such constraints are enforced. Also, if desired, the featurelist
itself can be passed as this parameter.}
}
\value{
An integer value that can be used as the modelJobId parameter
  in subsequent calls to the GetDatetimeModelFromJobId function.
}
\description{
This function requests the creation of a new datetime model in the DataRobot
modeling project defined by the project parameter.  The function also
allows the user to specify alternatives to the project default for
featurelist, samplePct, and scoringType.  This function returns an
integer modelJobId value, which can be used by the GetDatetimeModelFromJobId
function to return the full model object.
}
\details{
Motivation for this function is the fact that some models - e.g., very
complex machine learning models fit to large datasets - may take a long
time to complete.  Splitting the model creation request from model
retrieval in these cases allows the user to perform other interactive R
session tasks between the time the model creation/update request is made
and the time the final model is available.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  blueprints <- ListBlueprints(projectId)
  blueprint <- blueprints[[1]]
  RequestNewDatetimeModel(projectId, blueprint)
}
}
