% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RequestNewModel}
\alias{RequestNewModel}
\title{RequestNewModel}
\usage{
RequestNewModel(
  project,
  blueprint,
  featurelist = NULL,
  samplePct = NULL,
  trainingRowCount = NULL,
  scoringType = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  monotonicDecreasingFeaturelistId = NULL,
  ...
)
}
\description{
Train a new model
}
\details{
Train a new model. To specify the amount of data to use to train the model, use either `samplePct` to express a percentage of the rows of the dataset to use or `trainingRowCount` to express the number of rows to use. If neither `samplePct` or `trainingRowCount` is specified, the model will be trained on the maximum available training data that can be used to train an in-memory model. For projects using smart sampling, samplePct and trainingRowCount will be interpreted as a percent or number of rows of the minority class.   When configuring retraining sample sizes for models in projects with large row counts, DataRobot recommends requesting sample sizes using integer row counts instead of percentages. This is because percentages map to many actual possible row counts and only one of which is the actual sample size for up to validation. For example, if a project has 199,408 rows and you request a 64% sample size, any number of rows between 126,625 rows and 128,618 rows maps to 64% of the data. Using actual integer row counts (or `project.max_training_rows`) avoids ambiguity around how many rows of data you want the model to use.
}
\seealso{
The method ProjectsModelsCreate in [datarobot.apicore::ModelsApi], which this function wraps.

Other Models: 
\code{\link{AddEureqaSolution}()},
\code{\link{CreateCustomInferenceImagesFeatureImpact}()},
\code{\link{CreateCustomModelTests}()},
\code{\link{CreateCustomModelsFromCustomModel}()},
\code{\link{CreateCustomModelsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersionsConversions}()},
\code{\link{CreateCustomModelsVersionsDependencyBuild}()},
\code{\link{CreateCustomModelsVersionsFeatureImpact}()},
\code{\link{CreateCustomModelsVersionsFromRepository}()},
\code{\link{CreateCustomModelsVersionsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersions}()},
\code{\link{CreateCustomModels}()},
\code{\link{CreateCustomTrainingBlueprints}()},
\code{\link{CreateFromLatestCustomModelsVersion}()},
\code{\link{CreateModelExports}()},
\code{\link{CreateModelPackagesArchive}()},
\code{\link{CreateModelPackagesFromLearningModel}()},
\code{\link{CreatePrimeCode}()},
\code{\link{CreateProjectsBlenderBuilders}()},
\code{\link{CreateProjectsDatetimeModelsFromModel}()},
\code{\link{CreateProjectsDeploymentReadyModels}()},
\code{\link{CreateProjectsModelsFromModel}()},
\code{\link{CreateProjectsModelsPredictionIntervals}()},
\code{\link{CreateRatingTable}()},
\code{\link{DeleteCustomModelTests}()},
\code{\link{DeleteCustomModelsVersionsConversions}()},
\code{\link{DeleteCustomModels}()},
\code{\link{DeleteManyCustomModelsVersionsDependencyBuild}()},
\code{\link{DeleteModelJob}()},
\code{\link{DownloadPrimeCode}()},
\code{\link{DownloadRatingTable}()},
\code{\link{DownloadScoringCode}()},
\code{\link{DownloadTransferableModel}()},
\code{\link{GetBlenderModel}()},
\code{\link{GetDatetimeModel}()},
\code{\link{GetFrozenModel}()},
\code{\link{GetMissingValuesReport}()},
\code{\link{GetModelJob}()},
\code{\link{GetModelParameters}()},
\code{\link{GetModel}()},
\code{\link{GetPrimeEligibility}()},
\code{\link{GetPrimeFile}()},
\code{\link{GetPrimeModel}()},
\code{\link{GetRatingTableModel}()},
\code{\link{GetRatingTable}()},
\code{\link{GetRulesets}()},
\code{\link{IsBlenderEligible}()},
\code{\link{ListCustomInferenceImagesFeatureImpact}()},
\code{\link{ListCustomModelDeployments}()},
\code{\link{ListCustomModelLimits}()},
\code{\link{ListCustomModelTestsLog}()},
\code{\link{ListCustomModelTestsTail}()},
\code{\link{ListCustomModelTests}()},
\code{\link{ListCustomModelsAccessControl}()},
\code{\link{ListCustomModelsDownload}()},
\code{\link{ListCustomModelsVersionsConversions}()},
\code{\link{ListCustomModelsVersionsDependencyBuildLog}()},
\code{\link{ListCustomModelsVersionsDependencyBuild}()},
\code{\link{ListCustomModelsVersionsDownload}()},
\code{\link{ListCustomModelsVersionsFeatureImpact}()},
\code{\link{ListCustomModelsVersions}()},
\code{\link{ListCustomModels}()},
\code{\link{ListCustomTrainingBlueprints}()},
\code{\link{ListModelJobs}()},
\code{\link{ListModelPackagesFeatures}()},
\code{\link{ListModelPackagesSharedRoles}()},
\code{\link{ListModelPackages}()},
\code{\link{ListModelRecommendations}()},
\code{\link{ListModels}()},
\code{\link{ListPrimeFiles}()},
\code{\link{ListPrimeModels}()},
\code{\link{ListProjectsBlenderBuilders}()},
\code{\link{ListProjectsBlenderModels}()},
\code{\link{ListProjectsCombinedModelsSegmentsDownload}()},
\code{\link{ListProjectsCombinedModelsSegments}()},
\code{\link{ListProjectsCombinedModels}()},
\code{\link{ListProjectsDatetimeModels}()},
\code{\link{ListProjectsFrozenModels}()},
\code{\link{ListProjectsModelsClusterNames}()},
\code{\link{ListProjectsModelsNumIterationsTrained}()},
\code{\link{ListProjectsModelsPredictionIntervals}()},
\code{\link{ListProjectsRecommendedModelsRecommendedModel}()},
\code{\link{ListRatingTableModels}()},
\code{\link{ListRatingTables}()},
\code{\link{PatchCustomModelsVersions}()},
\code{\link{PatchCustomModels}()},
\code{\link{PatchManyCustomModelsAccessControl}()},
\code{\link{PatchManyCustomModelsTrainingData}()},
\code{\link{PatchManyCustomModelsVersionsFromRepository}()},
\code{\link{PatchManyProjectsModelsClusterNames}()},
\code{\link{PutManyProjectsSegmentChampion}()},
\code{\link{RenameRatingTable}()},
\code{\link{RequestApproximation}()},
\code{\link{RequestBlender}()},
\code{\link{RequestNewDatetimeModel}()},
\code{\link{RequestNewRatingTableModel}()},
\code{\link{RequestPrimeModel}()},
\code{\link{RetrieveCustomModelTests}()},
\code{\link{RetrieveCustomModelsVersionsConversions}()},
\code{\link{RetrieveCustomModelsVersions}()},
\code{\link{RetrieveCustomModels}()},
\code{\link{RetrieveModelPackages}()},
\code{\link{RetrieveProjectsBlenderBuilders}()},
\code{\link{RetrieveProjectsCombinedModels}()},
\code{\link{RetrieveProjectsEureqaDistributionPlot}()},
\code{\link{RetrieveProjectsEureqaModelDetail}()}
}
\concept{Models}
