% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RatingTables.R
\name{RequestNewRatingTableModel}
\alias{RequestNewRatingTableModel}
\title{RequestNewRatingTableModel}
\usage{
RequestNewRatingTableModel(project, ratingTableId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{ratingTableId}{character. The ID of the rating table.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An integer value that can be used as the modelJobId parameter
  in subsequent calls to the GetModelFromJobId function.
}
\description{
Create New Models From A Rating Table
}
\details{
Create a new rating table model from a validated rating table record

This method invokes \verb{POST /projects/\{projectId\}/ratingTableModels/} in the DataRobot Public API.

Create a new model from a rating table.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ratingTableId <- "5984b4d7100d2b31c1166529"
RequestNewModel(projectId, ratingTableId)
}
}
\seealso{
The method ProjectsRatingTableModelsCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
