% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RequestPrimeModel.R
\name{RequestPrimeModel}
\alias{RequestPrimeModel}
\title{RequestPrimeModel}
\usage{
RequestPrimeModel(project, ruleset, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{ruleset}{list. A list specifying ruleset parameters (see GetRulesets)}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
job Id
}
\description{
Create a Prime Model from a Ruleset
}
\details{
Create a Prime model using a particular ruleset.  DataRobot Prime is not available for multiclass projects.  Once rulesets approximating a parent model have been created, using POST /api/v2/projects/(projectId)/models/(modelId)/primeRulesets/, this route will allow creation of a Prime model using one of those rulesets.  Available rulesets can be retrieved via GET /api/v2/projects/(projectId)/models/(modelId)/primeRulesets/.

This method invokes \verb{POST /projects/\{projectId\}/primeModels/} in the DataRobot Public API.

Request training for a DataRobot Prime model using a specified ruleset

Training a model using a ruleset is a necessary prerequisite for being able to download the code
for a ruleset.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
rulesets <- GetRulesets(projectId, modelId)
ruleset <- rulesets[[1]]
RequestPrimeModel(projectId, ruleset)
}
}
\seealso{
The method ProjectsPrimeModelsCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
