% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReasonCodes.R
\name{RequestReasonCodesInitialization}
\alias{RequestReasonCodesInitialization}
\title{Request reason codes initialization for specified model}
\usage{
RequestReasonCodesInitialization(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModelObject, or each element of the list returned by
the function GetAllModels.}
}
\value{
job Id
}
\description{
Reason codes initializations are a prerequisite for computing reason codes, and include
a sample what the computed reason codes for a prediction dataset would look like.
}
