% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferrableModels.R
\name{RequestTransferrableModel}
\alias{RequestTransferrableModel}
\title{Request generation of an transferrable model file for use in an on-premise DataRobot standalone
prediction environment.
This function can only be used if model export is enabled, and will only be useful
if you have an on-premise environment in which to import it.
This function does not download the exported file. Use DownloadTransferrableModel for that.
Example:
jobId<-RequestTransferrrableModel(projectObject, modelId) 
WaitForJobToComplete(projectObject, jobId)
DownloadTransferrableModel(projectObject, modelId, file)}
\usage{
RequestTransferrableModel(project, modelId)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{modelId}{Unique alphanumeric identifier for the model of interest.}
}
\value{
jobId
}
\description{
Request generation of an transferrable model file for use in an on-premise DataRobot standalone
prediction environment.
This function can only be used if model export is enabled, and will only be useful
if you have an on-premise environment in which to import it.
This function does not download the exported file. Use DownloadTransferrableModel for that.
Example:
jobId<-RequestTransferrrableModel(projectObject, modelId) 
WaitForJobToComplete(projectObject, jobId)
DownloadTransferrableModel(projectObject, modelId, file)
}

