% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{RetrieveBatchPredictions}
\alias{RetrieveBatchPredictions}
\title{RetrieveBatchPredictions}
\usage{
RetrieveBatchPredictions(predictionJobId, partNumber = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Batch Prediction job
}
\details{
Retrieve a Batch Prediction job.

This method invokes \verb{GET /batchPredictions/\{predictionJobId\}/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsRetrieve in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
