% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{RetrieveComments}
\alias{RetrieveComments}
\title{RetrieveComments}
\usage{
RetrieveComments(commentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a comment
}
\details{
Retrieve a comment

This method invokes \verb{GET /comments/\{commentId\}/} in the DataRobot Public API.
}
\seealso{
The method CommentsRetrieve in \link[datarobot.apicore:CommentsApi]{datarobot.apicore::CommentsApi}, which invokes the same underlying API endpoint.
}
