% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{RetrieveCredentials}
\alias{RetrieveCredentials}
\title{RetrieveCredentials}
\usage{
RetrieveCredentials(credentialId, ...)
}
\description{
Retrieve the credentials set.
}
\details{
Return the credential set matching the specified ID.
}
\seealso{
The method CredentialsRetrieve in [datarobot.apicore::CredentialsApi], which this function wraps.

Other Credentials: 
\code{\link{CreateCredentials}()},
\code{\link{DeleteCredentials}()},
\code{\link{ListCredentialsAssociations}()},
\code{\link{ListCredentials}()},
\code{\link{PatchCredentials}()},
\code{\link{PatchManyCredentialsAssociations}()}
}
\concept{Credentials}
