% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{RetrieveCustomTasks}
\alias{RetrieveCustomTasks}
\title{RetrieveCustomTasks}
\usage{
RetrieveCustomTasks(customTaskId, ...)
}
\description{
Get custom task.
}
\details{
Retrieve metadata for a custom task
}
\seealso{
The method CustomTasksRetrieve in [datarobot.apicore::CustomTasksApi], which this function wraps.

Other CustomTasks: 
\code{\link{CreateCustomTasksFromCustomTask}()},
\code{\link{CreateCustomTasksVersionsDependencyBuild}()},
\code{\link{CreateCustomTasksVersionsFromRepository}()},
\code{\link{CreateCustomTasksVersions}()},
\code{\link{CreateCustomTasks}()},
\code{\link{CreateFromLatestCustomTaskVersion}()},
\code{\link{DeleteCustomTasks}()},
\code{\link{DeleteManyCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksAccessControl}()},
\code{\link{ListCustomTasksDownload}()},
\code{\link{ListCustomTasksVersionsDependencyBuildLog}()},
\code{\link{ListCustomTasksVersionsDependencyBuild}()},
\code{\link{ListCustomTasksVersionsDownload}()},
\code{\link{ListCustomTasksVersions}()},
\code{\link{ListCustomTasks}()},
\code{\link{PatchCustomTasksVersions}()},
\code{\link{PatchCustomTasks}()},
\code{\link{PatchManyCustomTasksAccessControl}()},
\code{\link{PatchManyCustomTasksVersionsFromRepository}()},
\code{\link{RetrieveCustomTasksVersions}()}
}
\concept{CustomTasks}
