% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveDataEngineQueryGenerators}
\alias{RetrieveDataEngineQueryGenerators}
\title{RetrieveDataEngineQueryGenerators}
\usage{
RetrieveDataEngineQueryGenerators(dataEngineQueryGeneratorId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a data engine query generator given ID.
}
\details{
Retrieve a data engine query generator

This method invokes \verb{GET /dataEngineQueryGenerators/\{dataEngineQueryGeneratorId\}/} in the DataRobot Public API.
}
\seealso{
The method DataEngineQueryGeneratorsRetrieve in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
