% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveDatasets}
\alias{RetrieveDatasets}
\title{RetrieveDatasets}
\usage{
RetrieveDatasets(datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get dataset details
}
\details{
Retrieves the details of the dataset with given ID.

This method invokes \verb{GET /datasets/\{datasetId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsRetrieve in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
