% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveDatasetsFeatureHistograms}
\alias{RetrieveDatasetsFeatureHistograms}
\title{RetrieveDatasetsFeatureHistograms}
\usage{
RetrieveDatasetsFeatureHistograms(
  featureName,
  datasetId,
  key,
  binLimit = 60,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get dataset feature histogram
}
\details{
Get histogram chart data for a specific feature in the specified dataset.

This method invokes \verb{GET /datasets/\{datasetId\}/featureHistograms/\{featureName\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFeatureHistogramsRetrieve in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
