% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveDatasetsVersionsFeaturelists}
\alias{RetrieveDatasetsVersionsFeaturelists}
\title{RetrieveDatasetsVersionsFeaturelists}
\usage{
RetrieveDatasetsVersionsFeaturelists(
  datasetId,
  featurelistId,
  datasetVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get dataset featurelist
}
\details{
Retrieves the specified featurelist of the dataset with given ID and the latest dataset version.

This method invokes \verb{GET /datasets/\{datasetId\}/versions/\{datasetVersionId\}/featurelists/\{featurelistId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsFeaturelistsRetrieve in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.
}
