% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{RetrieveDeploymentsRetrainingPolicies}
\alias{RetrieveDeploymentsRetrainingPolicies}
\title{RetrieveDeploymentsRetrainingPolicies}
\usage{
RetrieveDeploymentsRetrainingPolicies(retrainingPolicyId, deploymentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for fetching a deployment retraining policy.
}
\details{
Retrieve a deployment retraining policy.

This method invokes \verb{GET /deployments/\{deploymentId\}/retrainingPolicies/\{retrainingPolicyId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingPoliciesRetrieve in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
