% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{RetrieveDeploymentsRetrainingPoliciesRuns}
\alias{RetrieveDeploymentsRetrainingPoliciesRuns}
\title{RetrieveDeploymentsRetrainingPoliciesRuns}
\usage{
RetrieveDeploymentsRetrainingPoliciesRuns(
  retrainingPolicyId,
  deploymentId,
  runId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for fetching a single deployment retraining policy run.
}
\details{
Retrieve a single deployment retraining policy run.

This method invokes \verb{GET /deployments/\{deploymentId\}/retrainingPolicies/\{retrainingPolicyId\}/runs/\{runId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingPoliciesRunsRetrieve in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
