% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R,
%   R/RestoreDiscardedFeatures.R
\name{RetrieveDiscardedFeaturesInformation}
\alias{RetrieveDiscardedFeaturesInformation}
\title{RetrieveDiscardedFeaturesInformation}
\usage{
RetrieveDiscardedFeaturesInformation(projectId)

RetrieveDiscardedFeaturesInformation(projectId)
}
\arguments{
\item{projectId}{Character string, the ID of the project.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An object with information:
\itemize{
\item totalRestoreLimit Integer. The total limit indicating how many features can
be restored in this project.
\item remainingRestoreLimit Integer. The remaining available number of the features which can
be restored in this project.
\item features List of character strings giving discarded feature names. They can be used
to pass into \code{RestoreDiscardedFeatures}
\item count Integer. Discarded features count.
}
}
\description{
Get discarded features.
}
\details{
Get features which were discarded during feature reduction process.

This method invokes \verb{GET /projects/\{projectId\}/discardedFeatures/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
discardedInfo <- RetrieveDiscardedFeaturesInformation(projectId)
}
}
\seealso{
The method DiscardedFeaturesList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.

Other Restore Discarded Features functions: 
\code{\link{RestoreDiscardedFeatures}()}
}
\concept{Restore Discarded Features functions}
