% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{RetrieveEventLogs}
\alias{RetrieveEventLogs}
\title{RetrieveEventLogs}
\usage{
RetrieveEventLogs(recordId, includeIdentifyingFields = "true", ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get audit record by ID.
}
\details{
Get audit record by ID.

This method invokes \verb{GET /eventLogs/\{recordId\}/} in the DataRobot Public API.
}
\seealso{
The method EventLogsRetrieve in \link[datarobot.apicore:AnalyticsApi]{datarobot.apicore::AnalyticsApi}, which invokes the same underlying API endpoint.
}
