% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{RetrieveImageAugmentationLists}
\alias{RetrieveImageAugmentationLists}
\title{RetrieveImageAugmentationLists}
\usage{
RetrieveImageAugmentationLists(augmentationId, ...)
}
\description{
Returns a single augmentation list with the specified id.
}
\details{
Returns a single augmentation list with the specified id.
}
\seealso{
The method ImageAugmentationListsRetrieve in [datarobot.apicore::ImagesApi], which this function wraps.

Other Images: 
\code{\link{CreateImageAugmentationListsSamples}()},
\code{\link{CreateImageAugmentationLists}()},
\code{\link{CreateImageAugmentationSamples}()},
\code{\link{CreateProjectsModelsImageActivationMaps}()},
\code{\link{CreateProjectsModelsImageEmbeddings}()},
\code{\link{DeleteImageAugmentationLists}()},
\code{\link{ListDatasetsImagesDataQualityLogFile}()},
\code{\link{ListDatasetsImagesDataQualityLog}()},
\code{\link{ListImageAugmentationListsSamples}()},
\code{\link{ListImageAugmentationLists}()},
\code{\link{ListImageAugmentationSamples}()},
\code{\link{ListProjectsDuplicateImages}()},
\code{\link{ListProjectsImageActivationMaps}()},
\code{\link{ListProjectsImageBins}()},
\code{\link{ListProjectsImageEmbeddings}()},
\code{\link{ListProjectsImageSamples}()},
\code{\link{ListProjectsImagesDataQualityLogFile}()},
\code{\link{ListProjectsImagesDataQualityLog}()},
\code{\link{ListProjectsImagesFile}()},
\code{\link{ListProjectsImages}()},
\code{\link{ListProjectsModelsImageActivationMaps}()},
\code{\link{ListProjectsModelsImageEmbeddings}()},
\code{\link{PatchImageAugmentationLists}()},
\code{\link{RetrieveImageAugmentationOptions}()},
\code{\link{RetrieveProjectsImages}()}
}
\concept{Images}
