% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{RetrieveModelComplianceDocsInitializations}
\alias{RetrieveModelComplianceDocsInitializations}
\title{RetrieveModelComplianceDocsInitializations}
\usage{
RetrieveModelComplianceDocsInitializations(entityId, ...)
}
\description{
Check if compliance documentation pre-processing is initialized for the model
}
\details{
Check if compliance documentation pre-processing is initialized for the current model. This is only required for custom models.
}
\seealso{
The method ModelComplianceDocsInitializationsRetrieve in [datarobot.apicore::DocumentationApi], which this function wraps.

Other Documentation: 
\code{\link{CreateAutomatedDocuments}()},
\code{\link{CreateOneModelComplianceDocsInitializations}()},
\code{\link{DeleteAutomatedDocuments}()},
\code{\link{DeleteComplianceDocTemplate}()},
\code{\link{GetComplianceDocTemplate}()},
\code{\link{ListAutomatedDocumentOptions}()},
\code{\link{ListAutomatedDocuments}()},
\code{\link{ListComplianceDocTemplatesDefault}()},
\code{\link{ListComplianceDocTemplatesSharedRoles}()},
\code{\link{ListComplianceDocTemplates}()},
\code{\link{PatchManyComplianceDocTemplatesSharedRoles}()},
\code{\link{RetrieveAutomatedDocuments}()},
\code{\link{UpdateComplianceDocTemplate}()},
\code{\link{UploadComplianceDocTemplate}()}
}
\concept{Documentation}
