% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveProjectsBatchTypeTransformFeaturesResult}
\alias{RetrieveProjectsBatchTypeTransformFeaturesResult}
\title{RetrieveProjectsBatchTypeTransformFeaturesResult}
\usage{
RetrieveProjectsBatchTypeTransformFeaturesResult(
  variableType,
  jobId,
  parentNames,
  prefix,
  suffix,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the result of a batch variable type transformation.
}
\details{
Retrieve the result of a batch variable type transformation.

This method invokes \verb{GET /projects/\{projectId\}/batchTypeTransformFeaturesResult/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method BatchTypeTransformFeaturesResultRetrieve in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
