% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RetrieveProjectsEureqaModelDetail}
\alias{RetrieveProjectsEureqaModelDetail}
\title{RetrieveProjectsEureqaModelDetail}
\usage{
RetrieveProjectsEureqaModelDetail(solutionId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Eureqa model details plot.
}
\details{
Retrieve Eureqa model details plot.  Available for regression projects only

This method invokes \verb{GET /projects/\{projectId\}/eureqaModelDetail/\{solutionId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsEureqaModelDetailRetrieve in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
