% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs}
\alias{RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs}
\title{RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs}
\usage{
RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs(
  baselineValidationJobId,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Baseline Validation Job
}
\details{
Retrieve information to confirm if the validation job triggered via /api/v2/projects/(projectId)/externalTimeSeriesBaselineDataValidationJobs/ is valid.

This method invokes \verb{GET /projects/\{projectId\}/externalTimeSeriesBaselineDataValidationJobs/\{baselineValidationJobId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalTimeSeriesBaselineDataValidationJobsRetrieve in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
