% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveProjectsFeatureLineages}
\alias{RetrieveProjectsFeatureLineages}
\title{RetrieveProjectsFeatureLineages}
\usage{
RetrieveProjectsFeatureLineages(featureLineageId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Feature Discovery Lineage
}
\details{
Retrieve single Feature Discovery feature lineage.

This method invokes \verb{GET /projects/\{projectId\}/featureLineages/\{featureLineageId\}/} in the DataRobot Public API.
}
\seealso{
The method FeatureLineagesRetrieve in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
