% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{RetrieveProjectsImages}
\alias{RetrieveProjectsImages}
\title{RetrieveProjectsImages}
\usage{
RetrieveProjectsImages(imageId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Returns a single image metadata.
}
\details{
Returns a single image metadata.

This method invokes \verb{GET /projects/\{projectId\}/images/\{imageId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsImagesRetrieve in \link[datarobot.apicore:ImagesApi]{datarobot.apicore::ImagesApi}, which invokes the same underlying API endpoint.
}
