% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveProjectsModelingFeatures}
\alias{RetrieveProjectsModelingFeatures}
\title{RetrieveProjectsModelingFeatures}
\usage{
RetrieveProjectsModelingFeatures(featureName, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve project modeling feature.
}
\details{
Retrieve the specified modeling feature with descriptive information.

This method invokes \verb{GET /projects/\{projectId\}/modelingFeatures/\{featureName\}/} in the DataRobot Public API.
}
\seealso{
The method ModelingFeaturesRetrieve in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
