% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsModelsMulticlassLiftChart}
\alias{RetrieveProjectsModelsMulticlassLiftChart}
\title{RetrieveProjectsModelsMulticlassLiftChart}
\usage{
RetrieveProjectsModelsMulticlassLiftChart(modelId, source, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the multiclass lift chart data from a single source.
}
\details{
Retrieve the multiclass lift chart data from a single source.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/multiclassLiftChart/\{source\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsMulticlassLiftChartRetrieve in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
