% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsModelsMultilabelLiftCharts}
\alias{RetrieveProjectsModelsMultilabelLiftCharts}
\title{RetrieveProjectsModelsMultilabelLiftCharts}
\usage{
RetrieveProjectsModelsMultilabelLiftCharts(
  modelId,
  source,
  projectId,
  labels,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve labelwise lift charts for model and given source.
}
\details{
Retrieve labelwise lift charts for model and given source.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/multilabelLiftCharts/\{source\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsMultilabelLiftChartsRetrieve in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
