% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsModelsResiduals}
\alias{RetrieveProjectsModelsResiduals}
\title{RetrieveProjectsModelsResiduals}
\usage{
RetrieveProjectsModelsResiduals(modelId, source, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the residuals chart data from a single source.
}
\details{
Retrieve the residuals chart data from a single source.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/residuals/\{source\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsResidualsRetrieve in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.
}
