% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveProjectsSegmentationTaskJobResults}
\alias{RetrieveProjectsSegmentationTaskJobResults}
\title{RetrieveProjectsSegmentationTaskJobResults}
\usage{
RetrieveProjectsSegmentationTaskJobResults(segmentationTaskId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve segmentation task statuses.
}
\details{
Retrieve the statuses of segmentation task jobs associated with the ID.

This method invokes \verb{GET /projects/\{projectId\}/segmentationTaskJobResults/\{segmentationTaskId\}/} in the DataRobot Public API.
}
\seealso{
The method SegmentationTaskJobResultsRetrieve in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
