% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_configuration_apiWrapper.R
\name{RetrieveSsoConfigurations}
\alias{RetrieveSsoConfigurations}
\title{RetrieveSsoConfigurations}
\usage{
RetrieveSsoConfigurations(configurationId, ...)
}
\description{
Retrieve SSO configuration of a specific organization.
}
\details{
Retrieve SSO configuration of a specific organization.
}
\seealso{
The method SsoConfigurationsRetrieve in [datarobot.apicore::SsoConfigurationApi], which this function wraps.

Other SsoConfiguration: 
\code{\link{CreateSsoConfigurations}()},
\code{\link{ListSsoConfigurations}()},
\code{\link{PatchSsoConfigurations}()}
}
\concept{SsoConfiguration}
