% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_apiWrapper.R
\name{RetrieveStatus}
\alias{RetrieveStatus}
\title{RetrieveStatus}
\usage{
RetrieveStatus(statusId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get task status
}
\details{
Check the status of an asynchronous task such as project creation

This method invokes \verb{GET /status/\{statusId\}/} in the DataRobot Public API.
}
\seealso{
The method StatusRetrieve in \link[datarobot.apicore:JobsApi]{datarobot.apicore::JobsApi}, which invokes the same underlying API endpoint.
}
