% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{RetrieveUsageDataExports}
\alias{RetrieveUsageDataExports}
\title{RetrieveUsageDataExports}
\usage{
RetrieveUsageDataExports(artifactId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a prepared customer usage data artifact.
}
\details{
Retrieve a prepared customer usage data artifact. Requires \"CAN_ACCESS_USER_ACTIVITY\" permission.

This method invokes \verb{GET /usageDataExports/\{artifactId\}/} in the DataRobot Public API.
}
\seealso{
The method UsageDataExportsRetrieve in \link[datarobot.apicore:AnalyticsApi]{datarobot.apicore::AnalyticsApi}, which invokes the same underlying API endpoint.
}
