% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{RetrieveUseCasesAttachments}
\alias{RetrieveUseCasesAttachments}
\title{RetrieveUseCasesAttachments}
\usage{
RetrieveUseCasesAttachments(useCaseId, attachmentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a resource that is attached to a use case.
}
\details{
Get a resource that is attached to a use case.

This method invokes \verb{GET /useCases/\{useCaseId\}/attachments/\{attachmentId\}/} in the DataRobot Public API.
}
\seealso{
The method UseCasesAttachmentsRetrieve in \link[datarobot.apicore:UseCaseApi]{datarobot.apicore::UseCaseApi}, which invokes the same underlying API endpoint.
}
