% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{RetrieveUseCaseValueTemplates}
\alias{RetrieveUseCaseValueTemplates}
\title{RetrieveUseCaseValueTemplates}
\usage{
RetrieveUseCaseValueTemplates(templateType, ...)
}
\description{
Get an individual use case value template by its name.
}
\details{
Get an individual use case value template by its name.
}
\seealso{
The method UseCaseValueTemplatesRetrieve in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{CreateUseCases}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplatesCalculation}()},
\code{\link{ListUseCaseValueTemplates}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesAttachments}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCasesSharedRoles}()},
\code{\link{ListUseCases}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{PatchUseCases}()},
\code{\link{RetrieveUseCasesAttachments}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
