% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{RetrieveUserCleanupPreviewJobs}
\alias{RetrieveUserCleanupPreviewJobs}
\title{RetrieveUserCleanupPreviewJobs}
\usage{
RetrieveUserCleanupPreviewJobs(statusId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve users perma-delete preview job status.
}
\details{
Get async status of the users perma-delete preview building job.

This method invokes \verb{GET /userCleanupPreviewJobs/\{statusId\}/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupPreviewJobsRetrieve in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
