% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeUtils.R
\docType{data}
\name{RFC3339DateTimeFormat}
\alias{RFC3339DateTimeFormat}
\title{RFC 3339 datetime format}
\format{
An object of class \code{character} of length 1.
}
\usage{
RFC3339DateTimeFormat
}
\description{
The DataRobot API returns dates in RFC 3339 format. Since this comes from a
Python datetime object, we assume that the period returned is in the format
"%Y-%m-%dT%H:%M:%OSZ".
}
\seealso{
Other API datetime functions: 
\code{\link{formatRFC3339Timestamp}()},
\code{\link{parseRFC3339Timestamp}()},
\code{\link{transformRFC3339Period}()},
\code{\link{validateReportingPeriodTime}()}
}
\concept{API datetime functions}
\keyword{datasets}
