% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{SegmentAnalysisAttribute}
\alias{SegmentAnalysisAttribute}
\title{Segment analysis attributes}
\format{
An object of class \code{list} of length 3.
}
\usage{
SegmentAnalysisAttribute
}
\description{
Added in DataRobot API 2.20.
}
\details{
For usage, see \code{GetDeploymentServiceStats}.
}
\keyword{datasets}
