% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{SeriesAggregationType}
\alias{SeriesAggregationType}
\title{Series aggregation type}
\format{An object of class \code{list} of length 2.}
\usage{
SeriesAggregationType
}
\description{
For details, see "Calculating features across series" in the Time Series section of the
DataRobot user guide.
}
\keyword{datasets}
