% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartAutopilot.R
\name{SetTarget}
\alias{SetTarget}
\title{Set the target variable (and by default, start the DataRobot Autopilot)}
\usage{
SetTarget(project, target, metric = NULL, weights = NULL,
  partition = NULL, mode = NULL, seed = NULL, positiveClass = NULL,
  blueprintThreshold = NULL, responseCap = NULL, quickrun = NULL,
  featurelistId = NULL, smartDownsampled = NULL,
  majorityDownsamplingRate = NULL, maxWait = 600)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{target}{character. String giving the name of the response variable
to be predicted by all project models.}

\item{metric}{character. Optional. String specifying the model fitting metric
to be optimized; a list of valid options for this parameter, which depends on
both project and target, may be obtained with the function GetValidMetrics.}

\item{weights}{character. Optional. String specifying the name of the column
from the modeling dataset to be used as weights in model fitting.}

\item{partition}{Optional S3 object of class 'partition' whose elements specify
a valid partitioning scheme.  See help for functions CreateGroupPartition,
CreateRandomPartition, CreateStratifiedPartition, CreateUserPartition
and CreateDatetimePartitionSpecification.}

\item{mode}{character. Optional. Specifies the autopilot mode used to start the
modeling project; valid options are 'auto' (fully automatic,
the current DataRobot default, obtained when mode = NULL), 'manual' and 'quick'}

\item{seed}{integer. Optional. Seed for the random number generator used in
creating random partitions for model fitting.}

\item{positiveClass}{Optional target variable value corresponding to a positive
response in binary classification problems.}

\item{blueprintThreshold}{integer. Optional. The maximum time
(in hours) that any modeling blueprint is allowed to run before being
terminated.}

\item{responseCap}{numeric. Optional. Floating point value, between 0.5 and 1.0,
specifying a capping limit for the response variable. The default value
NULL corresponds to an uncapped response, equivalent to responseCap = 1.0.}

\item{quickrun}{logical. Optional. if TRUE then DR will perform
a quickrun, limiting the number of models evaluated during autopilot.
(quickrun flag is deprecated in 2.4, will be removed in 3.0)}

\item{featurelistId}{numeric. Specifies which feature list to use. If NULL (default),
a default featurelist is used.}

\item{smartDownsampled}{logical. Optional. Whether to use smart downsampling to throw
away excess rows of the majority class. Only applicable to classification and zero-boosted
regression projects.}

\item{majorityDownsamplingRate}{numeric. Optional. Floating point value, between 0.0 and 100.0.
The percentage of the majority rows that should be kept.  Specify only if using smart
downsampling. May not cause the majority class to become smaller than the minority class.}

\item{maxWait}{integer. Specifies how many seconds to wait for the server to finish
analyzing the target and begin the modeling process. If the process takes
longer than this parameter specifies, execution will stop (but the server
will continue to process the request).}
}
\description{
This function sets the target variable for the project defined by
project, starting the process of building models to predict the response
variable target.  Both of these parameters - project and target - are
required and they are sufficient to start a modeling project with
DataRobot default specifications for the other 10 optional parameters.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  SetTarget(projectId, "targetFeature")
  SetTarget(projectId, "targetFeature", metric = "LogLoss")
  SetTarget(projectId, "targetFeature", mode = AutopilotMode$Manual)
}
}
