% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{SharingRole}
\alias{SharingRole}
\title{Sharing role}
\format{
An object of class \code{list} of length 6.
}
\usage{
SharingRole
}
\description{
This is a list that contains the valid values for granting access to other users (see
\code{Share}). If you wish, you can specify access roles using the list values, e.g.,
\code{SharingRole$ReadWrite} instead of typing the string "READ_WRITE". This way you can
benefit from autocomplete and not have to remember the valid options.
}
\details{
\code{Owner} allows any action including deletion.

\code{ReadWrite} or \code{Editor} allows modifications to the state, e.g., renaming
and creating data sources from a data store, but *not* deleting the entity.

\code{ReadOnly} or \code{Consumer} - for data sources, enables creating projects and predictions;
for data stores, allows viewing them only.
}
\keyword{datasets}
