% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{SourceType}
\alias{SourceType}
\title{Source types}
\format{
An object of class \code{list} of length 2.
}
\usage{
SourceType
}
\description{
This is a list that contains the valid values for source type
}
\keyword{datasets}
