% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartAutopilot.R
\name{StartAutopilot}
\alias{StartAutopilot}
\title{(deprecated: Use SetTarget instead)}
\usage{
StartAutopilot(...)
}
\arguments{
\item{...}{arguments to SetTarget}
}
\description{
(deprecated: Use SetTarget instead)
}

