% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/StartAutopilot.R
\name{StartNewAutoPilot}
\alias{StartNewAutoPilot}
\title{StartNewAutoPilot}
\usage{
StartNewAutoPilot(project, featurelistId, mode = AutopilotMode$FullAuto, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{featurelistId}{numeric. Specifies which feature list to use.}

\item{mode}{character. Optional. Specifies the autopilot mode used to start the
modeling project; See \code{AutopilotMode} for valid options; \code{AutopilotMode$FullAuto} is
default.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Start autopilot
}
\details{
Start autopilot on provided featurelist.

This method invokes \verb{POST /projects/\{projectId\}/autopilots/} in the DataRobot Public API.

Starts autopilot on provided featurelist.
Only one autopilot can be running at the time.
That's why any ongoing autopilot on different featurelist will
be halted - modeling jobs in queue would not
be affected but new jobs would not be added to queue by
halted autopilot.

There is an error if autopilot is currently running on or has already
finished running on the provided featurelist and also if project's target was not selected
(via SetTarget).
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
featureList <- CreateFeaturelist(projectId, "myFeaturelist", c("feature1", "feature2"))
featurelistId <- featureList$featurelistId
StartNewAutoPilot(projectId, featurelistId)
}
}
\seealso{
The method AutopilotsCreate in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
