% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartAutopilot.R
\name{StartProject}
\alias{StartProject}
\title{Start a project, set the target, and run autopilot.}
\usage{
StartProject(
  dataSource,
  projectName = NULL,
  target,
  metric = NULL,
  weights = NULL,
  partition = NULL,
  mode = NULL,
  seed = NULL,
  targetType = NULL,
  positiveClass = NULL,
  blueprintThreshold = NULL,
  responseCap = NULL,
  featurelistId = NULL,
  smartDownsampled = NULL,
  majorityDownsamplingRate = NULL,
  scaleoutModelingMode = NULL,
  accuracyOptimizedBlueprints = NULL,
  offset = NULL,
  exposure = NULL,
  eventsCount = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  monotonicDecreasingFeaturelistId = NULL,
  onlyIncludeMonotonicBlueprints = FALSE,
  workerCount = NULL,
  wait = FALSE,
  checkInterval = 20,
  timeout = NULL,
  username = NULL,
  password = NULL,
  verbosity = 1,
  maxWait = 600
)
}
\arguments{
\item{projectName}{character. Optional. String specifying a project name.
The password is encrypted at server side and never saved or stored.}

\item{wait}{logical. If \code{TRUE}, invokes \code{WaitForAutopilot} to block execution until
the autopilot is complete.}

\item{checkInterval}{numeric. Optional. Maximum wait (in seconds) between checks that Autopilot
is finished. Defaults to 20.}

\item{timeout}{numeric. Optional. Time (in seconds) after which to give up (Default is no
timeout). There is an error if Autopilot is not finished before timing out.}

\item{username}{character. The username to use for authentication to the database.}

\item{password}{character. The password to use for authentication to the database.}

\item{verbosity}{numeric. Optional. 0 is silent, 1 or more displays information about progress.
Default is 1.}

\item{maxWait}{integer. The maximum time to wait for each of two steps: (1) The initial
project creation request, and (2) data processing that occurs after receiving the response
to this initial request.}
}
\description{
This function is a convenient shorthand to start a project and set the target.
See \code{SetupProject} and \code{SetTarget}.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
StartProject(iris,
  projectName = "iris",
  target = "Species",
  targetType = TargetType$Multiclass
)
}
}
