% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringify.R
\name{Stringify}
\alias{Stringify}
\title{Convert a function into a single string for DataRobot}
\usage{
Stringify(functionToConvert, dputFile = tempfile())
}
\arguments{
\item{functionToConvert}{function. The function to convert to a string.}

\item{dputFile}{character. Optional. A filepath to sink dput into.}
}
\description{
Convert a function into a single string for DataRobot
}
