% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deployment.R, R/deployments_apiWrapper.R
\name{SubmitActuals}
\alias{SubmitActuals}
\title{Submit actuals for processing.}
\usage{
SubmitActuals(actuals, deploymentId, batchSize = 10000, ...)

SubmitActuals(actuals, deploymentId, batchSize = 10000, ...)
}
\arguments{
\item{actuals}{dataframe. Data that describes actual values. Any strings stored as factors will
be coerced to characters with \code{as.character}. Allowed columns are:
\itemize{
   \item associationId string. A unique identifier used with a prediction. Max length 128
     characters.
   \item actualValue string or numeric. The actual value of a prediction;
     should be numeric for deployments with regression models or string for deployments with
     classification model.
   \item wasActedOn logical. Optional. Indicates if the prediction was acted on in a way that
     could have affected the actual outcome.
   \item timestamp POSIXt. Optional. If the datetime provided does not have a timezone, we assume
     it is UTC.
}}

\item{batchSize}{integer. Optional. The max number of actuals in each batch request. Cannot
exceed 10000.}
}
\description{
The actuals submitted will be used to calculate accuracy metrics.
Values are not processed immediately and may take some time to propagate through deployment
systems. Submission of actuals is limited to 10,000,000 actuals per hour. For time series
deployments, total actuals = number of actuals * number of forecast distances. For example,
submitting 10 actuals for a deployment with 50 forecast distances = 500 total actuals. For
multiclass deployments, a similar calculation is made where total actuals = number of actuals *
number of classes. For example, submitting 10 actuals for a deployment with 20 classes = 200
actuals.

Submit actuals values
}
\details{
Submit actuals values for processing. Values are not processed immediately and may take some time to propagate through deployment systems. Submission of actuals is limited to 10,000,000 actuals per hour. For time series deployments, total actuals = number of actuals * number of forecast distances. For example, submitting 10 actuals for a deployment with 50 forecast distances = 500 total actuals. For multiclass deployments, a similar calculation is made where total actuals = number of actuals * number of classes. For example, submitting 10 actuals for a deployment with 20 classes = 200 actuals.
}
\examples{
\dontrun{
deploymentId <- "5e319d2e422fbd6b58a5edad"
myActuals <- data.frame(
  associationId = c("439917"),
  actualValue = c("True"),
  wasActedOn = c(TRUE)
)
SubmitActuals(
  actuals = myActuals,
  deploymentId
)
}
}
\seealso{
The method DeploymentsActualsFromJSONCreate in [datarobot.apicore::DeploymentsApi], which this function wraps.

Other deployment accuracy functions: 
\code{\link{GetDeploymentAccuracyOverTime}()},
\code{\link{GetDeploymentAssociationId}()}

Other Deployments: 
\code{\link{CreateDeploymentsActualsFromDataset}()},
\code{\link{CreateDeploymentsChallengerPredictions}()},
\code{\link{CreateDeploymentsChallengers}()},
\code{\link{CreateDeploymentsFromProjectRecommendedModel}()},
\code{\link{CreateDeploymentsOnDemandReports}()},
\code{\link{CreateDeploymentsPredictionInputsFromDataset}()},
\code{\link{CreateDeploymentsPredictions}()},
\code{\link{CreateDeploymentsScoringCodeBuilds}()},
\code{\link{CreateDeployment}()},
\code{\link{DeleteDeploymentsChallengers}()},
\code{\link{DeleteDeployment}()},
\code{\link{GetDeploymentAccuracy}()},
\code{\link{GetDeploymentServiceStats}()},
\code{\link{GetDeploymentSettings}()},
\code{\link{GetDeployment}()},
\code{\link{ListDeletedDeployments}()},
\code{\link{ListDeploymentsAccuracyOverTime}()},
\code{\link{ListDeploymentsChallengers}()},
\code{\link{ListDeploymentsFeatureDrift}()},
\code{\link{ListDeploymentsFeatures}()},
\code{\link{ListDeploymentsHumilityStatsOverTime}()},
\code{\link{ListDeploymentsHumilityStats}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfigurationHistory}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{ListDeploymentsPredictionResults}()},
\code{\link{ListDeploymentsScoringCode}()},
\code{\link{ListDeploymentsServiceStatsOverTime}()},
\code{\link{ListDeploymentsSharedRoles}()},
\code{\link{ListDeploymentsTargetDrift}()},
\code{\link{ListDeployments}()},
\code{\link{ListPredictionServers}()},
\code{\link{PatchDeploymentsChallengers}()},
\code{\link{PatchDeployments}()},
\code{\link{PatchManyDeletedDeployments}()},
\code{\link{PatchManyDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{PatchManyDeploymentsSharedRoles}()},
\code{\link{ReplaceDeployedModel}()},
\code{\link{RetrieveDeploymentsChallengers}()},
\code{\link{ValidateReplaceDeployedModel}()}
}
\concept{Deployments}
\concept{deployment accuracy functions}
