% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethods.R
\name{summary.dataRobotModel}
\alias{summary.dataRobotModel}
\alias{summary.dataRobotProject}
\alias{summary.listOfBlueprints}
\alias{summary.listOfFeaturelists}
\alias{summary.listOfModels}
\alias{summary.projectSummaryList}
\title{DataRobot S3 object methods for R's generic summary function}
\usage{
\method{summary}{dataRobotModel}(object, ...)

\method{summary}{dataRobotProject}(object, ...)

\method{summary}{listOfBlueprints}(object, nList = 6, ...)

\method{summary}{listOfFeaturelists}(object, nList = 6, ...)

\method{summary}{listOfModels}(object, nList = 6, ...)

\method{summary}{projectSummaryList}(object, nList = 6, ...)
}
\arguments{
\item{object}{The S3 object to be summarized.}

\item{nList}{Integer: for the 'listOf' class objects, the first
nList elements of the list are summarized in the dataframe in
the second element of the list returned by the function.}

\item{\dots}{Not currently used.}
}
\value{
An object-specific summary: for objects of class
dataRobotModel and dataRobotProject, this summary is a
character vector giving key characteristics of the model or
project, respectively; for the other object classes, the
value is a two-element list where the first element is a
brief summary character string and the second element
is a more detailed dataframe with nList elements.
}
\description{
These functions extend R's generic summary function to the
DataRobot S3 object classes dataRobotModel, dataRobotProject,
listOfBlueprints, listOfFeaturelists, listOfModels, and
projectSummaryList.
}

