% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelDeployment.R
\name{summary.dataRobotModelDeploymentsList}
\alias{summary.dataRobotModelDeploymentsList}
\title{Summarize model deployments}
\usage{
\method{summary}{dataRobotModelDeploymentsList}(object, ...)
}
\arguments{
\item{object}{dataRobotModelDeploymentsList. The output from \code{ListModelDeployments}.}

\item{...}{list. Additional arguments to pass from summary. These are ignored.}
}
\description{
Summarize model deployments
}
