% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{TargetLeakageType}
\alias{TargetLeakageType}
\title{Target leakage report values}
\format{An object of class \code{list} of length 4.}
\usage{
TargetLeakageType
}
\description{
Target leakage report values
}
\keyword{datasets}
