% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{TargetType}
\alias{TargetType}
\title{Target Type modes}
\format{
An object of class \code{list} of length 3.
}
\usage{
TargetType
}
\description{
This is a list that contains the valid values for the Target Types
}
\keyword{datasets}
