% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{TaskStatus}
\alias{TaskStatus}
\title{Task statuses}
\format{
An object of class \code{list} of length 6.
}
\usage{
TaskStatus
}
\description{
This is a list that contains the valid values when getting the status of an async task.
}
\keyword{datasets}
