% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeploymentMonitoringUtils.R
\name{tidyServiceOverTimeObject}
\alias{tidyServiceOverTimeObject}
\title{Tidies a ServiceOverTime response object for use in a DF}
\usage{
tidyServiceOverTimeObject(df, valueColName)
}
\arguments{
\item{df}{A data frame that contains the following:
\itemize{
  \item period list, containing the following two items:
  \itemize{
     \item start POSIXct.
     \item end POSIXct.
  }
  \item value object.
}}

\item{valueColName}{character. The column in df currently named 'value' will be renamed to this.}
}
\description{
Tidies a ServiceOverTime response object for use in a DF
}
