% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeploymentMonitoringUtils.R
\name{transformRFC3339Period}
\alias{transformRFC3339Period}
\title{The DataRobot Monitoring APIs return dates formatted as RFC 3339
strings. This is the same as ISO 8601. Specifically, 'T' is the
date/time separator and 'Z' is used to denote UTC. Fractional seconds
are returned. e.g. 2020-01-01T05:00:00.000000Z}
\usage{
transformRFC3339Period(periodContainer)
}
\arguments{
\item{periodContainer}{an object containing the following:
\itemize{
  \item period list, containing the following two items:
  \itemize{
     \item start character. RFC 3339 formatted timestamp.
     \item end character. RFC 3339 formatted timestamp.
  }
}}
}
\description{
The DataRobot Monitoring APIs return dates formatted as RFC 3339
strings. This is the same as ISO 8601. Specifically, 'T' is the
date/time separator and 'Z' is used to denote UTC. Fractional seconds
are returned. e.g. 2020-01-01T05:00:00.000000Z
}
\seealso{
Other API datetime functions: 
\code{\link{RFC3339DateTimeFormat}},
\code{\link{formatRFC3339Timestamp}()},
\code{\link{parseRFC3339Timestamp}()},
\code{\link{validateReportingPeriodTime}()}
}
\concept{API datetime functions}
