% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{TreatAsExponential}
\alias{TreatAsExponential}
\title{Treat as exponential}
\format{An object of class \code{list} of length 3.}
\usage{
TreatAsExponential
}
\description{
Treat as exponential
}
\keyword{datasets}
