% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataRobotRequests.R
\name{TryingToSubmitNull}
\alias{TryingToSubmitNull}
\title{Checks to see if we are trying to submit `NULL` as a value.}
\usage{
TryingToSubmitNull(body)
}
\arguments{
\item{body}{list. The body to check for NULL.}
}
\description{
Checks to see if we are trying to submit `NULL` as a value.
}
