% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Calendars.R
\name{UpdateCalendar}
\alias{UpdateCalendar}
\title{UpdateCalendar}
\usage{
UpdateCalendar(calendarId, name = NULL, ...)
}
\arguments{
\item{calendarId}{character. The ID of the calendar to retrieve.}

\item{name}{character. The new name to name the calendar.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Update a calendar's name
}
\details{
Update a calendar's name

This method invokes \verb{PATCH /calendars/\{calendarId\}/} in the DataRobot Public API.

Update a calendar

Currently supports changing the name of a calendar.
}
\examples{
\dontrun{
calendarId <- "5da75da31fb4a45b8a815a53"
UpdateCalendar(calendarId, name = "New name for calendar")
}
}
\seealso{
The method CalendarsPatch in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
