% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{UpdateComplianceDocTemplate}
\alias{UpdateComplianceDocTemplate}
\title{UpdateComplianceDocTemplate}
\usage{
UpdateComplianceDocTemplate(templateId, name = NULL, sections = NULL, ...)
}
\description{
Update an existing model compliance documentation template
}
\details{
Update an existing model compliance documentation template with the given ``templateId``. The template must be accessible by the user. If the ``templateId`` is not found for the user, the update cannot be performed. For a description of the template ``sections`` object options, see the sample ``sections`` on the schema below.
}
\seealso{
The method ComplianceDocTemplatesPatch in [datarobot.apicore::DocumentationApi], which this function wraps.

Other Documentation: 
\code{\link{CreateAutomatedDocuments}()},
\code{\link{CreateOneModelComplianceDocsInitializations}()},
\code{\link{DeleteAutomatedDocuments}()},
\code{\link{DeleteComplianceDocTemplate}()},
\code{\link{GetComplianceDocTemplate}()},
\code{\link{ListAutomatedDocumentOptions}()},
\code{\link{ListAutomatedDocuments}()},
\code{\link{ListComplianceDocTemplatesDefault}()},
\code{\link{ListComplianceDocTemplatesSharedRoles}()},
\code{\link{ListComplianceDocTemplates}()},
\code{\link{PatchManyComplianceDocTemplatesSharedRoles}()},
\code{\link{RetrieveAutomatedDocuments}()},
\code{\link{RetrieveModelComplianceDocsInitializations}()},
\code{\link{UploadComplianceDocTemplate}()}
}
\concept{Documentation}
