% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStores.R, R/data_connectivity_apiWrapper.R
\name{UpdateDataStore}
\alias{UpdateDataStore}
\title{Update a data store.}
\usage{
UpdateDataStore(
  dataStoreId,
  canonicalName = NULL,
  driverId = NULL,
  jdbcUrl = NULL,
  ...
)

UpdateDataStore(
  dataStoreId,
  canonicalName = NULL,
  driverId = NULL,
  jdbcUrl = NULL,
  ...
)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}

\item{canonicalName}{character. The user-friendly name of the data store.}

\item{driverId}{character. The ID of the driver to use.}

\item{jdbcUrl}{character. The full JDBC url.}
}
\description{
Updates a data store configuration.
}
\details{
Updates a data store configuration.
}
\examples{
\dontrun{
dataStoreId <- "5c1303269300d900016b41a7"
UpdateDataStore(dataStoreId, canonicalName = "Different Name")
}
}
\seealso{
The method ExternalDataStoresPatch in [datarobot.apicore::DataConnectivityApi], which this function wraps.

Other DataConnectivity: 
\code{\link{CreateDataSource}()},
\code{\link{CreateDataStore}()},
\code{\link{CreateExternalDataDriverFile}()},
\code{\link{CreateExternalDataDrivers}()},
\code{\link{CreateExternalDataStoresColumns}()},
\code{\link{CreateExternalDataStoresVerifySQL}()},
\code{\link{DeleteDataSource}()},
\code{\link{DeleteDataStore}()},
\code{\link{DeleteExternalDataDrivers}()},
\code{\link{GetDataSource}()},
\code{\link{GetDataStoreSchemas}()},
\code{\link{GetDataStoreTables}()},
\code{\link{GetDataStore}()},
\code{\link{GetDriver}()},
\code{\link{ListDataSources}()},
\code{\link{ListDataStores}()},
\code{\link{ListDrivers}()},
\code{\link{ListExternalDataDriversConfiguration}()},
\code{\link{ListExternalDataSourcesAccessControl}()},
\code{\link{ListExternalDataSourcesPermissions}()},
\code{\link{ListExternalDataSourcesSharedRoles}()},
\code{\link{ListExternalDataStoresCredentials}()},
\code{\link{ListExternalDataStoresPermissions}()},
\code{\link{ListExternalDataStoresSharedRoles}()},
\code{\link{ListExternalDriverConfigurations}()},
\code{\link{PatchExternalDataDrivers}()},
\code{\link{PatchManyExternalDataSourcesAccessControl}()},
\code{\link{PatchManyExternalDataSourcesSharedRoles}()},
\code{\link{PatchManyExternalDataStoresAccessControl}()},
\code{\link{PatchManyExternalDataStoresSharedRoles}()},
\code{\link{RetrieveExternalDriverConfigurations}()},
\code{\link{TestDataStore}()},
\code{\link{UpdateDataSource}()}
}
\concept{DataConnectivity}
