% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deployment.R
\name{UpdateDeploymentSettings}
\alias{UpdateDeploymentSettings}
\title{Updates configuration settings for a deployed model.}
\usage{
UpdateDeploymentSettings(deployment, newSettings, maxWait)
}
\arguments{
\item{newSettings}{List containing the settings to be modified. Any settings
not explicitly defined will be unprocessed.}
}
\description{
Updates the deployment settings and returns all settings, including those not
changed, on success.
}
\details{
Marked as internal since we do not yet want to add this to the package index.
}
\concept{deployment configuration functions}
\keyword{internal}
