% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Featurelists.R
\name{UpdateFeaturelist}
\alias{UpdateFeaturelist}
\title{Update a featurelist}
\usage{
UpdateFeaturelist(featurelist, listName = NULL, description = NULL)
}
\arguments{
\item{featurelist}{list. The featurelist to delete.}

\item{description}{character. A user-friendly description to give a featurelist.}
}
\description{
Updates a featurelist to change the name or description.
}
