% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelDeployment.R
\name{UpdateModelDeployment}
\alias{UpdateModelDeployment}
\title{Update a model deployment.}
\usage{
UpdateModelDeployment(modelDeploymentId, label = NULL, description = NULL,
  status = NULL)
}
\arguments{
\item{modelDeploymentId}{character. ID of the model deployment to update}

\item{label}{character. A name for the model deployment.}

\item{description}{character. Optional. Description for the model deployment.}

\item{status}{character. Optional. status for the model deployment. Can be "active",
"inactive" or "archived".}
}
\value{
The model deployment data with the updated fields. See \code{GetModelDeployment}
   for details.
}
\description{
Update a model deployment.
}
\examples{
\dontrun{
  modelDeploymentId <- "59a5af20c80891534e3c2bde"
  UpdateModelDeployment(modelDeploymentId, label = "My Different Label for Deployment")
}
}
