% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Featurelists.R
\name{UpdateModelingFeaturelist}
\alias{UpdateModelingFeaturelist}
\title{Update a modeling featurelist}
\usage{
UpdateModelingFeaturelist(featurelist, listName = NULL,
  description = NULL)
}
\arguments{
\item{featurelist}{list. The modeling featurelist to delete.}

\item{listName}{character. String identifying the new featurelist
to be created.}

\item{description}{character. A user-friendly description to give a featurelist.}
}
\description{
Updates a modeling featurelist to change the name or description.
}
