% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{UploadComplianceDocTemplate}
\alias{UploadComplianceDocTemplate}
\title{UploadComplianceDocTemplate}
\usage{
UploadComplianceDocTemplate(name, filename = NULL, sections = NULL, ...)
}
\description{
Create a new compliance documentation template
}
\details{
Create a new compliance documentation template. One can retrieve the default DataRobot template via ``GET /api/v2/complianceDocTemplates/default/`` endpoint.
}
\seealso{
The method ComplianceDocTemplatesCreate in [datarobot.apicore::DocumentationApi], which this function wraps.

Other Documentation: 
\code{\link{CreateAutomatedDocuments}()},
\code{\link{CreateOneModelComplianceDocsInitializations}()},
\code{\link{DeleteAutomatedDocuments}()},
\code{\link{DeleteComplianceDocTemplate}()},
\code{\link{GetComplianceDocTemplate}()},
\code{\link{ListAutomatedDocumentOptions}()},
\code{\link{ListAutomatedDocuments}()},
\code{\link{ListComplianceDocTemplatesDefault}()},
\code{\link{ListComplianceDocTemplatesSharedRoles}()},
\code{\link{ListComplianceDocTemplates}()},
\code{\link{PatchManyComplianceDocTemplatesSharedRoles}()},
\code{\link{RetrieveAutomatedDocuments}()},
\code{\link{RetrieveModelComplianceDocsInitializations}()},
\code{\link{UpdateComplianceDocTemplate}()}
}
\concept{Documentation}
