% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadData.R
\name{UploadData}
\alias{UploadData}
\title{Upload a data source.}
\usage{
UploadData(dataSource, fileName = NULL)
}
\arguments{
\item{dataSource}{character. The file to upload.}

\item{fileName}{character. The name of the file after it is uploaded. If not set, defaults
to the name of the uploaded file.}
}
\value{
An httr object specifying the form upload content of the file path.
}
\description{
Takes either a file path or a dataframe and returns output for POST that specifies
the file object via form upload. This function is meant to facilitate uploading
CSV data sources into DataRobot, such as through \code{SetupProject}.
}
\seealso{
SetupProject
}
