% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{UploadPredictionDatasetFromCatalog}
\alias{UploadPredictionDatasetFromCatalog}
\title{UploadPredictionDatasetFromCatalog}
\usage{
UploadPredictionDatasetFromCatalog(
  project,
  datasetId,
  datasetVersionId = NULL,
  forecastPoint = NULL,
  predictionsStartDate = NULL,
  predictionsEndDate = NULL,
  maxWait = 600,
  relaxKIAFeaturesCheck = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create prediction dataset
}
\details{
Create a prediction dataset from a Dataset Asset referenced by AI Catalog item/version ID.

This method invokes \verb{POST /projects/\{projectId\}/predictionDatasets/datasetUploads/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionDatasetsDatasetUploadsCreate in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
