% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{UploadTransferableModel}
\alias{UploadTransferableModel}
\title{UploadTransferableModel}
\usage{
UploadTransferableModel(modelFile, maxWait = 600, ...)
}
\description{
Import a previously exported DRX model to be served by SSE
}
\details{
This route can be used on-prem to put a given DRX model into file storage so that it can be served by a Standalone Scoring Engine (SSE). Please note however that for this to work SSE needs to be configured to use the same file storage as the main Datarbot application (namely, Public API).
}
\seealso{
The method ImportedModelsCreate in [datarobot.apicore::PredictionsApi], which this function wraps.

Other Predictions: 
\code{\link{CreateBatchPredictionJobDefinitions}()},
\code{\link{CreateBatchPredictionsCsvUploadFinalizeMultipart}()},
\code{\link{CreateBatchPredictionsFromExisting}()},
\code{\link{CreateBatchPredictionsFromJobDefinition}()},
\code{\link{CreateBatchPredictions}()},
\code{\link{CreateProjectsPredictionDatasetsFileUploads}()},
\code{\link{DeleteBatchPredictionJobDefinitions}()},
\code{\link{DeleteBatchPredictions}()},
\code{\link{DeletePredictJob}()},
\code{\link{DeletePredictionDataset}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteScheduledJobs}()},
\code{\link{DeleteTransferableModel}()},
\code{\link{GetPredictJobs}()},
\code{\link{GetPredictJob}()},
\code{\link{GetPredictionDataset}()},
\code{\link{GetPredictionExplanationsMetadata}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetTransferableModel}()},
\code{\link{ListBatchPredictionJobDefinitionsPortable}()},
\code{\link{ListBatchPredictionJobDefinitions}()},
\code{\link{ListBatchPredictionsDownload}()},
\code{\link{ListBatchPredictions}()},
\code{\link{ListPredictionDatasets}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListPredictions}()},
\code{\link{ListProjectsPredictionsMetadata}()},
\code{\link{ListScheduledJobs}()},
\code{\link{ListTrainingPredictions}()},
\code{\link{ListTransferableModels}()},
\code{\link{PatchBatchPredictionJobDefinitions}()},
\code{\link{PatchBatchPredictions}()},
\code{\link{PatchScheduledJobs}()},
\code{\link{PutBatchPredictionsCsvUploadPart}()},
\code{\link{PutImportedModels}()},
\code{\link{PutManyBatchPredictionsCsvUpload}()},
\code{\link{RetrieveBatchPredictionJobDefinitions}()},
\code{\link{RetrieveBatchPredictions}()},
\code{\link{RetrieveProjectsPredictionsMetadata}()},
\code{\link{RetrieveProjectsPredictions}()},
\code{\link{RetrieveScheduledJobs}()},
\code{\link{UpdateTransferableModel}()},
\code{\link{UploadPredictionDatasetFromCatalog}()},
\code{\link{UploadPredictionDatasetFromDataSource}()},
\code{\link{UploadPredictionDataset}()}
}
\concept{Predictions}
