% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{ValidateActuals}
\alias{ValidateActuals}
\title{Validate that the actuals are a dataframe and contain required columns.}
\usage{
ValidateActuals(actuals, error = TRUE)
}
\arguments{
\item{actuals}{dataframe. Contains all actuals to be submitted.}

\item{error}{logical. Should an error be raised if there is an issue?}
}
\value{
TRUE if the actuals dataframe has required properties, otherwise FALSE or raises error.
}
\description{
Validate that the actuals are a dataframe and contain required columns.
}
