% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{ValidateAndReturnModel}
\alias{ValidateAndReturnModel}
\title{Validate that an object is a valid DataRobot model and contains the fields projectId and modelId.}
\usage{
ValidateAndReturnModel(model)
}
\arguments{
\item{model}{An object representing a DataRobot model like that returned by the function GetModel, or each element of the list returned by the function ListModels, or any of the related functions in the datarobot.apicore package.}
}
\value{
the model object if it is valid, potentially with a new field `modelId` if it didn't previously exist; throws an error otherwise.
}
\description{
Validate that an object is a valid DataRobot model and contains the fields projectId and modelId.
}
