% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{ValidateCalendar}
\alias{ValidateCalendar}
\title{Get a calendar id from a calendar object.}
\usage{
ValidateCalendar(calendar)
}
\arguments{
\item{calendar}{object. Either list with calendarId element or calendarId value}
}
\description{
Get a calendar id from a calendar object.
}
