% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{ValidateModel}
\alias{ValidateModel}
\title{Validate that model belongs to class 'dataRobotModel' and includes
projectId and modelId.}
\usage{
ValidateModel(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModel, or each element of the list returned by
the function ListModels.}
}
\description{
Validate that model belongs to class 'dataRobotModel' and includes
projectId and modelId.
}
