% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{ValidateMultiSeriesProperties}
\alias{ValidateMultiSeriesProperties}
\title{Validate that the multiseries properties indicate a successful multiseries setup.}
\usage{
ValidateMultiSeriesProperties(properties, error = TRUE)
}
\arguments{
\item{properties}{list. List of multiseries properties.}

\item{error}{logical. Should an error be raised if there is an issue?}
}
\value{
TRUE if all properties verify, otherwise FALSE or raises error.
}
\description{
Validate that the multiseries properties indicate a successful multiseries setup.
}
