% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{ValidatePartition}
\alias{ValidatePartition}
\title{Checks if a partition is valid.}
\usage{
ValidatePartition(validationType, partition, reps = NULL,
  validationPct = NULL)
}
\arguments{
\item{validationType}{character. The type of partition to validate.}

\item{partition}{partition. The partition object.}

\item{reps}{numeric. The number of repetitions for a CV validation.}

\item{validationPct}{numeric. The size of the validation set for TVH validation.}
}
\description{
Checks if a partition is valid.
}
