% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{ValidateProject}
\alias{ValidateProject}
\title{Get a projectId from a project object.}
\usage{
ValidateProject(project)
}
\arguments{
\item{project}{object. Either list with projectId element or projectId value}
}
\description{
Get a projectId from a project object.
}
