% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R, R/Deployment.R
\name{ValidateReplaceDeployedModel}
\alias{ValidateReplaceDeployedModel}
\title{ValidateReplaceDeployedModel}
\usage{
ValidateReplaceDeployedModel(deploymentId, newModelId, ...)
}
\arguments{
\item{deploymentId}{character. The ID of the deployment.}

\item{newModelId}{character. The ID of the model to use in the deployment. This model
will replace the old model. You can also pass a dataRobotModel object.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A validation report with:
\itemize{
  \item status character. Either PASSED or FAILED depending on whether all checks passed
    or not.
  \item message character. A message explaining the status failure, if any.
  \item checks list. A list of each check and the individual status.
}
}
\description{
Model Replacement Validation.
}
\details{
Validate that a model can be used to replace the current model of the deployment.

This method invokes \verb{POST /deployments/\{deploymentId\}/model/validation/} in the DataRobot Public API.

Validate a potential deployment model replacement.
}
\examples{
\dontrun{
deploymentId <- "5e319d2e422fbd6b58a5edad"
newModelId <- "5996f820af07fc605e81ead4"
ValidateReplaceDeployedModel(deploymentId, newModelId)
}
}
\seealso{
The method DeploymentsModelValidationCreate in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
