% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeploymentMonitoringUtils.R
\name{validateReportingPeriodTime}
\alias{validateReportingPeriodTime}
\title{Helper function for validating reporting period objects used by
the deployment monitoring functions. See \code{GetDeploymentServiceStats},
\code{GetDeploymentAccuracy}, \code{GetDeploymentServiceStatsOverTime}, and
\code{GetDeploymentAccuracyOverTime}.}
\usage{
validateReportingPeriodTime(timestamp, tsName = "timestamp")
}
\arguments{
\item{timestamp}{character. A timestamp in RFC 3339 format.}

\item{tsName}{character. Optional. Explanation of the timestamp for error messages.}
}
\description{
Helper function for validating reporting period objects used by
the deployment monitoring functions. See \code{GetDeploymentServiceStats},
\code{GetDeploymentAccuracy}, \code{GetDeploymentServiceStatsOverTime}, and
\code{GetDeploymentAccuracyOverTime}.
}
\seealso{
Other API datetime functions: 
\code{\link{RFC3339DateTimeFormat}},
\code{\link{formatRFC3339Timestamp}()},
\code{\link{parseRFC3339Timestamp}()},
\code{\link{transformRFC3339Period}()}
}
\concept{API datetime functions}
