% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enums.R
\docType{data}
\name{VariableTransformTypes}
\alias{VariableTransformTypes}
\title{Types of variable transformations}
\format{An object of class \code{list} of length 4.}
\usage{
VariableTransformTypes
}
\description{
Types of variable transformations
}
\keyword{datasets}
