% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewWebProject.R
\name{ViewWebProject}
\alias{ViewWebProject}
\title{Retrieve a DataRobot web page that displays detailed project information}
\usage{
ViewWebProject(project)
}
\arguments{
\item{project}{object. Either list with projectId element or projectId value}
}
\description{
This function brings up a web page that displays detailed project
information like that available from the standard DataRobot user interface.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ViewWebProject(projectId)
}
}
