% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaitForAutopilot.R
\name{WaitForAutopilot}
\alias{WaitForAutopilot}
\title{This function periodically checks whether Autopilot is finished and returns only after it is.}
\usage{
WaitForAutopilot(project, checkInterval = 20, timeout = NULL,
  verbosity = 1)
}
\arguments{
\item{project}{The project for which you want to wait until autopilot is finished}

\item{checkInterval}{Maximum wait (in seconds) between checks that Autopilot is finished}

\item{timeout}{Time (in seconds) after which to give up (NULL implies no timeout). There is
an error if Autopilot is not finished before timing out.}

\item{verbosity}{Verbosity level: 0 is silent, 1 or more displays information about progress}
}
\description{
This function periodically checks whether Autopilot is finished and returns only after it is.
}
