% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplySchema.R
\name{ApplySchema}
\alias{ApplySchema}
\title{Apply a schema to DataRobot objects (lists, frames)}
\usage{
ApplySchema(inList, schema, mask = NULL)
}
\arguments{
\item{inList}{object. The DataRobot object to apply the schema to.}

\item{schema}{list. The schema to apply.}

\item{mask}{list. Search the schema and only apply values that match this with grep.
Defaults to NULL, or no masking.}
}
\description{
Apply a schema to DataRobot objects (lists, frames)
}
