% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{CreateApplications}
\alias{CreateApplications}
\title{CreateApplications}
\usage{
CreateApplications(
  sources,
  purpose,
  applicationTypeId,
  name,
  description,
  modelDeploymentId,
  cloudProvider = "drcloud",
  authenticationType = "invitedUsersOnly",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create an application
}
\details{
Create an application. Note that the number of active applications users can have at the same time is limited.

This method invokes \verb{POST /applications/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsCreate in \link[datarobot.apicore:ApplicationsApi]{datarobot.apicore::ApplicationsApi}, which invokes the same underlying API endpoint.
}
