% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{CreateApplicationsDuplicate}
\alias{CreateApplicationsDuplicate}
\title{CreateApplicationsDuplicate}
\usage{
CreateApplicationsDuplicate(
  name,
  description,
  applicationId,
  duplicatePredictions = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a duplicate of the application
}
\details{
Create a copy of App Builder application. Note that the number of active applications users can have at the same time is limited.

This method invokes \verb{POST /applications/\{applicationId\}/duplicate/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsDuplicateCreate in \link[datarobot.apicore:ApplicationsApi]{datarobot.apicore::ApplicationsApi}, which invokes the same underlying API endpoint.
}
