% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{CreateApprovalPolicies}
\alias{CreateApprovalPolicies}
\title{CreateApprovalPolicies}
\usage{
CreateApprovalPolicies(
  automaticAction,
  review,
  name,
  trigger,
  active = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new Approval Policy.
}
\details{
Create a new Approval Policy.

This method invokes \verb{POST /approvalPolicies/} in the DataRobot Public API.
}
\seealso{
The method ApprovalPoliciesCreate in \link[datarobot.apicore:GovernanceApi]{datarobot.apicore::GovernanceApi}, which invokes the same underlying API endpoint.
}
