% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{CreateBatchPredictionJobDefinitions}
\alias{CreateBatchPredictionJobDefinitions}
\title{CreateBatchPredictionJobDefinitions}
\usage{
CreateBatchPredictionJobDefinitions(
  columnNamesRemapping,
  timeseriesSettings,
  modelId,
  passthroughColumns,
  outputSettings,
  enabled,
  predictionWarningEnabled,
  intakeSettings,
  deploymentId,
  predictionInstance,
  csvSettings,
  chunkSize,
  modelPackageId,
  schedule,
  numConcurrent,
  passthroughColumnsSet,
  explanationAlgorithm,
  name,
  thresholdHigh,
  thresholdLow,
  pinnedModelId,
  includeProbabilitiesClasses = c(),
  includePredictionStatus = FALSE,
  skipDriftTracking = FALSE,
  disableRowLevelErrorHandling = FALSE,
  abortOnError = TRUE,
  maxExplanations = 0,
  includeProbabilities = TRUE,
  ...
)
}
\description{
Creates a new Batch Prediction job definition
}
\details{
Create a Batch Prediction Job definition. A configuration for a Batch Prediction job which can either be executed manually upon request or on scheduled intervals, if enabled. The API payload is the same as for `/batchPredictions` along with optional ``enabled`` and ``schedule`` items.
}
\seealso{
The method BatchPredictionJobDefinitionsCreate in [datarobot.apicore::PredictionsApi], which this function wraps.

Other Predictions: 
\code{\link{CreateBatchPredictionsCsvUploadFinalizeMultipart}()},
\code{\link{CreateBatchPredictionsFromExisting}()},
\code{\link{CreateBatchPredictionsFromJobDefinition}()},
\code{\link{CreateBatchPredictions}()},
\code{\link{CreateProjectsPredictionDatasetsFileUploads}()},
\code{\link{DeleteBatchPredictionJobDefinitions}()},
\code{\link{DeleteBatchPredictions}()},
\code{\link{DeletePredictJob}()},
\code{\link{DeletePredictionDataset}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteScheduledJobs}()},
\code{\link{DeleteTransferableModel}()},
\code{\link{GetPredictJobs}()},
\code{\link{GetPredictJob}()},
\code{\link{GetPredictionDataset}()},
\code{\link{GetPredictionExplanationsMetadata}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetTransferableModel}()},
\code{\link{ListBatchPredictionJobDefinitionsPortable}()},
\code{\link{ListBatchPredictionJobDefinitions}()},
\code{\link{ListBatchPredictionsDownload}()},
\code{\link{ListBatchPredictions}()},
\code{\link{ListPredictionDatasets}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListPredictions}()},
\code{\link{ListProjectsPredictionsMetadata}()},
\code{\link{ListScheduledJobs}()},
\code{\link{ListTrainingPredictions}()},
\code{\link{ListTransferableModels}()},
\code{\link{PatchBatchPredictionJobDefinitions}()},
\code{\link{PatchBatchPredictions}()},
\code{\link{PatchScheduledJobs}()},
\code{\link{PutBatchPredictionsCsvUploadPart}()},
\code{\link{PutImportedModels}()},
\code{\link{PutManyBatchPredictionsCsvUpload}()},
\code{\link{RetrieveBatchPredictionJobDefinitions}()},
\code{\link{RetrieveBatchPredictions}()},
\code{\link{RetrieveProjectsPredictionsMetadata}()},
\code{\link{RetrieveProjectsPredictions}()},
\code{\link{RetrieveScheduledJobs}()},
\code{\link{UpdateTransferableModel}()},
\code{\link{UploadPredictionDatasetFromCatalog}()},
\code{\link{UploadPredictionDatasetFromDataSource}()},
\code{\link{UploadPredictionDataset}()},
\code{\link{UploadTransferableModel}()}
}
\concept{Predictions}
