% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{CreateBatchPredictionsFromJobDefinition}
\alias{CreateBatchPredictionsFromJobDefinition}
\title{CreateBatchPredictionsFromJobDefinition}
\usage{
CreateBatchPredictionsFromJobDefinition(jobDefinitionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Launch a Batch Prediction job for scoring
}
\details{
Launches a one-time batch prediction job based off of the previously supplied definition referring to the job definition ID and puts it on the queue.

This method invokes \verb{POST /batchPredictions/fromJobDefinition/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsFromJobDefinitionCreate in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
