% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendars.R
\name{CreateCalendar}
\alias{CreateCalendar}
\title{Create a calendar from an uploaded CSV.}
\usage{
CreateCalendar(file, name = NULL, maxWait = 600)
}
\arguments{
\item{file}{character. The filename containing the calendar CSV to upload.}

\item{name}{character. Optional. The name of the calendar.}

\item{maxWait}{integer. The maximum time (in seconds) to wait for the retrieve to complete.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Create a calendar from an uploaded CSV.
}
\examples{
\dontrun{
   CreateCalendar("myRatingTable.csv", name = "myCalendar")
}
}
