% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Calendars.R
\name{CreateCalendar}
\alias{CreateCalendar}
\title{CreateCalendar}
\usage{
CreateCalendar(
  dataSource,
  name = NULL,
  multiSeriesIdColumn = NULL,
  maxWait = 600
)

CreateCalendar(
  dataSource,
  name = NULL,
  multiSeriesIdColumn = NULL,
  maxWait = 600
)
}
\arguments{
\item{dataSource}{object. Either (a) the name of a CSV file, or (b) a dataframe.
This parameter identifies the source of the calendar data.}

\item{name}{character. Optional. The name of the calendar.}

\item{multiSeriesIdColumn}{character. Optional. Added in 2.19. The column in
the calendar that defines which series an event belongs to. Only one
column is supported.}

\item{maxWait}{integer. The maximum time (in seconds) to wait for the retrieve to complete.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Create a calendar from a file.
}
\details{
Create a calendar from a file in a csv or xlsx format. The calendar file specifies the dates or events in a dataset such that DataRobot automatically derives and creates special features based on the calendar events (e.g., time until the next event, labeling the most recent event).

This method invokes \verb{POST /calendars/fileUpload/} in the DataRobot Public API.
}
\examples{
\dontrun{
CreateCalendar("inst/extdata/calendar.csv", name = "intlHolidayCalendar")
}
\dontrun{
holidayCalendarDF <- as.data.frame(myCalendar)
CreateCalendar(holidayCalendarDF, name = "intlHolidayCalendar")
}
\dontrun{
CreateCalendar("inst/extdata/calendar.csv",
  name = "intlHolidayCalendar",
  multiSeriesIdColumn = "Country"
)
}
}
\seealso{
The method CalendarsFileUploadCreate in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
