% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{CreateComments}
\alias{CreateComments}
\title{CreateComments}
\usage{
CreateComments(entityType, mentions, entityId, content, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Post a comment
}
\details{
Post a comment

This method invokes \verb{POST /comments/} in the DataRobot Public API.
}
\seealso{
The method CommentsCreate in \link[datarobot.apicore:CommentsApi]{datarobot.apicore::CommentsApi}, which invokes the same underlying API endpoint.
}
