% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{CreateCustomTasksVersionsDependencyBuild}
\alias{CreateCustomTasksVersionsDependencyBuild}
\title{CreateCustomTasksVersionsDependencyBuild}
\usage{
CreateCustomTasksVersionsDependencyBuild(
  customTaskVersionId,
  customTaskId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Start a custom task version's dependency build.
}
\details{
Start a custom task version's dependency build. This is required to test, deploy, or train custom tasks.

This method invokes \verb{POST /customTasks/\{customTaskId\}/versions/\{customTaskVersionId\}/dependencyBuild/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsDependencyBuildCreate in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
