% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/DataEngineQueryGenerator.R
\name{CreateDatasetFromDataEngineQueryGenerator}
\alias{CreateDatasetFromDataEngineQueryGenerator}
\title{CreateDatasetFromDataEngineQueryGenerator}
\usage{
CreateDatasetFromDataEngineQueryGenerator(
  generatorId,
  datasetId = NULL,
  datasetVersionId = NULL
)

CreateDatasetFromDataEngineQueryGenerator(
  generatorId,
  datasetId = NULL,
  datasetVersionId = NULL
)
}
\arguments{
\item{generatorId}{character. The ID of the query generator to use.}

\item{datasetId}{character. Optional. The ID of the dataset to apply the query to.}

\item{datasetVersionId}{character. Optional. The ID of the dataset version to apply the query to.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A dataset object:
\itemize{
\item datasetId. character. The ID of the created dataset.
\item versionId. character. The ID of the created dataset version.
\item name. character. The name of the created dataset.
\item categories. list of character. The categories of the created dataset.
\item creationDate. POSIXct. The timestamp the dataset was created.
\item createdBy. character. The user who created the dataset.
\item dataPersisted. logical. Whether the dataset has extended data profile available.
\item isDataEngineEligible. logical. Whether the dataset is eligible for data engine queries.
\item isLatestVersion. logical. Whether the dataset is the latest version.
\item isSnapshot. logical. Whether the dataset is a static snapshot.
\item datasetSize. numeric. Size of the dataset as a CSV in bytes.
\item rowCount. numeric. The number of rows in the dataset.
\item processingState. character. The current ingestion process state of the dataset.
}
}
\description{
Create Data Engine workspace state from a query generator
}
\details{
Create Data Engine workspace state in database from a query generator.

This method invokes \verb{POST /dataEngineWorkspaceStates/fromDataEngineQueryGenerator/} in the DataRobot Public API.

Create a Dataset from a DataEngineQueryGenerator for time series data prep.
}
\examples{
\dontrun{
generatorId <- "59a5af20c80891534e3c2bde"
dataset <- CreateDatasetFromDataEngineQueryGenerator(generatorId)
}
}
\seealso{
The method DataEngineWorkspaceStatesFromDataEngineQueryGeneratorCreate in \link[datarobot.apicore:AiCatalogApi]{datarobot.apicore::AiCatalogApi}, which invokes the same underlying API endpoint.

Other Data Engine Query Generator functions: 
\code{\link{CreateDataEngineQueryGenerator}()},
\code{\link{GetDataEngineQueryGenerator}()}
}
\concept{Data Engine Query Generator functions}
