% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStores.R, R/data_connectivity_apiWrapper.R
\name{CreateDataStore}
\alias{CreateDataStore}
\title{Create a data store.}
\usage{
CreateDataStore(type, canonicalName, driverId, jdbcUrl, ...)

CreateDataStore(type, canonicalName, driverId, jdbcUrl, ...)
}
\arguments{
\item{type}{character. The type of data store.}

\item{canonicalName}{character. The user-friendly name of the data store.}

\item{driverId}{character. The ID of the driver to use.}

\item{jdbcUrl}{character. The full JDBC url.}
}
\description{
Create a data store.
}
\details{
Create a data store which includes a name and a driver ID or a connector ID. The driver would be configured by a JDBC URL or by jdbc fields; The connector would be configured by connection fields.
}
\examples{
\dontrun{
CreateDataStore(
  type = "jdbc",
  canonicalName = "Demo DB",
  driverId = "57a7c978c808916f4a630f89",
  jdbcUrl = "jdbc:postgresql://my.db.address.org:5432/my_db"
)
}
}
\seealso{
The method ExternalDataStoresCreate in [datarobot.apicore::DataConnectivityApi], which this function wraps.

Other DataConnectivity: 
\code{\link{CreateDataSource}()},
\code{\link{CreateExternalDataDriverFile}()},
\code{\link{CreateExternalDataDrivers}()},
\code{\link{CreateExternalDataStoresColumns}()},
\code{\link{CreateExternalDataStoresVerifySQL}()},
\code{\link{DeleteDataSource}()},
\code{\link{DeleteDataStore}()},
\code{\link{DeleteExternalDataDrivers}()},
\code{\link{GetDataSource}()},
\code{\link{GetDataStoreSchemas}()},
\code{\link{GetDataStoreTables}()},
\code{\link{GetDataStore}()},
\code{\link{GetDriver}()},
\code{\link{ListDataSources}()},
\code{\link{ListDataStores}()},
\code{\link{ListDrivers}()},
\code{\link{ListExternalDataDriversConfiguration}()},
\code{\link{ListExternalDataSourcesAccessControl}()},
\code{\link{ListExternalDataSourcesPermissions}()},
\code{\link{ListExternalDataSourcesSharedRoles}()},
\code{\link{ListExternalDataStoresCredentials}()},
\code{\link{ListExternalDataStoresPermissions}()},
\code{\link{ListExternalDataStoresSharedRoles}()},
\code{\link{ListExternalDriverConfigurations}()},
\code{\link{PatchExternalDataDrivers}()},
\code{\link{PatchManyExternalDataSourcesAccessControl}()},
\code{\link{PatchManyExternalDataSourcesSharedRoles}()},
\code{\link{PatchManyExternalDataStoresAccessControl}()},
\code{\link{PatchManyExternalDataStoresSharedRoles}()},
\code{\link{RetrieveExternalDriverConfigurations}()},
\code{\link{TestDataStore}()},
\code{\link{UpdateDataSource}()},
\code{\link{UpdateDataStore}()}
}
\concept{DataConnectivity}
