% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{CreateExternalDataDrivers}
\alias{CreateExternalDataDrivers}
\title{CreateExternalDataDrivers}
\usage{
CreateExternalDataDrivers(
  baseNames,
  localJarUrls,
  className,
  configurationId,
  version,
  canonicalName,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new JDBC driver.
}
\details{
Create a new JDBC driver. To create connections from fields, rather than supplying a JDBC URL, use 'configurationId'. When using 'configurationId', do not include 'canonicalName' or 'className' as they are part of the driver configuration. Specifying a 'version' is required as it is used in the construction of the canonicalName.

This method invokes \verb{POST /externalDataDrivers/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataDriversCreate in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
