% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{CreateExternalDataStoresColumns}
\alias{CreateExternalDataStoresColumns}
\title{CreateExternalDataStoresColumns}
\usage{
CreateExternalDataStoresColumns(
  schema,
  password,
  dataStoreId,
  catalog,
  query,
  credentialId,
  user,
  table,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieves a data store's data columns.
}
\details{
Retrieves a data store's data columns.

This method invokes \verb{POST /externalDataStores/\{dataStoreId\}/columns/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresColumnsCreate in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
