% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateNotificationEmailChannelVerificationStatus}
\alias{CreateNotificationEmailChannelVerificationStatus}
\title{CreateNotificationEmailChannelVerificationStatus}
\usage{
CreateNotificationEmailChannelVerificationStatus(
  emailAddress,
  verificationCode,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the status of whether admin enter code correctly
}
\details{
Retrieve the status of whether admin enter code correctly

This method invokes \verb{POST /notificationEmailChannelVerificationStatus/} in the DataRobot Public API.
}
\seealso{
The method NotificationEmailChannelVerificationStatusCreate in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
