% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateNotificationPolicyMutes}
\alias{CreateNotificationPolicyMutes}
\title{CreateNotificationPolicyMutes}
\usage{
CreateNotificationPolicyMutes(policyId, entityId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new ignored notification
}
\details{
Create a new ignored notification

This method invokes \verb{POST /notificationPolicyMutes/} in the DataRobot Public API.
}
\seealso{
The method NotificationPolicyMutesCreate in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
