% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateNotifications}
\alias{CreateNotifications}
\title{CreateNotifications}
\usage{
CreateNotifications(notificationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Resends the notification
}
\details{
Resends the notification

This method invokes \verb{POST /notifications/} in the DataRobot Public API.
}
\seealso{
The method NotificationsCreate in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
