% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{CreateOneModelComplianceDocsInitializations}
\alias{CreateOneModelComplianceDocsInitializations}
\title{CreateOneModelComplianceDocsInitializations}
\usage{
CreateOneModelComplianceDocsInitializations(entityId, body, ...)
}
\description{
Initialize compliance documentation pre-processing for the model
}
\details{
Initialize compliance documentation pre-processing for the current model. This route must be called before generating documentation for a custom model.
}
\seealso{
The method ModelComplianceDocsInitializationsCreateOne in [datarobot.apicore::DocumentationApi], which this function wraps.

Other Documentation: 
\code{\link{CreateAutomatedDocuments}()},
\code{\link{DeleteAutomatedDocuments}()},
\code{\link{DeleteComplianceDocTemplate}()},
\code{\link{GetComplianceDocTemplate}()},
\code{\link{ListAutomatedDocumentOptions}()},
\code{\link{ListAutomatedDocuments}()},
\code{\link{ListComplianceDocTemplatesDefault}()},
\code{\link{ListComplianceDocTemplatesSharedRoles}()},
\code{\link{ListComplianceDocTemplates}()},
\code{\link{PatchManyComplianceDocTemplatesSharedRoles}()},
\code{\link{RetrieveAutomatedDocuments}()},
\code{\link{RetrieveModelComplianceDocsInitializations}()},
\code{\link{UpdateComplianceDocTemplate}()},
\code{\link{UploadComplianceDocTemplate}()}
}
\concept{Documentation}
