% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{CreateOrganizationsUsers}
\alias{CreateOrganizationsUsers}
\title{CreateOrganizationsUsers}
\usage{
CreateOrganizationsUsers(
  organizationId,
  lastName,
  firstName,
  password,
  accessRoleId,
  language,
  accessRoleIds,
  username,
  orgAdmin = FALSE,
  requireClickthrough = FALSE,
  create = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add user to an existing organization.
}
\details{
Add user to an existing organization. A user can only be part of one organization at a time

This method invokes \verb{POST /organizations/\{organizationId\}/users/} in the DataRobot Public API.
}
\seealso{
The method OrganizationsUsersCreate in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
