% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{CreateProjectsPredictionDatasetsFileUploads}
\alias{CreateProjectsPredictionDatasetsFileUploads}
\title{CreateProjectsPredictionDatasetsFileUploads}
\usage{
CreateProjectsPredictionDatasetsFileUploads(
  predictionsEndDate,
  file,
  credentials,
  relaxKnownInAdvanceFeaturesCheck,
  predictionsStartDate,
  actualValueColumn,
  secondaryDatasetsConfigId,
  forecastPoint,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Upload a file for predictions from an attached file.
}
\details{
Upload a file for predictions from an attached file.

This method invokes \verb{POST /projects/\{projectId\}/predictionDatasets/fileUploads/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionDatasetsFileUploadsCreate in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
