% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/RatingTables.R
\name{CreateRatingTable}
\alias{CreateRatingTable}
\title{CreateRatingTable}
\usage{
CreateRatingTable(
  project,
  parentModelId,
  dataSource,
  ratingTableName = "Uploaded Rating Table"
)

CreateRatingTable(
  project,
  parentModelId,
  dataSource,
  ratingTableName = "Uploaded Rating Table"
)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{parentModelId}{integer. The id of the model to validate the rating table against.}

\item{dataSource}{object. Either (a) the name of a CSV file, or (b) a
dataframe. This parameter identifies the source of the rating table.}

\item{ratingTableName}{character. Optional. The name of the rating table.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An integer value that can be used as the JobId parameter
  in subsequent calls representing this job.
}
\description{
Upload Modified Rating Table File
}
\details{
Create a new rating table from a rating table file This will create a new rating table, regardless of whether the validation succeeds. The rating table object will have a validationError which will be left blank in the case of successful validation.

This method invokes \verb{POST /projects/\{projectId\}/ratingTables/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
modelId <- "5984b4d7100d2b31c1166529"
CreateRatingTable(projectId, modelId, dataSource = "myRatingTable.csv")
}
}
\seealso{
The method ProjectsRatingTablesCreate in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
