% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateRemoteEvents}
\alias{CreateRemoteEvents}
\title{CreateRemoteEvents}
\usage{
CreateRemoteEvents(
  spoolerChannelData,
  predictionRequestData,
  data,
  deploymentId,
  predictionEnvironmentId,
  eventType,
  message,
  orgId,
  timestamp,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Post a remote deployment event
}
\details{
Post an event from a remote deployment.

This method invokes \verb{POST /remoteEvents/} in the DataRobot Public API.
}
\seealso{
The method RemoteEventsCreate in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
