% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_configuration_apiWrapper.R
\name{CreateSsoConfigurations}
\alias{CreateSsoConfigurations}
\title{CreateSsoConfigurations}
\usage{
CreateSsoConfigurations(
  autoGenerateUsers,
  certificate,
  securityParameters,
  entityId,
  issuer,
  signOnUrl,
  organizationId,
  idpMetadataUrl,
  roleMapping,
  idpMetadataHttpsVerify,
  idpMetadata,
  attributeMapping,
  signOutUrl,
  enableSso,
  groupMapping,
  name,
  enforceSso,
  configurationType,
  idpResponseMethod = "POST",
  sessionLengthSeconds = 604800,
  spRequestMethod = "REDIRECT",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create an SSO configuration for a specific organization
}
\details{
Create an SSO configuration for a specific organization

This method invokes \verb{POST /ssoConfigurations/} in the DataRobot Public API.
}
\seealso{
The method SsoConfigurationsCreate in \link[datarobot.apicore:SsoConfigurationApi]{datarobot.apicore::SsoConfigurationApi}, which invokes the same underlying API endpoint.
}
