% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{CreateUserBlueprintsFromUserBlueprintId}
\alias{CreateUserBlueprintsFromUserBlueprintId}
\title{CreateUserBlueprintsFromUserBlueprintId}
\usage{
CreateUserBlueprintsFromUserBlueprintId(
  userBlueprintId,
  description,
  modelType,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)

CreateUserBlueprintsFromUserBlueprintId(
  userBlueprintId,
  description,
  modelType,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)
}
\description{
Clone a user blueprint.

Clone a user blueprint.
}
\details{
Clone a user blueprint.

Clone a user blueprint.
}
\seealso{
The method UserBlueprintsFromUserBlueprintIdCreate in [datarobot.apicore::AiCatalogApi], which this function wraps.

The method UserBlueprintsFromUserBlueprintIdCreate in [datarobot.apicore::BlueprintsApi], which this function wraps.

Other AiCatalog: 
\code{\link{CreateDataEngineQueryGenerator}()},
\code{\link{CreateDataEngineWorkspaceStates}()},
\code{\link{CreateDatasetFromDataEngineQueryGenerator}()},
\code{\link{CreateDatasetsFeatureTransforms}()},
\code{\link{CreateDatasetsFeaturelists}()},
\code{\link{CreateDatasetsFromDataEngineWorkspaceState}()},
\code{\link{CreateDatasetsFromDataSource}()},
\code{\link{CreateDatasetsFromFile}()},
\code{\link{CreateDatasetsFromHDFS}()},
\code{\link{CreateDatasetsFromURL}()},
\code{\link{CreateDatasetsRefreshJobs}()},
\code{\link{CreateDatasetsRelationships}()},
\code{\link{CreateDatasetsVersionsFromDataEngineWorkspaceState}()},
\code{\link{CreateDatasetsVersionsFromDataSource}()},
\code{\link{CreateDatasetsVersionsFromFile}()},
\code{\link{CreateDatasetsVersionsFromHDFS}()},
\code{\link{CreateDatasetsVersionsFromLatestVersion}()},
\code{\link{CreateDatasetsVersionsFromURL}()},
\code{\link{CreateDatasetsVersionsFromVersion}()},
\code{\link{CreateUserBlueprintsBulkValidations}()},
\code{\link{CreateUserBlueprintsFromBlueprintId}()},
\code{\link{CreateUserBlueprintsFromCustomTaskVersionId}()},
\code{\link{CreateUserBlueprintsProjectBlueprints}()},
\code{\link{CreateUserBlueprintsTaskParameters}()},
\code{\link{CreateUserBlueprintsValidations}()},
\code{\link{CreateUserBlueprints}()},
\code{\link{DeleteDatasetsFeaturelists}()},
\code{\link{DeleteDatasetsRefreshJobs}()},
\code{\link{DeleteDatasetsRelationships}()},
\code{\link{DeleteDatasetsVersions}()},
\code{\link{DeleteDatasets}()},
\code{\link{DeleteManyUserBlueprints}()},
\code{\link{DeleteUserBlueprints}()},
\code{\link{DownloadDatasetAsCsv}()},
\code{\link{ListCatalogItems}()},
\code{\link{ListDatasetsAccessControl}()},
\code{\link{ListDatasetsAllFeaturesDetails}()},
\code{\link{ListDatasetsFeatureTransforms}()},
\code{\link{ListDatasetsFeaturelists}()},
\code{\link{ListDatasetsFile}()},
\code{\link{ListDatasetsPermissions}()},
\code{\link{ListDatasetsProjects}()},
\code{\link{ListDatasetsRefreshJobsExecutionResults}()},
\code{\link{ListDatasetsRefreshJobs}()},
\code{\link{ListDatasetsRelationships}()},
\code{\link{ListDatasetsSharedRoles}()},
\code{\link{ListDatasetsVersionsAllFeaturesDetails}()},
\code{\link{ListDatasetsVersionsFeaturelists}()},
\code{\link{ListDatasetsVersionsFile}()},
\code{\link{ListDatasetsVersionsProjects}()},
\code{\link{ListDatasetsVersions}()},
\code{\link{ListDatasets}()},
\code{\link{ListUserBlueprintsInputTypes}()},
\code{\link{ListUserBlueprintsSharedRoles}()},
\code{\link{ListUserBlueprintsTasks}()},
\code{\link{ListUserBlueprints}()},
\code{\link{PatchCatalogItems}()},
\code{\link{PatchDatasetsFeaturelists}()},
\code{\link{PatchDatasetsRefreshJobs}()},
\code{\link{PatchDatasetsRelationships}()},
\code{\link{PatchDatasets}()},
\code{\link{PatchManyDatasetsAccessControl}()},
\code{\link{PatchManyDatasetsDeleted}()},
\code{\link{PatchManyDatasetsSharedRoles}()},
\code{\link{PatchManyDatasetsVersionsDeleted}()},
\code{\link{PatchManyDatasets}()},
\code{\link{PatchManyUserBlueprintsSharedRoles}()},
\code{\link{PatchUserBlueprints}()},
\code{\link{RetrieveCatalogItems}()},
\code{\link{RetrieveDataEngineQueryGenerators}()},
\code{\link{RetrieveDataEngineWorkspaceStates}()},
\code{\link{RetrieveDatasetsFeatureHistograms}()},
\code{\link{RetrieveDatasetsFeatureTransforms}()},
\code{\link{RetrieveDatasetsFeaturelists}()},
\code{\link{RetrieveDatasetsRefreshJobs}()},
\code{\link{RetrieveDatasetsVersionsFeatureHistograms}()},
\code{\link{RetrieveDatasetsVersionsFeaturelists}()},
\code{\link{RetrieveDatasetsVersions}()},
\code{\link{RetrieveDatasets}()},
\code{\link{RetrieveUserBlueprints}()}

Other Blueprints: 
\code{\link{CreateUserBlueprintsBulkValidations}()},
\code{\link{CreateUserBlueprintsFromBlueprintId}()},
\code{\link{CreateUserBlueprintsFromCustomTaskVersionId}()},
\code{\link{CreateUserBlueprintsProjectBlueprints}()},
\code{\link{CreateUserBlueprintsTaskParameters}()},
\code{\link{CreateUserBlueprintsValidations}()},
\code{\link{CreateUserBlueprints}()},
\code{\link{DeleteManyUserBlueprints}()},
\code{\link{DeleteUserBlueprints}()},
\code{\link{GetBlueprintChart}()},
\code{\link{GetBlueprintDocumentation}()},
\code{\link{GetBlueprint}()},
\code{\link{GetModelBlueprintChart}()},
\code{\link{GetModelBlueprintDocumentation}()},
\code{\link{ListBlueprints}()},
\code{\link{ListProjectsModelsLogs}()},
\code{\link{ListProjectsModelsTrainingArtifact}()},
\code{\link{ListUserBlueprintsInputTypes}()},
\code{\link{ListUserBlueprintsSharedRoles}()},
\code{\link{ListUserBlueprintsTasks}()},
\code{\link{ListUserBlueprints}()},
\code{\link{PatchManyUserBlueprintsSharedRoles}()},
\code{\link{PatchUserBlueprints}()},
\code{\link{RetrieveUserBlueprints}()}
}
\concept{AiCatalog}
\concept{Blueprints}
