% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{CreateUserCleanupJobs}
\alias{CreateUserCleanupJobs}
\title{CreateUserCleanupJobs}
\usage{
CreateUserCleanupJobs(orphansOwner, userIds, orgId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Users permanent delete.
}
\details{
Schedules a job which permanently delete the users represented by the provided users ids or organization id. Number of users to be deleted in one go is restricted by DELETED_USERS_BATCH_LIMIT system setting.

This method invokes \verb{POST /userCleanupJobs/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupJobsCreate in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
