% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataRobot.R
\docType{package}
\name{datarobot-package}
\alias{datarobot}
\alias{datarobot-package}
\title{datarobot: 'DataRobot' Predictive Modeling API}
\description{
For working with the 'DataRobot' predictive modeling platform's API <https://www.datarobot.com/>.
}
