% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnomalyAssessment.R
\name{DeleteAnomalyAssessmentRecord}
\alias{DeleteAnomalyAssessmentRecord}
\title{Delete anomaly assessment record.}
\usage{
DeleteAnomalyAssessmentRecord(projectId, recordId)
}
\arguments{
\item{projectId}{character. The ID of the project.}

\item{recordId}{character. The ID of the anomaly assessment record.}
}
\description{
Record is deleted with preview and explanations.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
recordId <- "59a5af20c80891534e3c2bdb"
explanations <- DeleteAnomalyAssessmentRecord(projectId, recordId)
}
}
\seealso{
Other Anomaly Assessment functions: 
\code{\link{GetAnomalyAssessmentExplanations}()},
\code{\link{GetAnomalyAssessmentPredictionsPreview}()},
\code{\link{InitializeAnomalyAssessment}()},
\code{\link{ListAnomalyAssessmentRecords}()}
}
\concept{Anomaly Assessment functions}
