% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{DeleteApplications}
\alias{DeleteApplications}
\title{DeleteApplications}
\usage{
DeleteApplications(applicationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete an application
}
\details{
Delete an application

This method invokes \verb{DELETE /applications/\{applicationId\}/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsDelete in \link[datarobot.apicore:ApplicationsApi]{datarobot.apicore::ApplicationsApi}, which invokes the same underlying API endpoint.
}
