% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{DeleteApplicationsDeployments}
\alias{DeleteApplicationsDeployments}
\title{DeleteApplicationsDeployments}
\usage{
DeleteApplicationsDeployments(modelDeploymentId, applicationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete link between application and deployment.
}
\details{
Delete link between application and deployment.

This method invokes \verb{DELETE /applications/\{applicationId\}/deployments/\{modelDeploymentId\}/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsDeploymentsDelete in \link[datarobot.apicore:ApplicationsApi]{datarobot.apicore::ApplicationsApi}, which invokes the same underlying API endpoint.
}
