% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{DeleteApprovalPolicies}
\alias{DeleteApprovalPolicies}
\title{DeleteApprovalPolicies}
\usage{
DeleteApprovalPolicies(approvalPolicyId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete an Approval Policy.
}
\details{
Delete the policy with the given ID.

This method invokes \verb{DELETE /approvalPolicies/\{approvalPolicyId\}/} in the DataRobot Public API.
}
\seealso{
The method ApprovalPoliciesDelete in \link[datarobot.apicore:GovernanceApi]{datarobot.apicore::GovernanceApi}, which invokes the same underlying API endpoint.
}
