% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendars.R
\name{DeleteCalendar}
\alias{DeleteCalendar}
\title{Delete a calendar}
\usage{
DeleteCalendar(calendarId)
}
\arguments{
\item{calendarId}{character. The ID of the calendar to retrieve.}
}
\description{
Delete a calendar
}
\examples{
\dontrun{
   calendarId <- "5da75da31fb4a45b8a815a53"
   DeleteCalendar(calendarId)
}
}
