% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{DeleteComments}
\alias{DeleteComments}
\title{DeleteComments}
\usage{
DeleteComments(commentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a comment
}
\details{
Delete a comment

This method invokes \verb{DELETE /comments/\{commentId\}/} in the DataRobot Public API.
}
\seealso{
The method CommentsDelete in \link[datarobot.apicore:CommentsApi]{datarobot.apicore::CommentsApi}, which invokes the same underlying API endpoint.
}
