% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplianceDocumentation.R
\name{DeleteComplianceDocTemplate}
\alias{DeleteComplianceDocTemplate}
\title{Deletes a compliance doc template.}
\usage{
DeleteComplianceDocTemplate(templateId)
}
\arguments{
\item{templateId}{character. The ID of the template to update.}
}
\value{
Nothing returned, but deletes teh compliance doc template.
}
\description{
Note that default templates cannot be deleted.
}
\examples{
\dontrun{
  templateId <- "5cf85080d9436e5c310c796d"
  DeleteComplianceDocTemplate(templateId)
}
}
