% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{DeleteCustomTasks}
\alias{DeleteCustomTasks}
\title{DeleteCustomTasks}
\usage{
DeleteCustomTasks(customTaskId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete custom task.
}
\details{
Delete a custom task. Only users who have permission to edit custom task can delete it. Only custom tasks which are not currently deployed can be deleted. Relevant CustomTaskImage will be deleted also.

This method invokes \verb{DELETE /customTasks/\{customTaskId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksDelete in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
