% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStores.R
\name{DeleteDataStore}
\alias{DeleteDataStore}
\title{Delete a data store.}
\usage{
DeleteDataStore(dataStoreId)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}
}
\description{
Delete a data store.
}
\examples{
\dontrun{
 dataStoreId <- "5c1303269300d900016b41a7"
 DeleteDataStore(dataStoreId)
}
}
