% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataStores.R
\name{DeleteDataStore}
\alias{DeleteDataStore}
\title{DeleteDataStore}
\usage{
DeleteDataStore(dataStoreId)

DeleteDataStore(dataStoreId)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete the data store.
}
\details{
Delete the data store with given ID if it is not used by any data source.

This method invokes \verb{DELETE /externalDataStores/\{dataStoreId\}/} in the DataRobot Public API.

Delete a data store.
}
\examples{
\dontrun{
dataStoreId <- "5c1303269300d900016b41a7"
DeleteDataStore(dataStoreId)
}
}
\seealso{
The method ExternalDataStoresDelete in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
