% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deployment.R
\name{DeleteDeployment}
\alias{DeleteDeployment}
\title{Delete a deployment.}
\usage{
DeleteDeployment(deploymentId)
}
\arguments{
\item{deploymentId}{character. The ID of the deployment.}
}
\description{
Delete a deployment.
}
\examples{
\dontrun{
  deploymentId <- "5e319d2e422fbd6b58a5edad"
  DeleteDeployment(deploymentId)
}
}
