% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{DeleteExternalDataDrivers}
\alias{DeleteExternalDataDrivers}
\title{DeleteExternalDataDrivers}
\usage{
DeleteExternalDataDrivers(driverId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete the driver.
}
\details{
Delete the driver with given ID if it is not used by any data store or data source.

This method invokes \verb{DELETE /externalDataDrivers/\{driverId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataDriversDelete in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
