% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jobs.R
\name{DeleteJob}
\alias{DeleteJob}
\title{Cancel a running job}
\usage{
DeleteJob(job)
}
\arguments{
\item{job}{object. The job you want to cancel (one of the items in the list returned from
\code{ListJobs})}
}
\description{
Cancel a running job
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  initialJobs <- ListModelJobs(project)
  job <- initialJobs[[1]]
  DeleteJob(job)
}
}
