% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{DeleteModel}
\alias{DeleteModel}
\title{Delete a specified DataRobot model}
\usage{
DeleteModel(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModelObject, or each element of the list returned by
the function GetAllModels.}
}
\description{
This function removes the model specified by the parameter model from its
associated project.
}
