% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{DeleteNotificationChannels}
\alias{DeleteNotificationChannels}
\title{DeleteNotificationChannels}
\usage{
DeleteNotificationChannels(channelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete notification channel
}
\details{
Delete the notification channel.

This method invokes \verb{DELETE /notificationChannels/\{channelId\}/} in the DataRobot Public API.
}
\seealso{
The method NotificationChannelsDelete in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
