% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{DeleteNotificationPolicyMutes}
\alias{DeleteNotificationPolicyMutes}
\title{DeleteNotificationPolicyMutes}
\usage{
DeleteNotificationPolicyMutes(muteId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete the existing notification policy mute
}
\details{
Delete the existing notification policy mute

This method invokes \verb{DELETE /notificationPolicyMutes/\{muteId\}/} in the DataRobot Public API.
}
\seealso{
The method NotificationPolicyMutesDelete in \link[datarobot.apicore:NotificationsApi]{datarobot.apicore::NotificationsApi}, which invokes the same underlying API endpoint.
}
