% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionDatasets.R,
%   R/predictions_apiWrapper.R
\name{DeletePredictionDataset}
\alias{DeletePredictionDataset}
\title{Delete a specified prediction dataset}
\usage{
DeletePredictionDataset(project, datasetId, ...)

DeletePredictionDataset(project, datasetId, ...)
}
\arguments{
\item{datasetId}{The id of the dataset to delete}
}
\description{
This function removes a prediction dataset

Delete a dataset that was uploaded for prediction.
}
\details{
Delete a dataset that was uploaded for prediction.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
datasets <- ListPredictionDatasets(projectId)
dataset <- datasets[[1]]
datasetId <- dataset$id
DeletePredictionDataset(projectId, datasetId)
}
}
\seealso{
The method ProjectsPredictionDatasetsDelete in [datarobot.apicore::PredictionsApi], which this function wraps.

Other Predictions: 
\code{\link{CreateBatchPredictionJobDefinitions}()},
\code{\link{CreateBatchPredictionsCsvUploadFinalizeMultipart}()},
\code{\link{CreateBatchPredictionsFromExisting}()},
\code{\link{CreateBatchPredictionsFromJobDefinition}()},
\code{\link{CreateBatchPredictions}()},
\code{\link{CreateProjectsPredictionDatasetsFileUploads}()},
\code{\link{DeleteBatchPredictionJobDefinitions}()},
\code{\link{DeleteBatchPredictions}()},
\code{\link{DeletePredictJob}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteScheduledJobs}()},
\code{\link{DeleteTransferableModel}()},
\code{\link{GetPredictJobs}()},
\code{\link{GetPredictJob}()},
\code{\link{GetPredictionDataset}()},
\code{\link{GetPredictionExplanationsMetadata}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetTransferableModel}()},
\code{\link{ListBatchPredictionJobDefinitionsPortable}()},
\code{\link{ListBatchPredictionJobDefinitions}()},
\code{\link{ListBatchPredictionsDownload}()},
\code{\link{ListBatchPredictions}()},
\code{\link{ListPredictionDatasets}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListPredictions}()},
\code{\link{ListProjectsPredictionsMetadata}()},
\code{\link{ListScheduledJobs}()},
\code{\link{ListTrainingPredictions}()},
\code{\link{ListTransferableModels}()},
\code{\link{PatchBatchPredictionJobDefinitions}()},
\code{\link{PatchBatchPredictions}()},
\code{\link{PatchScheduledJobs}()},
\code{\link{PutBatchPredictionsCsvUploadPart}()},
\code{\link{PutImportedModels}()},
\code{\link{PutManyBatchPredictionsCsvUpload}()},
\code{\link{RetrieveBatchPredictionJobDefinitions}()},
\code{\link{RetrieveBatchPredictions}()},
\code{\link{RetrieveProjectsPredictionsMetadata}()},
\code{\link{RetrieveProjectsPredictions}()},
\code{\link{RetrieveScheduledJobs}()},
\code{\link{UpdateTransferableModel}()},
\code{\link{UploadPredictionDatasetFromCatalog}()},
\code{\link{UploadPredictionDatasetFromDataSource}()},
\code{\link{UploadPredictionDataset}()},
\code{\link{UploadTransferableModel}()}
}
\concept{Predictions}
