% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R, R/DeletePredictJob.R
\name{DeletePredictJob}
\alias{DeletePredictJob}
\title{DeletePredictJob}
\usage{
DeletePredictJob(project, predictJobId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{predictJobId}{integer. The integer ID \code{predictionJobId}
that is created by the call to \code{RequestPredictions}.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
Logical TRUE and displays a message to the user if the delete
request was successful; otherwise, execution halts and an error message
is displayed.
}
\description{
Cancel a queued prediction job
}
\details{
Cancel a queued prediction job

This method invokes \verb{DELETE /projects/\{projectId\}/predictJobs/\{jobId\}/} in the DataRobot Public API.

Function to delete one predict job from the DataRobot queue

This function deletes the predict job specified by predictJobId from
the DataRobot queue.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
initialJobs <- GetPredictJobs(project)
job <- initialJobs[[1]]
predictJobId <- job$predictJobId
DeletePredictJob(projectId, predictJobId)
}
}
\seealso{
The method ProjectsPredictJobsDelete in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
