% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReasonCodes.R
\name{DeleteReasonCodesInitialization}
\alias{DeleteReasonCodesInitialization}
\title{Delete the reason codes initialization for a model.}
\usage{
DeleteReasonCodesInitialization(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModel, or each element of the list returned by
the function ListModels.}
}
\description{
Delete the reason codes initialization for a model.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  modelId <- "5996f820af07fc605e81ead4"
  model <- GetModel(projectId, modelId)
  DeleteReasonCodesInitialization(model)
}
}
