% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferableModels.R
\name{DeleteTransferableModel}
\alias{DeleteTransferableModel}
\title{Delete this imported model.}
\usage{
DeleteTransferableModel(importId)
}
\arguments{
\item{importId}{character. Id of the import.}
}
\description{
Delete this imported model.
}
\examples{
\dontrun{
  id <- UploadTransferableModel("model.drmodel")
  DeleteTransferableModel(id)
}
}
