% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/Models.R
\name{DownloadScoringCode}
\alias{DownloadScoringCode}
\title{DownloadScoringCode}
\usage{
DownloadScoringCode(project, modelId, fileName, sourceCode = FALSE)

DownloadScoringCode(project, modelId, fileName, sourceCode = FALSE)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{fileName}{character. File path where scoring code will be saved.}

\item{sourceCode}{logical. Optional. Set to TRUE to download source code archive.
It will not be executable.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Scoring Code
}
\details{
Retrieve Scoring Code for making new predictions from an existing model offline. You need the \"Scoring Code\" feature enabled to use this route.  By default, returns a compiled executable JAR that can be executed locally to calculate model predictions, or it can be used as a library for a Java application. Execute it with the '--help\\` parameters to learn how to use it as a command-line utility. See model API documentation (\url{https://javadoc.io/doc/com.datarobot/datarobot-prediction/latest/index.html}) to be able to use it inside an existing Java application.  With the sourceCode query parameter set to 'true', returns a source code archive that can be used to review internal calculations of the model. This JAR is NOT executable.  See \"\url{https://app.datarobot.com/docs/predictions/scoring-code/scorecode.html}\" in DataRobot application for more information.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/scoringCode/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
file <- file.path(tempdir(), "scoringCode.jar")
DownloadScoringCode(projectId, modelId, file)
}
}
\seealso{
The method ProjectsModelsScoringCodeList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
