% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureLogs.R
\name{DownloadTimeSeriesFeatureDerivationLog}
\alias{DownloadTimeSeriesFeatureDerivationLog}
\title{Download the time series feature derivation log as a text file.}
\usage{
DownloadTimeSeriesFeatureDerivationLog(project, file)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{file}{character. The name or path of the file to download to.}
}
\value{
Nothing, but writes the output to the desired file.
}
\description{
Download the time series feature derivation log as a text file.
}
\examples{
\dontrun{
 projectId <- "5984b4d7100d2b31c1166529"
 DownloadTimeSeriesFeatureDerivationLog(projectId, "featureLog.txt")
}
}
\seealso{
\code{\link{GetTimeSeriesFeatureDerivationLog}}
}
