% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_partitioning_apiWrapper.R,
%   R/FeatureLogs.R
\name{DownloadTimeSeriesFeatureDerivationLog}
\alias{DownloadTimeSeriesFeatureDerivationLog}
\title{DownloadTimeSeriesFeatureDerivationLog}
\usage{
DownloadTimeSeriesFeatureDerivationLog(project, file)

DownloadTimeSeriesFeatureDerivationLog(project, file)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{file}{character. The name or path of the file to download to.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
Nothing, but writes the output to the desired file.
}
\description{
Retrieve a text file containing the time series project feature log
}
\details{
Retrieve a text file containing the time series project feature log.  The Time Series Feature Log provides details about the feature generation process for a time series project. It includes information about which features are generated and their priority,as well as the detected properties of the time series data such as whether the series is stationary, and periodicities detected.  This route is only supported for time series projects that have finished partitioning.  The feature derivation log will include information about:  -   Maximum number of feature to be generated, e.g.,     \verb{Limit on the maximum number of feature in this project is 500} -   Number of derived features tested during the feature generation     process, e.g.,     \verb{Total number of derived features during the feature generation process is 571} -   Number of generated features removed during the feature reduction     process e.g.     \verb{Total number of features removed during the feature reduction process is 472} -   Number of remaining features after the combined feature generation     and reduction process, e.g.,     \verb{The finalized number of features is 99} -   Detected stationarity of the series, e.g.,     \verb{Series detected as non-stationary} -   Detected presence of multiplicative trend in the series, e.g.,     \verb{Multiplicative trend detected} -   Detected periodicities in the series, e.g.,     \verb{Detected periodicities: 7 day} -   Window sizes used in rolling statistics / lag extractors, e.g.,     \verb{The window sizes chosen to be: 2 months (because the time step is 1 month and Feature Derivation Window is 2 months)} -   Features that are specified as known-in-advance, e.g.,     \verb{Variables treated as apriori: holiday} -   Details about why certain variables are transformed in the input     data, e.g.,     \verb{Generating variable \\"y (log)\\" from \\"y\\" because multiplicative trend is detected} -   Details about features generated as time series features, and their     priority, e.g.,     \verb{Generating feature \\"date (actual)\\" from \\"date\\" (priority: 1)}

This method invokes \verb{GET /projects/\{projectId\}/timeSeriesFeatureLog/file/} in the DataRobot Public API.

Download the time series feature derivation log as a text file.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
DownloadTimeSeriesFeatureDerivationLog(projectId, "featureLog.txt")
}
}
\seealso{
The method ProjectsTimeSeriesFeatureLogFileList in \link[datarobot.apicore:DatetimePartitioningApi]{datarobot.apicore::DatetimePartitioningApi}, which invokes the same underlying API endpoint.

\code{\link{GetTimeSeriesFeatureDerivationLog}}
}
