% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blueprints.R
\name{GetBlueprintChart}
\alias{GetBlueprintChart}
\title{Retrieve a blueprint chart}
\usage{
GetBlueprintChart(project, blueprintId)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{blueprintId}{Character strint Id of blueprint to retrieve}
}
\value{
List with the following two components:
\describe{
  \item{nodes}{list each element contains information about one node of a blueprint : id and label}
  \item{edges}{Two column matrix, identifying blueprint nodes connections}
}
}
\description{
A Blueprint chart can be used to understand data flow in blueprint.
}
