% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R, R/Blueprints.R
\name{GetBlueprintChart}
\alias{GetBlueprintChart}
\title{GetBlueprintChart}
\usage{
GetBlueprintChart(project, blueprintId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{blueprintId}{character. Id of blueprint to retrieve.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
List with the following two components:
\itemize{
  \item nodes. list each element contains information about one node
     of a blueprint : id and label.
  \item edges. Two column matrix, identifying blueprint nodes connections.
}
}
\description{
Retrieve a blueprint chart by blueprint id.
}
\details{
Retrieve a blueprint chart by blueprint id.

This method invokes \verb{GET /projects/\{projectId\}/blueprints/\{blueprintId\}/blueprintChart/} in the DataRobot Public API.

Retrieve a blueprint chart

A Blueprint chart can be used to understand data flow in blueprint.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
model <- GetModel(projectId, modelId)
blueprintId <- model$blueprintId
GetBlueprintChart(projectId, blueprintId)
}
}
\seealso{
The method ProjectsBlueprintsBlueprintChartList in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.
}
