% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blueprints.R
\name{GetBlueprintDocuments}
\alias{GetBlueprintDocuments}
\title{Get documentation for tasks used in the blueprint}
\usage{
GetBlueprintDocuments(project, blueprintId)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{blueprintId}{character. Id of blueprint to retrieve.}
}
\description{
(Deprecated in 2.8, will be removed in 2.10. Use GetBlueprintDocumentation instead.)
}
