% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendars.R
\name{GetCalendar}
\alias{GetCalendar}
\title{Retrieve a calendar}
\usage{
GetCalendar(calendarId)
}
\arguments{
\item{calendarId}{character. The ID of the calendar to retrieve.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Retrieve a calendar
}
\examples{
\dontrun{
   calendarId <- "5da75da31fb4a45b8a815a53"
   GetCalendar(calendarId)
}
}
