% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Calendars.R
\name{GetCalendar}
\alias{GetCalendar}
\title{GetCalendar}
\usage{
GetCalendar(calendarId)

GetCalendar(calendarId)
}
\arguments{
\item{calendarId}{character. The ID of the calendar to retrieve.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Retrieve information about a calendar.
}
\details{
List all the information about a calendar such as the total number of event dates, the earliest calendar event date, the IDs of projects currently using this calendar and the others.

This method invokes \verb{GET /calendars/\{calendarId\}/} in the DataRobot Public API.
}
\examples{
\dontrun{
calendarId <- "5da75da31fb4a45b8a815a53"
GetCalendar(calendarId)
}
}
\seealso{
The method CalendarsRetrieve in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
