% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendars.R
\name{GetCalendarFromProject}
\alias{GetCalendarFromProject}
\title{Retrieve the calendar for a particular project.}
\usage{
GetCalendarFromProject(project)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Retrieve the calendar for a particular project.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
   GetCalendar(projectId)
}
}
