% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataSources.R
\name{GetDataSource}
\alias{GetDataSource}
\title{GetDataSource}
\usage{
GetDataSource(dataSourceId, ...)
}
\arguments{
\item{dataSourceId}{character. The id of the data source}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list containing information on the particular data source:
\itemize{
  \item className character. The Java class name of the driver.
  \item baseNames character. A vector of the file name(s) of the jar files.
  \item canonicalName character. The user-friendly name of the driver.
  \item id character. The dataSourceId of the driver.
  \item creator character. The userId of the user who created the driver.
}
}
\description{
Data source details.
}
\details{
Return details of the existing data source with given ID, including SQL query or selected table and schema data, which fully describe which data to extract and from which location.

This method invokes \verb{GET /externalDataSources/\{dataSourceId\}/} in the DataRobot Public API.

Returns information about a particular data source.
}
\examples{
\dontrun{
dataSourceId <- "57a7c978c808916f4a630f89"
GetDataSource(dataSourceId)
}
}
\seealso{
The method ExternalDataSourcesRetrieve in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
