% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataStores.R
\name{GetDataStore}
\alias{GetDataStore}
\title{GetDataStore}
\usage{
GetDataStore(dataStoreId, ...)
}
\arguments{
\item{dataStoreId}{character. The id of the data store.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list containing information on the particular data store:
\itemize{
  \item id character. The dataStoreId of the data store.
  \item canonicalName character. The user-friendly name of the data store.
  \item type character. The type of data store.
  \item updated datetime. A timestamp for the last time the data store was updated.
  \item creator character. The userId of the user who created the data store.
  \item params list. A list specifying the data store parameters.
}
}
\description{
Data store details.
}
\details{
A configured connection to a database - it has a name and a specified driver. The driver may be specified by a JDBC URL or connection parameters if the driver was created with the parameter configuration.

This method invokes \verb{GET /externalDataStores/\{dataStoreId\}/} in the DataRobot Public API.

Returns information about a particular data store.
}
\examples{
\dontrun{
dataStoreId <- "5c1303269300d900016b41a7"
GetDataStore(dataStoreId)
}
}
\seealso{
The method ExternalDataStoresRetrieve in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
