% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStores.R
\name{GetDataStoreSchemas}
\alias{GetDataStoreSchemas}
\title{Get the schemas associated with a data store.}
\usage{
GetDataStoreSchemas(dataStoreId, username, password)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}

\item{username}{character. The username to use for authentication to the database.}

\item{password}{character. The password to use for authentication to the database.
The password is encrypted at server side and never saved or stored.}
}
\value{
A list with the name of the catalog and the name of the schemas.
}
\description{
Get the schemas associated with a data store.
}
\examples{
\dontrun{
 dataStoreId <- "5c1303269300d900016b41a7"
 GetDataStoreSchemas(dataStoreId, username = "myUser", password = "mySecurePass129")
}
}
