% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataStores.R
\name{GetDataStoreSchemas}
\alias{GetDataStoreSchemas}
\title{GetDataStoreSchemas}
\usage{
GetDataStoreSchemas(dataStoreId, username, password, ...)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}

\item{username}{character. The username to use for authentication to the database.}

\item{password}{character. The password to use for authentication to the database.
The password is encrypted at server side and never saved or stored.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list with the name of the catalog and the name of the schemas.
}
\description{
Retrieves a data store's data schemas.
}
\details{
Retrieves a data store's data schemas.

This method invokes \verb{POST /externalDataStores/\{dataStoreId\}/schemas/} in the DataRobot Public API.

Get the schemas associated with a data store.
}
\examples{
\dontrun{
dataStoreId <- "5c1303269300d900016b41a7"
GetDataStoreSchemas(dataStoreId, username = "myUser", password = "mySecurePass129")
}
}
\seealso{
The method ExternalDataStoresSchemasCreate in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
