% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.R
\name{GetDatetimePartition}
\alias{GetDatetimePartition}
\title{Retrieve the DatetimePartitioning from a project}
\usage{
GetDatetimePartition(project)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}
}
\value{
list describing datetime partition. See GeneratetDatetimePartition
}
\description{
Only available if the project has already set the target as a datetime project.
}
