% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R, R/Features.R
\name{GetFeatureDiscoveryRelationships}
\alias{GetFeatureDiscoveryRelationships}
\title{GetFeatureDiscoveryRelationships}
\usage{
GetFeatureDiscoveryRelationships(project)

GetFeatureDiscoveryRelationships(project)
}
\arguments{
\item{project}{object. Either list with projectId element or projectId value}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve relationships configuration for a project

Gets the feature discovery relationships for the specified Project.
}
\details{
Retrieve relationships configuration for a project

This method invokes \verb{GET /projects/\{projectId\}/relationshipsConfiguration/} in the DataRobot Public API.
}
\seealso{
The method RelationshipsConfigurationList in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.
}
\concept{Features}
