% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{GetFeatureInfo}
\alias{GetFeatureInfo}
\title{Details about a feature}
\usage{
GetFeatureInfo(project, featureId)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{featureId}{id of the feature to be retrieved}
}
\value{
A named list which contains:
\describe{
 \item{id}{feature id}
 \item{name}{feature name}
 \item{feature_type}{feature type: 'Numeric', 'Categorical', etc.}
 \item{importance}{numeric measure of the strength of relationship between the feature and
 target (independent of any model or other features).}
 \item{low_information}{whether feature has too few values to be informative}
 \item{unique_count}{number of unique values}
 \item{na_count}{number of missing values}}
}
\description{
Details about a feature
}

