% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R, R/MissingValuesReport.R
\name{GetMissingValuesReport}
\alias{GetMissingValuesReport}
\title{GetMissingValuesReport}
\usage{
GetMissingValuesReport(project, modelId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
A list containing:
\itemize{
  \item feature character. The name of the feature.
  \item type character. Feature type (numeric or categorical).
  \item missingCount numeric. The number of missing values in the training data for
    that feature.
  \item missingPercentage numeric. The percentage of missing values in the training
    data for the feature.
  \item tasks list. A list of information on each task that was applied to that feature
    to handle missing values. This information contains:
    \itemize{
      \item id character. The id of the node in the model blueprint chart for this task.
        (See \link{GetBlueprintChart} for more information on blueprint charts.)
      \item name character. The name of the task.
      \item descriptions character. Aggregated information about how the task handles
        missing values.
    }
}
}
\description{
Retrieve a summary of how the model's subtasks handle missing values.
}
\details{
Retrieve a summary of how the model's subtasks handle missing values   Only models built after the missing value report feature was added will have reports, and only models with at least one imputation or encoding task, e.g. ordinal encoding, missing value imputation. Blenders and scaleout models do not support Missing Value reports.  The report will describe how each feature's missing values were treated, and report how many missing values were present in the training data. Features which were not processed by a given blueprint task will not mention it: for instance, a categorical feature with many unique values may not be considered eligible for processing by a One-Hot Encoding  Report is collected for those features which are considered eligible by given blueprint task. For instance, categorical feature with a lot of unique values may not be considered as eligible in One-Hot Encoding task.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/missingReport/} in the DataRobot Public API.

Get a report on missing values for the model.

The missing values report is a list of items, one per feature, sorted by
missing count in descending order. Each item in the report contains details on the number
of missing values for that feature and how they were handled by the model.
}
\examples{
\dontrun{
projectId <- "5984b4d7100d2b31c1166529"
modelId <- "5984b4d7100d2b31c1166529"
GetMissingValuesReport(projectId, modelId)
}
}
\seealso{
The method ProjectsModelsMissingReportList in \link[datarobot.apicore:ModelsApi]{datarobot.apicore::ModelsApi}, which invokes the same underlying API endpoint.
}
