% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R, R/Blueprints.R
\name{GetModelBlueprintChart}
\alias{GetModelBlueprintChart}
\title{GetModelBlueprintChart}
\usage{
GetModelBlueprintChart(project, modelId, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{modelId}{character. Unique alphanumeric identifier for the model of interest.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
List with the following two components:
\itemize{
  \item nodes. list each element contains information about one node
     of a blueprint : id and label.
  \item edges. Two column matrix, identifying blueprint nodes connections.
}
}
\description{
Retrieve a reduced model blueprint chart by model id.
}
\details{
Retrieve a reduced model blueprint chart by model id. The model blueprint charts are reduced from the full blueprint charts to show only those sections of the blueprint that were actually used in the model, given the selected featurelist.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/blueprintChart/} in the DataRobot Public API.

Retrieve a model blueprint chart

A model blueprint is a "pruned down" blueprint representing what was actually run for the model.
This is solely the branches of the blueprint that were executed based on the featurelist.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
modelId <- "5996f820af07fc605e81ead4"
GetModelBlueprintChart(projectId, modelId)
}
}
\seealso{
The method ProjectsModelsBlueprintChartList in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.
}
