% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{GetModelBlueprintDocumentation}
\alias{GetModelBlueprintDocumentation}
\title{GetModelBlueprintDocumentation}
\usage{
GetModelBlueprintDocumentation(project, modelId, ...)
}
\description{
Retrieve task documentation for a reduced model blueprint.
}
\details{
Retrieve task documentation for a reduced model blueprint. The model blueprint is reduced from the full blueprint to show only those sections of the blueprint that were actually used in the model, given the selected featurelist.
}
\seealso{
The method ProjectsModelsBlueprintDocsList in [datarobot.apicore::BlueprintsApi], which this function wraps.

Other Blueprints: 
\code{\link{CreateUserBlueprintsBulkValidations}()},
\code{\link{CreateUserBlueprintsFromBlueprintId}()},
\code{\link{CreateUserBlueprintsFromCustomTaskVersionId}()},
\code{\link{CreateUserBlueprintsFromUserBlueprintId}()},
\code{\link{CreateUserBlueprintsProjectBlueprints}()},
\code{\link{CreateUserBlueprintsTaskParameters}()},
\code{\link{CreateUserBlueprintsValidations}()},
\code{\link{CreateUserBlueprints}()},
\code{\link{DeleteManyUserBlueprints}()},
\code{\link{DeleteUserBlueprints}()},
\code{\link{GetBlueprintChart}()},
\code{\link{GetBlueprintDocumentation}()},
\code{\link{GetBlueprint}()},
\code{\link{GetModelBlueprintChart}()},
\code{\link{ListBlueprints}()},
\code{\link{ListProjectsModelsLogs}()},
\code{\link{ListProjectsModelsTrainingArtifact}()},
\code{\link{ListUserBlueprintsInputTypes}()},
\code{\link{ListUserBlueprintsSharedRoles}()},
\code{\link{ListUserBlueprintsTasks}()},
\code{\link{ListUserBlueprints}()},
\code{\link{PatchManyUserBlueprintsSharedRoles}()},
\code{\link{PatchUserBlueprints}()},
\code{\link{RetrieveUserBlueprints}()}
}
\concept{Blueprints}
