% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{GetModelParameters}
\alias{GetModelParameters}
\title{Retrieve model parameters}
\usage{
GetModelParameters(project, modelId)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{modelId}{Unique alphanumeric identifier for the model of interest.}
}
\value{
List with the following components:
\describe{
  \item{parameters}{List of model parameters that are related to the whole model with following components: name, value}
  \item{derivedFeatures}{List containing preprocessing information about derived features with following components: 
  originalFeature, derivedFeature, type, coefficient and list of applied transformations}
}
}
\description{
Retrieve model parameters
}
