% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R,
%   R/predictions_apiWrapper.R
\name{GetPredictionExplanationsMetadata}
\alias{GetPredictionExplanationsMetadata}
\title{GetPredictionExplanationsMetadata}
\usage{
GetPredictionExplanationsMetadata(
  project,
  predictionExplanationId,
  datasetId = NULL,
  finishTime = NULL,
  id = NULL,
  maxExplanations = NULL,
  modelId = NULL,
  numColumns = NULL,
  predictionExplanationsLocation = NULL,
  predictionThreshold = NULL,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  ...
)

GetPredictionExplanationsMetadata(
  project,
  predictionExplanationId,
  datasetId = NULL,
  finishTime = NULL,
  id = NULL,
  maxExplanations = NULL,
  modelId = NULL,
  numColumns = NULL,
  predictionExplanationsLocation = NULL,
  predictionThreshold = NULL,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  ...
)
}
\description{
Retrieve a PredictionExplanationsRecord object.

Retrieve a PredictionExplanationsRecord object.
}
\details{
Retrieve a PredictionExplanationsRecord object. A PredictionExplanationsRecord contains metadata about the computed prediction explanations and the location at which the PredictionExplanations can be retrieved.

Retrieve a PredictionExplanationsRecord object. A PredictionExplanationsRecord contains metadata about the computed prediction explanations and the location at which the PredictionExplanations can be retrieved.
}
\seealso{
The method ProjectsPredictionExplanationsRecordsRetrieve in [datarobot.apicore::InsightsApi], which this function wraps.

The method ProjectsPredictionExplanationsRecordsRetrieve in [datarobot.apicore::PredictionsApi], which this function wraps.

Other Insights: 
\code{\link{CreateMultilabelInsightsPairwiseManualSelections}()},
\code{\link{CreateProjectsDatetimeModelsFeatureEffects}()},
\code{\link{CreateProjectsDatetimeModelsFeatureFit}()},
\code{\link{CreateProjectsDatetimeModelsMulticlassFeatureEffects}()},
\code{\link{CreateProjectsExternalScores}()},
\code{\link{CreateProjectsFeatureAssociationMatrix}()},
\code{\link{CreateProjectsGeometryFeaturePlots}()},
\code{\link{CreateProjectsModelsClusterInsights}()},
\code{\link{CreateProjectsModelsCrossClassAccuracyScores}()},
\code{\link{CreateProjectsModelsDataDisparityInsights}()},
\code{\link{CreateProjectsModelsFairnessInsights}()},
\code{\link{CreateProjectsModelsFeatureEffects}()},
\code{\link{CreateProjectsModelsFeatureFit}()},
\code{\link{CreateProjectsModelsImageActivationMaps}()},
\code{\link{CreateProjectsModelsImageEmbeddings}()},
\code{\link{CreateProjectsModelsMulticlassFeatureEffects}()},
\code{\link{CreateProjectsModelsShapImpact}()},
\code{\link{CreateProjectsPayoffMatrices}()},
\code{\link{CreateProjectsShapMatrices}()},
\code{\link{DeleteAnomalyAssessmentRecord}()},
\code{\link{DeleteMultilabelInsightsPairwiseManualSelections}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteProjectsPayoffMatrices}()},
\code{\link{GetAnomalyAssessmentExplanations}()},
\code{\link{GetAnomalyAssessmentPredictionsPreview}()},
\code{\link{GetFeatureAssociationMatrixDetails}()},
\code{\link{GetFeatureAssociationMatrix}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetWordCloud}()},
\code{\link{InitializeAnomalyAssessment}()},
\code{\link{ListAnomalyAssessmentRecords}()},
\code{\link{ListMultilabelInsightsHistogram}()},
\code{\link{ListMultilabelInsightsPairwiseManualSelections}()},
\code{\link{ListMultilabelInsightsPairwiseStatistics}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListProjectsBiasVsAccuracyInsights}()},
\code{\link{ListProjectsDatetimeModelsBacktestStabilityPlot}()},
\code{\link{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadata}()},
\code{\link{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsPreview}()},
\code{\link{ListProjectsDatetimeModelsFeatureEffectsMetadata}()},
\code{\link{ListProjectsDatetimeModelsFeatureEffects}()},
\code{\link{ListProjectsDatetimeModelsFeatureFitMetadata}()},
\code{\link{ListProjectsDatetimeModelsFeatureFit}()},
\code{\link{ListProjectsDatetimeModelsForecastDistanceStabilityPlot}()},
\code{\link{ListProjectsDatetimeModelsMulticlassFeatureEffects}()},
\code{\link{ListProjectsDatetimeModelsMultiseriesHistograms}()},
\code{\link{ListProjectsDatetimeModelsMultiseriesScoresFile}()},
\code{\link{ListProjectsExternalScores}()},
\code{\link{ListProjectsFeatureAssociationFeaturelists}()},
\code{\link{ListProjectsFeaturesFrequentValues}()},
\code{\link{ListProjectsImageActivationMaps}()},
\code{\link{ListProjectsImageEmbeddings}()},
\code{\link{ListProjectsModelsAnomalyInsightsFile}()},
\code{\link{ListProjectsModelsAnomalyInsightsTable}()},
\code{\link{ListProjectsModelsClusterInsightsDownload}()},
\code{\link{ListProjectsModelsClusterInsights}()},
\code{\link{ListProjectsModelsConfusionChartsClassDetails}()},
\code{\link{ListProjectsModelsConfusionChartsMetadata}()},
\code{\link{ListProjectsModelsConfusionCharts}()},
\code{\link{ListProjectsModelsCrossClassAccuracyScores}()},
\code{\link{ListProjectsModelsDataDisparityInsights}()},
\code{\link{ListProjectsModelsDatasetConfusionChartsClassDetails}()},
\code{\link{ListProjectsModelsDatasetConfusionChartsMetadata}()},
\code{\link{ListProjectsModelsDatasetConfusionCharts}()},
\code{\link{ListProjectsModelsDatasetLiftCharts}()},
\code{\link{ListProjectsModelsDatasetMulticlassLiftCharts}()},
\code{\link{ListProjectsModelsDatasetResidualsCharts}()},
\code{\link{ListProjectsModelsDatasetRocCurves}()},
\code{\link{ListProjectsModelsFairnessInsights}()},
\code{\link{ListProjectsModelsFeatureEffectsMetadata}()},
\code{\link{ListProjectsModelsFeatureEffects}()},
\code{\link{ListProjectsModelsFeatureFitMetadata}()},
\code{\link{ListProjectsModelsFeatureFit}()},
\code{\link{ListProjectsModelsFeatureListsClusterInsights}()},
\code{\link{ListProjectsModelsImageActivationMaps}()},
\code{\link{ListProjectsModelsImageEmbeddings}()},
\code{\link{ListProjectsModelsLabelwiseRocCurves}()},
\code{\link{ListProjectsModelsLiftChart}()},
\code{\link{ListProjectsModelsMulticlassFeatureEffects}()},
\code{\link{ListProjectsModelsMulticlassFeatureImpact}()},
\code{\link{ListProjectsModelsMulticlassLiftChart}()},
\code{\link{ListProjectsModelsResiduals}()},
\code{\link{ListProjectsModelsRocCurve}()},
\code{\link{ListProjectsModelsShapImpact}()},
\code{\link{ListProjectsMulticategoricalInvalidFormatFile}()},
\code{\link{ListProjectsMulticategoricalInvalidFormat}()},
\code{\link{ListProjectsPayoffMatrices}()},
\code{\link{ListProjectsShapMatrices}()},
\code{\link{PatchMultilabelInsightsPairwiseManualSelections}()},
\code{\link{PutProjectsPayoffMatrices}()},
\code{\link{RetrieveProjectsDatetimeModelsDatasetAccuracyOverTimePlots}()},
\code{\link{RetrieveProjectsGeometryFeaturePlots}()},
\code{\link{RetrieveProjectsModelsDatasetConfusionCharts}()},
\code{\link{RetrieveProjectsModelsLiftChart}()},
\code{\link{RetrieveProjectsModelsMulticlassLiftChart}()},
\code{\link{RetrieveProjectsModelsMultilabelLiftCharts}()},
\code{\link{RetrieveProjectsModelsResiduals}()},
\code{\link{RetrieveProjectsModelsRocCurve}()},
\code{\link{RetrieveProjectsShapMatrices}()}

Other Predictions: 
\code{\link{CreateBatchPredictionJobDefinitions}()},
\code{\link{CreateBatchPredictionsCsvUploadFinalizeMultipart}()},
\code{\link{CreateBatchPredictionsFromExisting}()},
\code{\link{CreateBatchPredictionsFromJobDefinition}()},
\code{\link{CreateBatchPredictions}()},
\code{\link{CreateProjectsPredictionDatasetsFileUploads}()},
\code{\link{DeleteBatchPredictionJobDefinitions}()},
\code{\link{DeleteBatchPredictions}()},
\code{\link{DeletePredictJob}()},
\code{\link{DeletePredictionDataset}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteScheduledJobs}()},
\code{\link{DeleteTransferableModel}()},
\code{\link{GetPredictJobs}()},
\code{\link{GetPredictJob}()},
\code{\link{GetPredictionDataset}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetTransferableModel}()},
\code{\link{ListBatchPredictionJobDefinitionsPortable}()},
\code{\link{ListBatchPredictionJobDefinitions}()},
\code{\link{ListBatchPredictionsDownload}()},
\code{\link{ListBatchPredictions}()},
\code{\link{ListPredictionDatasets}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListPredictions}()},
\code{\link{ListProjectsPredictionsMetadata}()},
\code{\link{ListScheduledJobs}()},
\code{\link{ListTrainingPredictions}()},
\code{\link{ListTransferableModels}()},
\code{\link{PatchBatchPredictionJobDefinitions}()},
\code{\link{PatchBatchPredictions}()},
\code{\link{PatchScheduledJobs}()},
\code{\link{PutBatchPredictionsCsvUploadPart}()},
\code{\link{PutImportedModels}()},
\code{\link{PutManyBatchPredictionsCsvUpload}()},
\code{\link{RetrieveBatchPredictionJobDefinitions}()},
\code{\link{RetrieveBatchPredictions}()},
\code{\link{RetrieveProjectsPredictionsMetadata}()},
\code{\link{RetrieveProjectsPredictions}()},
\code{\link{RetrieveScheduledJobs}()},
\code{\link{UpdateTransferableModel}()},
\code{\link{UploadPredictionDatasetFromCatalog}()},
\code{\link{UploadPredictionDatasetFromDataSource}()},
\code{\link{UploadPredictionDataset}()},
\code{\link{UploadTransferableModel}()}
}
\concept{Insights}
\concept{Predictions}
