% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R,
%   R/predictions_apiWrapper.R
\name{GetPredictionExplanationsPage}
\alias{GetPredictionExplanationsPage}
\title{GetPredictionExplanationsPage}
\usage{
GetPredictionExplanationsPage(
  project,
  predictionExplanationId,
  limit = NULL,
  offset = 0,
  excludeAdjustedPredictions = TRUE,
  adjustmentMethod = NULL,
  count = NULL,
  data = NULL,
  id = NULL,
  next_ = NULL,
  predictionExplanationsRecordLocation = NULL,
  previous = NULL,
  ...
)

GetPredictionExplanationsPage(
  project,
  predictionExplanationId,
  limit = NULL,
  offset = 0,
  excludeAdjustedPredictions = TRUE,
  adjustmentMethod = NULL,
  count = NULL,
  data = NULL,
  id = NULL,
  next_ = NULL,
  predictionExplanationsRecordLocation = NULL,
  previous = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve stored Prediction Explanations.

Retrieve stored Prediction Explanations.
}
\details{
Retrieve stored Prediction Explanations. Each PredictionExplanationsRow retrieved corresponds to a row of the prediction dataset, although some rows may not have had prediction explanations computed depending on the thresholds selected.

This method invokes \verb{GET /projects/\{projectId\}/predictionExplanations/\{predictionExplanationsId\}/} in the DataRobot Public API.

Retrieve stored Prediction Explanations. Each PredictionExplanationsRow retrieved corresponds to a row of the prediction dataset, although some rows may not have had prediction explanations computed depending on the thresholds selected.

This method invokes \verb{GET /projects/\{projectId\}/predictionExplanations/\{predictionExplanationsId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionExplanationsList in \link[datarobot.apicore:InsightsApi]{datarobot.apicore::InsightsApi}, which invokes the same underlying API endpoint.

The method ProjectsPredictionExplanationsList in \link[datarobot.apicore:PredictionsApi]{datarobot.apicore::PredictionsApi}, which invokes the same underlying API endpoint.
}
