% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrimeEligibility.R
\name{GetPrimeEligibility}
\alias{GetPrimeEligibility}
\title{Check if model can be approximated with DataRobot Prime}
\usage{
GetPrimeEligibility(project, modelId)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{modelId}{Unique alphanumeric identifier for the model of interest.}
}
\value{
list with two members: 
canMakePrime : logical value. TRUE if model can be approximated using DataRobot Prime, FALSE if model can not be approximated
message : character. Provides information why model may not be approximated with DataRobot Prime
}
\description{
This function returns list with two members: 
canMakePrime : logical value. TRUE if model can be approximated using DataRobot Prime, FALSE if model can not be approximated
message : character. Provides information why model may not be approximated with DataRobot Prime
}

