% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrimeModels.R
\name{GetPrimeModel}
\alias{GetPrimeModel}
\title{Retrieve information about specified DataRobot Prime model}
\usage{
GetPrimeModel(project, modelId)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{modelId}{Unique alphanumeric identifier for the model of interest.}
}
\value{
list containing informatione about specified DataRobot Prime model
}
\description{
This function requests the DataRobot Prime model information for the DataRobot
project specified by the project argument, and modelId
The function returns list containing informatione about specified DataRobot Prime model
}

