% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blueprints.R
\name{GetRecommendedBlueprints}
\alias{GetRecommendedBlueprints}
\title{Retrieve the list of available blueprints for a project}
\usage{
GetRecommendedBlueprints(project)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}
}
\description{
(Deprecated in 2.3, will be removed in 3.0. Use ListBlueprints instead.)
}
