% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetValidMetrics.R, R/projects_apiWrapper.R
\name{GetValidMetrics}
\alias{GetValidMetrics}
\title{Retrieve the valid fitting metrics for a specified project and target}
\usage{
GetValidMetrics(project, target, ...)

GetValidMetrics(project, target, ...)
}
\arguments{
\item{target}{character. String giving the name of the response variable
to be predicted by all project models.}
}
\value{
Character vector containing the names of the metric values
that are valid for a subsequent call to the SetTarget function.
}
\description{
For the response variable defined by the character string target
and the project defined by the parameter project, return the vector
of metric names that can be specified for fitting models in this project.
This function is intended for use after SetupProject has been run but
before SetTarget, allowing the user to specify valid non-default
values for the metric parameter.

List feature metrics
}
\details{
List the appropriate metrics if a feature were chosen as the target. The metrics listed will include both weighted and unweighted metrics - which are appropriate will depend on whether a weights column is used.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
GetValidMetrics(projectId, "targetFeature")
}
}
\seealso{
The method FeaturesMetricsList in [datarobot.apicore::ProjectsApi], which this function wraps.

Other Projects: 
\code{\link{BatchFeaturesTypeTransform}()},
\code{\link{CloneProject}()},
\code{\link{CreateCalendarsFromCountryCode}()},
\code{\link{CreateCalendarsFromDataset}()},
\code{\link{CreateCalendar}()},
\code{\link{CreateFeaturelist}()},
\code{\link{CreateHdfsProjects}()},
\code{\link{CreateModelingFeaturelist}()},
\code{\link{CreateProjectCleanupJobs}()},
\code{\link{CreateProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{CreateProjectsFeatureDiscoveryRecipeSqlExports}()},
\code{\link{CreateProjectsSecondaryDatasetsConfigurations}()},
\code{\link{CreateProjectsSegmentationTasks}()},
\code{\link{CreateRelationshipsConfigurations}()},
\code{\link{DeleteCalendar}()},
\code{\link{DeleteProjectCleanupJobs}()},
\code{\link{DeleteProjectsJobs}()},
\code{\link{DeleteProjectsSecondaryDatasetsConfigurations}()},
\code{\link{DeleteProject}()},
\code{\link{DeleteRelationshipsConfigurations}()},
\code{\link{GetCalendar}()},
\code{\link{GetFeatureHistogram}()},
\code{\link{GetFeatureInfo}()},
\code{\link{GetFeaturelist}()},
\code{\link{GetJob}()},
\code{\link{GetModelingFeaturelist}()},
\code{\link{GetProjectStatus}()},
\code{\link{GetProject}()},
\code{\link{ListCalendarCountryCodes}()},
\code{\link{ListCalendarsAccessControl}()},
\code{\link{ListCalendars}()},
\code{\link{ListDeletedProjectsCount}()},
\code{\link{ListDeletedProjects}()},
\code{\link{ListFeatureInfo}()},
\code{\link{ListFeaturelists}()},
\code{\link{ListJobs}()},
\code{\link{ListModelingFeaturelists}()},
\code{\link{ListProjectCleanupJobsDownload}()},
\code{\link{ListProjectCleanupJobsSummary}()},
\code{\link{ListProjectCleanupJobs}()},
\code{\link{ListProjectsAccessControl}()},
\code{\link{ListProjectsCalendarEvents}()},
\code{\link{ListProjectsFeatureDiscoveryDatasetDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogsDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogs}()},
\code{\link{ListProjectsFeatureDiscoveryRecipeSQLsDownload}()},
\code{\link{ListProjectsModelingFeatures}()},
\code{\link{ListProjectsMultiseriesIdsCrossSeriesProperties}()},
\code{\link{ListProjectsMultiseriesNames}()},
\code{\link{ListProjectsRelationshipsConfiguration}()},
\code{\link{ListProjectsSecondaryDatasetsConfigurations}()},
\code{\link{ListProjectsSegmentationTasksMappings}()},
\code{\link{ListProjectsSegmentationTasks}()},
\code{\link{ListProjects}()},
\code{\link{PatchDeletedProjects}()},
\code{\link{PatchManyCalendarsAccessControl}()},
\code{\link{PatchManyProjectsAccessControl}()},
\code{\link{PatchProjectsSegments}()},
\code{\link{PauseQueue}()},
\code{\link{PutRelationshipsConfigurations}()},
\code{\link{RequestCrossSeriesDetection}()},
\code{\link{RequestMultiSeriesDetection}()},
\code{\link{RestoreDiscardedFeatures}()},
\code{\link{RetrieveDiscardedFeaturesInformation}()},
\code{\link{RetrieveProjectCleanupJobs}()},
\code{\link{RetrieveProjectsBatchTypeTransformFeaturesResult}()},
\code{\link{RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{RetrieveProjectsFeatureLineages}()},
\code{\link{RetrieveProjectsModelingFeatures}()},
\code{\link{RetrieveProjectsSecondaryDatasetsConfigurations}()},
\code{\link{RetrieveProjectsSegmentationTaskJobResults}()},
\code{\link{RetrieveProjectsSegmentationTasks}()},
\code{\link{RetrieveRelationshipsConfigurations}()},
\code{\link{SetTarget}()},
\code{\link{SetupProject}()},
\code{\link{StartNewAutoPilot}()},
\code{\link{UpdateCalendar}()},
\code{\link{UpdateProject}()}
}
\concept{Projects}
