% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{ListApplicationsAccessControl}
\alias{ListApplicationsAccessControl}
\title{ListApplicationsAccessControl}
\usage{
ListApplicationsAccessControl(
  applicationId,
  userId,
  username,
  offset = 0,
  limit = 0,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
A list of users with access to this application
}
\details{
A list of users who have access to this application and their roles

This method invokes \verb{GET /applications/\{applicationId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsAccessControlList in \link[datarobot.apicore:ApplicationsApi]{datarobot.apicore::ApplicationsApi}, which invokes the same underlying API endpoint.
}
