% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListApprovalPolicyMatch}
\alias{ListApprovalPolicyMatch}
\title{ListApprovalPolicyMatch}
\usage{
ListApprovalPolicyMatch(fieldName, entityType, action, fieldValue, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get policy ID matching the query
}
\details{
Get policy ID matching the query

This method invokes \verb{GET /approvalPolicyMatch/} in the DataRobot Public API.
}
\seealso{
The method ApprovalPolicyMatchList in \link[datarobot.apicore:GovernanceApi]{datarobot.apicore::GovernanceApi}, which invokes the same underlying API endpoint.
}
