% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R, R/Blueprints.R
\name{ListBlueprints}
\alias{ListBlueprints}
\title{ListBlueprints}
\usage{
ListBlueprints(project, ...)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An S3 object of class 'listOfBlueprints', a list
with one element for each recommended blueprint in the
associated project. For more information see GetBlueprint()
}
\description{
List blueprints
}
\details{
List appropriate blueprints for the project.

This method invokes \verb{GET /projects/\{projectId\}/blueprints/} in the DataRobot Public API.

Retrieve the list of available blueprints for a project

This function returns the list of available blueprints
for a specified modeling project, as an S3 object of class
listOfBlueprints; see Value.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ListBlueprints(projectId)
}
}
\seealso{
The method ProjectsBlueprintsList in \link[datarobot.apicore:BlueprintsApi]{datarobot.apicore::BlueprintsApi}, which invokes the same underlying API endpoint.
}
