% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendars.R
\name{ListCalendars}
\alias{ListCalendars}
\title{List all available calendars.}
\usage{
ListCalendars()
}
\value{
A list of S3 objects of class "dataRobotCalendar"
}
\description{
List all available calendars.
}
\examples{
\dontrun{
   ListCalendars()
}
}
