% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListChangeRequestsReviews}
\alias{ListChangeRequestsReviews}
\title{ListChangeRequestsReviews}
\usage{
ListChangeRequestsReviews(
  changeRequestId,
  changeVersionId,
  status,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List Change Request reviews.
}
\details{
List Change Request reviews.

This method invokes \verb{GET /changeRequests/\{changeRequestId\}/reviews/} in the DataRobot Public API.
}
\seealso{
The method ChangeRequestsReviewsList in \link[datarobot.apicore:GovernanceApi]{datarobot.apicore::GovernanceApi}, which invokes the same underlying API endpoint.
}
