% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{ListComments}
\alias{ListComments}
\title{ListComments}
\usage{
ListComments(entityType, orderBy, entityId, offset = 0, limit = 100, ...)
}
\description{
List comments
}
\details{
List comments
}
\seealso{
The method CommentsList in [datarobot.apicore::CommentsApi], which this function wraps.

Other Comments: 
\code{\link{CreateComments}()},
\code{\link{DeleteComments}()},
\code{\link{PatchComments}()},
\code{\link{RetrieveComments}()}
}
\concept{Comments}
