% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{ListComplianceDocTemplatesSharedRoles}
\alias{ListComplianceDocTemplatesSharedRoles}
\title{ListComplianceDocTemplatesSharedRoles}
\usage{
ListComplianceDocTemplatesSharedRoles(
  shareRecipientType,
  name,
  id,
  templateId,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get template's access control list
}
\details{
Get a list of users, groups and organizations who have access to this template and their roles on the template.

This method invokes \verb{GET /complianceDocTemplates/\{templateId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method ComplianceDocTemplatesSharedRolesList in \link[datarobot.apicore:DocumentationApi]{datarobot.apicore::DocumentationApi}, which invokes the same underlying API endpoint.
}
