% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{ListCustomTasksAccessControl}
\alias{ListCustomTasksAccessControl}
\title{ListCustomTasksAccessControl}
\usage{
ListCustomTasksAccessControl(customTaskId, offset = 0, limit = 1000, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of users who have access to this custom task and their roles on it.
}
\details{
Get a list of users who have access to this custom task and their roles on it.

This method invokes \verb{GET /customTasks/\{customTaskId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksAccessControlList in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
