% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{ListCustomTasksDownload}
\alias{ListCustomTasksDownload}
\title{ListCustomTasksDownload}
\usage{
ListCustomTasksDownload(customTaskId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download the latest custom task version content.
}
\details{
Download the latest item bundle from a custom task as a zip compressed archive.

This method invokes \verb{GET /customTasks/\{customTaskId\}/download/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksDownloadList in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
