% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{ListCustomTasksVersionsDependencyBuildLog}
\alias{ListCustomTasksVersionsDependencyBuildLog}
\title{ListCustomTasksVersionsDependencyBuildLog}
\usage{
ListCustomTasksVersionsDependencyBuildLog(
  customTaskVersionId,
  customTaskId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the custom task version's dependency build log.
}
\details{
Retrieve the custom task version's dependency build log.

This method invokes \verb{GET /customTasks/\{customTaskId\}/versions/\{customTaskVersionId\}/dependencyBuildLog/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsDependencyBuildLogList in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
