% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSources.R
\name{ListDataSources}
\alias{ListDataSources}
\title{Returns a dataframe with information on available data sources.}
\usage{
ListDataSources()
}
\value{
data.frame containing information on possible data sources.
}
\description{
Returns a dataframe with information on available data sources.
}
\examples{
\dontrun{
 ListDataSources()
}
}
