% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStores.R
\name{ListDataStores}
\alias{ListDataStores}
\title{Returns a dataframe with information on available data stores.}
\usage{
ListDataStores()
}
\value{
data.frame containing information on possible data stores.
}
\description{
Returns a dataframe with information on available data stores.
}
\examples{
\dontrun{
 ListDataStores()
}
}
