% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsAccuracyOverTime}
\alias{ListDeploymentsAccuracyOverTime}
\title{ListDeploymentsAccuracyOverTime}
\usage{
ListDeploymentsAccuracyOverTime(
  segmentAttribute,
  targetClasses,
  metric,
  modelId,
  deploymentId,
  start,
  end,
  bucketSize,
  segmentValue = "",
  ...
)
}
\description{
Retrieve accuracy over time data for one single metric.
}
\details{
Retrieve accuracy over time data for one single metric.          The following metrics can be retrieved.              For classification deployments:              - Accuracy             - AUC             - Balanced Accuracy             - FVE Binomial             - Gini Norm             - Kolmogorov-Smirnov             - LogLoss (default)             - Rate@Top5%             - Rate@Top10%              For regression deployments:              - Gamma Deviance (default)             - FVE Gamma             - FVE Poisson             - FVE Tweedie             - MAD             - MAE             - MAPE             - Poisson Deviance             - R Squared             - RMSE             - RMSLE             - Tweedie Deviance
}
\seealso{
The method DeploymentsAccuracyOverTimeList in [datarobot.apicore::DeploymentsApi], which this function wraps.

Other Deployments: 
\code{\link{CreateDeploymentsActualsFromDataset}()},
\code{\link{CreateDeploymentsChallengerPredictions}()},
\code{\link{CreateDeploymentsChallengers}()},
\code{\link{CreateDeploymentsFromProjectRecommendedModel}()},
\code{\link{CreateDeploymentsOnDemandReports}()},
\code{\link{CreateDeploymentsPredictionInputsFromDataset}()},
\code{\link{CreateDeploymentsPredictions}()},
\code{\link{CreateDeploymentsScoringCodeBuilds}()},
\code{\link{CreateDeployment}()},
\code{\link{DeleteDeploymentsChallengers}()},
\code{\link{DeleteDeployment}()},
\code{\link{GetDeploymentAccuracy}()},
\code{\link{GetDeploymentServiceStats}()},
\code{\link{GetDeploymentSettings}()},
\code{\link{GetDeployment}()},
\code{\link{ListDeletedDeployments}()},
\code{\link{ListDeploymentsChallengers}()},
\code{\link{ListDeploymentsFeatureDrift}()},
\code{\link{ListDeploymentsFeatures}()},
\code{\link{ListDeploymentsHumilityStatsOverTime}()},
\code{\link{ListDeploymentsHumilityStats}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfigurationHistory}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{ListDeploymentsPredictionResults}()},
\code{\link{ListDeploymentsScoringCode}()},
\code{\link{ListDeploymentsServiceStatsOverTime}()},
\code{\link{ListDeploymentsSharedRoles}()},
\code{\link{ListDeploymentsTargetDrift}()},
\code{\link{ListDeployments}()},
\code{\link{ListPredictionServers}()},
\code{\link{PatchDeploymentsChallengers}()},
\code{\link{PatchDeployments}()},
\code{\link{PatchManyDeletedDeployments}()},
\code{\link{PatchManyDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{PatchManyDeploymentsSharedRoles}()},
\code{\link{ReplaceDeployedModel}()},
\code{\link{RetrieveDeploymentsChallengers}()},
\code{\link{SubmitActuals}()},
\code{\link{ValidateReplaceDeployedModel}()}
}
\concept{Deployments}
