% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsScoringCode}
\alias{ListDeploymentsScoringCode}
\title{ListDeploymentsScoringCode}
\usage{
ListDeploymentsScoringCode(
  deploymentId,
  includeAgent = "false",
  sourceCode = "false",
  includePredictionExplanations = "false",
  includePredictionIntervals = "false",
  includePe = "false",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Scoring Code
}
\details{
Retrieve Scoring Code of the current deployed model for making predictions outside of a DataRobot prediction server. You need the \"Scoring Code\" feature enabled to use this route.  By default, returns a compiled executable JAR that can be executed locally to calculate model predictions, or it can be used as a library for a Java application. Execute it with the '--help\\` parameters to learn how to use it as a command-line utility. See model API documentation at \url{https://javadoc.io/doc/com.datarobot/datarobot-prediction/latest/index.html} to be able to use it inside an existing Java application.  With sourceCode query parameter set to 'true', returns a source code archive that can be used to review internal calculations of the model. This JAR is NOT executable.  See \"\url{https://app.datarobot.com/docs/predictions/scoring-code/scorecode.html}\" documentation in DataRobot application for more information.  \emph{Note} Cannot retrieve source code if agent is included.

This method invokes \verb{GET /deployments/\{deploymentId\}/scoringCode/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsScoringCodeList in \link[datarobot.apicore:DeploymentsApi]{datarobot.apicore::DeploymentsApi}, which invokes the same underlying API endpoint.
}
