% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers.R
\name{ListDrivers}
\alias{ListDrivers}
\title{Returns a dataframe with information on available drivers.}
\usage{
ListDrivers()
}
\value{
data.frame containing information on possible drivers.
}
\description{
Returns a dataframe with information on available drivers.
}
\examples{
\dontrun{
 ListDrivers()
}
}
