% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/Drivers.R
\name{ListDrivers}
\alias{ListDrivers}
\title{ListDrivers}
\usage{
ListDrivers(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
data.frame containing information on possible drivers.
}
\description{
List drivers
}
\details{
Fetch all drivers a user has access to.

This method invokes \verb{GET /externalDataDrivers/} in the DataRobot Public API.

Returns a dataframe with information on available drivers.
}
\examples{
\dontrun{
ListDrivers()
}
}
\seealso{
The method ExternalDataDriversList in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
