% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListEventLogsPredictionUsage}
\alias{ListEventLogsPredictionUsage}
\title{ListEventLogsPredictionUsage}
\usage{
ListEventLogsPredictionUsage(
  maxTimestamp,
  userId,
  projectId,
  minTimestamp,
  offset = 0,
  includeIdentifyingFields = "true",
  order = "desc",
  ...
)
}
\description{
Retrieve prediction usage data.
}
\details{
Retrieve prediction usage data. ``CAN_ACCESS_USER_ACTIVITY`` permission is required.
}
\seealso{
The method EventLogsPredictionUsageList in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{CreateUsageDataExports}()},
\code{\link{DeleteManyUsersRateLimitUsage}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsEvents}()},
\code{\link{ListEventLogs}()},
\code{\link{ListUsageDataExportsSupportedEvents}()},
\code{\link{ListUsersRateLimitUsage}()},
\code{\link{RetrieveEventLogs}()},
\code{\link{RetrieveUsageDataExports}()}
}
\concept{Analytics}
