% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{ListExternalDataSourcesPermissions}
\alias{ListExternalDataSourcesPermissions}
\title{ListExternalDataSourcesPermissions}
\usage{
ListExternalDataSourcesPermissions(dataSourceId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Describe data source permissions.
}
\details{
Describe what permissions current user has for given data source.

This method invokes \verb{GET /externalDataSources/\{dataSourceId\}/permissions/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataSourcesPermissionsList in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
