% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{ListExternalDataStoresSharedRoles}
\alias{ListExternalDataStoresSharedRoles}
\title{ListExternalDataStoresSharedRoles}
\usage{
ListExternalDataStoresSharedRoles(
  shareRecipientType,
  dataStoreId,
  name,
  id,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get data store's access control list
}
\details{
Get a list of users who have access to this data store and their roles on the data store.

This method invokes \verb{GET /externalDataStores/\{dataStoreId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresSharedRolesList in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
