% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{ListFeatureInfo}
\alias{ListFeatureInfo}
\title{Details about all features for this project}
\usage{
ListFeatureInfo(project)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}
}
\value{
A list of lists with one element for each feature The named list for
each feature contains:
\describe{
 \item{id}{feature id}
 \item{name}{feature name}
 \item{feature_type}{feature type: 'Numeric', 'Categorical', etc.}
 \item{importance}{numeric measure of the strength of relationship between the feature and
 target (independent of any model or other features).}
 \item{low_information}{whether feature has too few values to be informative}
 \item{unique_count}{number of unique values}
 \item{na_count}{number of missing values}}
}
\description{
Details about all features for this project
}

