% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlops_apiWrapper.R
\name{ListMlopsPortablePredictionServerImageMetadata}
\alias{ListMlopsPortablePredictionServerImageMetadata}
\title{ListMlopsPortablePredictionServerImageMetadata}
\usage{
ListMlopsPortablePredictionServerImageMetadata(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Fetches currently active PPS Docker image metadata
}
\details{
Fetches currently active PPS Docker image metadata

This method invokes \verb{GET /mlops/portablePredictionServerImage/metadata/} in the DataRobot Public API.
}
\seealso{
The method MlopsPortablePredictionServerImageMetadataList in \link[datarobot.apicore:MlopsApi]{datarobot.apicore::MlopsApi}, which invokes the same underlying API endpoint.
}
