% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R, R/models_apiWrapper.R
\name{ListModels}
\alias{ListModels}
\title{Retrieve all available model information for a DataRobot project}
\usage{
ListModels(
  project,
  orderBy = NULL,
  filter = NULL,
  bulkOperationId = NULL,
  name = NULL,
  samplePct = NULL,
  withMetric = NULL,
  isStarred = NULL,
  ...
)

ListModels(
  project,
  orderBy = NULL,
  filter = NULL,
  bulkOperationId = NULL,
  name = NULL,
  samplePct = NULL,
  withMetric = NULL,
  isStarred = NULL,
  ...
)
}
\arguments{
\item{orderBy}{character. Optional. A vector of keys to order the list by. You can
order by \code{metric} or \code{samplePct}. If the sort attribute is preceded by a
hyphen, models will be sorted in descending order, otherwise in ascending order.
Multiple sort attributes can be included as a comma-delimited string or in a vector.}

\item{filter}{list. Optional. A named list of parameters to search a model
by, such as \code{name}, \code{samplePct}, or \code{isStarred}.}
}
\value{
An S3 object of class listOfModels, which may be characterized
  using R's generic summary function or converted to a dataframe with
  the as.data.frame method.
}
\description{
This function requests the model information for the DataRobot
project specified by the project argument, described under Arguments.
This parameter may be obtained in several ways, including: (1), from
the projectId element of the list returned by ListProjects; (2), as
the object returned by the GetProject function; or (3), as the list
returned by the SetupProject function. The function returns an S3
object of class 'listOfModels'.

List project models
}
\details{
Lists all the models from a project.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
ListModels(projectId)
ListModels(projectId, orderBy = c("samplePct", "-metric"))
ListModels(projectId, filter = list("sample_pct__gt" = 64, "name" = "Ridge"))
ListModels(projectId, filter = list("isStarred" = TRUE))
}
}
\seealso{
The method ProjectsModelsList in [datarobot.apicore::ModelsApi], which this function wraps.

Other Models: 
\code{\link{AddEureqaSolution}()},
\code{\link{CreateCustomInferenceImagesFeatureImpact}()},
\code{\link{CreateCustomModelTests}()},
\code{\link{CreateCustomModelsFromCustomModel}()},
\code{\link{CreateCustomModelsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersionsConversions}()},
\code{\link{CreateCustomModelsVersionsDependencyBuild}()},
\code{\link{CreateCustomModelsVersionsFeatureImpact}()},
\code{\link{CreateCustomModelsVersionsFromRepository}()},
\code{\link{CreateCustomModelsVersionsPredictionExplanationsInitialization}()},
\code{\link{CreateCustomModelsVersions}()},
\code{\link{CreateCustomModels}()},
\code{\link{CreateCustomTrainingBlueprints}()},
\code{\link{CreateFromLatestCustomModelsVersion}()},
\code{\link{CreateModelExports}()},
\code{\link{CreateModelPackagesArchive}()},
\code{\link{CreateModelPackagesFromLearningModel}()},
\code{\link{CreatePrimeCode}()},
\code{\link{CreateProjectsBlenderBuilders}()},
\code{\link{CreateProjectsDatetimeModelsFromModel}()},
\code{\link{CreateProjectsDeploymentReadyModels}()},
\code{\link{CreateProjectsModelsFromModel}()},
\code{\link{CreateProjectsModelsPredictionIntervals}()},
\code{\link{CreateRatingTable}()},
\code{\link{DeleteCustomModelTests}()},
\code{\link{DeleteCustomModelsVersionsConversions}()},
\code{\link{DeleteCustomModels}()},
\code{\link{DeleteManyCustomModelsVersionsDependencyBuild}()},
\code{\link{DeleteModelJob}()},
\code{\link{DownloadPrimeCode}()},
\code{\link{DownloadRatingTable}()},
\code{\link{DownloadScoringCode}()},
\code{\link{DownloadTransferableModel}()},
\code{\link{GetBlenderModel}()},
\code{\link{GetDatetimeModel}()},
\code{\link{GetFrozenModel}()},
\code{\link{GetMissingValuesReport}()},
\code{\link{GetModelJob}()},
\code{\link{GetModelParameters}()},
\code{\link{GetModel}()},
\code{\link{GetPrimeEligibility}()},
\code{\link{GetPrimeFile}()},
\code{\link{GetPrimeModel}()},
\code{\link{GetRatingTableModel}()},
\code{\link{GetRatingTable}()},
\code{\link{GetRulesets}()},
\code{\link{IsBlenderEligible}()},
\code{\link{ListCustomInferenceImagesFeatureImpact}()},
\code{\link{ListCustomModelDeployments}()},
\code{\link{ListCustomModelLimits}()},
\code{\link{ListCustomModelTestsLog}()},
\code{\link{ListCustomModelTestsTail}()},
\code{\link{ListCustomModelTests}()},
\code{\link{ListCustomModelsAccessControl}()},
\code{\link{ListCustomModelsDownload}()},
\code{\link{ListCustomModelsVersionsConversions}()},
\code{\link{ListCustomModelsVersionsDependencyBuildLog}()},
\code{\link{ListCustomModelsVersionsDependencyBuild}()},
\code{\link{ListCustomModelsVersionsDownload}()},
\code{\link{ListCustomModelsVersionsFeatureImpact}()},
\code{\link{ListCustomModelsVersions}()},
\code{\link{ListCustomModels}()},
\code{\link{ListCustomTrainingBlueprints}()},
\code{\link{ListModelJobs}()},
\code{\link{ListModelPackagesFeatures}()},
\code{\link{ListModelPackagesSharedRoles}()},
\code{\link{ListModelPackages}()},
\code{\link{ListModelRecommendations}()},
\code{\link{ListPrimeFiles}()},
\code{\link{ListPrimeModels}()},
\code{\link{ListProjectsBlenderBuilders}()},
\code{\link{ListProjectsBlenderModels}()},
\code{\link{ListProjectsCombinedModelsSegmentsDownload}()},
\code{\link{ListProjectsCombinedModelsSegments}()},
\code{\link{ListProjectsCombinedModels}()},
\code{\link{ListProjectsDatetimeModels}()},
\code{\link{ListProjectsFrozenModels}()},
\code{\link{ListProjectsModelsClusterNames}()},
\code{\link{ListProjectsModelsNumIterationsTrained}()},
\code{\link{ListProjectsModelsPredictionIntervals}()},
\code{\link{ListProjectsRecommendedModelsRecommendedModel}()},
\code{\link{ListRatingTableModels}()},
\code{\link{ListRatingTables}()},
\code{\link{PatchCustomModelsVersions}()},
\code{\link{PatchCustomModels}()},
\code{\link{PatchManyCustomModelsAccessControl}()},
\code{\link{PatchManyCustomModelsTrainingData}()},
\code{\link{PatchManyCustomModelsVersionsFromRepository}()},
\code{\link{PatchManyProjectsModelsClusterNames}()},
\code{\link{PutManyProjectsSegmentChampion}()},
\code{\link{RenameRatingTable}()},
\code{\link{RequestApproximation}()},
\code{\link{RequestBlender}()},
\code{\link{RequestNewDatetimeModel}()},
\code{\link{RequestNewModel}()},
\code{\link{RequestNewRatingTableModel}()},
\code{\link{RequestPrimeModel}()},
\code{\link{RetrieveCustomModelTests}()},
\code{\link{RetrieveCustomModelsVersionsConversions}()},
\code{\link{RetrieveCustomModelsVersions}()},
\code{\link{RetrieveCustomModels}()},
\code{\link{RetrieveModelPackages}()},
\code{\link{RetrieveProjectsBlenderBuilders}()},
\code{\link{RetrieveProjectsCombinedModels}()},
\code{\link{RetrieveProjectsEureqaDistributionPlot}()},
\code{\link{RetrieveProjectsEureqaModelDetail}()}
}
\concept{Models}
