% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deployment.R
\name{ListPredictionServers}
\alias{ListPredictionServers}
\title{List all available prediction servers.}
\usage{
ListPredictionServers()
}
\value{
A list of DataRobotPredictionServer objects containing:
\itemize{
 \item id character. The ID of the prediction server.
 \item url character. The URL of the prediction server.
 \item dataRobotKey character. The key used to access the prediction server.
}
}
\description{
List all available prediction servers.
}
\examples{
\dontrun{
  ListPredictionServers()
}
}
