% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsAnomalyInsightsTable}
\alias{ListProjectsModelsAnomalyInsightsTable}
\title{ListProjectsModelsAnomalyInsightsTable}
\usage{
ListProjectsModelsAnomalyInsightsTable(
  modelId,
  anomalyScoreRounding,
  projectId,
  ...
)
}
\description{
Retrieve a table of the raw data displayed with the anomaly score from the specific model.
}
\details{
Retrieve a table of the raw data displayed with the anomaly score from the specific model. The number of rows displayed is limited to 100 rows by the ANOMALY_INSIGHT_SAMPLE_ROW_COUNT configuration setting. Additionally, feature column count and the size of data in text fields is also limited. Only models built from anomaly detection blueprints have those insights.
}
\seealso{
The method ProjectsModelsAnomalyInsightsTableList in [datarobot.apicore::InsightsApi], which this function wraps.

Other Insights: 
\code{\link{CreateMultilabelInsightsPairwiseManualSelections}()},
\code{\link{CreateProjectsDatetimeModelsFeatureEffects}()},
\code{\link{CreateProjectsDatetimeModelsFeatureFit}()},
\code{\link{CreateProjectsDatetimeModelsMulticlassFeatureEffects}()},
\code{\link{CreateProjectsExternalScores}()},
\code{\link{CreateProjectsFeatureAssociationMatrix}()},
\code{\link{CreateProjectsGeometryFeaturePlots}()},
\code{\link{CreateProjectsModelsClusterInsights}()},
\code{\link{CreateProjectsModelsCrossClassAccuracyScores}()},
\code{\link{CreateProjectsModelsDataDisparityInsights}()},
\code{\link{CreateProjectsModelsFairnessInsights}()},
\code{\link{CreateProjectsModelsFeatureEffects}()},
\code{\link{CreateProjectsModelsFeatureFit}()},
\code{\link{CreateProjectsModelsImageActivationMaps}()},
\code{\link{CreateProjectsModelsImageEmbeddings}()},
\code{\link{CreateProjectsModelsMulticlassFeatureEffects}()},
\code{\link{CreateProjectsModelsShapImpact}()},
\code{\link{CreateProjectsPayoffMatrices}()},
\code{\link{CreateProjectsShapMatrices}()},
\code{\link{DeleteAnomalyAssessmentRecord}()},
\code{\link{DeleteMultilabelInsightsPairwiseManualSelections}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteProjectsPayoffMatrices}()},
\code{\link{GetAnomalyAssessmentExplanations}()},
\code{\link{GetAnomalyAssessmentPredictionsPreview}()},
\code{\link{GetFeatureAssociationMatrixDetails}()},
\code{\link{GetFeatureAssociationMatrix}()},
\code{\link{GetPredictionExplanationsMetadata}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetWordCloud}()},
\code{\link{InitializeAnomalyAssessment}()},
\code{\link{ListAnomalyAssessmentRecords}()},
\code{\link{ListMultilabelInsightsHistogram}()},
\code{\link{ListMultilabelInsightsPairwiseManualSelections}()},
\code{\link{ListMultilabelInsightsPairwiseStatistics}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListProjectsBiasVsAccuracyInsights}()},
\code{\link{ListProjectsDatetimeModelsBacktestStabilityPlot}()},
\code{\link{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadata}()},
\code{\link{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsPreview}()},
\code{\link{ListProjectsDatetimeModelsFeatureEffectsMetadata}()},
\code{\link{ListProjectsDatetimeModelsFeatureEffects}()},
\code{\link{ListProjectsDatetimeModelsFeatureFitMetadata}()},
\code{\link{ListProjectsDatetimeModelsFeatureFit}()},
\code{\link{ListProjectsDatetimeModelsForecastDistanceStabilityPlot}()},
\code{\link{ListProjectsDatetimeModelsMulticlassFeatureEffects}()},
\code{\link{ListProjectsDatetimeModelsMultiseriesHistograms}()},
\code{\link{ListProjectsDatetimeModelsMultiseriesScoresFile}()},
\code{\link{ListProjectsExternalScores}()},
\code{\link{ListProjectsFeatureAssociationFeaturelists}()},
\code{\link{ListProjectsFeaturesFrequentValues}()},
\code{\link{ListProjectsImageActivationMaps}()},
\code{\link{ListProjectsImageEmbeddings}()},
\code{\link{ListProjectsModelsAnomalyInsightsFile}()},
\code{\link{ListProjectsModelsClusterInsightsDownload}()},
\code{\link{ListProjectsModelsClusterInsights}()},
\code{\link{ListProjectsModelsConfusionChartsClassDetails}()},
\code{\link{ListProjectsModelsConfusionChartsMetadata}()},
\code{\link{ListProjectsModelsConfusionCharts}()},
\code{\link{ListProjectsModelsCrossClassAccuracyScores}()},
\code{\link{ListProjectsModelsDataDisparityInsights}()},
\code{\link{ListProjectsModelsDatasetConfusionChartsClassDetails}()},
\code{\link{ListProjectsModelsDatasetConfusionChartsMetadata}()},
\code{\link{ListProjectsModelsDatasetConfusionCharts}()},
\code{\link{ListProjectsModelsDatasetLiftCharts}()},
\code{\link{ListProjectsModelsDatasetMulticlassLiftCharts}()},
\code{\link{ListProjectsModelsDatasetResidualsCharts}()},
\code{\link{ListProjectsModelsDatasetRocCurves}()},
\code{\link{ListProjectsModelsFairnessInsights}()},
\code{\link{ListProjectsModelsFeatureEffectsMetadata}()},
\code{\link{ListProjectsModelsFeatureEffects}()},
\code{\link{ListProjectsModelsFeatureFitMetadata}()},
\code{\link{ListProjectsModelsFeatureFit}()},
\code{\link{ListProjectsModelsFeatureListsClusterInsights}()},
\code{\link{ListProjectsModelsImageActivationMaps}()},
\code{\link{ListProjectsModelsImageEmbeddings}()},
\code{\link{ListProjectsModelsLabelwiseRocCurves}()},
\code{\link{ListProjectsModelsLiftChart}()},
\code{\link{ListProjectsModelsMulticlassFeatureEffects}()},
\code{\link{ListProjectsModelsMulticlassFeatureImpact}()},
\code{\link{ListProjectsModelsMulticlassLiftChart}()},
\code{\link{ListProjectsModelsResiduals}()},
\code{\link{ListProjectsModelsRocCurve}()},
\code{\link{ListProjectsModelsShapImpact}()},
\code{\link{ListProjectsMulticategoricalInvalidFormatFile}()},
\code{\link{ListProjectsMulticategoricalInvalidFormat}()},
\code{\link{ListProjectsPayoffMatrices}()},
\code{\link{ListProjectsShapMatrices}()},
\code{\link{PatchMultilabelInsightsPairwiseManualSelections}()},
\code{\link{PutProjectsPayoffMatrices}()},
\code{\link{RetrieveProjectsDatetimeModelsDatasetAccuracyOverTimePlots}()},
\code{\link{RetrieveProjectsGeometryFeaturePlots}()},
\code{\link{RetrieveProjectsModelsDatasetConfusionCharts}()},
\code{\link{RetrieveProjectsModelsLiftChart}()},
\code{\link{RetrieveProjectsModelsMulticlassLiftChart}()},
\code{\link{RetrieveProjectsModelsMultilabelLiftCharts}()},
\code{\link{RetrieveProjectsModelsResiduals}()},
\code{\link{RetrieveProjectsModelsRocCurve}()},
\code{\link{RetrieveProjectsShapMatrices}()}
}
\concept{Insights}
