% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R, R/insights_apiWrapper.R
\name{ListProjectsModelsImageActivationMaps}
\alias{ListProjectsModelsImageActivationMaps}
\title{ListProjectsModelsImageActivationMaps}
\usage{
ListProjectsModelsImageActivationMaps(modelId, featureName, projectId, ...)

ListProjectsModelsImageActivationMaps(modelId, featureName, projectId, ...)
}
\description{
Retrieve Image Activation Maps for a feature of a model.

Retrieve Image Activation Maps for a feature of a model.
}
\details{
Retrieve Image Activation Maps for a feature of a model. Image Activation maps are a technique to get the discriminative image regions used by a CNN to identify a specific class in the image. In other words, an image activation map lets us see which regions in the image were relevant to this class.  The higher the value in the activation map the greater the effect the region had on the prediction.

Retrieve Image Activation Maps for a feature of a model. Image Activation maps are a technique to get the discriminative image regions used by a CNN to identify a specific class in the image. In other words, an image activation map lets us see which regions in the image were relevant to this class.  The higher the value in the activation map the greater the effect the region had on the prediction.
}
\seealso{
The method ProjectsModelsImageActivationMapsList in [datarobot.apicore::ImagesApi], which this function wraps.

The method ProjectsModelsImageActivationMapsList in [datarobot.apicore::InsightsApi], which this function wraps.

Other Images: 
\code{\link{CreateImageAugmentationListsSamples}()},
\code{\link{CreateImageAugmentationLists}()},
\code{\link{CreateImageAugmentationSamples}()},
\code{\link{CreateProjectsModelsImageActivationMaps}()},
\code{\link{CreateProjectsModelsImageEmbeddings}()},
\code{\link{DeleteImageAugmentationLists}()},
\code{\link{ListDatasetsImagesDataQualityLogFile}()},
\code{\link{ListDatasetsImagesDataQualityLog}()},
\code{\link{ListImageAugmentationListsSamples}()},
\code{\link{ListImageAugmentationLists}()},
\code{\link{ListImageAugmentationSamples}()},
\code{\link{ListProjectsDuplicateImages}()},
\code{\link{ListProjectsImageActivationMaps}()},
\code{\link{ListProjectsImageBins}()},
\code{\link{ListProjectsImageEmbeddings}()},
\code{\link{ListProjectsImageSamples}()},
\code{\link{ListProjectsImagesDataQualityLogFile}()},
\code{\link{ListProjectsImagesDataQualityLog}()},
\code{\link{ListProjectsImagesFile}()},
\code{\link{ListProjectsImages}()},
\code{\link{ListProjectsModelsImageEmbeddings}()},
\code{\link{PatchImageAugmentationLists}()},
\code{\link{RetrieveImageAugmentationLists}()},
\code{\link{RetrieveImageAugmentationOptions}()},
\code{\link{RetrieveProjectsImages}()}

Other Insights: 
\code{\link{CreateMultilabelInsightsPairwiseManualSelections}()},
\code{\link{CreateProjectsDatetimeModelsFeatureEffects}()},
\code{\link{CreateProjectsDatetimeModelsFeatureFit}()},
\code{\link{CreateProjectsDatetimeModelsMulticlassFeatureEffects}()},
\code{\link{CreateProjectsExternalScores}()},
\code{\link{CreateProjectsFeatureAssociationMatrix}()},
\code{\link{CreateProjectsGeometryFeaturePlots}()},
\code{\link{CreateProjectsModelsClusterInsights}()},
\code{\link{CreateProjectsModelsCrossClassAccuracyScores}()},
\code{\link{CreateProjectsModelsDataDisparityInsights}()},
\code{\link{CreateProjectsModelsFairnessInsights}()},
\code{\link{CreateProjectsModelsFeatureEffects}()},
\code{\link{CreateProjectsModelsFeatureFit}()},
\code{\link{CreateProjectsModelsImageActivationMaps}()},
\code{\link{CreateProjectsModelsImageEmbeddings}()},
\code{\link{CreateProjectsModelsMulticlassFeatureEffects}()},
\code{\link{CreateProjectsModelsShapImpact}()},
\code{\link{CreateProjectsPayoffMatrices}()},
\code{\link{CreateProjectsShapMatrices}()},
\code{\link{DeleteAnomalyAssessmentRecord}()},
\code{\link{DeleteMultilabelInsightsPairwiseManualSelections}()},
\code{\link{DeletePredictionExplanations}()},
\code{\link{DeleteProjectsPayoffMatrices}()},
\code{\link{GetAnomalyAssessmentExplanations}()},
\code{\link{GetAnomalyAssessmentPredictionsPreview}()},
\code{\link{GetFeatureAssociationMatrixDetails}()},
\code{\link{GetFeatureAssociationMatrix}()},
\code{\link{GetPredictionExplanationsMetadata}()},
\code{\link{GetPredictionExplanationsPage}()},
\code{\link{GetWordCloud}()},
\code{\link{InitializeAnomalyAssessment}()},
\code{\link{ListAnomalyAssessmentRecords}()},
\code{\link{ListMultilabelInsightsHistogram}()},
\code{\link{ListMultilabelInsightsPairwiseManualSelections}()},
\code{\link{ListMultilabelInsightsPairwiseStatistics}()},
\code{\link{ListPredictionExplanationsMetadata}()},
\code{\link{ListProjectsBiasVsAccuracyInsights}()},
\code{\link{ListProjectsDatetimeModelsBacktestStabilityPlot}()},
\code{\link{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsMetadata}()},
\code{\link{ListProjectsDatetimeModelsDatasetAccuracyOverTimePlotsPreview}()},
\code{\link{ListProjectsDatetimeModelsFeatureEffectsMetadata}()},
\code{\link{ListProjectsDatetimeModelsFeatureEffects}()},
\code{\link{ListProjectsDatetimeModelsFeatureFitMetadata}()},
\code{\link{ListProjectsDatetimeModelsFeatureFit}()},
\code{\link{ListProjectsDatetimeModelsForecastDistanceStabilityPlot}()},
\code{\link{ListProjectsDatetimeModelsMulticlassFeatureEffects}()},
\code{\link{ListProjectsDatetimeModelsMultiseriesHistograms}()},
\code{\link{ListProjectsDatetimeModelsMultiseriesScoresFile}()},
\code{\link{ListProjectsExternalScores}()},
\code{\link{ListProjectsFeatureAssociationFeaturelists}()},
\code{\link{ListProjectsFeaturesFrequentValues}()},
\code{\link{ListProjectsImageActivationMaps}()},
\code{\link{ListProjectsImageEmbeddings}()},
\code{\link{ListProjectsModelsAnomalyInsightsFile}()},
\code{\link{ListProjectsModelsAnomalyInsightsTable}()},
\code{\link{ListProjectsModelsClusterInsightsDownload}()},
\code{\link{ListProjectsModelsClusterInsights}()},
\code{\link{ListProjectsModelsConfusionChartsClassDetails}()},
\code{\link{ListProjectsModelsConfusionChartsMetadata}()},
\code{\link{ListProjectsModelsConfusionCharts}()},
\code{\link{ListProjectsModelsCrossClassAccuracyScores}()},
\code{\link{ListProjectsModelsDataDisparityInsights}()},
\code{\link{ListProjectsModelsDatasetConfusionChartsClassDetails}()},
\code{\link{ListProjectsModelsDatasetConfusionChartsMetadata}()},
\code{\link{ListProjectsModelsDatasetConfusionCharts}()},
\code{\link{ListProjectsModelsDatasetLiftCharts}()},
\code{\link{ListProjectsModelsDatasetMulticlassLiftCharts}()},
\code{\link{ListProjectsModelsDatasetResidualsCharts}()},
\code{\link{ListProjectsModelsDatasetRocCurves}()},
\code{\link{ListProjectsModelsFairnessInsights}()},
\code{\link{ListProjectsModelsFeatureEffectsMetadata}()},
\code{\link{ListProjectsModelsFeatureEffects}()},
\code{\link{ListProjectsModelsFeatureFitMetadata}()},
\code{\link{ListProjectsModelsFeatureFit}()},
\code{\link{ListProjectsModelsFeatureListsClusterInsights}()},
\code{\link{ListProjectsModelsImageEmbeddings}()},
\code{\link{ListProjectsModelsLabelwiseRocCurves}()},
\code{\link{ListProjectsModelsLiftChart}()},
\code{\link{ListProjectsModelsMulticlassFeatureEffects}()},
\code{\link{ListProjectsModelsMulticlassFeatureImpact}()},
\code{\link{ListProjectsModelsMulticlassLiftChart}()},
\code{\link{ListProjectsModelsResiduals}()},
\code{\link{ListProjectsModelsRocCurve}()},
\code{\link{ListProjectsModelsShapImpact}()},
\code{\link{ListProjectsMulticategoricalInvalidFormatFile}()},
\code{\link{ListProjectsMulticategoricalInvalidFormat}()},
\code{\link{ListProjectsPayoffMatrices}()},
\code{\link{ListProjectsShapMatrices}()},
\code{\link{PatchMultilabelInsightsPairwiseManualSelections}()},
\code{\link{PutProjectsPayoffMatrices}()},
\code{\link{RetrieveProjectsDatetimeModelsDatasetAccuracyOverTimePlots}()},
\code{\link{RetrieveProjectsGeometryFeaturePlots}()},
\code{\link{RetrieveProjectsModelsDatasetConfusionCharts}()},
\code{\link{RetrieveProjectsModelsLiftChart}()},
\code{\link{RetrieveProjectsModelsMulticlassLiftChart}()},
\code{\link{RetrieveProjectsModelsMultilabelLiftCharts}()},
\code{\link{RetrieveProjectsModelsResiduals}()},
\code{\link{RetrieveProjectsModelsRocCurve}()},
\code{\link{RetrieveProjectsShapMatrices}()}
}
\concept{Images}
\concept{Insights}
