% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_partitioning_apiWrapper.R
\name{ListProjectsOptimizedDatetimePartitionings}
\alias{ListProjectsOptimizedDatetimePartitionings}
\title{ListProjectsOptimizedDatetimePartitionings}
\usage{
ListProjectsOptimizedDatetimePartitionings(
  projectId,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all created optimized datetime partitioning configurations
}
\details{
List all created optimized datetime partitioning configurations

This method invokes \verb{GET /projects/\{projectId\}/optimizedDatetimePartitionings/} in the DataRobot Public API.
}
\seealso{
The method ProjectsOptimizedDatetimePartitioningsList in \link[datarobot.apicore:DatetimePartitioningApi]{datarobot.apicore::DatetimePartitioningApi}, which invokes the same underlying API endpoint.
}
