% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatingTables.R
\name{ListRatingTables}
\alias{ListRatingTables}
\title{Retrieve information about all rating tables.}
\usage{
ListRatingTables(project)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}
}
\value{
data.frame containing information about each rating table in a
  project (one row per model with a rating table).
}
\description{
Retrieve information about all rating tables.
}
\examples{
\dontrun{
 projectId <- "5984b4d7100d2b31c1166529"
 ListRatingTables(projectId)
}
}
