% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReasonCodes.R
\name{ListReasonCodesMetadata}
\alias{ListReasonCodesMetadata}
\title{Retrieve metadata for reason codes in specified project}
\usage{
ListReasonCodesMetadata(project, modelId = NULL, limit = NULL,
  offset = NULL)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{modelId}{(optional) Character string If specified, only reason codes computed for this model will be returned}

\item{limit}{(optional) Integer At most this many results are returned, default: no limit}

\item{offset}{(optional) Integer This many results will be skipped, default: 0}
}
\value{
List of metadata for all reason codes in the project. 
Each element of list is metadata for one reason codes (for format see GetReasonCodesMetadata)
}
\description{
Retrieve metadata for reason codes in specified project
}
