% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_configuration_apiWrapper.R
\name{ListSsoConfigurations}
\alias{ListSsoConfigurations}
\title{ListSsoConfigurations}
\usage{
ListSsoConfigurations(orgId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List sso configurations.
}
\details{
List the sso configurations that correspond to provided conditions.

This method invokes \verb{GET /ssoConfigurations/} in the DataRobot Public API.
}
\seealso{
The method SsoConfigurationsList in \link[datarobot.apicore:SsoConfigurationApi]{datarobot.apicore::SsoConfigurationApi}, which invokes the same underlying API endpoint.
}
