% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_apiWrapper.R
\name{ListStatus}
\alias{ListStatus}
\title{ListStatus}
\usage{
ListStatus(offset = 0, limit = 0, ...)
}
\description{
List tasks
}
\details{
List the async tasks that are currently running for your account
}
\seealso{
The method StatusList in [datarobot.apicore::JobsApi], which this function wraps.

Other Jobs: 
\code{\link{DeleteStatus}()},
\code{\link{RetrieveStatus}()}
}
\concept{Jobs}
