% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainingPredictions.R
\name{ListTrainingPredictions}
\alias{ListTrainingPredictions}
\title{Retrieve information about all training prediction datasets in a project.}
\usage{
ListTrainingPredictions(project)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}
}
\value{
data.frame containing information about each training prediction.
}
\description{
Retrieve information about all training prediction datasets in a project.
}
\examples{
\dontrun{
 projectId <- "5984b4d7100d2b31c1166529"
 ListTrainingPredictions(projectId)
}
}
