% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCasesActivities}
\alias{ListUseCasesActivities}
\title{ListUseCasesActivities}
\usage{
ListUseCasesActivities(useCaseId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the activities of a use case.
}
\details{
Retrieve the activities of a use case.

This method invokes \verb{GET /useCases/\{useCaseId\}/activities/} in the DataRobot Public API.
}
\seealso{
The method UseCasesActivitiesList in \link[datarobot.apicore:UseCaseApi]{datarobot.apicore::UseCaseApi}, which invokes the same underlying API endpoint.
}
