% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCasesSharedRoles}
\alias{ListUseCasesSharedRoles}
\title{ListUseCasesSharedRoles}
\usage{
ListUseCasesSharedRoles(
  shareRecipientType,
  useCaseId,
  name,
  id,
  offset = 0,
  limit = 10,
  ...
)
}
\description{
Get a list of users, groups and organizations that have an access to this use case
}
\details{
Get a list of users, groups and organizations that have an access to this use case
}
\seealso{
The method UseCasesSharedRolesList in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{CreateUseCases}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplatesCalculation}()},
\code{\link{ListUseCaseValueTemplates}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesAttachments}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCases}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{PatchUseCases}()},
\code{\link{RetrieveUseCaseValueTemplates}()},
\code{\link{RetrieveUseCasesAttachments}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
