% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCaseValueTemplatesCalculation}
\alias{ListUseCaseValueTemplatesCalculation}
\title{ListUseCaseValueTemplatesCalculation}
\usage{
ListUseCaseValueTemplatesCalculation(
  incorrectDecisionsCount,
  templateType,
  incorrectDecisionCost,
  accuracyImprovement,
  targetValue,
  decisionsCount,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Calculate value of template with given template parameters.
}
\details{
Calculate value of template with given template parameters.

This method invokes \verb{GET /useCaseValueTemplates/\{templateType\}/calculation/} in the DataRobot Public API.
}
\seealso{
The method UseCaseValueTemplatesCalculationList in \link[datarobot.apicore:UseCaseApi]{datarobot.apicore::UseCaseApi}, which invokes the same underlying API endpoint.
}
