% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListUserCleanupSummariesDeleteParams}
\alias{ListUserCleanupSummariesDeleteParams}
\title{ListUserCleanupSummariesDeleteParams}
\usage{
ListUserCleanupSummariesDeleteParams(reportId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Users permanent delete report parameters.
}
\details{
Retrieve users permanent delete report parameters.

This method invokes \verb{GET /userCleanupSummaries/\{reportId\}/deleteParams/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupSummariesDeleteParamsList in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
