% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListUsersLimits}
\alias{ListUsersLimits}
\title{ListUsersLimits}
\usage{
ListUsersLimits(userId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get the rate limits and account limits for a user
}
\details{
User limits are those limits on an account that can be updated per-user. Rate limits that can't be modified per-user are not returned here.

This method invokes \verb{GET /users/\{userId\}/limits/} in the DataRobot Public API.
}
\seealso{
The method UsersLimitsList in \link[datarobot.apicore:UserManagementApi]{datarobot.apicore::UserManagementApi}, which invokes the same underlying API endpoint.
}
