% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListUsersRateLimitUsage}
\alias{ListUsersRateLimitUsage}
\title{ListUsersRateLimitUsage}
\usage{
ListUsersRateLimitUsage(userId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List resource usage for a user
}
\details{
List the rate limit resource usage for a user. The usage array returned will have one object corresponding to each rate limit applied to the user.

This method invokes \verb{GET /users/\{userId\}/rateLimitUsage/} in the DataRobot Public API.
}
\seealso{
The method UsersRateLimitUsageList in \link[datarobot.apicore:AnalyticsApi]{datarobot.apicore::AnalyticsApi}, which invokes the same underlying API endpoint.
}
