% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{PatchApplications}
\alias{PatchApplications}
\title{PatchApplications}
\usage{
PatchApplications(name, description, applicationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update an application's name and/or description
}
\details{
Update an application's name and/or description

This method invokes \verb{PATCH /applications/\{applicationId\}/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsPatch in \link[datarobot.apicore:ApplicationsApi]{datarobot.apicore::ApplicationsApi}, which invokes the same underlying API endpoint.
}
