% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{PatchChangeRequests}
\alias{PatchChangeRequests}
\title{PatchChangeRequests}
\usage{
PatchChangeRequests(changeRequestId, autoApply, change, comment, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update Change Request.
}
\details{
Update Change Request with the given ID.

This method invokes \verb{PATCH /changeRequests/\{changeRequestId\}/} in the DataRobot Public API.
}
\seealso{
The method ChangeRequestsPatch in \link[datarobot.apicore:GovernanceApi]{datarobot.apicore::GovernanceApi}, which invokes the same underlying API endpoint.
}
