% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{PatchComments}
\alias{PatchComments}
\title{PatchComments}
\usage{
PatchComments(mentions, commentId, content, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update a comment
}
\details{
Update a comment

This method invokes \verb{PATCH /comments/\{commentId\}/} in the DataRobot Public API.
}
\seealso{
The method CommentsPatch in \link[datarobot.apicore:CommentsApi]{datarobot.apicore::CommentsApi}, which invokes the same underlying API endpoint.
}
