% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{PatchCredentials}
\alias{PatchCredentials}
\title{PatchCredentials}
\usage{
PatchCredentials(
  gcpKey,
  clientId,
  oauthIssuerUrl,
  awsAccessKeyId,
  oauthIssuerType,
  tokenName,
  description,
  oauthScopes,
  token,
  password,
  name,
  credentialId,
  clientSecret,
  personalAccessToken,
  snowflakeAccountName,
  user,
  azureConnectionString,
  awsSessionToken,
  refreshToken,
  awsSecretAccessKey,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update specified credentials
}
\details{
Update specified credentials

This method invokes \verb{PATCH /credentials/\{credentialId\}/} in the DataRobot Public API.
}
\seealso{
The method CredentialsPatch in \link[datarobot.apicore:CredentialsApi]{datarobot.apicore::CredentialsApi}, which invokes the same underlying API endpoint.
}
