% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{PatchCustomTasksVersions}
\alias{PatchCustomTasksVersions}
\title{PatchCustomTasksVersions}
\usage{
PatchCustomTasksVersions(
  requiredMetadata,
  customTaskVersionId,
  customTaskId,
  description,
  requiredMetadataValues,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update custom task version.
}
\details{
Edit metadata of a specific task version.

This method invokes \verb{PATCH /customTasks/\{customTaskId\}/versions/\{customTaskVersionId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsPatch in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
