% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{PatchManyComplianceDocTemplatesSharedRoles}
\alias{PatchManyComplianceDocTemplatesSharedRoles}
\title{PatchManyComplianceDocTemplatesSharedRoles}
\usage{
PatchManyComplianceDocTemplatesSharedRoles(roles, templateId, operation, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update template's access controls
}
\details{
Set roles for users on this template.

This method invokes \verb{PATCH /complianceDocTemplates/\{templateId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method ComplianceDocTemplatesSharedRolesPatchMany in \link[datarobot.apicore:DocumentationApi]{datarobot.apicore::DocumentationApi}, which invokes the same underlying API endpoint.
}
