% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{PatchManyCredentialsAssociations}
\alias{PatchManyCredentialsAssociations}
\title{PatchManyCredentialsAssociations}
\usage{
PatchManyCredentialsAssociations(
  credentialsToRemove,
  credentialId,
  credentialsToAdd,
  ...
)
}
\description{
Add objects associated with credentials
}
\details{
Add objects associated with credentials
}
\seealso{
The method CredentialsAssociationsPatchMany in [datarobot.apicore::CredentialsApi], which this function wraps.

Other Credentials: 
\code{\link{CreateCredentials}()},
\code{\link{DeleteCredentials}()},
\code{\link{ListCredentialsAssociations}()},
\code{\link{ListCredentials}()},
\code{\link{PatchCredentials}()},
\code{\link{RetrieveCredentials}()}
}
\concept{Credentials}
