% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{PatchManyCustomTasksAccessControl}
\alias{PatchManyCustomTasksAccessControl}
\title{PatchManyCustomTasksAccessControl}
\usage{
PatchManyCustomTasksAccessControl(data, customTaskId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Grant access or update roles for users on this custom task and appropriate learning data.
}
\details{
Grant access or update roles for users on this custom task and appropriate learning data. Up to 100 user roles may be set in a single request.

This method invokes \verb{PATCH /customTasks/\{customTaskId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksAccessControlPatchMany in \link[datarobot.apicore:CustomTasksApi]{datarobot.apicore::CustomTasksApi}, which invokes the same underlying API endpoint.
}
