% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_configuration_apiWrapper.R
\name{PatchSsoConfigurations}
\alias{PatchSsoConfigurations}
\title{PatchSsoConfigurations}
\usage{
PatchSsoConfigurations(
  autoGenerateUsers,
  certificate,
  securityParameters,
  entityId,
  issuer,
  signOnUrl,
  organizationId,
  idpMetadataUrl,
  roleMapping,
  idpMetadataHttpsVerify,
  idpMetadata,
  attributeMapping,
  idpResponseMethod,
  signOutUrl,
  sessionLengthSeconds,
  enableSso,
  groupMapping,
  name,
  enforceSso,
  configurationId,
  spRequestMethod,
  configurationType,
  advancedConfiguration,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update an SSO configuration for a specific organization.
}
\details{
Update an SSO configuration for a specific organization.

This method invokes \verb{PATCH /ssoConfigurations/\{configurationId\}/} in the DataRobot Public API.
}
\seealso{
The method SsoConfigurationsPatch in \link[datarobot.apicore:SsoConfigurationApi]{datarobot.apicore::SsoConfigurationApi}, which invokes the same underlying API endpoint.
}
