% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{PatchUseCases}
\alias{PatchUseCases}
\title{PatchUseCases}
\usage{
PatchUseCases(
  owner,
  feasibility,
  predictionTargets,
  notes,
  stage,
  realizedValue,
  useCaseId,
  name,
  description,
  businessImpact,
  potentialValueTemplate,
  potentialValue,
  targetDates,
  ...
)
}
\description{
Update a use case.
}
\details{
Update a use case.
}
\seealso{
The method UseCasesPatch in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{CreateUseCases}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplatesCalculation}()},
\code{\link{ListUseCaseValueTemplates}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesAttachments}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCasesSharedRoles}()},
\code{\link{ListUseCases}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{RetrieveUseCaseValueTemplates}()},
\code{\link{RetrieveUseCasesAttachments}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
