% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PauseQueue.R
\name{PauseQueue}
\alias{PauseQueue}
\title{Pause the DataRobot modeling queue}
\usage{
PauseQueue(project)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}
}
\description{
This function pauses the DataRobot modeling queue for a specified project
}

