% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{PutApprovalPolicies}
\alias{PutApprovalPolicies}
\title{PutApprovalPolicies}
\usage{
PutApprovalPolicies(
  automaticAction,
  review,
  name,
  trigger,
  approvalPolicyId,
  active = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update an Approval Policy.
}
\details{
Update the policy with the given ID.

This method invokes \verb{PUT /approvalPolicies/\{approvalPolicyId\}/} in the DataRobot Public API.
}
\seealso{
The method ApprovalPoliciesPut in \link[datarobot.apicore:GovernanceApi]{datarobot.apicore::GovernanceApi}, which invokes the same underlying API endpoint.
}
