% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_apiWrapper.R
\name{PutManyClusterLicense}
\alias{PutManyClusterLicense}
\title{PutManyClusterLicense}
\usage{
PutManyClusterLicense(licenseKey, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create or replace license
}
\details{
Create or replace the cluster license.

This method invokes \verb{PUT /clusterLicense/} in the DataRobot Public API.
}
\seealso{
The method ClusterLicensePutMany in \link[datarobot.apicore:InfrastructureApi]{datarobot.apicore::InfrastructureApi}, which invokes the same underlying API endpoint.
}
