% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ReplaceDeployedModel}
\alias{ReplaceDeployedModel}
\title{ReplaceDeployedModel}
\usage{
ReplaceDeployedModel(
  deploymentId,
  newModelId,
  replacementReason,
  maxWait = 600,
  ...
)
}
\description{
Model Replacement.
}
\details{
Replace the model used to make predictions for the deployment. A validation process will be performed to make sure the new model is eligible as a replacement. If the validation fails, the model replacement will not occur.The Model Replacement Validation endpoint can be used to confirm the new model is eligible as a replacement.
}
\seealso{
The method DeploymentsModelPatchMany in [datarobot.apicore::DeploymentsApi], which this function wraps.

Other Deployments: 
\code{\link{CreateDeploymentsActualsFromDataset}()},
\code{\link{CreateDeploymentsChallengerPredictions}()},
\code{\link{CreateDeploymentsChallengers}()},
\code{\link{CreateDeploymentsFromProjectRecommendedModel}()},
\code{\link{CreateDeploymentsOnDemandReports}()},
\code{\link{CreateDeploymentsPredictionInputsFromDataset}()},
\code{\link{CreateDeploymentsPredictions}()},
\code{\link{CreateDeploymentsScoringCodeBuilds}()},
\code{\link{CreateDeployment}()},
\code{\link{DeleteDeploymentsChallengers}()},
\code{\link{DeleteDeployment}()},
\code{\link{GetDeploymentAccuracy}()},
\code{\link{GetDeploymentServiceStats}()},
\code{\link{GetDeploymentSettings}()},
\code{\link{GetDeployment}()},
\code{\link{ListDeletedDeployments}()},
\code{\link{ListDeploymentsAccuracyOverTime}()},
\code{\link{ListDeploymentsChallengers}()},
\code{\link{ListDeploymentsFeatureDrift}()},
\code{\link{ListDeploymentsFeatures}()},
\code{\link{ListDeploymentsHumilityStatsOverTime}()},
\code{\link{ListDeploymentsHumilityStats}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfigurationHistory}()},
\code{\link{ListDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{ListDeploymentsPredictionResults}()},
\code{\link{ListDeploymentsScoringCode}()},
\code{\link{ListDeploymentsServiceStatsOverTime}()},
\code{\link{ListDeploymentsSharedRoles}()},
\code{\link{ListDeploymentsTargetDrift}()},
\code{\link{ListDeployments}()},
\code{\link{ListPredictionServers}()},
\code{\link{PatchDeploymentsChallengers}()},
\code{\link{PatchDeployments}()},
\code{\link{PatchManyDeletedDeployments}()},
\code{\link{PatchManyDeploymentsModelSecondaryDatasetConfiguration}()},
\code{\link{PatchManyDeploymentsSharedRoles}()},
\code{\link{RetrieveDeploymentsChallengers}()},
\code{\link{SubmitActuals}()},
\code{\link{ValidateReplaceDeployedModel}()}
}
\concept{Deployments}
