% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSeries.R
\name{RequestMultiSeriesDetection}
\alias{RequestMultiSeriesDetection}
\title{Format a multiseries.}
\usage{
RequestMultiSeriesDetection(project, dateColumn)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}

\item{dateColumn}{character. The name of the column containing the date that defines the
time series.}
}
\description{
Call this function to request the project be formatted as a multiseries project, with the
\code{dateColumn} specifying the time series.
}
