% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestPrimeModel.R
\name{RequestPrimeModel}
\alias{RequestPrimeModel}
\title{Request training for a DataRobot Prime model using a specified ruleset}
\usage{
RequestPrimeModel(project, ruleset)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{ruleset}{A list specifying rulest parameters (see GetRulesets)}
}
\value{
job Id
}
\description{
Training a model using a ruleset is a necessary prerequisite for being able to download the code for a ruleset
}
