% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReasonCodes.R
\name{RequestReasonCodesInitialization}
\alias{RequestReasonCodesInitialization}
\title{Request reason codes initialization for specified model (deprecated)}
\usage{
RequestReasonCodesInitialization(model)
}
\arguments{
\item{model}{An S3 object of class dataRobotModel like that returned by
the function GetModel, or each element of the list returned by
the function ListModels.}
}
\value{
job Id
}
\description{
Use \link{RequestPredictionExplanationsInitialization} instead.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  modelId <- "5996f820af07fc605e81ead4"
  model <- GetModel(projectId, modelId)
  RequestReasonCodesInitialization(model)
}
}
