% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R,
%   R/RestoreDiscardedFeatures.R
\name{RestoreDiscardedFeatures}
\alias{RestoreDiscardedFeatures}
\title{RestoreDiscardedFeatures}
\usage{
RestoreDiscardedFeatures(projectId, featuresToRestore, maxWait = 60 * 10)

RestoreDiscardedFeatures(projectId, featuresToRestore, maxWait = 60 * 10)
}
\arguments{
\item{projectId}{Character string, the ID of the project.}

\item{featuresToRestore}{List of character strings giving discarded feature names to restore.}

\item{maxWait}{integer. The maximum time (in seconds) to wait for the restoration to complete.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
An object with restoration information:
\itemize{
\item restoredFeatures. List of character strings giving restored feature names.
\item warnings. List of character strings with the warnings about failed to restore features.
}
}
\description{
Restore discarded time series features.
}
\details{
Restore discarded time series features.

This method invokes \verb{POST /projects/\{projectId\}/modelingFeatures/fromDiscardedFeatures/} in the DataRobot Public API.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
restoredInfo <- RestoreDiscardedFeatures(projectId, featuresToRestore = list("Sales (7 day mean)"))
}
}
\seealso{
The method ModelingFeaturesFromDiscardedFeaturesCreate in \link[datarobot.apicore:ProjectsApi]{datarobot.apicore::ProjectsApi}, which invokes the same underlying API endpoint.

Other Restore Discarded Features functions: 
\code{\link{RetrieveDiscardedFeaturesInformation}()}
}
\concept{Restore Discarded Features functions}
