% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{RetrieveApprovalPolicies}
\alias{RetrieveApprovalPolicies}
\title{RetrieveApprovalPolicies}
\usage{
RetrieveApprovalPolicies(approvalPolicyId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve an Approval Policy.
}
\details{
Retrieve the policy with the given ID.

This method invokes \verb{GET /approvalPolicies/\{approvalPolicyId\}/} in the DataRobot Public API.
}
\seealso{
The method ApprovalPoliciesRetrieve in \link[datarobot.apicore:GovernanceApi]{datarobot.apicore::GovernanceApi}, which invokes the same underlying API endpoint.
}
