% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{RetrieveAutomatedDocuments}
\alias{RetrieveAutomatedDocuments}
\title{RetrieveAutomatedDocuments}
\usage{
RetrieveAutomatedDocuments(documentId, filename, ...)
}
\arguments{
\item{documentId}{character. The ID of the automated documentation to download}

\item{filename}{character. Filename of file to save the automated documentation to.}
}
\value{
Nothing returned, but downloads the file to the stated filename.
}
\description{
Download generated document.
}
\details{
Download a generated Automated Documentation file.  \verb{text GET https://app.datarobot.com/api/v2/automatedDocuments/5ec4ea7e41054c158c5b002f/ HTTP/1.1 Authorization: Bearer DnwzBUSTOtKBO5Sp1hoUByG4YgZwCCw4}  In response, you will get a file containing the generated documentation.
}
\examples{
\dontrun{
documentId <- "5ec4ea7e41054c158c5b002f"
RetrieveAutomatedDocuments(documentId, filename = "documentation.docx")
}
}
