% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_apiWrapper.R
\name{RetrieveComments}
\alias{RetrieveComments}
\title{RetrieveComments}
\usage{
RetrieveComments(commentId, ...)
}
\description{
Retrieve a comment
}
\details{
Retrieve a comment
}
\seealso{
The method CommentsRetrieve in [datarobot.apicore::CommentsApi], which this function wraps.

Other Comments: 
\code{\link{CreateComments}()},
\code{\link{DeleteComments}()},
\code{\link{ListComments}()},
\code{\link{PatchComments}()}
}
\concept{Comments}
