% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoreDiscardedFeatures.R,
%   R/projects_apiWrapper.R
\name{RetrieveDiscardedFeaturesInformation}
\alias{RetrieveDiscardedFeaturesInformation}
\title{Retrieve the information about discarded during time series feature generation process features.}
\usage{
RetrieveDiscardedFeaturesInformation(projectId, search = NULL, ...)

RetrieveDiscardedFeaturesInformation(projectId, search = NULL, ...)
}
\arguments{
\item{projectId}{Character string, the ID of the project.}
}
\value{
An object with information:
\itemize{
\item totalRestoreLimit Integer. The total limit indicating how many features can
be restored in this project.
\item remainingRestoreLimit Integer. The remaining available number of the features which can
be restored in this project.
\item features List of character strings giving discarded feature names. They can be used
to pass into \code{RestoreDiscardedFeatures}
\item count Integer. Discarded features count.
}
}
\description{
Get discarded features.
}
\details{
Get features which were discarded during feature reduction process.
}
\examples{
\dontrun{
projectId <- "59a5af20c80891534e3c2bde"
discardedInfo <- RetrieveDiscardedFeaturesInformation(projectId)
}
}
\seealso{
The method DiscardedFeaturesList in [datarobot.apicore::ProjectsApi], which this function wraps.

Other Restore Discarded Features functions: 
\code{\link{RestoreDiscardedFeatures}()}

Other Projects: 
\code{\link{BatchFeaturesTypeTransform}()},
\code{\link{CloneProject}()},
\code{\link{CreateCalendarsFromCountryCode}()},
\code{\link{CreateCalendarsFromDataset}()},
\code{\link{CreateCalendar}()},
\code{\link{CreateFeaturelist}()},
\code{\link{CreateHdfsProjects}()},
\code{\link{CreateModelingFeaturelist}()},
\code{\link{CreateProjectCleanupJobs}()},
\code{\link{CreateProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{CreateProjectsFeatureDiscoveryRecipeSqlExports}()},
\code{\link{CreateProjectsSecondaryDatasetsConfigurations}()},
\code{\link{CreateProjectsSegmentationTasks}()},
\code{\link{CreateRelationshipsConfigurations}()},
\code{\link{DeleteCalendar}()},
\code{\link{DeleteProjectCleanupJobs}()},
\code{\link{DeleteProjectsJobs}()},
\code{\link{DeleteProjectsSecondaryDatasetsConfigurations}()},
\code{\link{DeleteProject}()},
\code{\link{DeleteRelationshipsConfigurations}()},
\code{\link{GetCalendar}()},
\code{\link{GetFeatureHistogram}()},
\code{\link{GetFeatureInfo}()},
\code{\link{GetFeaturelist}()},
\code{\link{GetJob}()},
\code{\link{GetModelingFeaturelist}()},
\code{\link{GetProjectStatus}()},
\code{\link{GetProject}()},
\code{\link{GetValidMetrics}()},
\code{\link{ListCalendarCountryCodes}()},
\code{\link{ListCalendarsAccessControl}()},
\code{\link{ListCalendars}()},
\code{\link{ListDeletedProjectsCount}()},
\code{\link{ListDeletedProjects}()},
\code{\link{ListFeatureInfo}()},
\code{\link{ListFeaturelists}()},
\code{\link{ListJobs}()},
\code{\link{ListModelingFeaturelists}()},
\code{\link{ListProjectCleanupJobsDownload}()},
\code{\link{ListProjectCleanupJobsSummary}()},
\code{\link{ListProjectCleanupJobs}()},
\code{\link{ListProjectsAccessControl}()},
\code{\link{ListProjectsCalendarEvents}()},
\code{\link{ListProjectsFeatureDiscoveryDatasetDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogsDownload}()},
\code{\link{ListProjectsFeatureDiscoveryLogs}()},
\code{\link{ListProjectsFeatureDiscoveryRecipeSQLsDownload}()},
\code{\link{ListProjectsModelingFeatures}()},
\code{\link{ListProjectsMultiseriesIdsCrossSeriesProperties}()},
\code{\link{ListProjectsMultiseriesNames}()},
\code{\link{ListProjectsRelationshipsConfiguration}()},
\code{\link{ListProjectsSecondaryDatasetsConfigurations}()},
\code{\link{ListProjectsSegmentationTasksMappings}()},
\code{\link{ListProjectsSegmentationTasks}()},
\code{\link{ListProjects}()},
\code{\link{PatchDeletedProjects}()},
\code{\link{PatchManyCalendarsAccessControl}()},
\code{\link{PatchManyProjectsAccessControl}()},
\code{\link{PatchProjectsSegments}()},
\code{\link{PauseQueue}()},
\code{\link{PutRelationshipsConfigurations}()},
\code{\link{RequestCrossSeriesDetection}()},
\code{\link{RequestMultiSeriesDetection}()},
\code{\link{RestoreDiscardedFeatures}()},
\code{\link{RetrieveProjectCleanupJobs}()},
\code{\link{RetrieveProjectsBatchTypeTransformFeaturesResult}()},
\code{\link{RetrieveProjectsExternalTimeSeriesBaselineDataValidationJobs}()},
\code{\link{RetrieveProjectsFeatureLineages}()},
\code{\link{RetrieveProjectsModelingFeatures}()},
\code{\link{RetrieveProjectsSecondaryDatasetsConfigurations}()},
\code{\link{RetrieveProjectsSegmentationTaskJobResults}()},
\code{\link{RetrieveProjectsSegmentationTasks}()},
\code{\link{RetrieveRelationshipsConfigurations}()},
\code{\link{SetTarget}()},
\code{\link{SetupProject}()},
\code{\link{StartNewAutoPilot}()},
\code{\link{UpdateCalendar}()},
\code{\link{UpdateProject}()}
}
\concept{Projects}
\concept{Restore Discarded Features functions}
