% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{RetrieveEventLogs}
\alias{RetrieveEventLogs}
\title{RetrieveEventLogs}
\usage{
RetrieveEventLogs(recordId, includeIdentifyingFields = "true", ...)
}
\description{
Get audit record by ID.
}
\details{
Get audit record by ID.
}
\seealso{
The method EventLogsRetrieve in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{CreateUsageDataExports}()},
\code{\link{DeleteManyUsersRateLimitUsage}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsEvents}()},
\code{\link{ListEventLogsPredictionUsage}()},
\code{\link{ListEventLogs}()},
\code{\link{ListUsageDataExportsSupportedEvents}()},
\code{\link{ListUsersRateLimitUsage}()},
\code{\link{RetrieveUsageDataExports}()}
}
\concept{Analytics}
