% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_apiWrapper.R
\name{RetrieveStatus}
\alias{RetrieveStatus}
\title{RetrieveStatus}
\usage{
RetrieveStatus(statusId, ...)
}
\description{
Get task status
}
\details{
Check the status of an asynchronous task such as project creation
}
\seealso{
The method StatusRetrieve in [datarobot.apicore::JobsApi], which this function wraps.

Other Jobs: 
\code{\link{DeleteStatus}()},
\code{\link{ListStatus}()}
}
\concept{Jobs}
