% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{RetrieveUsageDataExports}
\alias{RetrieveUsageDataExports}
\title{RetrieveUsageDataExports}
\usage{
RetrieveUsageDataExports(artifactId, ...)
}
\description{
Retrieve a prepared customer usage data artifact.
}
\details{
Retrieve a prepared customer usage data artifact. Requires \"CAN_ACCESS_USER_ACTIVITY\" permission.
}
\seealso{
The method UsageDataExportsRetrieve in [datarobot.apicore::AnalyticsApi], which this function wraps.

Other Analytics: 
\code{\link{CreateUsageDataExports}()},
\code{\link{DeleteManyUsersRateLimitUsage}()},
\code{\link{DeleteUsersRateLimitUsage}()},
\code{\link{ListEventLogsEvents}()},
\code{\link{ListEventLogsPredictionUsage}()},
\code{\link{ListEventLogs}()},
\code{\link{ListUsageDataExportsSupportedEvents}()},
\code{\link{ListUsersRateLimitUsage}()},
\code{\link{RetrieveEventLogs}()}
}
\concept{Analytics}
