% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{RetrieveUseCasesAttachments}
\alias{RetrieveUseCasesAttachments}
\title{RetrieveUseCasesAttachments}
\usage{
RetrieveUseCasesAttachments(useCaseId, attachmentId, ...)
}
\description{
Get a resource that is attached to a use case.
}
\details{
Get a resource that is attached to a use case.
}
\seealso{
The method UseCasesAttachmentsRetrieve in [datarobot.apicore::UseCaseApi], which this function wraps.

Other UseCase: 
\code{\link{CreateUseCasesAttachments}()},
\code{\link{CreateUseCases}()},
\code{\link{DeleteUseCasesAttachments}()},
\code{\link{DeleteUseCases}()},
\code{\link{ListUseCaseValueTemplatesCalculation}()},
\code{\link{ListUseCaseValueTemplates}()},
\code{\link{ListUseCasesActivities}()},
\code{\link{ListUseCasesAttachments}()},
\code{\link{ListUseCasesRealizedValueOverTime}()},
\code{\link{ListUseCasesSharedRoles}()},
\code{\link{ListUseCases}()},
\code{\link{PatchManyUseCasesSharedRoles}()},
\code{\link{PatchUseCases}()},
\code{\link{RetrieveUseCaseValueTemplates}()},
\code{\link{RetrieveUseCases}()}
}
\concept{UseCase}
