% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStores.R
\name{TestDataStore}
\alias{TestDataStore}
\title{Test the database connection to the data store.}
\usage{
TestDataStore(dataStoreId, username, password)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}

\item{username}{character. The username to use for authentication to the database.}

\item{password}{character. The password to use for authentication to the database.
The password is encrypted at server side and never saved or stored.}
}
\value{
TRUE if successful, otherwise it will error.
}
\description{
Test the database connection to the data store.
}
\examples{
\dontrun{
 dataStoreId <- "5c1303269300d900016b41a7"
 TestDataStore(dataStoreId, username = "myUser", password = "mySecurePass129")
}
}
