% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataStores.R
\name{TestDataStore}
\alias{TestDataStore}
\title{TestDataStore}
\usage{
TestDataStore(dataStoreId, username, password, ...)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}

\item{username}{character. The username to use for authentication to the database.}

\item{password}{character. The password to use for authentication to the database.
The password is encrypted at server side and never saved or stored.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
TRUE if successful, otherwise it will error.
}
\description{
Tests data store connection.
}
\details{
Tests the ability to connect to a data store with specified authentication.

This method invokes \verb{POST /externalDataStores/\{dataStoreId\}/test/} in the DataRobot Public API.

Test the database connection to the data store.
}
\examples{
\dontrun{
dataStoreId <- "5c1303269300d900016b41a7"
TestDataStore(dataStoreId, username = "myUser", password = "mySecurePass129")
}
}
\seealso{
The method ExternalDataStoresTestCreate in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
