% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnpauseQueue.R
\name{UnpauseQueue}
\alias{UnpauseQueue}
\title{Re-start the DataRobot modeling queue}
\usage{
UnpauseQueue(project)
}
\arguments{
\item{project}{character. Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this
identifier.}
}
\description{
This function unpauses the modeling queue for a specified
DataRobot project.
}
\examples{
\dontrun{
  projectId <- "59a5af20c80891534e3c2bde"
  UnpauseQueue(projectId)
}
}
