% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendars.R
\name{UpdateCalendar}
\alias{UpdateCalendar}
\title{Update a calendar}
\usage{
UpdateCalendar(calendarId, name = NULL)
}
\arguments{
\item{calendarId}{character. The ID of the calendar to retrieve.}

\item{name}{character. The new name to name the calendar.}
}
\value{
An S3 object of class "dataRobotCalendar"
}
\description{
Currently supports changing the name of a calendar.
}
\examples{
\dontrun{
   calendarId <- "5da75da31fb4a45b8a815a53"
   UpdateCalendar(calendarId, name = "New name for calendar")
}
}
