% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataSources.R
\name{UpdateDataSource}
\alias{UpdateDataSource}
\title{UpdateDataSource}
\usage{
UpdateDataSource(
  dataSourceId,
  canonicalName = NULL,
  dataStoreId = NULL,
  query = NULL,
  table = NULL,
  schema = NULL,
  partitionColumn = NULL,
  fetchSize = NULL
)

UpdateDataSource(
  dataSourceId,
  canonicalName = NULL,
  dataStoreId = NULL,
  query = NULL,
  table = NULL,
  schema = NULL,
  partitionColumn = NULL,
  fetchSize = NULL
)
}
\arguments{
\item{dataSourceId}{character. The ID of the data store to update.}

\item{canonicalName}{character. The user-friendly name of the data source.}

\item{dataStoreId}{character. The ID of the data store to connect to.}

\item{query}{character. A query to execute on the data store to get the data. Optional.}

\item{table}{character. The specified database table. Optional.}

\item{schema}{character. The specified database schema. Optional.}

\item{partitionColumn}{character. The name of the partition column. Optional.}

\item{fetchSize}{integer. a user specified fetch size in the range [1, 20000]. Optional.
By default a fetchSize will be assigned to balance throughput and memory usage}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update the data source.
}
\details{
Update the data source with given ID. To test the SQL query before updating the data source,use :\url{http:post}:/api/v2/externalDataStores/(dataStoreId)/verifySQL/.

This method invokes \verb{PATCH /externalDataSources/\{dataSourceId\}/} in the DataRobot Public API.

Update a data store.
}
\examples{
\dontrun{
dataSourceId <- "5c1303269300d900016b41a7"
UpdateDataSource(dataSourceId, canonicalName = "Different Name")
}
}
\seealso{
The method ExternalDataSourcesPatch in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
