% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStores.R
\name{UpdateDataStore}
\alias{UpdateDataStore}
\title{Update a data store.}
\usage{
UpdateDataStore(dataStoreId, canonicalName = NULL, driverId = NULL,
  jdbcUrl = NULL)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}

\item{canonicalName}{character. The user-friendly name of the data store.}

\item{driverId}{character. The ID of the driver to use.}

\item{jdbcUrl}{character. The full JDBC url.}
}
\description{
Update a data store.
}
\examples{
\dontrun{
 dataStoreId <- "5c1303269300d900016b41a7"
 UpdateDataStore(dataStoreId, canonicalName = "Different Name")
}
}
