% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R, R/DataStores.R
\name{UpdateDataStore}
\alias{UpdateDataStore}
\title{UpdateDataStore}
\usage{
UpdateDataStore(
  dataStoreId,
  canonicalName = NULL,
  driverId = NULL,
  jdbcUrl = NULL
)

UpdateDataStore(
  dataStoreId,
  canonicalName = NULL,
  driverId = NULL,
  jdbcUrl = NULL
)
}
\arguments{
\item{dataStoreId}{character. The ID of the data store to update.}

\item{canonicalName}{character. The user-friendly name of the data store.}

\item{driverId}{character. The ID of the driver to use.}

\item{jdbcUrl}{character. The full JDBC url.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Updates a data store configuration.
}
\details{
Updates a data store configuration.

This method invokes \verb{PATCH /externalDataStores/\{dataStoreId\}/} in the DataRobot Public API.

Update a data store.
}
\examples{
\dontrun{
dataStoreId <- "5c1303269300d900016b41a7"
UpdateDataStore(dataStoreId, canonicalName = "Different Name")
}
}
\seealso{
The method ExternalDataStoresPatch in \link[datarobot.apicore:DataConnectivityApi]{datarobot.apicore::DataConnectivityApi}, which invokes the same underlying API endpoint.
}
