% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R,
%   R/ComplianceDocumentation.R
\name{UploadComplianceDocTemplate}
\alias{UploadComplianceDocTemplate}
\title{UploadComplianceDocTemplate}
\usage{
UploadComplianceDocTemplate(name, filename = NULL, sections = NULL, ...)
}
\arguments{
\item{name}{character. A name to identify the compliance doc template by.}

\item{filename}{character. Optional. Filename of file to save the compliance doc template to.}

\item{sections}{list. Optional. Section definitions for the compliance template.}

\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\value{
Nothing returned, but uploads the compliance doc template.
}
\description{
Create a new compliance documentation template
}
\details{
Create a new compliance documentation template. One can retrieve the default DataRobot template via \verb{GET /api/v2/complianceDocTemplates/default/} endpoint.

This method invokes \verb{POST /complianceDocTemplates/} in the DataRobot Public API.

The structure of the compliance doc template can be specified by either a file specified by
\code{filename} or by specifying it with a list via \code{sections}.
}
\examples{
\dontrun{
## Create a compliance documentation template from uploading a file
DownloadComplianceDocTemplate("template.json")
# Edit template.json in your favorite editor
UploadComplianceDocTemplate("myTemplate", "template.json")

## Create a compliance documentation template from a list.
sections <- list(
  list(
    "title" = "Missing Values Report",
    "highlightedText" = "NOTICE",
    "regularText" = paste(
      "This dataset had a lot of Missing Values.",
      "See the chart below: {{missingValues}}"
    ),
    "type" = "user"
  ),
  list(
    "title" = "Blueprints",
    "regularText" = "{{blueprintDiagram}} /n Blueprint for this model",
    "type" = "user"
  )
)
}
}
\seealso{
The method ComplianceDocTemplatesCreate in \link[datarobot.apicore:DocumentationApi]{datarobot.apicore::DocumentationApi}, which invokes the same underlying API endpoint.
}
