% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewWebProject.R
\name{ViewWebProject}
\alias{ViewWebProject}
\title{Retrieve a DataRobot web page that displays detailed project information}
\usage{
ViewWebProject(project)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}
}
\description{
This function brings up a web page that displays detailed project
information like that available from the standard DataRobot user interface.
}
