% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaitForJobToComplete.R
\name{WaitForJobToComplete}
\alias{WaitForJobToComplete}
\title{Wait for specified job to complete}
\usage{
WaitForJobToComplete(project, jobId, maxWait = 60)
}
\arguments{
\item{project}{Either (1) a character string giving the unique alphanumeric
identifier for the project, or (2) a list containing the element projectId with this identifier.}

\item{jobId}{integer identifier (returned for example by RequestPrimeModel)}

\item{maxWait}{maximum time to wait (in seconds) for the job to complete}
}
\description{
Wait for specified job to complete
}

